Create spline formula for high-dimension applications

Create a formula that contains spline terms for high-dimension applications using spline_df

create_HD_formula(formula, data, spl_df, rm_overlap = TRUE, verbose = T)

Arguments

formula
an optional object of class "formula": a symbolic description of the parametric part of the model

data
an optional data frame, containing the parametric variables in the model, which will directly translate to a formula

spl_df
a data frame of three columns ("Var","Func", "Args") that is used to create spline terms in the model. Please see example or details or vignettes

rm_overlap
a logical variable: when there are overlapping variable(s) in both parametric part and spline part of the model, if the overlapping variables are removed from the parametric part.

verbose
a logical variable: if the result formula would be printed

Value

an object of class "formula": the final spline formula that could be to bgam or gam

Details

One of the [formula] and [data] arguments are required to create the parametric part of the formula. When both arguments are supplied, [data] arguments would be ignored. When [data] is supplied, DF2formula will be called to generate [formula].

[spl_df] is a data frame of three string columns: "Var" column contains the variable for splines, "Func" column contains the corresponding spline of choice (see also mgcv[smooth.terms]), "Args" column contains the arguments for the spline function.

See also

Examples

# spl_df template
spline_df <- tribble(
 ~Var, ~Func, ~Args,
 "x0", "s", "k=1, bs='cr'",
 "x1", "s", "k=2, bs='cr'",
 "x2", "", "k=2" # will be ignored
)

#> Error in tribble(~Var, ~Func, ~Args, "x0", "s", "k=1, bs='cr'", "x1",     "s", "k=2, bs='cr'", "x2", "", "k=2"): could not find function "tribble"