Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>         x1.pen1      x1.pen2     x1.pen3     x1.null1       x2.pen1     x2.pen2
#> 1    0.10935941 -0.876031744 -1.06745734  0.641763795  0.1259196499 -0.17570305
#> 2   -0.18896593 -0.531654802 -1.15692341  0.395731534  0.3117738874  1.29672178
#> 3   -0.34384321  0.821317660  5.13582414  3.098576859 -0.1142755471 -0.76977271
#> 4    0.21656634  1.020872130  0.51508373 -0.793904726  0.4083359573 -1.25940907
#> 5    0.29656186  0.840825431  1.58655863 -1.152077499  0.4109455347 -0.82750624
#> 6    0.10414840  0.985308189  0.11509990 -0.637648910  0.2143551897  1.25125563
#> 7   -0.14705769  0.786267603 -0.39224113 -0.403153616  0.3416158451  1.27841245
#> 8    0.44156214 -1.150189167 -0.49179339  1.110057374  0.3168547510 -1.19717443
#> 9   -0.09202021  0.839664323 -0.29796691 -0.451139154 -0.2862395228  0.61163844
#> 10  -0.17336809  0.758612898 -0.43573311 -0.380100084  0.3396588612  1.28107818
#> 11  -0.17139315  0.760740288 -0.43249658 -0.381837395  0.0775714267 -0.97681172
#> 12   0.29575286  0.924971370  1.26027889 -1.049441320  0.3592690706  1.18590531
#> 13  -0.33617414  0.543814727 -0.69684532 -0.225223246 -0.3395340263  0.50875458
#> 14   0.29638953  0.838512757  1.59459804 -1.154553901  0.3572928936  1.14538655
#> 15   0.20844661  1.020826281  0.47652685 -0.779584375 -0.1622764880 -0.71193803
#> 16  -0.46576380  0.105398035 -1.00743938  0.034606747 -0.4710188002  0.03037737
#> 17   0.48463798 -1.138791468 -0.16921832  1.280602782  0.5039240416 -1.26831075
#> 18  -0.36125638 -0.259428899 -1.13215206  0.236744527  0.4734621919 -1.28069105
#> 19  -0.24951054  0.669281991 -0.55814733 -0.311490195 -0.1582986241 -0.71686048
#> 20  -0.36810363 -0.245698117 -1.12930987  0.229041427 -0.4643295918  0.11108918
#> 21  -0.43553635 -0.075605436 -1.08221119  0.134697297 -0.4659358191 -0.09605578
#> 22   0.22784682  0.505373836  2.54412012 -1.433875502 -0.4714754610 -0.01588813
#> 23   0.24690520  0.576000058  2.36441463 -1.382673752 -0.0476381817 -0.84529727
#> 24  -0.38127000 -0.218094996 -1.12315992  0.213609261 -0.4335193221 -0.24234089
#> 25  -0.09248225  0.839239565 -0.29878103 -0.450735429 -0.0105451880 -0.88544000
#> 26  -0.28118561 -0.398945314 -1.15264837  0.316396773  0.1414249058 -1.04024281
#> 27   0.44869064 -1.151638224 -0.45602492  1.130478359  0.3316903063  0.97953179
#> 28  -0.46517973  0.155506654 -0.98194156  0.006582221  0.3518665169  1.25591292
#> 29   0.11254366  0.139340069  3.39662920 -1.669973783  0.2830943203  1.29258356
#> 30  -0.46013798  0.037534188 -1.03856876  0.072270011  0.3462639773  1.27042877
#> 31   0.37876571 -0.719329289  1.40110361  1.908213999 -0.3262774197  0.53579432
#> 32  -0.01071155  0.908584583 -0.14644531 -0.523475088  0.3386391180  1.01484063
#> 33  -0.45237624 -0.008108697 -1.05738871  0.097481098  0.0183923250 -0.91599370
#> 34   0.29681474 -1.056076836 -0.87646288  0.846726273  0.4345057519 -1.27176972
#> 35  -0.14903986  0.784234630 -0.39554793 -0.401422428  0.4137229644 -1.26223775
#> 36  -0.26175104  0.653385388 -0.57758319 -0.300015037 -0.1494387496  0.82918102
#> 37  -0.40495225  0.408813560 -0.81622074 -0.140422737  0.3375244292 -0.64801299
#> 38  -0.14931188  0.783955010 -0.39600132 -0.401184796 -0.0015167435 -0.89503908
#> 39   0.24550680  1.014687260  0.67683330 -0.852580716  0.3305763630  0.97412423
#> 40  -0.06824319  0.861009825 -0.25544923 -0.471986394  0.4981443220 -1.27349942
#> 41   0.44292486 -1.150517628 -0.48522740  1.113840812  0.3107810055 -1.19236916
#> 42  -0.12700665 -0.611228497 -1.15129806  0.445916914  0.0668807505 -0.04885821
#> 43  -0.32190212  0.566903736 -0.67362363 -0.240444693 -0.4651576405  0.10438576
#> 44  -0.44509801  0.288382029 -0.90323085 -0.069125724 -0.1888393682  0.49543247
#> 45   0.25848676 -0.443557628  2.12483184  2.151568511  0.3975332561 -1.25336568
#> 46   0.28985935  0.780248852  1.78669678 -1.213067687  0.1730008436  1.21549001
#> 47  -0.27092389  0.641140143 -0.59214452 -0.291299317  0.4642553273 -0.97275462
#> 48  -0.46583746  0.137865617 -0.99116774  0.016473820 -0.3099138091  0.56775168
#> 49  -0.36660119 -0.248744811 -1.12995300  0.230749036  0.2346650635 -0.41350616
#> 50  -0.38416832  0.455128663 -0.77815559 -0.168813990  0.5053913648 -1.11292680
#> 51   0.27161337  0.679873079  2.08426840 -1.301488069 -0.0063466906 -0.88991167
#> 52  -0.24730666  0.672092822 -0.55464561 -0.313539033 -0.3504851736 -0.43452214
#> 53   0.29269750  0.943725867  1.17384298 -1.021475609  0.3377696220  1.01025022
#> 54  -0.04062208 -0.295513345  4.33222216 -1.921809304 -0.3746890683  0.43067117
#> 55  -0.35931683  0.853551392  5.21180008  3.122295946 -0.0696664726 -0.82086001
#> 56  -0.15032797 -0.581978191 -1.15415346  0.427184318  0.0306064508  0.02862758
#> 57   0.16961888 -0.937257185 -1.02374157  0.699818982  0.0950839979  1.13599353
#> 58   0.26145665  1.004724611  0.79109066 -0.892801237  0.4857794420 -1.03992685
#> 59  -0.06117788  0.867160932 -0.24256247 -0.478215873  0.5012291680 -1.27100891
#> 60   0.23196701  0.520186478  2.50701681 -1.423354618  0.4498114414 -1.27702140
#> 61  -0.42457983  0.357365604 -0.85538991 -0.109569117  0.1189605539 -1.01820672
#> 62   0.30249774 -1.060912305 -0.86738313  0.854397123  0.4415026120 -1.27438894
#> 63  -0.25778845 -0.434700828 -1.15534355  0.337346880  0.4663152161 -1.28031058
#> 64  -0.04002434  0.885059609 -0.20320029 -0.497000095  0.3428484560  1.27654383
#> 65   0.21520614  1.020908346  0.50845653 -0.791453002 -0.0385569022  0.97772116
#> 66   0.22495306  1.020216496  0.55760205 -0.809541450  0.2306910551 -1.12409604
#> 67  -0.29280388  0.610663728 -0.62694019 -0.270034925  0.0902281109  1.13065355
#> 68  -0.04922256  0.877371766 -0.22046609 -0.488804906  0.0634065719 -0.96242976
#> 69   0.17927544 -0.946831375 -1.01548365  0.709677109  0.4891016014 -1.27820137
#> 70  -0.30261953 -0.364560457 -1.14904598  0.296486548  0.0860226534  1.12599852
#> 71   0.41980784 -1.142394102 -0.58313656  1.055628256  0.4669096736 -1.28037016
#> 72   0.12094051  0.994006705  0.16187128 -0.656916464 -0.4638517866  0.11479117
#> 73   0.28109517  0.727243746  1.94798886 -1.261301448 -0.2857514369  0.61251731
#> 74   0.29714064  0.849552221  1.55588121 -1.142606320  0.1557160223 -1.05408683
#> 75   0.08640914 -0.852125982 -1.08100367  0.621027612 -0.1112369539 -0.77332743
#> 76  -0.33983488 -0.300125869 -1.13972179  0.259696689 -0.2827435093  0.61791206
#> 77  -0.25522624 -0.438516900 -1.15556641  0.339599148  0.3587439390  1.16807829
#> 78   0.22843289  0.507467696  2.53889276 -1.432394778 -0.0002951761  0.09446641
#> 79   0.48065712 -1.143644265 -0.21909449  1.255953666 -0.4270213498 -0.26203517
#> 80   0.48726280 -1.134242339 -0.12920578  1.300012033  0.3204379988 -0.60805179
#> 81  -0.00530991 -0.753393357 -1.12187547  0.543443805  0.5129259305 -1.14952318
#> 82   0.32184178 -1.076946401 -0.83423221  0.881501888  0.0627366943  1.09974347
#> 83  -0.23167103  0.691610430 -0.52976347 -0.327940169 -0.2623876792  0.65350663
#> 84  -0.41596746 -0.134638877 -1.10104649  0.167283685 -0.3450056208 -0.44456005
#> 85   0.49309597 -1.078128287  0.20266747  1.450763701 -0.1999658496 -0.66395295
#> 86   0.49368095 -1.106346960  0.05720586  1.386705719  0.4489800568 -0.92893759
#> 87   0.27302484 -1.035292474 -0.91160795  0.815883622  0.5144566492 -1.24879440
#> 88   0.25385999  1.010278800  0.73357134 -0.872671679 -0.0037950212  1.02089766
#> 89   0.25580206  1.009015789  0.74767483 -0.877629047 -0.2638555380  0.65099037
#> 90  -0.15902079 -0.570869642 -1.15499194  0.420161407  0.3545514015  1.11734225
#> 91  -0.25140906 -0.881916157  5.57364092 -2.253901146  0.2554731987 -1.14607576
#> 92  -0.45919533  0.030860433 -1.04142499  0.075960737  0.0768187369 -0.97605012
#> 93  -0.45742175  0.019405654 -1.04624354  0.082291465 -0.3444155729  0.49850536
#> 94  -0.16102663 -0.568289616 -1.15516786  0.418537085  0.3144989492  0.90129483
#> 95  -0.02774467 -0.728294077 -1.12921500  0.525259575 -0.4654195893  0.10218009
#> 96  -0.43834211  0.313720558 -0.88625092 -0.083871489 -0.4670862409 -0.08606284
#> 97   0.09790994  0.981876457  0.09849063 -0.630731495  0.1515395412  0.32522039
#> 98   0.20847747 -0.975319634 -0.98805713  0.740575732 -0.4509104307 -0.17967441
#> 99   0.49124169 -1.123062639 -0.04592886  1.339455050  0.0812384783  1.12066959
#> 100 -0.07713684 -0.671393013 -1.14224494  0.485705603  0.0858990173 -0.98521775
#>         x2.pen3      x2.null1       x3.pen1     x3.pen2     x3.pen3
#> 1    4.16067867  2.2514984152  0.3460577280 -1.13413100 -0.64640015
#> 2    0.39689178 -1.1532072508  0.4375146015 -1.23350734 -0.42539840
#> 3   -0.83218377  0.3257917910  0.2499124665  1.10908816  1.10820339
#> 4   -0.03686836  0.8564095327  0.1033935810  1.01092244  0.37001728
#> 5    2.41141658  1.7466059306  0.3038980578 -1.08237684 -0.70690910
#> 6   -0.07484448 -0.9402086239  0.1736716486  1.07476537  0.63636146
#> 7    0.68685384 -1.2692390128 -0.3533169257  0.15378224 -0.61370686
#> 8   -0.30254540  0.7181222322 -0.3466526407  0.22186983 -0.57505757
#> 9   -0.85232927 -0.4011406276 -0.2796972476  0.45475798 -0.41537075
#> 10   0.66122817 -1.2593218825  0.1458703925  0.37601551  4.05683996
#> 11  -0.67341875  0.4740756309  0.3888828866 -1.18358526 -0.56324324
#> 12   1.22691588 -1.4665476458  0.0532773742 -0.74353269 -0.84953926
#> 13  -0.89519002 -0.3454039292  0.3201833832 -1.10266046 -0.68556928
#> 14   1.39601072 -1.5246055227 -0.3514547721  0.18057101 -0.59888564
#> 15  -0.86299863  0.2896453218 -0.0130114321 -0.64889006 -0.85233400
#> 16  -0.99659998 -0.1035632930 -0.1025963170 -0.51775136 -0.84200639
#> 17   0.55661338  1.1156041869 -0.3256837184  0.32246429 -0.51171414
#> 18   0.28184311  1.0019073091  0.3917467473 -1.18674411 -0.55664993
#> 19  -0.86054603  0.2926613482 -0.0469935522 -0.59982928 -0.85018188
#> 20  -0.98941937 -0.1434715892 -0.1014564832 -0.51946996 -0.84223115
#> 21  -1.00017508 -0.0409030260  0.0915227661  0.99863054  0.33292348
#> 22  -0.99899839 -0.0806793531  0.3346955040 -1.12043538 -0.66446398
#> 23  -0.78447905  0.3757679005 -0.0007250165 -0.66653665 -0.85254605
#> 24  -0.99232812  0.0325678171 -0.3273818713 -0.06957993 -0.71937421
#> 25  -0.75496192  0.4039599806  0.2587345498  1.10722297  1.19335832
#> 26  -0.60048838  0.5297453626  0.0723333671  0.97807900  0.27652804
#> 27   1.96647613 -1.7110843000 -0.3479407431  0.21252112 -0.58055764
#> 28   0.85959832 -1.3346328054  0.2898530325  1.04726252  1.77318751
#> 29   0.21629236 -1.0759345313  0.0649373544  0.96995351  0.25586250
#> 30   0.75517336 -1.2953964175 -0.3538322648  0.14148412 -0.62034805
#> 31  -0.88478544 -0.3598369877 -0.0918699363 -0.53385270 -0.84402593
#> 32   1.85544001 -1.6757622608 -0.2988272679  0.40634078 -0.45252482
#> 33  -0.73016266  0.4263673514  0.3983945031 -1.19398958 -0.54073949
#> 34   0.06976404  0.9070914786  0.2334941696  1.10760779  0.97564843
#> 35  -0.01651807  0.8662614140 -0.3462377837  0.01662473 -0.68223450
#> 36  -0.72636039 -0.5294172480  0.2338206988 -0.99187213 -0.77630953
#> 37   2.92725755  1.9003531773 -0.2068454914  0.60112044 -0.28697275
#> 38  -0.74740367  0.4109037274 -0.2256289984 -0.31450095 -0.80136998
#> 39   1.98313581 -1.7163485848  0.3946210299 -1.18989211 -0.54987715
#> 40   0.48849800  1.0882286789 -0.0612294146  0.81787491 -0.03012987
#> 41  -0.31639526  0.7103168274 -0.3132951203 -0.11556049 -0.73740104
#> 42   4.47705172  2.3393905005  0.0805511568 -0.78203220 -0.84532605
#> 43  -0.99016110 -0.1401546055  0.2505587143  1.10902777  1.11404217
#> 44   5.81855303  2.7096156684  0.5095410888 -1.26317375  0.08688145
#> 45  -0.07558954  0.8374114026 -0.0064271068 -0.65835250 -0.85248648
#> 46  -0.20080402 -0.8757290615  0.2241879145  1.10477459  0.91095274
#> 47   1.95289668  1.6044797013 -0.0283296662  0.85981774  0.03502946
#> 48  -0.87171590 -0.3770834695 -0.3479021872  0.21281375 -0.58038643
#> 49   3.55450741  2.0810980250  0.4788096874 -1.26498977 -0.23242784
#> 50   1.44098122  1.4374001696 -0.3533085191  0.08403978 -0.65005020
#> 51  -0.75146649  0.4071842304  0.0502839318 -0.73929045 -0.84988536
#> 52  -0.96087870  0.1323828274 -0.1508161376  0.69226080 -0.19141534
#> 53   1.87011109 -1.6804534371  0.0681527746  0.97349929  0.26477764
#> 54  -0.92207359 -0.3044318310 -0.1881163357 -0.38141737 -0.81766787
#> 55  -0.80094859  0.3592004951 -0.0683418827 -0.56867956 -0.84771420
#> 56   4.66890644  2.3924742150  0.2481215122 -1.01070427 -0.76467106
#> 57  -0.38348822 -0.7733181852 -0.0270769136  0.86138016  0.03760584
#> 58   1.71977517  1.5296690965  0.3452453521 -1.13315871 -0.64773985
#> 59   0.52328415  1.1022700736 -0.3037440132 -0.14280030 -0.74750512
#> 60   0.14325883  0.9407746948  0.1174431093 -0.83363600 -0.83633743
#> 61  -0.62772000  0.5095886333 -0.0654126595  0.81240669 -0.03811340
#> 62   0.10213205  0.9220430216  0.3100943520 -1.09013379 -0.69906367
#> 63   0.23586488  0.9819451027 -0.2972965595 -0.15994357 -0.75364570
#> 64   0.70384969 -1.2757840149 -0.1906470370 -0.37708316 -0.81669521
#> 65  -0.59870125 -0.6318407755 -0.1053390983  0.75844003 -0.11170269
#> 66  -0.47139272  0.6174257046 -0.0309843670 -0.62299659 -0.85147433
#> 67  -0.39313326 -0.7675433294 -0.1706788257  0.66137452 -0.22543101
#> 68  -0.68784795  0.4623613717 -0.0156436318  0.87553197  0.06149895
#> 69   0.40115256  1.0523789531 -0.3211441721  0.33864235 -0.50077535
#> 70  -0.40134961 -0.7625897493  0.4901049721 -1.26956256 -0.15153787
#> 71   0.23953198  0.9835484037  0.1955346722 -0.94071184 -0.80229646
#> 72  -0.98899842 -0.1453037733  0.0354218966 -0.71818178 -0.85127983
#> 73  -0.85192331 -0.4016271800 -0.3370342069 -0.03128325 -0.70341914
#> 74  -0.58218160  0.5429253378 -0.1110775017  0.75039238 -0.12195649
#> 75  -0.83013971  0.3280678723 -0.3462674729  0.01681386 -0.68214813
#> 76  -0.84941569 -0.4046181306  0.3415423296 -0.72857862  2.13114769
#> 77   1.30370492 -1.4930973505 -0.3535579970  0.14865605 -0.61648742
#> 78   4.83138636  2.4373474447  0.0989995210  1.00641269  0.35607647
#> 79  -0.99025050  0.0425900934  0.5029814532 -1.26922796 -0.02071414
#> 80   3.03711575  1.9324337933 -0.0111949708 -0.65150180 -0.85238493
#> 81   1.28574873  1.3844773614  0.1219770568 -0.83993725 -0.83494913
#> 82  -0.44467914 -0.7359172789 -0.3276640754  0.31503114 -0.51666714
#> 83  -0.83217596 -0.4245489371  0.4309208562 -1.22730197 -0.44800881
#> 84  -0.95851254  0.1377467493 -0.1904521266 -0.37741779 -0.81677072
#> 85  -0.88539060  0.2607635809 -0.2015364865  0.61034055 -0.27793278
#> 86   2.09664946  1.6496957615 -0.1131256733 -0.50178306 -0.83981478
#> 87   0.73736422  1.1860516387 -0.3541510295  0.10985998 -0.63696507
#> 88  -0.55120428 -0.6657397056  0.3142391887 -1.09529600 -0.69363063
#> 89  -0.83343526 -0.4231282817  0.0218240982 -0.02156079  5.07815499
#> 90   1.50350168 -1.5607756647 -0.0912404988 -0.53479278 -0.84413783
#> 91  -0.42823853  0.6444236003 -0.1604223067  0.67749446 -0.20790520
#> 92  -0.67420026  0.4734477983  0.5116233500 -1.25867044  0.13900307
#> 93  -0.89898225 -0.3399685800  0.2560319050  1.10805115  1.16592683
#> 94   2.20004202 -1.7841083338 -0.3225810687  0.33364049 -0.50418079
#> 95  -0.99039938 -0.1390633642  0.3013241428 -1.07914107 -0.71007365
#> 96  -1.00023615 -0.0458740920 -0.0937859289 -0.53098800 -0.84368083
#> 97   3.66708656 -2.2155207015  0.2906101541  1.03827319  1.82671806
#> 98  -0.99730392  0.0009076602  0.4790667187 -1.08683052  1.01010905
#> 99  -0.41054623 -0.7570070088  0.2767405693  1.09230648  1.42532132
#> 100 -0.66465886  0.4810630478  0.1011387445  1.00861429  0.36283181
#>         x3.null1      x4.pen1      x4.pen2     x4.pen3     x4.null1
#> 1    0.850571842 -0.308270375  0.638894534  1.05126781 -0.313777477
#> 2    1.056047499  0.366794541 -1.112200059 -0.87105243  1.175663049
#> 3   -0.981031499  0.323638377  1.207345119  0.25565408 -0.839846578
#> 4   -0.677608667  0.350320566 -0.890059205 -1.65882799  1.415388919
#> 5    0.777453504  0.368393411 -1.011853519 -1.27023648  1.300479912
#> 6   -0.794187921 -0.179296789  0.793584389  0.96631702 -0.406730405
#> 7   -0.066517707  0.092389635  1.052352627  0.70689429 -0.603180923
#> 8   -0.103581492  0.052483186 -1.067965955  0.46287352  0.654729056
#> 9   -0.235101815  0.293414881  0.545043063 -2.51822145 -1.748195662
#> 10  -1.911910943 -0.556144242  0.141974019  1.15728431 -0.054626850
#> 11   0.936171883 -0.572797592 -0.033029794  1.14922247  0.031426835
#> 12   0.456145911  0.452931119  1.057899931 -1.03836267 -1.317568226
#> 13   0.804635656  0.406334277  1.221032007 -0.08986734 -0.985538447
#> 14  -0.081053225  0.309824478  0.587319278 -2.41223536 -1.719231053
#> 15   0.389400599 -0.554458111 -0.182852665  1.12499189  0.105284872
#> 16   0.305507090 -0.571696408 -0.058526205  1.14623462  0.043950157
#> 17  -0.159267390 -0.190544959  0.781137440  0.97447885 -0.398839538
#> 18   0.942435304  0.341150558 -1.173466581 -0.50911233  1.054361912
#> 19   0.357031573 -0.129784320  0.846455743  0.92848975 -0.441301964
#> 20   0.306557789 -0.519101003 -0.306807184  1.09240847  0.167353438
#> 21  -0.660523056  0.398605297  0.837492596 -1.74419269 -1.531148925
#> 22   0.829948965  0.367633871 -1.003427294 -1.29947299  1.309327018
#> 23   0.401390412 -0.573017740 -0.006877195  1.15180810  0.018588001
#> 24   0.053531113 -0.251558218 -0.763218634  0.84883492  0.418788535
#> 25  -1.012738160 -0.310954567 -0.686547894  0.90720929  0.372241906
#> 26  -0.634081396 -0.360567480  0.566504110  1.08061953 -0.273238961
#> 27  -0.098467695 -0.558883214 -0.160884756  1.12956030  0.094399323
#> 28  -1.216824534  0.111731876 -1.113027606  0.35454478  0.708387272
#> 29  -0.624243564  0.092313508  1.052289843  0.70699122 -0.603120092
#> 30  -0.059862179  0.304235153 -1.193382217 -0.25449333  0.962686103
#> 31   0.315394076 -0.532856468  0.228696382  1.15316161 -0.097673702
#> 32  -0.206918366  0.052319974 -1.067834846  0.46314681  0.654588824
#> 33   0.957290577  0.457558064  1.108582891 -0.83744922 -1.252693611
#> 34  -0.930580601 -0.001028521 -1.023116032  0.54687133  0.610303747
#> 35   0.007298592  0.347091002 -1.166076801 -0.56641412  1.074198009
#> 36   0.672196889 -0.411416924 -0.534726745  0.99909155  0.286462022
#> 37  -0.324565335  0.399022488  1.222349809 -0.04816313 -0.968958405
#> 38   0.187191359 -0.151497139  0.823640140  0.94547075 -0.426160976
#> 39   0.948802704 -0.274769027 -0.734157745  0.87206854  0.400824597
#> 40  -0.478131690  0.078472827  1.040788714  0.72428116 -0.592165478
#> 41   0.078264548 -0.572097798  0.023247867  1.15418577  0.003800281
#> 42   0.485207132  0.129023704 -1.125110163  0.31881746  0.725320762
#> 43  -0.983222380  0.333558596  0.650024271 -2.25206512 -1.675057859
#> 44   1.394485904  0.025091148 -1.045467806  0.50719181  0.631631792
#> 45   0.395805690 -0.557040897 -0.170380086  1.12762916  0.099101187
#> 46  -0.905429842  0.368107361 -1.008554062 -1.28173908  1.303964946
#> 47  -0.513286792 -0.572524456 -0.041587495  1.14827108  0.035629212
#> 48  -0.098627636  0.100682523  1.059160649  0.69620834 -0.609847930
#> 49   1.197482945  0.214762684  1.145457738  0.51700694 -0.712207650
#> 50  -0.028890128  0.439453527  1.198758593 -0.35070783 -1.084491293
#> 51   0.453017026  0.177092499  1.118694507  0.58382480 -0.675873258
#> 52  -0.384952405 -0.536597098  0.216809439  1.15405097 -0.091745428
#> 53  -0.628497719  0.017183710 -1.038792139  0.51945038  0.625110992
#> 54   0.225057186 -0.399400933  0.507400944  1.10052358 -0.241141123
#> 55   0.337132069  0.286839965  1.189774284  0.35875179 -0.791811134
#> 56   0.692358714  0.130075170  1.082779171  0.65614175 -0.634197103
#> 57  -0.514651355  0.211520729  1.143236117  0.52311771 -0.708961185
#> 58   0.849072194 -0.122884075 -0.424717790 -4.75363175 -2.332681102
#> 59   0.092966199  0.454515238  1.159799323 -0.59653505 -1.171669027
#> 60   0.526234630  0.104217475 -1.107612417  0.36942855  0.701228473
#> 61  -0.473735301 -0.564257878  0.098179016  1.15730807 -0.033029457
#> 62   0.787652776  0.194950778  1.131632616  0.55325369 -0.692731836
#> 63   0.102242928  0.148059363 -1.137742590  0.27687839  0.744774143
#> 64   0.222577967  0.425404983  1.212722207 -0.22014021 -1.035911225
#> 65  -0.432208554 -0.530910479 -0.271538217  1.10288265  0.149559724
#> 66   0.372147989  0.195745989 -1.165647590  0.15722067  0.798051220
#> 67  -0.363965449  0.342786995 -1.171619841 -0.52416024  1.059596713
#> 68  -0.527218691 -0.427542454 -0.506603667  1.01323186  0.271261442
#> 69  -0.168358363 -0.560817008  0.118447753  1.15746723 -0.043016000
#> 70   1.250846363 -0.422525244 -0.515501040  1.00884599  0.276051406
#> 71   0.620920251  0.280279030 -1.194211385 -0.13645027  0.917885594
#> 72   0.437651672 -0.257497810  0.703204847  1.02001468 -0.351199410
#> 73   0.032981915 -0.549649407  0.169986741  1.15654877 -0.068484498
#> 74  -0.426269932 -0.407644174 -0.541116855  0.99576392  0.289941200
#> 75   0.007197192 -0.394138783 -0.563457488  0.98378572  0.302183270
#> 76   2.296500772  0.445454127  1.188614649 -0.42491902 -1.111356025
#> 77  -0.063742265 -0.532767309 -0.265591066  1.10455280  0.146571048
#> 78  -0.671214897  0.109549491  1.066370855  0.68449440 -0.617070033
#> 79   1.331841392 -0.368758040 -0.603423453  0.96096603  0.324412120
#> 80   0.391162983 -0.447903760 -0.468952673  1.03092211  0.251173720
#> 81   0.531424925 -0.267594209  0.690814506  1.02644856 -0.343870977
#> 82  -0.155104675 -0.138456955  0.837409378  0.93534834 -0.435255496
#> 83   1.037746834 -0.160941073  0.622297747 -5.10940091  2.317877124
#> 84   0.222769233 -0.572017158 -0.052831006  1.14694208  0.041151951
#> 85  -0.330476803 -0.572404597 -0.044609763  1.14792266  0.037113552
#> 86   0.295794973  0.369714671 -1.030529412 -1.20370087  1.280209075
#> 87  -0.042790560  0.178842551  1.119981018  0.58091108 -0.677498105
#> 88   0.794571923 -0.083066230  0.893744217  0.88969514 -0.473940458
#> 89  -2.198444782  0.455370874  1.154724749 -0.62312509 -1.180807488
#> 90   0.315974368  0.168315123  1.112183910  0.59818213 -0.667808865
#> 91  -0.374844389  0.388224870  0.805263865 -1.83544106 -1.557496485
#> 92   1.423753459  0.303423918 -1.193519067 -0.25007160  0.961036879
#> 93  -1.002581094 -0.176318822 -0.850734597  0.76917946  0.475860332
#> 94  -0.165543020 -0.459460100  0.402627163  1.12782135 -0.185951631
#> 95   0.773266070  0.358682040 -1.143633330 -0.70953086  1.122649926
#> 96   0.313627848  0.210077034  1.142241512  0.52581563 -0.707523207
#> 97  -1.234803149 -0.510961363 -0.329047497  1.08529405  0.178640873
#> 98   1.846562254  0.411122644  1.219694637 -0.11927171 -0.997090404
#> 99  -1.096650068 -0.158981999 -0.869619805  0.74975096  0.488892489
#> 100 -0.674317214  0.219741479  1.148836964  0.50747855 -0.717241066
#>           x5.pen1     x5.pen2      x5.pen3    x5.null1
#> 1   -0.3413840155  0.68561896 -0.899284317 -0.39444484
#> 2    0.4403391717 -0.83473812  2.511523725  1.61456309
#> 3    0.3011385692  1.34946472  0.783891130 -1.37564500
#> 4   -0.4693329542 -0.41009014 -0.869804536  0.14984870
#> 5   -0.5032946135  0.30347896 -0.971772523 -0.20152410
#> 6   -0.5571389497 -0.03510151 -0.962140028 -0.03828427
#> 7    0.3063614517  1.20131835  1.403059412 -1.59540885
#> 8    0.1243115774  1.33123678 -0.293348612 -0.89562651
#> 9   -0.5481838449 -0.16307064 -0.941138137  0.02421584
#> 10   0.4926177680 -1.20695477  0.895909875  1.10849929
#> 11   0.4578994067 -1.22720249  0.608237654  1.00339676
#> 12  -0.3964578852  0.57718201 -0.930040941 -0.33761757
#> 13  -0.0203351604  0.18067065  5.227232481  2.33825080
#> 14   0.2726688933  1.39862466  0.404662368 -1.22632950
#> 15  -0.4850173828  0.36089014 -0.966881187 -0.22949331
#> 16   0.2166933129  1.39851377  0.041300220 -1.06655190
#> 17  -0.1927430016 -0.78289751 -0.655570488  0.36672752
#> 18   0.2599102898  0.90030240  2.290265838 -1.87890398
#> 19   0.1769573019  0.53930696  3.177174480 -2.14104107
#> 20   0.3898485892 -1.21963940  0.284765300  0.87533403
#> 21   0.1648618051  1.36642933 -0.168823227 -0.96282423
#> 22  -0.0006209128  1.19286409 -0.558703739 -0.73000225
#> 23  -0.1025949502  1.06027262 -0.698138064 -0.62175813
#> 24  -0.2478790088 -0.72436123 -0.701010005  0.32905830
#> 25   0.2690530738  1.40077523  0.372666769 -1.21302495
#> 26   0.2890669895  1.06015137  1.848310758 -1.74118679
#> 27   0.1948988108  1.38730494 -0.056069743 -1.01978002
#> 28  -0.2007045476  0.91963015 -0.796057081 -0.52816476
#> 29   0.1668581700  1.36797459 -0.161952533 -0.96639155
#> 30  -0.4658914183  0.41447928 -0.960534473 -0.25581618
#> 31  -0.5048945958  0.29806993 -0.972133560 -0.19889906
#> 32  -0.2219876887 -0.75240801 -0.679955827  0.34686822
#> 33   0.2769582923  1.39538464  0.445339366 -1.24306493
#> 34   0.3574624453 -0.62534598  3.136012304  1.78883061
#> 35  -0.2079126501  0.90869580 -0.802289813 -0.52142936
#> 36  -0.4820625399  0.36954047 -0.965974483 -0.23372677
#> 37  -0.4400073607 -0.46505137 -0.847551475  0.17918679
#> 38  -0.0551372915  1.12360915 -0.639442049 -0.66999896
#> 39   0.2917899102 -0.47512539  3.550968135  1.90101560
#> 40  -0.1792389990  0.50506288  6.032416324  2.54521715
#> 41   0.2208204616  0.72122674  2.743614151 -2.01482532
#> 42   0.2081812759  1.39461179  0.001392252 -1.04762695
#> 43  -0.3590023366 -0.58898915 -0.787043553  0.24803691
#> 44   0.1434726103  1.34870382 -0.237816102 -0.92622097
#> 45  -0.5457648685  0.11418631 -0.974433448 -0.11024750
#> 46   0.2327094393 -1.13737059 -0.132482019  0.68652406
#> 47  -0.1647223970  0.97289720 -0.763147945 -0.56194511
#> 48   0.1880468759  1.38302904 -0.083711186 -1.00611525
#> 49  -0.5569337950 -0.05655711 -0.959310019 -0.02787832
#> 50  -0.3996748424  0.57040347 -0.931669845 -0.33413716
#> 51   0.0309139006 -0.01355884  4.411988209 -2.48821727
#> 52   0.3031176631  1.16558888  1.523119474 -1.63554381
#> 53   0.1490854030 -1.07856203 -0.278071162  0.61083685
#> 54  -0.1932157701  0.93089382 -0.789458856 -0.53516969
#> 55   0.1561172875  1.35943190 -0.198014160 -0.94751511
#> 56   0.1131919901 -1.05146751 -0.331136785  0.58140609
#> 57   0.3934470177 -0.71236163  2.884710487  1.71960028
#> 58   0.1338020274  1.34005846 -0.266515983 -0.91054451
#> 59  -0.3399223761  0.68832924 -0.898395911 -0.39589667
#> 60  -0.1377974382  1.01143260 -0.736373339 -0.58754879
#> 61   0.0362941358 -0.99054472 -0.430290348  0.52305848
#> 62  -0.5188088900 -0.28954488 -0.910089963  0.08746246
#> 63   0.3443045023 -0.59450446  3.222935688  1.81253466
#> 64   0.1735954372 -1.09648096 -0.238870335  0.63189511
#> 65   0.2664251276  1.40204888  0.350555737 -1.20375543
#> 66  -0.2903943141 -0.67580515 -0.734645967  0.29913624
#> 67   0.0240792284  1.22279055 -0.516389622 -0.75919521
#> 68  -0.5473111850  0.10323969 -0.973964994 -0.10498489
#> 69  -0.4765713214  0.38523196 -0.964214691 -0.24142078
#> 70   0.4992347547 -1.19710923  0.981194190  1.13840653
#> 71  -0.4820265812  0.36964481 -0.965963272 -0.23377787
#> 72  -0.3392276111 -0.61529545 -0.772127662  0.26323230
#> 73  -0.2798654869  0.79392076 -0.858672997 -0.45396981
#> 74   0.2190331581  1.39946572  0.052747005 -1.07192095
#> 75  -0.4441590718 -0.45771765 -0.850672317  0.17523465
#> 76  -0.2638847998 -0.70647176 -0.713800125  0.31790695
#> 77  -0.5535064876  0.04735420 -0.970511941 -0.07809811
#> 78   0.2995847397 -1.17845482  0.014197146  0.75686067
#> 79  -0.5481266637  0.09713817 -0.973674387 -0.10205127
#> 80  -0.0740317400 -0.89631577 -0.548109355  0.44641614
#> 81  -0.2617447591  0.82387725 -0.845419913 -0.47107114
#> 82   0.2971426050  1.11549212  1.682183459 -1.68774068
#> 83   0.4157644934 -1.22626423  0.388903096  0.91792711
#> 84  -0.4328596649  0.49672073 -0.947311185 -0.29674866
#> 85   0.4054448381 -1.22403092  0.345376813  0.90029906
#> 86  -0.3993422050 -0.53112929 -0.817193265  0.21537659
#> 87   0.1397683929  1.34543519 -0.248978691 -0.92015697
#> 88   0.5025880131 -1.19025261  1.033912994  1.15664278
#> 89   0.2670142012  0.93603708  2.195273163 -1.84979079
#> 90   0.3038147907  1.17245345  1.500533131 -1.62804400
#> 91   0.0890238551  1.29590457 -0.383119898 -0.84365769
#> 92   0.0387908857  0.05966938  4.926112083  2.26074781
#> 93  -0.1317179634 -0.84305161 -0.602246036  0.40769417
#> 94   0.2880558809 -1.17185842 -0.013563177  0.74393297
#> 95  -0.4125450046  0.54268444 -0.937992220 -0.31998014
#> 96   0.4807766539 -1.21787741  0.777359524  1.06602628
#> 97   0.2876406158  1.38292448  0.563988089 -1.29081460
#> 98   0.1502378871 -1.07941574 -0.276284611  0.61180864
#> 99  -0.2164823753  0.89557392 -0.809552930 -0.51342675
#> 100  0.2660334285  1.40221953  0.347335875 -1.20240034