Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                 x1           x2          x3          x4 y
#>   [1,]  1.43307642 -0.285760926 -0.32270793 -0.55717067 0
#>   [2,] -0.07942711  0.071481265  0.75448186 -0.44011183 0
#>   [3,]  2.17773518  1.011235496  0.71963316 -0.16726746 1
#>   [4,] -0.99773355  0.065908329  1.06587242  1.99172109 0
#>   [5,]  0.18555918  1.355655209  1.53819495 -0.58320396 1
#>   [6,] -0.16719254  0.962794011 -0.88176584 -0.57993291 0
#>   [7,]  1.44138605  0.639064480 -0.75238896  1.23955700 0
#>   [8,] -0.64955786  0.615883793 -0.31178934  2.17560578 0
#>   [9,] -0.11011904  1.641269214  0.75752538 -0.10459253 1
#>  [10,] -0.15005802  1.884044217 -0.26065530 -0.84858819 0
#>  [11,]  0.37276529  0.080834037 -0.96946274  0.63302432 0
#>  [12,]  1.76717120  0.442291143  0.25767407 -0.29310605 0
#>  [13,]  0.90033527  0.845487577 -0.71080571 -0.63830992 0
#>  [14,] -0.78387101 -1.549187340  0.44277684  0.33554330 1
#>  [15,]  0.77810164  0.557474452 -0.23401321 -1.09139853 0
#>  [16,]  0.65988526  1.063621698  0.59413197 -0.92169457 0
#>  [17,] -0.67712093  0.008494965  3.12581245  0.69367681 1
#>  [18,]  0.64966879  0.014835608 -0.44205513  0.70101487 0
#>  [19,]  2.21272447 -2.372326977 -0.41617076  0.85254951 1
#>  [20,] -0.21493117  1.073435598  0.36262443  1.92859008 0
#>  [21,] -2.03787450 -0.420630039 -0.07654127  0.91886969 0
#>  [22,] -0.65258352  1.224355901 -0.88985795 -0.06337534 0
#>  [23,] -2.64955472  0.801716275  0.07642624  1.20018800 0
#>  [24,] -1.29800693 -0.935576168 -0.55353893  1.06375003 0
#>  [25,]  0.65278293 -1.279095674 -0.18277760  0.80935677 0
#>  [26,] -0.17743629 -1.403188525 -0.61895058  0.25101796 0
#>  [27,] -0.76563586  0.681958910  0.19875148 -0.54639020 1
#>  [28,] -0.82785135 -2.211093198  1.85643821  1.48299312 1
#>  [29,]  0.56765008  0.557938916 -1.55357567  0.54857280 0
#>  [30,] -2.25130526 -1.133451363 -0.37004945 -1.51686211 0
#>  [31,] -0.72845214 -1.134628744  0.08780082 -2.29892886 0
#>  [32,]  0.59010174 -1.963406292 -2.01875791 -1.43018354 0
#>  [33,] -1.17151623 -0.736273609 -0.13943299 -0.48764538 0
#>  [34,]  0.46700358  1.110827227 -1.68279081  1.62882824 0
#>  [35,]  0.85553753 -1.287934406  1.36906216 -2.47884272 0
#>  [36,] -0.01405020  0.128392769  2.27699809 -0.53923353 1
#>  [37,] -0.44834628  0.635466840  0.70566997  0.34927415 1
#>  [38,]  0.31945343  0.167506238  0.11423616 -0.76227588 0
#>  [39,]  0.22056194 -0.192494607 -1.03498550  0.57966404 0
#>  [40,]  0.52868485 -2.647441282  0.51368229 -0.01069061 1
#>  [41,] -0.33190381 -0.078021853 -0.35445308  0.66022393 0
#>  [42,]  0.69852850 -0.454353215  0.42592238  1.71544648 0
#>  [43,] -0.74954167 -0.248316076 -1.17569141  0.08586321 0
#>  [44,] -1.36357343  0.088654621  1.23380953  1.16094899 0
#>  [45,]  0.53224742 -0.553423804  0.74945325  0.58825588 1
#>  [46,]  0.79244348  0.247234446  1.11685600  1.00592977 0
#>  [47,] -0.61714300  0.196844235 -3.14831105  0.43261741 0
#>  [48,] -0.19164368 -1.919254796  1.00284450  0.16374859 0
#>  [49,] -1.07881827  0.255894913 -1.76709769  0.27453296 0
#>  [50,]  2.23256030  1.382088312  1.17519470 -0.31006253 1
#>  [51,]  0.44979280 -1.699530668  1.78988839 -1.42122619 0
#>  [52,]  1.31710147  1.232680640 -0.80204616  0.84238697 0
#>  [53,] -0.73982939  1.416766275 -0.72633277 -0.36261960 1
#>  [54,]  0.89274116  0.907816980 -0.07739945 -0.69602811 0
#>  [55,]  0.66701245  0.205236028  0.12645274 -0.35659491 0
#>  [56,]  1.12197715  0.208841511 -0.51903505  0.17942446 0
#>  [57,]  1.71182272 -0.090379328 -0.10084135  1.17619144 0
#>  [58,] -1.07023168 -0.549051486  0.78176653  0.21406629 1
#>  [59,] -0.25088523  1.370907609  0.17715961 -0.66289609 0
#>  [60,] -1.55798566 -0.521252330  0.47956116  0.94129929 0
#>  [61,] -0.28356555  1.048175232 -0.30352198 -0.54044428 0
#>  [62,]  0.38097272 -0.691812824  0.70850879 -0.64143268 1
#>  [63,] -0.04086524 -0.776676660 -0.23537807  1.27933229 0
#>  [64,]  0.59604897 -0.896311373 -0.15965723  0.81024919 0
#>  [65,]  0.88780121  1.118025298 -0.35150089  0.28363971 0
#>  [66,] -0.59079653  0.661044791 -0.03724441  0.86239468 1
#>  [67,] -1.46414620 -0.764995908 -0.45903912 -0.12979677 0
#>  [68,]  0.99597717 -1.513301650  0.36425019  0.86483806 1
#>  [69,]  0.69702088 -0.949588236 -1.76290486 -0.38836608 0
#>  [70,]  0.60966587 -0.003150622  0.68191137 -0.10005694 0
#>  [71,] -0.70858554 -1.019687199  0.15780867  0.72791173 0
#>  [72,] -0.34006061  0.412101382 -0.76623613 -0.21142073 0
#>  [73,]  1.17024180 -0.789543717 -0.48328008 -0.25569641 0
#>  [74,] -1.87145128  1.619632100  0.41690371  0.01464090 1
#>  [75,] -1.27185565 -0.665749137 -1.11518065 -1.01448599 0
#>  [76,] -0.63979338  0.457959477 -0.05379157  0.59880592 1
#>  [77,]  0.50899045  0.277520520  1.43993432 -0.01795871 1
#>  [78,]  1.53340199 -0.952145411 -1.79411760  0.66001540 0
#>  [79,] -0.24109430  0.914801367 -0.22122913 -0.28918760 0
#>  [80,]  0.31727763  1.599793512 -0.36066708 -1.37048989 0
#>  [81,] -1.35058287  0.536653239 -0.67560418  1.85390222 0
#>  [82,]  0.60121311 -1.129704576 -0.89369903 -1.79978846 0
#>  [83,] -0.43625567  2.683946579  0.28668993 -0.70303371 1
#>  [84,]  0.27995151  0.959130539  0.73693760 -0.41168736 1
#>  [85,] -1.81822534 -0.100627522 -0.36223557 -0.63256300 0
#>  [86,] -1.67971910  0.404849101 -1.96771697  0.36618020 0
#>  [87,] -1.20541655 -0.012945101 -1.21722416 -1.07654490 0
#>  [88,] -0.37259563  0.172381075  2.12780984  1.10607926 0
#>  [89,] -0.81367661 -0.565344290 -0.70421979 -0.38709299 1
#>  [90,]  1.14968911 -0.217687676 -0.85902270  0.49903411 0
#>  [91,]  0.57728952 -1.522547053 -0.56898721  0.74961905 0
#>  [92,]  0.11331782  0.856935565 -0.15063235 -0.09765408 0
#>  [93,]  0.69379862 -0.345267363  0.35324446  0.75555994 0
#>  [94,]  0.62070042  1.813562896  1.92163305  0.53713659 1
#>  [95,] -0.06161397 -0.226845893 -0.17288860 -0.14412288 0
#>  [96,] -0.74658239  1.158181256  0.02271560 -1.42927306 0
#>  [97,]  1.76885551 -0.560895303  1.05082911 -0.22492835 1
#>  [98,]  0.49278109  0.574502356  0.10383783 -0.27531281 1
#>  [99,]  0.09588208  0.263163396  0.02594697 -0.62033495 0
#> [100,]  1.37269224 -1.380650986  0.17485182 -0.17314498 1
#> 
#> $eta
#>   [1]  -0.29576833  -4.32940764   3.53443636 -18.85358623  11.27673332
#>   [6] -15.78118401  -8.17294163 -19.10376006   3.50369788 -12.05293175
#>  [11]  -9.73542418  -2.75256571 -11.13162941   8.24164827  -9.82060670
#>  [16] -10.38243517  15.72618238 -14.32199037   1.05100723 -26.79768550
#>  [21]  -3.35952223  -5.22299898  -3.51019463 -19.23519489  -7.44389754
#>  [26]  -6.04143872   6.37621466   4.06701551 -10.42489560 -21.14533435
#>  [31] -23.33635175 -30.37650801  -9.53806946 -27.75609139 -25.25706493
#>  [36]   6.45397619   4.28553941  -2.59340380  -3.92229601   1.26080877
#>  [41] -12.35260801 -14.48960158  -1.71057682  -8.60969401   2.94304812
#>  [46]  -6.65999961 -20.93136287  -4.28453657 -16.64199511  11.33641407
#>  [51]  -0.03631033  -7.58580360   0.55765431  -9.39509379  -8.52933103
#>  [56]  -4.05906925 -15.80698416   3.58920264  -6.60701699   1.95254943
#>  [61] -13.59314088   3.55828882 -14.84194041 -12.53226012 -11.13515139
#>  [66]   0.74545114  -7.69480534   0.15739797 -21.49012398  -4.10929573
#>  [71]  -1.61764788  -8.55917287  -3.06344193   8.48782275 -18.17248466
#>  [76]   3.12072082   5.57057034 -19.40528067 -10.27870357  -4.52755381
#>  [81] -22.91728453 -27.04028618  -1.03030324   4.07774089  -2.82495537
#>  [86]  -7.63432859 -22.74611044  -1.77777185   0.58378386  -2.68945647
#>  [91]  -6.85297432   0.13105552  -3.36411849   5.55028821  -3.14030451
#>  [96] -10.10096331   2.12758863   2.96871879  -2.05624726   6.86139764
#> 
#> $mu
#>   [1] 4.265923e-01 1.300402e-02 9.716519e-01 6.486218e-09 9.999873e-01
#>   [6] 1.400615e-07 2.821071e-04 5.050594e-09 9.707928e-01 5.827412e-06
#>  [11] 5.914708e-05 5.994191e-02 1.464160e-05 9.997366e-01 5.431767e-05
#>  [16] 3.097079e-05 9.999999e-01 6.026128e-07 7.409683e-01 2.300981e-12
#>  [21] 3.358473e-02 5.362228e-03 2.902355e-02 4.428544e-09 5.846587e-04
#>  [26] 2.372493e-03 9.983013e-01 9.831600e-01 2.968332e-05 6.556891e-10
#>  [31] 7.330797e-11 2.220446e-16 7.205063e-05 8.824349e-13 1.073980e-11
#>  [36] 9.984282e-01 9.864207e-01 6.956415e-02 1.941133e-02 7.791653e-01
#>  [41] 4.318457e-06 5.096192e-07 1.530889e-01 1.822965e-04 9.499339e-01
#>  [46] 1.279508e-03 8.121282e-10 1.359270e-02 5.922067e-08 9.999881e-01
#>  [51] 4.909234e-01 5.073492e-04 6.359096e-01 8.312401e-05 1.975481e-04
#>  [56] 1.697206e-02 1.364941e-07 9.731220e-01 1.349033e-03 8.757244e-01
#>  [61] 1.249031e-06 9.723015e-01 3.582837e-07 3.608336e-06 1.459012e-05
#>  [66] 6.781867e-01 4.549785e-04 5.392685e-01 4.644701e-10 1.615409e-02
#>  [71] 1.655295e-01 1.917411e-04 4.464068e-02 9.997941e-01 1.281711e-08
#>  [76] 9.577394e-01 9.962061e-01 3.735887e-09 3.435586e-05 1.069154e-02
#>  [81] 1.114679e-10 1.805315e-12 2.630253e-01 9.833367e-01 5.599044e-02
#>  [86] 4.833295e-04 1.322786e-10 1.445785e-01 6.419376e-01 6.359838e-02
#>  [91] 1.055195e-03 5.327171e-01 3.343587e-02 9.961287e-01 4.147501e-02
#>  [96] 4.103832e-05 8.935559e-01 9.511408e-01 1.134227e-01 9.989536e-01
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                  x1            x2           x3           x4   y
#>   [1,] -1.172040302  0.7014462114  0.103601443 -0.005079490   4
#>   [2,] -0.539524126  0.0007091975  0.012061457 -1.431838770   0
#>   [3,] -0.254511415  0.9109973891  0.061206925  0.066648983   2
#>   [4,]  0.962589027  1.6288737450 -0.583569945 -1.373551133   0
#>   [5,] -1.626325711  0.7969899979 -0.644666498  2.260927851   0
#>   [6,]  0.971461357 -1.3633499862  0.927871516 -1.987304806   0
#>   [7,]  0.173487770 -0.3479077703 -2.085837368  0.371957154  16
#>   [8,]  0.836317677  2.2603808528  0.189905200 -1.011299242   0
#>   [9,] -1.956331802 -1.7851319434 -1.800910705  0.801650503   0
#>  [10,]  0.559811483 -1.4154312771  1.410434007  0.159296895  25
#>  [11,]  2.324811593 -0.0945872464  0.477129644 -0.522107463   4
#>  [12,] -0.654030026  1.6359843095  0.483555183 -1.113300360   4
#>  [13,] -1.735726490 -1.8877113513  1.019152583 -0.710208880   3
#>  [14,] -0.752839212  2.1924682535  1.450393631  0.140178985  14
#>  [15,]  0.430030947  1.1889587981  0.348742178 -0.307190075   4
#>  [16,]  0.727195032 -0.6516464862 -0.886292240  1.619563715   0
#>  [17,] -0.005368364  0.8515863668  0.379536368 -0.385435008   1
#>  [18,]  0.411187945  0.7367143534 -1.325831663  0.284136074   2
#>  [19,] -0.629170913 -0.5631922006  2.228086277 -0.197059235 424
#>  [20,] -0.823549394 -0.7872515812 -0.087816758  0.112126232   7
#>  [21,]  0.693999172 -0.1183611619  0.344614714 -0.063221681   1
#>  [22,] -1.209129632 -0.7790281061  0.083631940  1.130918766   0
#>  [23,] -0.767846030 -0.2671850766  0.393212510 -1.435866080   0
#>  [24,] -1.235746617 -0.6784100380  0.464179829  0.926078878   1
#>  [25,] -0.829155700  1.7201118245 -0.518722929  0.641074757   7
#>  [26,]  0.180803202 -0.5359012760  0.676304140  0.969183889   4
#>  [27,]  1.239328581  2.4696853528 -1.458650299 -2.129270211   0
#>  [28,] -0.237281027 -0.4719493672  0.440032931 -0.187029362   6
#>  [29,]  1.058040952 -0.8223207510 -0.524099726 -1.425264213   1
#>  [30,] -0.337003020 -0.4292354857 -1.082860503  0.323116750   2
#>  [31,] -0.983339437 -0.1475455287 -0.779491649 -0.664985829   1
#>  [32,]  0.727466691  0.2916616868  1.782387125  0.276434032   7
#>  [33,]  1.100140823 -1.1337764155 -0.528804186  0.172150840   4
#>  [34,]  0.211183372 -0.4476569468 -0.557035451 -1.988887021   0
#>  [35,]  0.691645846 -0.2455422792  0.485522826 -0.008220698   6
#>  [36,] -0.420791604 -0.0214018334 -0.201177856 -0.657072656   0
#>  [37,]  0.536221722 -1.8913345214  0.968509136  0.326374217   2
#>  [38,] -0.071161283  0.5787843836  0.102685211  1.076196748   1
#>  [39,] -0.691694200  0.9442596807  0.548515096  0.470316365   4
#>  [40,]  0.219304511  0.4432024745  0.573881931 -1.391020234   0
#>  [41,]  0.324586857  0.0389640382 -0.005181904 -0.347390315   3
#>  [42,]  0.911314518 -1.5379491911 -0.358309469 -1.063422312   0
#>  [43,] -0.573876660  0.8711547326 -0.220079694 -1.759530914   0
#>  [44,] -0.855027963  0.4690965385 -0.262607440  0.165880840   9
#>  [45,] -1.484680916  0.8951755776 -0.111432057  0.439125717   1
#>  [46,]  0.030836351 -1.8359327819 -0.728357401  0.651340362   0
#>  [47,]  0.274265761 -0.5062446291 -0.231489382 -0.675675151  10
#>  [48,] -0.396743870  0.8212170907  1.098041322 -1.145938187   1
#>  [49,]  0.027228222  0.2793759629  0.594407829 -0.014699343   3
#>  [50,] -0.783083254 -0.8234300820  1.297321730 -0.349429249  11
#>  [51,] -0.391300842  2.0740996725 -0.164380706  1.478264534   0
#>  [52,]  0.816358195  1.6421053094  0.307096066  1.777279836   0
#>  [53,] -0.061809009  1.7318579256  0.113043215 -0.468692172   5
#>  [54,]  0.015446661  1.8147600460  0.111712019  0.489073677   6
#>  [55,]  0.819963719  0.2145911601 -1.370597808 -1.012757062   0
#>  [56,] -0.316710175 -0.0879923321  0.344299513  1.533087325   0
#>  [57,] -0.762568075  1.7124451347  1.289714436  0.514665218  36
#>  [58,] -0.883046321 -0.4139270865 -0.135918757 -0.699294190   9
#>  [59,]  0.957587431 -0.6946110931 -0.152165547  0.452127565   3
#>  [60,] -1.071553321 -0.8178527884 -1.732653868 -1.051730393   0
#>  [61,] -0.157777890  0.4954912880  0.352883082 -1.374598945   1
#>  [62,]  0.905272201  1.5038264478  1.157204951  0.965606985   3
#>  [63,] -0.501142048 -0.8276882566 -0.581827971 -2.932871275   0
#>  [64,] -1.409692945  0.1104052878 -0.024781457 -0.104735874   1
#>  [65,] -1.862067972  1.0318305017  1.133256159  1.011445225   3
#>  [66,]  0.200849784 -0.0975571343  0.596844346 -0.140762436   6
#>  [67,] -0.519237173  1.7671655112 -0.178083586  0.284497823   2
#>  [68,]  2.666214448  3.1572896327 -0.333360938 -0.798962288   1
#>  [69,] -1.463303963 -0.6697998994 -0.742086594 -1.261890862   0
#>  [70,] -0.222526012 -0.5448399600 -0.822919905  1.212526942   1
#>  [71,]  0.942865589 -0.0711848396  0.839508459 -1.301515498   0
#>  [72,]  0.352345876  0.3761718593 -0.733274885 -1.704556952   0
#>  [73,]  0.658890397 -0.7968966386  1.670573618  0.286607781   4
#>  [74,]  0.783706578  1.0850040627  1.934146756  0.694145249   2
#>  [75,] -0.294730508  0.4228885176  0.624790666 -0.292731121   1
#>  [76,]  0.018589853 -1.4435523181 -0.789231886  0.665252234   7
#>  [77,] -1.573679968 -0.8658824304  2.432802581 -0.084243267 153
#>  [78,]  0.768280567 -0.9646725381 -0.708764076  1.060577316   0
#>  [79,]  2.007472414 -0.8527624861 -0.605286439  0.035896143   2
#>  [80,] -0.922675219  1.5550654054  0.778925377  0.669467780  12
#>  [81,] -1.035873054  0.0293740072 -0.472829930 -0.421634259   1
#>  [82,] -0.372315447  0.7063839710  1.762434353 -0.156040603  31
#>  [83,] -1.204509926  0.8136741820 -0.402574767 -1.260957511   0
#>  [84,] -0.117195044 -0.9963435589 -0.493459773 -1.382643747   0
#>  [85,] -1.195435866  0.8471483163 -0.355027233  0.509279139   1
#>  [86,] -0.158654961 -0.0617886847  0.068805896  0.892544934   0
#>  [87,]  0.512010875 -0.1701606389 -0.218567558  1.481457317   0
#>  [88,] -0.405558367 -0.5557254761 -0.648673727  0.687938534   1
#>  [89,]  0.759714224  1.1583370344 -0.881795409 -0.882284125   0
#>  [90,] -0.458187006  1.6454205909  0.908906270 -1.026665898   0
#>  [91,] -0.936275312  0.1218842578 -0.904379235  0.083999479   3
#>  [92,] -1.003934126  0.5016193976 -0.364690909  0.025416933  12
#>  [93,] -0.798761292 -0.7082098974  0.183740879 -0.369844394   8
#>  [94,]  0.904618580  1.4038451737  0.809883623 -1.253320944   0
#>  [95,] -1.916439972  0.8015702959  1.643608345 -1.201079560   3
#>  [96,]  0.792029075  0.3180034562 -1.161238767  0.903873546   0
#>  [97,] -0.479494651  0.6554840359  0.221522574 -0.078180748   8
#>  [98,]  1.329924331  0.6550708372 -0.765352960 -1.605532664   0
#>  [99,] -1.061733198 -1.0107333491 -1.070658299 -1.022051420   0
#> [100,]  1.079838350  0.4005022072 -0.072078592  0.526644906   4
#> 
#> $eta
#>   [1]   1.282082186  -3.329388189  -0.072787361  -1.830612017  -8.097356852
#>   [6]  -5.129076869   2.557616482  -1.839502950  -0.608223170   3.220827680
#>  [11]   1.532267428   0.931346050   1.116573263   2.907648715   1.030800119
#>  [16]  -4.953152734   1.115952748   2.025204294   5.984481741   1.566445727
#>  [21]   0.296543608  -2.720920554  -0.965801322  -1.125118996   1.929841865
#>  [26]   1.563086287  -6.205865192   1.502383204  -3.285238345   1.070841046
#>  [31]   0.260287697   1.998800926   1.536727594  -4.726380865   1.171001838
#>  [36]  -0.819156292   0.631592538  -0.305524310   1.523707158  -0.523550886
#>  [41]   1.081137113  -0.771507970  -4.681502143   2.811215469   0.475459976
#>  [46]  -0.325237070   2.250767291  -0.978740615   1.584376669   2.378085187
#>  [51]  -4.661677271  -4.764178099   1.205992020   1.158758325  -2.515943133
#>  [56]  -4.544824902   3.591124934   2.040777664   0.519959807  -2.375559110
#>  [61]  -2.182892746   0.735804490 -16.807740169  -0.150484053   0.003088779
#>  [66]   2.197616194   1.675619900  -1.737819157  -2.218907775  -2.040231257
#>  [71]  -2.397022398  -3.530032722   1.407600897   0.826504115   1.168269137
#>  [76]   1.402552771   5.046839791  -3.007756138   0.351734455   2.430397343
#>  [81]  -0.280410544   3.262046881  -2.772262091  -4.139161708  -0.290250404
#>  [86]  -1.635337749  -3.145004332   0.363281183  -2.269407633   0.540487872
#>  [91]   0.572709591   2.134171331   1.925344097  -1.356746067   0.860638154
#>  [96]  -1.546727114   2.246889050  -2.203271838  -2.173572766   1.729366154
#> 
#> $mu
#>   [1] 3.604136e+00 3.581501e-02 9.297985e-01 1.603154e-01 3.043425e-04
#>   [6] 5.922025e-03 1.290502e+01 1.588964e-01 5.443172e-01 2.504884e+01
#>  [11] 4.628660e+00 2.537923e+00 3.054370e+00 1.831369e+01 2.803308e+00
#>  [16] 7.061112e-03 3.052475e+00 7.577659e+00 3.972166e+02 4.789594e+00
#>  [21] 1.345201e+00 6.581414e-02 3.806780e-01 3.246138e-01 6.888421e+00
#>  [26] 4.773531e+00 2.017562e-03 4.492383e+00 3.743166e-02 2.917833e+00
#>  [31] 1.297303e+00 7.380201e+00 4.649351e+00 8.858473e-03 3.225222e+00
#>  [36] 4.408034e-01 1.880603e+00 7.367370e-01 4.589207e+00 5.924132e-01
#>  [41] 2.948030e+00 4.623154e-01 9.265086e-03 1.663012e+01 1.608754e+00
#>  [46] 7.223561e-01 9.495018e+00 3.757841e-01 4.876251e+00 1.078423e+01
#>  [51] 9.450598e-03 8.529896e-03 3.340071e+00 3.185975e+00 8.078668e-02
#>  [56] 1.062203e-02 3.627486e+01 7.696592e+00 1.681960e+00 9.296250e-02
#>  [61] 1.127150e-01 2.087160e+00 5.017544e-08 8.602914e-01 1.003094e+00
#>  [66] 9.003525e+00 5.342106e+00 1.759036e-01 1.087278e-01 1.299986e-01
#>  [71] 9.098848e-02 2.930396e-02 4.086141e+00 2.285316e+00 3.216421e+00
#>  [76] 4.065565e+00 1.555302e+02 4.940241e-02 1.421531e+00 1.136340e+01
#>  [81] 7.554735e-01 2.610291e+01 6.252042e-02 1.593621e-02 7.480762e-01
#>  [86] 1.948865e-01 4.306674e-02 1.438040e+00 1.033734e-01 1.716844e+00
#>  [91] 1.773065e+00 8.450041e+00 6.857508e+00 2.574973e-01 2.364669e+00
#>  [96] 2.129438e-01 9.458266e+00 1.104412e-01 1.137704e-01 5.637080e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                 x1           x2          x3          x4           y
#>   [1,] -3.02292204 -0.221573450  0.62320414 -0.77754003   0.9053320
#>   [2,]  0.59664912 -0.262722165  0.89539336  0.64814176   2.2876949
#>   [3,]  0.29936238 -2.042738430  0.10931541 -2.18230031 -24.0990041
#>   [4,]  0.50100161 -1.239704420  0.06392608  0.92198091  -4.2818922
#>   [5,] -1.94119338  0.549337377  0.68139001  0.53826820   5.8609577
#>   [6,]  1.20694070 -0.962252744  0.36515973  1.00617389  -1.5927432
#>   [7,] -1.15023725  1.001945728 -1.84432167 -0.45717223 -19.5961394
#>   [8,] -0.03313802 -0.040987307  0.59126567  2.03278814 -23.3621940
#>   [9,]  0.76304264  1.463714711 -0.30819787  0.03684490  -4.2944530
#>  [10,] -0.03807070  0.606856162  0.15941927 -0.68258282   0.1599139
#>  [11,] -0.96150972  0.722073408 -0.79623152  1.40482401 -12.2787315
#>  [12,] -0.97491486  0.037885503  0.25259436 -1.53042777 -14.8599498
#>  [13,] -0.17992511 -0.709692293 -0.74945842  0.35373825 -13.1399715
#>  [14,] -0.42548355 -0.332137963  0.66010099 -1.81211183 -14.8403616
#>  [15,]  0.25276004 -1.066484706 -0.71027310 -0.28259720  -4.6739612
#>  [16,] -0.52097322 -1.188501488 -0.21989006  1.10573327  -6.9855455
#>  [17,] -0.02097976  0.198974634 -0.15359290 -1.05743949 -12.0566230
#>  [18,] -0.04722104 -1.291822213  0.58032646  2.20376590 -20.3374066
#>  [19,]  0.81948309  1.048159160  0.23706677  0.06176255  -7.8741741
#>  [20,]  1.09462228 -0.365430486 -0.80850672 -0.09415973   1.2044069
#>  [21,]  0.38320850  0.903303053 -1.21369646 -0.71715814 -11.5792698
#>  [22,]  1.25881622  0.699437807  0.08425759  0.35613894   7.7527365
#>  [23,] -0.72057943 -0.823738829  0.51976098 -0.69547176   0.9903209
#>  [24,]  0.84631112  1.211556485  0.39518189 -0.64421559  -8.3845230
#>  [25,]  1.17651538  2.008841372 -0.94129862 -0.48613132  -5.7390719
#>  [26,]  3.05378294 -0.448360270  0.91566657  0.94326220   5.0610012
#>  [27,]  0.36586293 -0.848738345  0.68581436 -0.43650463   1.8181023
#>  [28,] -0.09023079 -0.438251903 -0.10734285  0.88540169  -4.1466718
#>  [29,] -0.81658863  1.276351983 -0.36453915  0.45984564  -0.2929023
#>  [30,]  0.22280058 -0.659007900  0.55803281  0.26936584   5.4402487
#>  [31,]  0.49447107  1.906570829  0.82650144  0.85817624  -3.5307177
#>  [32,]  1.16695484 -0.926302154 -0.82087225  1.24732435 -13.9491098
#>  [33,]  0.21601407 -0.232561074 -2.03074556  0.40960268  -7.9998133
#>  [34,] -1.41920975 -0.778889892  0.38206941 -0.43755756  -4.6738105
#>  [35,]  1.13172396 -1.068536434 -0.31249035 -1.86665656 -17.7637628
#>  [36,]  1.52023368 -1.172680382 -0.20116707 -0.50774945  -6.3483475
#>  [37,] -0.55256480  0.135099745  0.81412888 -0.85046892  -1.1684274
#>  [38,]  0.72179646 -0.410217583  0.03280745  0.42437963  -4.4246902
#>  [39,] -0.19950942 -0.241676815 -1.08454179  0.74736913 -14.0310463
#>  [40,] -0.55087792  1.581737312  0.97086226 -2.06620803 -10.9517943
#>  [41,] -2.32634011  0.564783701  0.38995473  1.57716857 -10.8739050
#>  [42,] -0.37305840 -0.979726524 -1.30498788 -0.26314549 -16.7496974
#>  [43,]  0.69442555 -0.166423409  0.56893742  0.68489061  -4.0675900
#>  [44,] -0.93877495 -1.351514134  1.31353582  0.14670214  12.1998703
#>  [45,] -0.73073953  0.156663387 -0.10980248  0.53128376  -2.8889644
#>  [46,] -0.08067286  0.186271803 -1.34202160 -0.09837734 -15.8975677
#>  [47,]  1.28589419 -1.488541678 -0.71481233 -1.00872654  -2.4922507
#>  [48,] -0.17015334 -0.254675371 -1.48778692  0.99620264 -18.1345533
#>  [49,]  1.25289553 -0.089895298  0.63990607  1.07384246   0.3488609
#>  [50,]  1.53442235 -0.354719160  1.77341260 -0.49702809  11.3726045
#>  [51,] -0.13084939 -1.489204644  0.46270958 -1.36012422  -6.4152792
#>  [52,] -1.24713086  2.243045456  1.07018047 -1.51125757 -13.0553824
#>  [53,]  0.93613488 -0.203741082 -1.25120255 -1.15267352 -16.2419652
#>  [54,]  0.41483024  0.405525928  1.40900955  1.29575836   4.4784778
#>  [55,] -0.24038271 -0.065371595 -0.48830463  0.04923462 -12.2869657
#>  [56,]  0.03242107 -0.965049681 -1.27426075 -0.53854180 -13.0833811
#>  [57,] -0.92313332  1.030137680  0.69292813 -0.54763140  -2.0549091
#>  [58,] -1.35734773 -0.289512495 -0.36854504 -0.51610119  -5.0802888
#>  [59,]  0.92459799  0.257685761  0.29029827  1.08902987 -10.3472005
#>  [60,] -0.30678578 -1.777986539 -1.57369223  0.69920320 -20.8974615
#>  [61,] -0.38561203  2.251331068  0.58166212  0.17055253  -4.5492953
#>  [62,]  0.93432429 -0.694968821  0.82599632 -0.05462794   1.7051550
#>  [63,]  1.08705845  1.477941831  0.85875259 -0.42054791   7.3127987
#>  [64,]  0.27927445 -2.471690086  0.83187728  1.10961060   7.3371778
#>  [65,]  0.20958244 -2.127720162 -0.40029211  0.73746534  -2.1028111
#>  [66,] -0.74287043 -0.176543215  0.82925460  0.64637403   5.8467913
#>  [67,]  0.35612404 -0.844157028 -0.49787407 -0.19164813  -4.1333225
#>  [68,] -0.11601959 -0.892960085  0.62976470 -0.37609493  -6.3346616
#>  [69,] -2.24365575 -0.747156357 -0.12389107 -0.01237899  -8.4575240
#>  [70,] -1.96664215  0.135217544 -1.21367798  0.99352826 -16.7969523
#>  [71,]  0.30938248 -1.067725184  0.48862987 -0.72193270   3.8775356
#>  [72,]  1.47243181 -1.418945000  0.23769458  0.11543307   4.6048813
#>  [73,] -1.49412380 -0.004604868  0.17960531  0.68229152  -6.6707294
#>  [74,] -0.23559954  1.113249230 -1.72109576 -0.64595046 -21.4690827
#>  [75,]  1.27344049  1.080871268 -0.64899584 -0.26174355  -5.9376380
#>  [76,]  0.22674058  0.275063834  1.02209338  0.17842203   7.7823937
#>  [77,]  1.79289129 -1.097306196  1.59373551  0.98169115  -3.3924861
#>  [78,] -1.05051045 -0.388075322  0.44915710  0.43292222   2.1413257
#>  [79,] -1.68123049 -1.184417160 -0.13724890 -0.58469739  -1.1030914
#>  [80,] -1.92277248  1.580300636 -0.15597480 -0.86120200  -0.8369439
#>  [81,] -1.09699546  0.058772332  0.92339752 -1.65724519 -17.2922577
#>  [82,]  0.62514219 -0.411500997 -0.13170658  1.36732384 -11.7872175
#>  [83,] -0.62156464  0.204369778  1.06801895 -0.50317096   5.9146181
#>  [84,]  0.87497484 -1.146969563 -1.70523403 -0.37242118 -16.1335145
#>  [85,]  2.01385180 -0.132464598 -1.07403833  0.01626614  -6.8805696
#>  [86,]  0.62839346 -0.153862556 -1.56247205 -1.29145701 -24.1731634
#>  [87,]  0.09418507 -0.131788195 -1.88018227 -0.69373315 -14.1775509
#>  [88,] -1.22270601 -0.558939000  2.35994008  0.14391888  10.9263773
#>  [89,]  0.43703045  2.185015238  0.25635893  1.28933930  -9.4027879
#>  [90,]  0.61299061  1.616208707  0.25505062 -0.35923254   0.3497996
#>  [91,]  0.17253001  0.994511949 -1.85692269  1.10058936 -17.7910913
#>  [92,] -0.75645924  0.218911233 -0.02412973  0.85629281  -2.4953269
#>  [93,] -1.36414614  0.309416073  1.64539945 -1.90697966 -13.2964736
#>  [94,]  1.38966100  0.294683693 -1.23990188 -0.66121079 -10.0063117
#>  [95,]  0.40033588 -0.592052743 -0.97081352  0.52224980  -4.1818590
#>  [96,] -1.09275465  2.167225421  0.81429526  0.95811323  -7.4431640
#>  [97,] -1.13393049  0.790030830 -0.41500224  0.17897704  -6.6945849
#>  [98,]  1.10836159  0.830142797  0.03419532  1.79394198 -12.5876369
#>  [99,] -0.73158016 -2.284954321  2.01912330 -1.37590557   8.1232766
#> [100,]  0.22318677  0.618231473 -1.02553939  0.15090515   1.6311049
#> 
#> $eta
#>   [1]  -0.23770236   1.11068703 -22.77056941  -3.71138431   6.87340348
#>   [6]  -3.41548766 -21.19393565 -22.55323941  -5.35313440  -0.11634413
#>  [11] -12.44653622 -14.77420430 -12.12251021 -12.81433060  -3.59217658
#>  [16]  -7.82218552 -11.59586527 -20.99905902  -8.55001190   0.16881477
#>  [21]  -9.79255681   6.28197783   0.83062320  -8.01780809  -7.46435901
#>  [26]   5.14147808   1.79369350  -4.77406383  -1.48870764   6.69920350
#>  [31]  -1.91011863 -13.86694803  -8.11457394  -5.11525924 -19.50459682
#>  [36]  -4.57325475  -2.36925731  -3.13356394 -13.81896380 -11.44998078
#>  [41] -12.05060448 -16.98006610  -5.23050953  11.78086283  -2.14661554
#>  [46] -15.16213252  -2.36395934 -17.75269195  -0.36569504  10.49625444
#>  [51]  -8.00425505 -13.56681453 -16.77218122   2.94607372 -11.88546171
#>  [56] -13.42797109  -0.32843827  -6.22674263  -9.71535368 -20.57390918
#>  [61]  -3.33585762   0.82138051   8.58060284   6.04500276  -2.84178864
#>  [66]   6.53530640  -4.29133037  -5.69976669  -9.09482052 -16.43529154
#>  [71]   1.07618889   4.72112685  -6.38792458 -22.80319473  -4.79052252
#>  [76]   8.07664414  -3.34935730   2.61517421   0.85888099   0.18784587
#>  [81] -16.51983845 -11.18308594   4.26849467 -16.54829106  -8.45590366
#>  [86] -23.23969355 -13.35969548  10.20773552  -9.50065294   0.02063764
#>  [91] -17.72088124  -2.87765780 -14.08405828  -8.80005862  -3.86721210
#>  [96]  -7.36454666  -6.89470902 -14.25288351   8.75211903   1.04471889
#> 
#> $mu
#>   [1]  -0.23770236   1.11068703 -22.77056941  -3.71138431   6.87340348
#>   [6]  -3.41548766 -21.19393565 -22.55323941  -5.35313440  -0.11634413
#>  [11] -12.44653622 -14.77420430 -12.12251021 -12.81433060  -3.59217658
#>  [16]  -7.82218552 -11.59586527 -20.99905902  -8.55001190   0.16881477
#>  [21]  -9.79255681   6.28197783   0.83062320  -8.01780809  -7.46435901
#>  [26]   5.14147808   1.79369350  -4.77406383  -1.48870764   6.69920350
#>  [31]  -1.91011863 -13.86694803  -8.11457394  -5.11525924 -19.50459682
#>  [36]  -4.57325475  -2.36925731  -3.13356394 -13.81896380 -11.44998078
#>  [41] -12.05060448 -16.98006610  -5.23050953  11.78086283  -2.14661554
#>  [46] -15.16213252  -2.36395934 -17.75269195  -0.36569504  10.49625444
#>  [51]  -8.00425505 -13.56681453 -16.77218122   2.94607372 -11.88546171
#>  [56] -13.42797109  -0.32843827  -6.22674263  -9.71535368 -20.57390918
#>  [61]  -3.33585762   0.82138051   8.58060284   6.04500276  -2.84178864
#>  [66]   6.53530640  -4.29133037  -5.69976669  -9.09482052 -16.43529154
#>  [71]   1.07618889   4.72112685  -6.38792458 -22.80319473  -4.79052252
#>  [76]   8.07664414  -3.34935730   2.61517421   0.85888099   0.18784587
#>  [81] -16.51983845 -11.18308594   4.26849467 -16.54829106  -8.45590366
#>  [86] -23.23969355 -13.35969548  10.20773552  -9.50065294   0.02063764
#>  [91] -17.72088124  -2.87765780 -14.08405828  -8.80005862  -3.86721210
#>  [96]  -7.36454666  -6.89470902 -14.25288351   8.75211903   1.04471889
#>