Title

Title

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>          x1.pen1      x1.pen2     x1.pen3    x1.null1      x2.pen1
#> 1    0.482601094  1.135747007 -0.07733336 -1.05347802  0.040870971
#> 2    0.502994576  1.127122638  0.07370019 -1.12011946  0.216122692
#> 3    0.170468483  0.954544420 -0.78922780 -0.64048809 -0.437076921
#> 4    0.332862191  1.076704719 -0.56128183 -0.80131205  0.029365868
#> 5   -0.450770823 -0.336681088 -0.88396035  0.12644003 -0.326803835
#> 6    0.391277544  1.110912772 -0.42956400 -0.87802884  0.304461672
#> 7    0.111220373  0.087543049  3.32835527 -2.17569621  0.205783750
#> 8    0.279247653 -0.994061418  1.25717875  1.11400101 -0.443747412
#> 9   -0.144049884  0.659440859 -0.99135123 -0.41225023  0.098378794
#> 10   0.285764646  1.044401668 -0.64335745 -0.74861160 -0.204973020
#> 11  -0.234610080 -0.738097059 -0.58750760  0.36039212 -0.133691653
#> 12   0.359531005  1.093303550 -0.50616121 -0.83443388  0.065054821
#> 13   0.210799485 -1.111022861  0.46811096  0.84758241  0.537656153
#> 14   0.296320501  1.051923981 -0.62640401 -0.75987610  0.211952977
#> 15   0.162024681  0.187135380  3.08618195 -2.10706411  0.158399691
#> 16  -0.486741198 -0.195324815 -0.94738284  0.05215369  0.138635993
#> 17  -0.336245304 -0.585151879 -0.72577261  0.26502018 -0.324315703
#> 18   0.425122274  1.126106987 -0.32962467 -0.93144090 -0.014029187
#> 19   0.042618984 -1.030871518 -0.11179187  0.61227650 -0.398540019
#> 20  -0.081702102 -0.917511813 -0.35376014  0.49451517 -0.177066060
#> 21   0.208719296 -1.111221860  0.45653716  0.84331396  0.202702908
#> 22  -0.388490894 -0.488009418 -0.79574259  0.20912555  0.542676408
#> 23   0.333598711  1.077181650 -0.55985412 -0.80219090  0.033227955
#> 24   0.310414235  1.061726715 -0.60253884 -0.77538255 -0.360106690
#> 25  -0.405628471 -0.451595746 -0.81910893  0.18881878 -0.175088207
#> 26   0.194455548 -1.111115549  0.38261442  0.81572341  0.207675631
#> 27   0.158373195 -1.102322922  0.22713033  0.75565318  0.545173838
#> 28  -0.497351485 -0.008842877 -1.00672969 -0.04393263  0.130200958
#> 29  -0.116046432 -0.881253895 -0.41046607  0.46424130  0.490925366
#> 30  -0.462622567 -0.298693826 -0.90267779  0.10626718 -0.222969998
#> 31   0.185192807 -0.445966701  2.79362766  1.55656810  0.114171935
#> 32  -0.491625466  0.044850902 -1.01897305 -0.07153827  0.164921212
#> 33  -0.174137063 -0.814863754 -0.50021265  0.41359517  0.004522933
#> 34   0.445397382  1.132634613 -0.25660415 -0.96842778  0.078786499
#> 35  -0.393394403  0.322848102 -1.04756705 -0.21724739 -0.370062718
#> 36   0.391677900  1.111116787 -0.42850581 -0.87861361  0.153343219
#> 37  -0.341110265 -0.576797612 -0.73226424  0.26009885 -0.174570755
#> 38  -0.412000270 -0.437272809 -0.82790931  0.18091154  0.224255218
#> 39   0.157752949  0.943957123 -0.80185569 -0.62987624 -0.134919825
#> 40  -0.339409359 -0.579732163 -0.72999482  0.26182480  0.196842532
#> 41   0.449090703  1.133543243 -0.24187028 -0.97570883  0.079453877
#> 42   0.109844461 -0.177879789  3.39225905  1.71602928 -0.022756502
#> 43   0.482052672  1.135832024 -0.08064459 -1.05197061  0.157705696
#> 44  -0.400467805 -0.462866127 -0.81203127  0.19507166 -0.138164340
#> 45  -0.290004267 -0.659265861 -0.66372619  0.30984587  0.052329548
#> 46   0.264420520 -0.800381591  1.89602541  1.30615583 -0.032487502
#> 47   0.105214565 -1.075373663  0.04980785  0.68312104  0.208230841
#> 48   0.293937422  1.050238752 -0.63029780 -0.75730784  0.143667141
#> 49   0.177274472 -0.416298248  2.86206619  1.57503172  0.111873408
#> 50  -0.498978562 -0.059055668 -0.99334058 -0.01813963 -0.075451821
#> 51   0.276528826 -1.021113463  1.14299306  1.07785629  0.120282768
#> 52  -0.279458210 -0.675013565 -0.64941801  0.31967748  0.153342241
#> 53   0.218460976  0.993427094 -0.73606417 -0.68262817 -0.225209078
#> 54  -0.051380196  0.754893373 -0.94965361 -0.47598104  0.073118166
#> 55   0.289734415  1.047247897 -0.63707005 -0.75281436 -0.013079941
#> 56  -0.020418280  0.784890733 -0.93308817 -0.49737651  0.410815911
#> 57   0.125964719  0.917032060 -0.83108897 -0.60423343  0.205717663
#> 58  -0.498980065 -0.057173453 -0.99387642 -0.01910669 -0.369805225
#> 59  -0.376906547  0.352145748 -1.04702653 -0.23311267 -0.150203414
#> 60  -0.438977784  0.228493412 -1.04454733 -0.16698288 -0.353956843
#> 61  -0.031362339 -0.966846203 -0.26412021  0.54009236  0.366067729
#> 62   0.523010893  1.067045858  0.48555932 -1.28555292 -0.049942462
#> 63   0.275518708  0.414755549  2.51894492 -1.94378940 -0.429568691
#> 64   0.499433517  1.129646263  0.04212917 -1.10651046  0.210132793
#> 65  -0.430368079 -0.392990754 -0.85379098  0.15671985 -0.063995860
#> 66  -0.249848322  0.536333234 -1.02583395 -0.33694630 -0.353895343
#> 67   0.456390606  0.817548235  1.41043639 -1.60598265  0.544310816
#> 68  -0.481224231 -0.224243130 -0.93576343  0.06719730  0.346067941
#> 69   0.278469414 -1.004357487  1.21505609  1.10074268  0.184741578
#> 70   0.103059331  0.897259829 -0.85026349 -0.58647257 -0.100054139
#> 71  -0.012044271 -0.984573200 -0.22713482  0.55818048  0.118054565
#> 72   0.225871976 -1.107530239  0.55955378  0.88085007  0.211639287
#> 73   0.164402851  0.191823464  3.07471227 -2.10380078  0.436488679
#> 74  -0.311798199 -0.625440905 -0.69307854  0.28911117  0.519556921
#> 75   0.276881379 -0.897707597  1.60110183  1.21932982  0.162470706
#> 76   0.240000084 -1.099922493  0.66129122  0.91697774  0.550373774
#> 77  -0.489084559 -0.181156078 -0.95283175  0.04480442  0.149721780
#> 78  -0.004301957 -0.991485175 -0.21182881  0.56555433  0.070209124
#> 79   0.412922207  1.121154016 -0.36832699 -0.91117163  0.137982871
#> 80   0.436535959  1.130078226 -0.29003324 -0.95168757  0.181656427
#> 81   0.218010197 -0.575515052  2.48562743  1.47252995  0.440725713
#> 82   0.279847229 -0.946567380  1.43590362  1.16934984 -0.227970625
#> 83  -0.080572909  0.725805219 -0.96401229 -0.45592312 -0.090218006
#> 84   0.198034147 -1.111363889  0.40034115  0.82239265 -0.442707636
#> 85  -0.461306293  0.169582371 -1.03912056 -0.13610400 -0.125055264
#> 86   0.460180723  1.135621016 -0.19430550 -0.99883266  0.144876164
#> 87   0.397218714  1.113887425 -0.41359072 -0.88680909  0.506305072
#> 88  -0.445955608  0.211406776 -1.04324815 -0.15799305  0.548217891
#> 89   0.482227459  1.135805532 -0.07959235 -1.05244986 -0.018913619
#> 90  -0.498617558 -0.034443916 -1.00013811 -0.03078258  0.208821525
#> 91  -0.481765963 -0.221614617 -0.93684746  0.06582728  0.364821367
#> 92   0.056877246 -0.001071667  3.77001117  1.81483672 -0.341636466
#> 93  -0.277181542  0.501054067 -1.03245844 -0.31638870  0.148489546
#> 94  -0.077212645 -0.922094507 -0.34610641  0.49851069  0.126895585
#> 95  -0.342752107  0.407694356 -1.04397700 -0.26361050 -0.417639196
#> 96   0.005932535 -1.000448786 -0.19112848  0.57542746  0.198153186
#> 97  -0.430774136  0.247436269 -1.04572202 -0.17698550  0.102158040
#> 98  -0.362783648 -0.538005375 -0.76119731  0.23754624 -0.410311035
#> 99   0.269315818 -1.053930255  0.98407801  1.02640344  0.203288941
#> 100  0.215204010 -1.110393783  0.49338496  0.85685728 -0.440120864
#>           x2.pen2      x2.pen3    x2.null1      x3.pen1     x3.pen2     x3.pen3
#> 1   -1.0649284519 -0.234453566 -0.68061167  0.414641463 -1.21132973 -0.26396723
#> 2    0.8999820198  0.697665885  0.63030708 -0.485269378 -0.26337478 -1.16476276
#> 3    0.0032956064  0.801148897  0.04145584 -0.416570069 -0.40453411 -1.15432495
#> 4   -0.0001921403 -4.196112105 -1.93974348  0.253746378  0.60260041  2.43446018
#> 5   -0.5600054794  0.510082844 -0.26444619 -0.410135446  0.51867277 -0.95326785
#> 6    0.9766010655  0.600919357  0.72340210  0.213441760  0.45872306  2.77482731
#> 7   -0.8586793653 -2.087443532 -1.34001398 -0.537957138  0.15046018 -1.11080437
#> 8   -0.0952163211  0.767812768 -0.01047861 -0.501529559 -0.22021104 -1.16490290
#> 9   -1.1059857761 -0.452318816 -0.77391416  0.275404698 -1.13668625 -0.62042486
#> 10   0.9347537863 -6.218437403 -2.48795914  0.327200983  0.97351562  1.40704668
#> 11   0.5548875884  0.857469050  0.35412471  0.023993660 -0.93156302 -0.92678789
#> 12  -0.1464184597 -3.873571597 -1.85161105 -0.061788739  0.97246509 -0.51531446
#> 13   1.0981814677 -0.148721574  1.18206383 -0.522677832 -0.15084464 -1.16226495
#> 14  -0.9924815321 -1.620272486 -1.19352466  0.329356258  1.01862841  1.24870636
#> 15  -0.5634413160 -2.898923603 -1.57929395  0.071534089  1.08728426 -0.27802129
#> 16  -1.1203567867 -0.659308571 -0.85660873  0.292707327  0.76141487  2.03153486
#> 17  -0.5652010726  0.506036382 -0.26750893  0.166862382  1.14835395 -0.03730715
#> 18  -1.0128684173 -0.075236016 -0.60709869 -0.338412878  0.63934121 -0.87289711
#> 19  -0.3778190115  0.633383923 -0.16135731 -0.091672735  0.94305211 -0.56019373
#> 20   0.8235946174 -5.978291506 -2.42289298  0.418180124 -1.21232761 -0.25055775
#> 21  -0.8308676978 -2.172866726 -1.36599179  0.324188274  0.94207302  1.50996795
#> 22   1.0927718773 -0.214036378  1.21312202 -0.469546984  0.39322758 -1.02016372
#> 23  -1.0582894516 -0.209980560 -0.66964135  0.261921654 -1.12724803 -0.64424987
#> 24  -0.4848150511  0.565127268 -0.22098773  0.382633542 -0.70291957  2.03082455
#> 25  -0.8168036874  0.259053013 -0.43072174  0.202336330 -1.08283067 -0.73644123
#> 26   0.8923646482  0.705134220  0.62217317 -0.431365073 -0.37774617 -1.15752437
#> 27   1.0886560679 -0.255256807  1.23233248  0.328857206  1.00070429  1.31337715
#> 28  -1.1188039897 -0.610409767 -0.83751807  0.420413432 -1.21291802 -0.24190428
#> 29   1.1002181261  0.163242283  1.02047682 -0.427649266  0.48492912 -0.97279094
#> 30  -0.7458260480  0.340997728 -0.38084732  0.327871846  0.98273440  1.37582462
#> 31   0.8055044087  0.772238795  0.53884102 -0.543381591 -0.03443438 -1.15006561
#> 32  -1.1175132485 -0.840810401 -0.92539026  0.007928244  1.03595580 -0.40047554
#> 33  -1.0315142999 -0.125045737 -0.63066639  0.122508288  1.12290303 -0.16040419
#> 34  -1.0940540728 -0.370299920 -0.73964030 -0.498228113 -0.22949615 -1.16498935
#> 35  -0.4597663571  0.582104188 -0.20682462  0.323163087  0.93324073  1.53799772
#> 36  -0.5382665381 -2.961518745 -1.59717139 -0.333613768 -0.53472093 -1.12991054
#> 37  -0.8175338175  0.258141938 -0.43125766  0.185667082  1.15704111  0.02275879
#> 38   0.9072739131  0.690219332  0.63824852 -0.506066401 -0.20691615 -1.16466878
#> 39   0.5535215717  0.857662293  0.35324634  0.317935126  0.89481984  1.65614500
#> 40  -0.7854946941 -2.306811400 -1.40629930 -0.225205672  0.79426446 -0.73919941
#> 41   0.7722579834  0.791058516  0.51052581 -0.402819597  0.95400103  5.99918297
#> 42   0.2089264210 -4.650340490 -2.06308832  0.414919144 -0.78393701  1.80524397
#> 43  -0.5599541297 -2.907635939 -1.58178628 -0.546561254  0.03962601 -1.13727997
#> 44  -0.8671360792  0.192289013 -0.46903831 -0.470948711 -0.29716022 -1.16367851
#> 45  -1.0744398399 -0.272819445 -0.69759313  0.142527890  1.13514825 -0.10768110
#> 46   0.2477068355 -4.734153219 -2.08580052 -0.456886184 -0.32758945 -1.16195432
#> 47   0.8928666512  0.704651468  0.62270430  0.293963629  0.76714918  2.01624553
#> 48  -1.1208032737 -0.690293177 -0.86857566 -0.074474217 -0.83762633 -0.99963310
#> 49   0.8033180704  0.773575183  0.53692744 -0.132874250  0.90029108 -0.61876012
#> 50  -0.9451127437  0.068049927 -0.53585127 -0.496544998  0.32098366 -1.05212246
#> 51  -0.3835749734 -3.333222832 -1.70207621  0.273771481  1.16709708  0.41895304
#> 52  -0.5382617176 -2.961530666 -1.59717479  0.252446824 -1.12046700 -0.66027335
#> 53  -0.7423276334  0.344732813 -0.37848891  0.499323814 -1.18854652  0.28269292
#> 54   0.7661415012  0.794190718  0.50548875  0.503552467 -1.07171467  0.86973365
#> 55   0.1703200373 -4.566833927 -2.04045127  0.493212503 -1.02200519  1.05694480
#> 56   1.0575700494  0.415366652  0.86523877 -0.299046525 -0.58209393 -1.11705050
#> 57  -0.8580377323 -2.089446707 -1.34062571 -0.532446751  0.18377740 -1.10109132
#> 58  -0.4604330005  0.581660631 -0.20719977 -0.466651812  0.40021414 -1.01683105
#> 59   0.7165951626 -5.747132167 -2.36026168  0.002824463 -0.91205221 -0.94384323
#> 60   0.2526754208  0.855900282  0.17568835  0.287658275 -1.14503171 -0.59764932
#> 61   1.0255756562  0.506137472  0.80009055 -0.444201248  0.45084506 -0.99133721
#> 62  -0.9742993551  0.011927855 -0.56441789 -0.274705590 -0.61372724 -1.10715229
#> 63   0.0478179937  0.814056822  0.06502512 -0.402749950 -0.42835119 -1.15097786
#> 64  -1.0206531427 -1.501652673 -1.15492806 -0.525885538 -0.13803818 -1.16139731
#> 65  -0.9583995715  0.043342293 -0.54853995  0.480841318 -1.20940901  0.08650329
#> 66  -0.4997552762  0.554690325 -0.22950437  0.320723420  1.10528404  0.88165823
#> 67   1.0902339230 -0.240059374  1.22528345 -0.438976571  0.46186278 -0.98546745
#> 68   0.5488377044 -2.286364252  1.99139380 -0.064173667 -0.84785241 -0.99273205
#> 69  -1.1024087160 -1.026804183 -0.99292972 -0.539398788 -0.06801930 -1.15457773
#> 70   0.5915887219  0.851426076  0.37808562  0.265357222  0.64706308  2.32511704
#> 71  -0.3736051241 -3.356546094 -1.70859503  0.341866270 -1.17873635 -0.48134190
#> 72  -0.9373998805 -1.827047990 -1.25936269 -0.284436525 -0.60123607 -1.11119115
#> 73   1.0739131351  0.350959853  0.90792688 -0.527507488  0.20858319 -1.09332001
#> 74   1.1047700879  0.007636467  1.10417354  0.323156849  0.93318880  1.53816146
#> 75  -1.1184142307 -0.821478806 -0.91820886 -0.449198715 -0.34330720 -1.16078123
#> 76   1.0456194904 -0.552240193  1.36329267  0.169061475  1.14944388 -0.03056003
#> 77  -0.5205449354 -3.005179630 -1.60960203  0.479028553 -0.96912527  1.23968392
#> 78  -0.1678757346 -3.825704806 -1.83847290 -0.533013161 -0.10555476 -1.15866912
#> 79  -1.1202711489 -0.655391206 -0.85508870  0.200589055 -1.08147026 -0.73886442
#> 80  -0.6884240457 -2.576169873 -1.48594906  0.266914736  1.16929516  0.37695602
#> 81   1.0764257594  0.339198986  0.91545941  0.495519574 -1.19509228  0.23141761
#> 82   0.4430280522  0.866584788  0.28484530  0.143105365 -1.03520513 -0.81128476
#> 83  -0.9275562981  0.098848687 -0.51977348 -0.479300782  0.36874236 -1.03150057
#> 84  -0.1380662220  0.751231398 -0.03306148  0.022750385 -0.93042682 -0.92781326
#> 85  -0.8841872229  0.167604480 -0.48274728  0.036575765  1.05990545 -0.34793362
#> 86  -1.1208514881 -0.697962981 -0.87152286 -0.511377512 -0.19042988 -1.16419887
#> 87   1.1040163423  0.087946694  1.06186436 -0.481155216 -0.27342092 -1.16453060
#> 88   1.0285333503 -0.643099758  1.40117798  0.315107567  1.12218322  0.79071416
#> 89   0.6756164218  0.830620846  0.43608654  0.230433492  0.51779284  2.63727579
#> 90  -1.0321937211 -1.449386014 -1.13771076 -0.179595635 -0.72685483 -1.06202504
#> 91   1.0246343226  0.508358152  0.79840847 -0.122668923 -0.78837691 -1.02998120
#> 92  -0.5279198719  0.534355075 -0.24571272  0.223270551  1.16915518  0.16256189
#> 93  -1.1208488337 -0.721441901 -0.88050858 -0.221085002 -0.85651099  5.56080555
#> 94  -1.1179494286 -0.592181041 -0.83033478  0.446845341 -1.21707634 -0.12579917
#> 95  -0.3081224686  0.672296303 -0.12351057  0.086311801  1.09818580 -0.24607825
#> 96   0.8837264661  0.713240564  0.61313848 -0.442480852 -0.35658905 -1.15963773
#> 97  -1.1079793000 -0.469307561 -0.78089794  0.117288660 -0.11719184  3.49712687
#> 98   0.1215414428  0.832475864  0.10433581 -0.377240413  0.57699637 -0.91660126
#> 99   0.8883919051  0.708908958  0.61799377  0.269045749  0.66159161  2.28886922
#> 100 -0.0226606271  0.793005505  0.02775334  0.284806405  0.72653905  2.12324285
#>        x3.null1      x4.pen1     x4.pen2      x4.pen3      x4.null1     x5.pen1
#> 1    0.93381824 -0.425553154 -0.18976441  0.716973005  0.1128973500  0.31663042
#> 2    0.11246032  0.063898966 -1.04692790  0.097299068  0.6426388274 -0.14346997
#> 3    0.18851854  0.494760318  1.12594385 -1.527268093 -1.6131412589 -0.40526862
#> 4   -1.66185280 -0.431648642  0.09630804  0.774969016 -0.0300332571  0.10405594
#> 5   -0.30189710 -0.304698427 -0.52544861  0.581830372  0.2879725678  0.37522997
#> 6   -1.75692462  0.332142042 -1.23829501 -0.958223105  1.1192472040 -0.05766666
#> 7   -0.10306038  0.162319244 -1.14995593 -0.130380011  0.7637597759  0.33443954
#> 8    0.08964893 -0.064338797  0.79094440  0.697951937 -0.4087748655  0.37165749
#> 9    0.74412667 -0.251291048 -0.62236842  0.525713957  0.3423311480 -0.26093140
#> 10  -1.35841345 -0.112507121 -0.83030129  0.366901687  0.4707255039  0.29436169
#> 11   0.52425222  0.358537385 -1.09473175 -1.851930464  1.4321498376  0.03731884
#> 12  -0.60965603  0.249964496  1.14751977  0.404053959 -0.7062115468  0.03863809
#> 13   0.05329078  0.335376547 -1.23604351 -0.993124654  1.1324619593  0.17887330
#> 14  -1.30851916  0.236913432 -1.21259513 -0.378073064  0.8808620521  0.27993769
#> 15  -0.73230245  0.384686323  1.27049223  0.110744847 -0.8998105369 -0.41824248
#> 16  -1.54631883  0.317903202 -0.87467546 -2.621341044  1.6729236137 -0.44182072
#> 17  -0.84182330 -0.438423084 -0.10000645  0.740353471  0.0678872229  0.19832718
#> 18  -0.37278246  0.205926006  1.10172465  0.467987805 -0.6556089428  0.37415577
#> 19  -0.58419124  0.302433963 -1.24344949 -0.713170962  1.0233032670  0.36755877
#> 20   0.94022603  0.059229891  0.93834940  0.619162874 -0.5115795029 -0.36504399
#> 21  -1.39029439 -0.292462637 -0.54891705  0.569078148  0.3009102783  0.04739860
#> 22  -0.23200891 -0.441808665 -0.02145412  0.756885573  0.0287034555  0.04957227
#> 23   0.72965820  0.203868567 -1.18717563 -0.257016805  0.8251707740  0.38888910
#> 24   1.75717494 -0.126053889 -0.81189313  0.383681944  0.4584414122 -0.42516002
#> 25   0.67049577  0.497859274  1.31247057 -0.485633977 -1.2013871831 -0.28714850
#> 26   0.17387637  0.462965911  1.31529912 -0.211751110 -1.0726722147  0.37525160
#> 27  -1.32902761  0.442099750  0.95731232 -2.100685451 -1.8106292805  0.37170487
#> 28   0.94433958  0.305823835  1.20237923  0.304536283 -0.7777107391 -0.34520190
#> 29  -0.28278109  0.174646209  1.06812738  0.507293736 -0.6220905322  0.16796320
#> 30  -1.34865909  0.234639670 -1.21099643 -0.369012004  0.8767837135 -0.15787868
#> 31  -0.00719007 -0.178388147 -0.73745863  0.445517332  0.4108622622  0.31924067
#> 32  -0.67125808  0.451008785  1.31123555 -0.146620823 -1.0399676334 -0.25582437
#> 33  -0.78731704  0.218261238 -1.19878984 -0.307102930  0.8485380241 -0.42332059
#> 34   0.09454226 -0.213787651 -0.68353432  0.485154605  0.3781102898 -0.43796325
#> 35  -1.39890700 -0.037235356 -0.92739444  0.265884184  0.5401006242 -0.22916981
#> 36   0.26157976 -0.183473925  0.63371872  0.749038907 -0.3131684179 -0.08945643
#> 37  -0.86745084 -0.301624776  0.44657398  0.780259940 -0.2096034498  0.36557940
#> 38   0.08265439 -0.373562579  0.29674123  0.787219116 -0.1313310184 -0.29955181
#> 39  -1.43490012 -0.066823272 -0.89019161  0.307370926  0.5124809727 -0.18388265
#> 40  -0.47226556  0.360007561 -1.17323846 -1.480980028  1.3078847396 -0.39652739
#> 41   2.86813061 -0.150251912 -0.77817345  0.412819326  0.4365023159 -0.32011132
#> 42   1.68872490 -0.142576732  0.69012514  0.733755320 -0.3463008594 -0.12711481
#> 43  -0.04552924  0.181547467 -1.16778906 -0.186025123  0.7911827448  0.38390857
#> 44   0.13044058  0.403651226  1.28434126  0.049698082 -0.9349073154  0.30402123
#> 45  -0.81101150  0.268789117  1.16646189  0.373110098 -0.7292771302 -0.08013828
#> 46   0.14674360  0.294196747 -1.24181241 -0.660138461  1.0017233134  0.40068123
#> 47  -1.54185896 -0.009453973 -0.96134229  0.224278402  0.5667439804 -0.04096695
#> 48   0.45298101  0.364993286  1.25486513  0.167022387 -0.8661901378 -0.14316855
#> 49  -0.54956721 -0.151024225 -0.77707842  0.413732949  0.4358011408  0.26825397
#> 50  -0.19297654  0.357675513 -1.08613929 -1.887340017  1.4436859041 -0.42073393
#> 51  -1.02422062 -0.054285853 -0.90609587  0.290110179  0.5241082467  0.21333146
#> 52   0.71975312 -0.260499124 -0.60658561  0.535496738  0.3333010378 -0.16031539
#> 53   1.16951225  0.255097693 -1.22440296 -0.455293384  0.9150797213  0.33221839
#> 54   1.38545960 -0.198319850 -0.70750156  0.468019557  0.3925125197  0.08787407
#> 55   1.44931783 -0.355782347 -0.41557089  0.635222535  0.2289226447 -0.03764445
#> 56   0.28914636  0.316866652 -1.24364604 -0.819086004  1.0654906933  0.38224654
#> 57  -0.12045283  0.080291311 -1.06517871  0.064684549  0.6610249565 -0.20192776
#> 58  -0.23582432 -0.271579362  0.49880627  0.774265141 -0.2377145483 -0.42020604
#> 59   0.50866290 -0.276414565 -0.57847169  0.552280093  0.3173975078 -0.38107844
#> 60   0.75768548 -0.320957033 -0.49280374  0.598747471  0.2701827158 -0.11865244
#> 61  -0.26372413  0.134081833  1.02345579  0.551978102 -0.5811196004  0.08126200
#> 62   0.30791847  0.505668622  1.30471876 -0.587558111 -1.2463234917  0.26323425
#> 63   0.20163591 -0.031992088  0.83060220  0.680332147 -0.4348158564  0.27934881
#> 64   0.04661071 -0.050915939  0.80751983  0.690859287 -0.4195416666  0.13963930
#> 65   1.09002783  0.241617675 -1.21582139 -0.397215276  0.8894330459  0.19959978
#> 66  -1.18820562  0.300346755  1.19719288  0.315393863 -0.7702635532 -0.34788453
#> 67  -0.26985837 -0.421382373 -0.21012229  0.710984385  0.1231648792 -0.22402534
#> 68   0.46032738  0.044218880  0.92092723  0.630611930 -0.4984374145  0.38873273
#> 69   0.01021444 -0.263550207 -0.60128071  0.538725745  0.3302827500 -0.33271102
#> 70  -1.63085437  0.338644472  1.23237874  0.233316265 -0.8247814856  0.36982766
#> 71   0.82334043 -0.441136914 -0.05933836  0.749363214  0.0475878443 -0.26492359
#> 72   0.30046834  0.132514238 -1.12060050 -0.052457629  0.7240581612  0.32702310
#> 73  -0.13344663 -0.324653486  0.40328409  0.783782708 -0.1866663307  0.37861181
#> 74  -1.39895723 -0.432483816 -0.14935131  0.728103279  0.0925863825 -0.21731750
#> 75   0.15520941 -0.053222050  0.80468461  0.692099383 -0.4176885031 -0.43505627
#> 76  -0.84473165  0.111148978 -1.09847446 -0.001867852  0.6973727239 -0.14159630
#> 77   1.50989755 -0.193256940 -0.71520590  0.462350912  0.3971932686  0.11539843
#> 78   0.02970211  0.202511684  1.09809720  0.472505571 -0.6518600058 -0.37971142
#> 79   0.66886311 -0.384804845  0.26783331  0.786925423 -0.1165382564  0.37356106
#> 80  -1.00845051 -0.255471958 -0.61524366  0.530163101  0.3382452589  0.37860459
#> 81   1.14917832 -0.194492189 -0.71333234  0.463736805  0.3960526198 -0.11500493
#> 82   0.61781317  0.503449904  1.30751882 -0.555189283 -1.2322024026  0.37408670
#> 83  -0.21869728  0.309508677 -1.24403253 -0.762721965  1.0431857752  0.37857993
#> 84   0.52333092 -0.076730439 -0.87747006  0.320707557  0.5033567640  0.20146489
#> 85  -0.69797555 -0.439196206  0.03641039  0.766770184 -0.0001336427  0.17178958
#> 86   0.07399921 -0.266674154  0.50696035  0.773150234 -0.2421510909  0.36395348
#> 87   0.11779442 -0.360464749  0.32786643  0.786960222 -0.1473540268  0.34269503
#> 88  -1.15720812  0.166340093 -1.15376298 -0.141637058  0.7693666019  0.39946883
#> 89  -1.71874377 -0.114407146 -0.82773860  0.369276994  0.4690013914 -0.42412007
#> 90   0.37804070  0.515809751  1.27205449 -0.848883508 -1.3558775288 -0.14944165
#> 91   0.41870481 -0.290866816  0.46576417  0.778284241 -0.2198722947 -0.44123528
#> 92  -0.92499388 -0.428136433  0.11588532  0.777200608 -0.0398277720 -0.43854548
#> 93  -2.49938242  0.405758266  1.28579456  0.042428491 -0.9390000464  0.20317826
#> 94   0.99801803  0.286777978  1.18414922  0.341178403 -0.7522562441 -0.16673075
#> 95  -0.74756251 -0.118458070 -0.82225420  0.374317165  0.4653270678 -0.43441494
#> 96   0.16238877  0.381544443  1.26807706  0.120168728 -0.8942698992  0.33287228
#> 97   2.17833252 -0.001483529  0.86720184  0.661947871 -0.4597848367 -0.42157422
#> 98  -0.33561289  0.110866860  0.99740148  0.574761386 -0.5587794104  0.29619015
#> 99  -1.62052342  0.279835320  1.17737388  0.353792933 -0.7432722115 -0.18227904
#> 100 -1.57294512 -0.441338067 -0.05410593  0.750451948  0.0449785646  0.32184946
#>         x5.pen2     x5.pen3     x5.null1
#> 1   -1.00989562 -1.05178257  0.922865827
#> 2    0.69126581  1.16792380 -0.571500381
#> 3    0.11042531  0.97933372 -0.212228937
#> 4    1.08151309  1.13740097 -0.892735796
#> 5   -0.93772027 -1.68996802  1.153398366
#> 6    0.83250270  1.17868548 -0.672587643
#> 7    1.36430822  0.73769604 -1.397837516
#> 8   -0.95267859 -1.60120789  1.122552052
#> 9    0.47598916  1.12241382 -0.432162482
#> 10   1.32890696  0.87262401 -1.273229134
#> 11   0.98129900  1.16645696 -0.795145542
#> 12   0.98331920  1.16606842 -0.796970284
#> 13   1.18773818  1.07580169 -1.018448927
#> 14  -0.49520102 -3.29584877  1.668976817
#> 15  -0.35033993  0.66852962  0.071625883
#> 16  -0.09062122  0.86473015 -0.091897111
#> 17  -0.98973209 -0.54660985  0.721267962
#> 18   1.37974230  0.51084020 -1.573652628
#> 19  -0.96434910 -1.52530144  1.095895350
#> 20  -0.47852724  0.54419616  0.159451683
#> 21  -0.90631817 -0.17353865  0.555426591
#> 22  -0.90774846 -0.17792987  0.557491234
#> 23   1.37120012  0.35904503 -1.676875276
#> 24  -0.32516206  0.69049246  0.055071964
#> 25  -0.59947636  0.40306404  0.249623141
#> 26  -0.83964069 -2.15318882  1.309429564
#> 27   1.37992508  0.53022134 -1.559779418
#> 28   0.28881203  1.05915957 -0.318335593
#> 29  -0.97609617 -0.45690074  0.682981539
#> 30   0.66653506  1.16433161 -0.554775404
#> 31  -1.00921014 -1.06826611  0.929111670
#> 32   0.48616970  1.12524810 -0.438496645
#> 33   0.03522636  0.93971115 -0.167505034
#> 34  -0.25933423  0.74462115  0.012651781
#> 35  -0.66971641  0.30674383  0.306731047
#> 36  -0.80427849  0.07661134  0.432206478
#> 37  -0.78111739 -2.38043776  1.383402617
#> 38   0.39538772  1.09774411 -0.382615777
#> 39   0.62092474  1.15652890 -0.524506026
#> 40  -0.41168864  0.61179336  0.112854152
#> 41   0.34925934  1.08187161 -0.354655693
#> 42   0.71892549  1.17138217 -0.590500041
#> 43   1.17791947 -0.51962723 -2.164467634
#> 44   1.33847044  0.84499190 -1.300515899
#> 45   0.79633297  1.17765842 -0.645629056
#> 46   1.32557596  0.05367180 -1.862799894
#> 47   0.85910785  1.17856754 -0.693012444
#> 48  -0.75698153  0.16624258  0.384947042
#> 49  -1.01069820 -0.80248592  0.826038714
#> 50  -0.34174780  0.67610726  0.065954353
#> 51  -0.99562927 -0.59491031  0.741524620
#> 52   0.66231657  1.16367315 -0.551945569
#> 53  -0.65139092 -2.82367392  1.523830220
#> 54  -0.93199311 -0.25868700  0.594931308
#> 55  -0.84578340 -0.01327831  0.477840515
#> 56   1.37705831  0.43699662 -1.624991716
#> 57   0.58843022  1.15007988 -0.503342691
#> 58  -0.34360790  0.67447412  0.067180178
#> 59  -0.44664816  0.57724367  0.136986186
#> 60  -0.77915521  0.12566424  0.406574282
#> 61  -0.92793884 -0.24424067  0.588305727
#> 62  -0.44990941 -3.42493734  1.708018051
#> 63  -1.01206280 -0.85234808  0.845782367
#> 64   1.13299669  1.11266779 -0.949961398
#> 65  -0.28760152 -3.86838041  1.840469466
#> 66  -0.50915420  0.51094824  0.181493289
#> 67  -0.67542969  0.29832890  0.311575181
#> 68   1.37139525  0.36115135 -1.675501491
#> 69  -0.53397344  0.48284373  0.199718403
#> 70  -0.95837297 -1.56503394  1.109881355
#> 71  -0.62780378  0.36571043  0.272147621
#> 72   1.35873204  0.76728605 -1.372182188
#> 73  -0.88000426 -1.97965342  1.251871146
#> 74  -0.68277144  0.28737230  0.317849882
#> 75  -0.03119247  0.90155199 -0.127767119
#> 76  -0.75843498  0.16365342  0.386338941
#> 77   1.09809119  1.13031015 -0.910531352
#> 78   0.19462000  1.01942677 -0.262207653
#> 79  -0.94554962 -1.64452983  1.137649987
#> 80  -0.90688432 -1.85269476  1.209106073
#> 81   0.73914973  1.17352110 -0.604607667
#> 82  -0.94328199 -1.65790472  1.142294720
#> 83  -0.90743844 -1.84995317  1.208176076
#> 84  -0.99101566 -0.55647279  0.725423993
#> 85  -0.97792665 -0.46767619  0.687627955
#> 86  -0.97213193 -1.47025600  1.076396349
#> 87   1.36980275  0.70138048 -1.428293125
#> 88   1.28254798 -0.14115462 -1.970781265
#> 89  -0.32923445  0.68698956  0.057736133
#> 90  -0.75113775  0.17655793  0.379385061
#> 91  -0.23189552  0.76588155 -0.004711628
#> 92  -0.05780749  0.88538463 -0.111746710
#> 93  -0.99170566 -0.56191001  0.727710765
#> 94   0.65115743  1.16186833 -0.544490633
#> 95  -0.28177315  0.72668118  0.026983039
#> 96  -1.00393918 -1.16233397  0.964419224
#> 97   0.04349146  0.94424818 -0.172431140
#> 98   1.33075762  0.86758499 -1.278286683
#> 99   0.62377711  1.15706026 -0.526378873
#> 100 -1.00843127 -1.08518326  0.935503178