Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>         x1.pen1      x1.pen2     x1.pen3    x1.null1      x2.pen1      x2.pen2
#> 1    0.39068576  1.201391973  0.19558824 -1.18295686 -0.157434222 -0.838110130
#> 2   -0.07558865 -0.971591368 -0.72393227  0.51455007  0.376610654  0.936135600
#> 3   -0.12564187 -0.927512486 -0.78399687  0.47947058 -0.077343684 -0.927804381
#> 4   -0.41855111 -0.584322293 -1.08881391  0.25802212  0.190634218 -1.176693956
#> 5    0.20427510  1.163153991 -0.68039460 -0.79591926  0.403795148  1.336596282
#> 6    0.37786153 -0.805003384  1.91531048  1.46138142 -0.278833319  0.638984452
#> 7   -0.53918182 -0.323003822 -1.21379751  0.11719698  0.364819648  1.334360547
#> 8   -0.37742371  0.482370264 -1.26222524 -0.28837187 -0.013160056 -0.994273260
#> 9    0.24414260  1.191723681 -0.58202131 -0.84806760 -0.522018789 -0.008245174
#> 10  -0.34118832 -0.694997857 -1.01301954  0.32289018  0.083371236  0.108829275
#> 11  -0.37432212 -0.650278230 -1.04561973  0.29616506 -0.497066333 -0.237505241
#> 12  -0.56488198 -0.196260871 -1.25261833  0.05263556 -0.200357209 -0.786289039
#> 13  -0.55097808  0.056492920 -1.29426863 -0.07292155  0.271082775 -1.229235470
#> 14   0.42531242 -1.109236442  0.89852121  1.16079799 -0.522393431 -0.059925497
#> 15   0.36426846  0.896140880  1.37966266 -1.57843578  0.314538956 -1.247645606
#> 16  -0.55301138 -0.268372119 -1.23210008  0.08916137  0.108157966  1.122721598
#> 17  -0.28173415 -0.767779305 -0.95334885  0.36822865 -0.227361775 -0.752067844
#> 18  -0.53838633  0.110208460 -1.29736159 -0.09946944  0.253120132 -1.219157590
#> 19   0.38541090  1.210666548  0.12561313 -1.15644995 -0.131999274  0.843424765
#> 20   0.04883371 -1.068938357 -0.55537616  0.60461056 -0.186277828 -0.803615455
#> 21  -0.07094918  0.896669879 -1.06652470 -0.53552369  0.280041982  1.284683063
#> 22  -0.31094541 -0.733078024 -0.98289548  0.34629791 -0.494032713 -0.249212712
#> 23  -0.49294383  0.248106133 -1.29616513 -0.16806139  0.405735006 -1.077132504
#> 24  -0.48938476  0.257049010 -1.29562644 -0.17254685 -0.010345738 -0.997090993
#> 25   0.14730512 -0.182436485  3.42429431  1.86201775 -0.217823085 -0.764311099
#> 26   0.39174543  1.022175187  0.98346724 -1.45465530  0.260103205 -1.223213685
#> 27   0.38384260  1.212860152  0.10726869 -1.14941592 -0.358550589 -0.560480842
#> 28  -0.23379316  0.697620411 -1.18850167 -0.40805621  0.382040927 -1.240858002
#> 29  -0.39705754 -0.617401058 -1.06783948  0.27697801 -0.148680759 -0.848329771
#> 30  -0.36901359  1.012942567  6.00622853  2.51294631  0.179850375 -1.168461781
#> 31  -0.34398572  0.537363098 -1.24776478 -0.31782947 -0.069906751  0.920501217
#> 32  -0.34760139 -0.686605686 -1.01935729  0.31781540 -0.002296748  1.000168615
#> 33   0.20072324  1.160411399 -0.68829812 -0.79156957 -0.025334831  0.973451821
#> 34  -0.27172757  0.645433696 -1.21087374 -0.37781635 -0.512932940 -0.160494565
#> 35  -0.52634807  0.152652540 -1.29840132 -0.12049215  0.388433868 -1.234508632
#> 36   0.42728853 -1.079789298  1.02507897  1.20044815  0.253081037 -1.219134416
#> 37  -0.40378078 -0.607280073 -1.07439988  0.27114271  0.218265629 -1.196683260
#> 38   0.33191999  1.231464151 -0.26436968 -0.99767219 -0.503654314 -0.209553718
#> 39   0.37973599  1.217713223  0.06312072 -1.13233532 -0.035300077  0.961764372
#> 40  -0.44599194 -0.538582828 -1.11569041  0.23232304  0.225993940  1.239362018
#> 41   0.41819316 -1.142978674  0.73270559  1.10756272  0.364372767 -0.881955311
#> 42  -0.49487833  0.243168133 -1.29643812 -0.16558715 -0.357595201 -0.562086418
#> 43  -0.45797168 -0.517089212 -1.12751676  0.22043156 -0.475987596 -0.309270352
#> 44  -0.10562255 -0.427328611  4.52549672 -2.44581444 -0.493710972 -0.250419237
#> 45   0.05726523  1.031756183 -0.92949121 -0.64274636  0.436361210  1.199147970
#> 46  -0.45281388 -0.526472513 -1.12241498  0.22560924  0.248994597 -1.216684536
#> 47  -0.56574719  1.465653248  6.97893313  2.75754418  0.292391841  1.293928687
#> 48   0.32027259 -1.203974629  0.06860218  0.87465247  0.218105193  1.232208813
#> 49  -0.56021894  0.004389326 -1.28937087 -0.04718661 -0.362945647  0.498465424
#> 50   0.34670974  0.830437846  1.56810807 -1.63525125  0.033734477  1.041182849
#> 51   0.07079600  0.033537445  3.50158898 -2.17426782  0.209553621 -1.190564771
#> 52   0.42735198 -1.071197839  1.05980517  1.21119188 -0.511130262 -0.171377846
#> 53   0.37819235  1.219257415  0.04774771 -1.12633547  0.044115730 -1.050143575
#> 54   0.40267337  1.125489386  0.59382445 -1.32554024 -0.452873193  0.305329408
#> 55  -0.53327139  0.129025196 -1.29797577 -0.10878268  0.268544456 -1.227884048
#> 56  -0.46591232  0.312130710 -1.29103567 -0.20031825  0.405671302  1.042311370
#> 57   0.38902553 -0.844999801  1.80273897  1.42978055  0.147666178  1.163864064
#> 58   0.14853889  1.117190076 -0.79164717 -0.73205930 -0.522617604 -0.048441606
#> 59   0.38757715  1.207254640  0.15261827 -1.16673950 -0.284432347 -0.674739853
#> 60  -0.13096463  0.827362039 -1.11684940 -0.48836800 -0.272444854  0.648763707
#> 61  -0.35553240  0.518793230 -1.25295243 -0.30781290  0.270411522 -1.228880559
#> 62   0.40233662  1.145320711  0.50541338 -1.29495196  0.378598480  1.337748294
#> 63   0.07235556  1.046558861 -0.90965794 -0.65649491 -0.519106298 -0.112393307
#> 64   0.42294988 -1.124279803  0.82807160  1.13836802  0.404382402 -1.201855142
#> 65  -0.48739490 -0.459019995 -1.15706478  0.18882477  0.282190826  0.645675041
#> 66   0.15519041  1.122971677 -0.77966313 -0.73923530  0.380828715 -1.241838511
#> 67  -0.11618445 -0.936072087 -0.77291234  0.48608563  0.335387252 -1.252145473
#> 68   0.14595550  1.114925511 -0.79621850 -0.72930084 -0.446257478 -0.387997911
#> 69   0.39627802  1.186275538  0.29282796 -1.21899468 -0.495441532  0.170831698
#> 70  -0.28754275  0.622836496 -1.21958974 -0.36500474 -0.512092517 -0.165679538
#> 71   0.33254035  1.231550590 -0.26127642 -0.99902151  0.410415194 -1.132373828
#> 72   0.42728791 -1.065501653  1.08237418  1.21814448 -0.177641269  0.783753157
#> 73   0.42286231 -1.007631657  1.29549150  1.28271295  0.119312038  1.134507724
#> 74   0.40174322  1.154301067  0.46291109 -1.28004919 -0.200011256 -0.786718844
#> 75   0.41030948 -1.161949909  0.62285149  1.07139506  0.410602861 -1.149176016
#> 76  -0.56716225 -0.172418989 -1.25852208  0.04065889  0.385764108  1.338575413
#> 77   0.40265452  1.135471010  0.55016757 -1.31050354 -0.345171077 -0.582598372
#> 78  -0.56335566 -0.208764557 -1.24934946  0.05893491 -0.433750955 -0.416735427
#> 79  -0.22906468 -0.826023653 -0.89864660  0.40656339 -0.287219854  0.625986993
#> 80   0.35817020 -0.740028382  2.09152278  1.51021850  0.346628722 -1.252826245
#> 81   0.40174413  1.102096382  0.69069339 -1.35845934 -0.409793668 -0.467045305
#> 82   0.12128759  1.092799237 -0.83767569 -0.70371607 -0.473844412 -0.315611150
#> 83   0.35742724 -1.202498047  0.23369392  0.93618879 -0.499875663 -0.226062041
#> 84   0.36867327  0.913779055  1.32734596 -1.56244701  0.164095609 -1.156051198
#> 85  -0.48282964  0.273073294 -1.29451778 -0.18059944 -0.090072397  0.895928758
#> 86   0.20917189 -1.167408661 -0.26071138  0.74140640  0.382526920  0.956571332
#> 87  -0.47120793 -0.492009256 -1.14069888  0.20669183  0.261102157  1.269620340
#> 88   0.37808295  0.953819449  1.20528324 -1.52474498  0.359958920 -0.865728826
#> 89  -0.14831239 -0.906534884 -0.81011329  0.46361016  0.206041936 -1.188048738
#> 90  -0.51288897  0.194109578 -1.29821355 -0.14109180  0.433858345  1.179631173
#> 91   0.25262714  1.197194123 -0.55842535 -0.86008149  0.030376292  1.037398707
#> 92  -0.56363018 -0.206645885 -1.24991177  0.05786659 -0.430893248  0.359481892
#> 93   0.26970092  1.207406056 -0.50746727 -0.88545372  0.397624977  1.338068107
#> 94  -0.03058147 -1.008760743 -0.66659296  0.54644210  0.438844229  1.227047731
#> 95  -0.12225351 -0.930591906 -0.78003885  0.48184039  0.315074443 -1.247804365
#> 96   0.19335812 -1.159421647 -0.29642704  0.72588682 -0.267077897 -0.699058460
#> 97  -0.33101728  0.557742635 -1.24170112 -0.32891078 -0.487697382 -0.271910126
#> 98  -0.36565141  0.502165022 -1.25733167 -0.29890503 -0.117155243  0.862246027
#> 99  -0.18957350  0.755313041 -1.15974161 -0.44273284 -0.374083387  0.477736149
#> 100  0.28879711 -0.539406346  2.59888537  1.64723576 -0.374956490  0.476083868
#>          x2.pen3     x2.null1      x3.pen1     x3.pen2      x3.pen3
#> 1   -0.593159660  0.451392652  0.479493348  1.20269732  1.959737219
#> 2    2.047141341 -1.773786969 -0.189120621 -0.73825808 -0.800036933
#> 3   -0.503058212  0.514325827  0.270508885 -1.32651021 -0.160968137
#> 4   -0.053590071  0.765092561  0.006885861 -1.01359303 -0.623398335
#> 5    0.283705119 -1.141556339  0.465355795 -1.38563240  1.138391263
#> 6   -0.919849228 -0.326622716  0.346850681 -1.39585108  0.091285014
#> 7    0.040276525 -1.031849889  0.205993236  0.41353326  4.253495403
#> 8   -0.421695936  0.566290330  0.316463162 -0.76346266  3.354797168
#> 9   -0.961991902  0.002427583 -0.420587800  0.24678992 -0.949943496
#> 10   4.143570155 -2.388420771  0.365144548  1.38307442 -0.008049196
#> 11  -0.914831525  0.115909261 -0.084763569  0.87051219 -0.751314169
#> 12  -0.637612987  0.417672588 -0.043350088  0.22148083  5.826482442
#> 13   0.182002162  0.873114349  0.248298501 -1.30353951 -0.218799028
#> 14  -0.954190776  0.027812263  0.495894963  1.27962158  1.652428356
#> 15   0.355674712  0.946668879 -0.032716128 -0.96152809 -0.665552685
#> 16  -0.608422872 -0.659281369  0.317672081 -1.37149823 -0.017086857
#> 17  -0.664496679  0.396221969 -0.447221849  0.12003726 -0.960102093
#> 18   0.121701967  0.846473704  0.496013071  1.28029531  1.649528404
#> 19  -0.838206549 -0.447014428  0.476274161  1.43322626  0.611335762
#> 20  -0.623279546  0.428769371  0.083240370 -1.11103552 -0.527343572
#> 21  -0.272030759 -0.872845664 -0.425250983  0.22939657 -0.951837004
#> 22  -0.911526718  0.121791964 -0.299558489 -0.54848475 -0.874774449
#> 23   1.693619347  1.419929929 -0.359905329  0.41617126 -0.922761124
#> 24  -0.417887377  0.568628614 -0.118296120 -0.84377349 -0.744260557
#> 25  -0.655087568  0.403828750  0.287326930  0.62494788  3.704820061
#> 26   0.144500428  0.856619573 -0.234318953 -0.66531837 -0.832223247
#> 27  -0.786615033  0.284763409 -0.212196009 -0.70167600 -0.816768387
#> 28   0.790144244  1.114876921 -0.315898745 -0.51622988 -0.884781104
#> 29  -0.583801555  0.458243356 -0.008061751 -0.99407797 -0.639831077
#> 30  -0.079635368  0.752425832 -0.115196783  0.82929771 -0.774782391
#> 31  -0.793341150 -0.497617812 -0.427352520  0.22115375 -0.952678096
#> 32  -0.735081889 -0.554855504  0.422217293 -1.43442339  0.506528631
#> 33  -0.756224477 -0.534971089 -0.397127933 -0.32244321 -0.931059882
#> 34  -0.934340405  0.077479797  0.413711275  1.41760692  0.190291239
#> 35   0.857852804  1.139508589 -0.230822265  0.65852398 -0.852663042
#> 36   0.121576497  0.846417612 -0.377649885  0.37359496 -0.931133694
#> 37   0.018284187  0.799231793  0.352716587 -1.40032124  0.115438800
#> 38  -0.922356091  0.101910578  0.210635003  1.23077733 -0.395669174
#> 39  -0.764931751 -0.526507625  0.451502270 -1.23018012  1.887275762
#> 40  -0.404980622 -0.795729406  0.490378836  1.25083801  1.772486210
#> 41   2.355206239  1.620800110 -0.254135126 -0.63151311 -0.845608638
#> 42  -0.785760146  0.285650379  0.498392873  1.40822330  0.935260527
#> 43  -0.893116153  0.152174402 -0.134829131 -0.81992503 -0.757933934
#> 44  -0.911180949  0.122398898 -0.005747740  0.97284268 -0.682678769
#> 45   1.162280654 -1.480752319  0.504314816  1.35057614  1.314211497
#> 46   0.108590081  0.840596644  0.435757161 -1.43330034  0.625918854
#> 47  -0.236271881 -0.892438995  0.123148147 -1.16024777 -0.467586819
#> 48  -0.421777254 -0.785429972  0.490125853  1.24961479  1.777436791
#> 49  -0.953033535 -0.250978113 -0.311186297 -0.52568649 -0.881920532
#> 50  -0.698771424 -0.587086428 -0.119699591  0.82309447 -0.778136164
#> 51  -0.005233789  0.788187337 -0.224722396 -0.68125669 -0.825587226
#> 52  -0.931814457  0.082885713 -0.132902979 -0.82272543 -0.756362882
#> 53  -0.339037820  0.615407140  0.213933135  1.23441425 -0.389609011
#> 54  -0.975269930 -0.152227781  0.315297738 -1.36938206 -0.025157233
#> 55   0.173142734  0.869238894  0.432389449 -1.14592107  2.196224244
#> 56   1.722157085 -1.669899475  0.103250486  1.10683780 -0.561301125
#> 57  -0.550075317 -0.701413990 -0.223487714  0.67023045 -0.848183835
#> 58  -0.956063257  0.022165943  0.492946389  1.41952758  0.825975365
#> 59  -0.719065320  0.349695637 -0.034335902 -0.95937193 -0.667187339
#> 60  -0.916926657 -0.332056904 -0.173553270  0.74641618 -0.816116345
#> 61   0.179647810  0.872085727  0.017011916 -1.02673265 -0.611865362
#> 62   0.113746978 -1.066019052 -0.398048757  0.31889025 -0.940347551
#> 63  -0.944613936  0.053667087 -0.410992125  0.27950306 -0.945941663
#> 64   1.105025156  1.226594296 -0.328877305  0.48308136 -0.907375831
#> 65   2.838854685 -2.014471278  0.356304194  0.81220526  3.193840633
#> 66   0.778345535  1.110547024  0.281716606 -1.33771042 -0.129609905
#> 67   0.459160400  0.988562993 -0.264914165  0.60197399 -0.872754974
#> 68  -0.865123333  0.192626276  0.433751818 -1.43378682  0.606455651
#> 69  -0.976750561 -0.085480929 -0.092096437  0.86068921 -0.757101079
#> 70  -0.933146358  0.080054350 -0.339723848 -0.46619517 -0.898937715
#> 71   1.464431766  1.346644982  0.407341258 -1.43134664  0.399332831
#> 72  -0.866959310 -0.410154176  0.309261392 -0.74215774  3.413117964
#> 73  -0.592777136 -0.670873139 -0.430615207 -0.20504920 -0.948642255
#> 74  -0.637263507  0.417945850  0.040268943 -1.05667274 -0.584053143
#> 75   1.387346067  1.321464500 -0.056270094  0.90810151 -0.727962850
#> 76   0.156484746 -1.085449266 -0.431338353  0.20469087 -0.954250366
#> 77  -0.774615653  0.297039525 -0.313139864 -0.52178213 -0.883108861
#> 78  -0.853752763  0.207604913 -0.296727253  0.54535490 -0.890464349
#> 79  -0.923601851 -0.319439208 -0.361136295 -0.41731875 -0.911238382
#> 80   0.523084982  1.013814019  0.043767383 -1.06114651 -0.579702844
#> 81  -0.832270352  0.234151180 -0.420584743 -0.24530455 -0.943448189
#> 82  -0.891026981  0.155404185  0.277265193 -1.33329579 -0.142252563
#> 83  -0.917974019  0.110170645  0.492112639  1.25941471  1.737448113
#> 84  -0.115900580  0.734505295  0.380056127 -1.41869065  0.241723710
#> 85  -0.808709883 -0.481057788  0.120760545  1.12768261 -0.537936833
#> 86   1.986698132 -1.754733575  0.373901922 -0.94028959  2.851236443
#> 87  -0.322588684 -0.844367095  0.155840041 -1.19948672 -0.412655430
#> 88   2.404388927  1.635219533  0.313415584  1.33721758 -0.167251890
#> 89  -0.014481949  0.783811345  0.119164926 -1.15539717 -0.473892223
#> 90   1.240538315 -1.508128772  0.148136271  1.15982032 -0.498818120
#> 91  -0.702339815 -0.584015180  0.454451138  1.10963968  2.289051492
#> 92  -0.971521609 -0.179479683  0.027096108  1.01392628 -0.649991121
#> 93   0.236319878 -1.120944900  0.497747700  1.40995253  0.920077334
#> 94   1.043266614 -1.438423546  0.456240121  1.43463890  0.439720604
#> 95   0.358131754  0.947679087  0.244979602 -1.30002623 -0.226991432
#> 96  -0.702758158  0.364066575 -0.363495505  0.40785694 -0.924480760
#> 97  -0.904858185  0.133232218  0.464273111 -1.30990699  1.550540248
#> 98  -0.828121898 -0.459025571 -0.409161256 -0.28509912 -0.937462657
#> 99  -0.956641062 -0.240140820 -0.455453124  0.01245217 -0.962243599
#> 100 -0.956915293 -0.239279972 -0.413662698 -0.27001978 -0.939831402
#>         x3.null1      x4.pen1     x4.pen2     x4.pen3    x4.null1       x5.pen1
#> 1   -1.819060135  0.026374946  0.02739721  3.54993671  2.56865600 -0.1798155253
#> 2    0.396628211  0.016072289 -0.12694475  4.36743541 -1.88485354 -0.0152448611
#> 3    0.899385750  0.481427343 -1.18517109  0.27302209  1.43589919  0.4085239970
#> 4    0.578487364 -0.156464343 -0.66233967 -1.05518883  0.41968225  0.3237937536
#> 5    1.497663451  0.185260059  1.01263052  0.45978185 -0.73683177  0.0471099320
#> 6    1.036939214  0.042504457  0.92832000  0.03895580 -0.56552315 -0.1581610417
#> 7   -2.546652304 -0.399279420 -0.25760770 -1.01670688  0.19224547  0.4055077408
#> 8    2.250712762 -0.311417783  0.54669607 -0.56472975 -0.25174329  0.2802645395
#> 9   -0.121714607  0.182227308  0.35763934  3.21929780 -1.58202916  0.4526271916
#> 10  -1.004226645 -0.232578883 -0.55659101 -1.05778671  0.35684337 -0.5873380777
#> 11  -0.473302605  0.303277901  0.87528027  1.68520143 -1.14750840 -0.5123041339
#> 12   2.982801340  0.418003538 -0.92181659  1.20122945  1.80058647 -0.3474969821
#> 13   0.865233724 -0.088531163  0.81499041 -0.21443827 -0.44743214 -0.3964096616
#> 14  -1.710522969 -0.275327460  0.59948817 -0.51000102 -0.28589598 -0.5113463831
#> 15   0.540507243 -0.320887192 -0.41566850 -1.04663475  0.27754887 -0.0950738233
#> 16   0.979855498  0.027897108 -0.88646963 -1.00749209  0.57063192  0.1179964156
#> 17  -0.056981846  0.198367841  1.01655214  0.51460841 -0.75752397  0.0779298932
#> 18  -1.709478716  0.079407051  0.95515175  0.12641310 -0.60326952 -0.3748254231
#> 19  -1.301751272 -0.435505814 -0.15852340 -0.98910753  0.13980621  0.3811893675
#> 20   0.657145942 -0.472428364  0.07736789 -0.89561608  0.01532046 -0.3902618628
#> 21  -0.112778629  0.004761486  0.89844756 -0.04124240 -0.52966798  0.3948310790
#> 22   0.288413197 -0.395976286  0.39368882 -0.69911411 -0.15917481 -0.5421680733
#> 23  -0.210096297 -0.235458117  0.65248280 -0.44956516 -0.32177036 -0.5243108177
#> 24   0.461595561  0.481142839 -1.17740891  0.31007108  1.45195965 -0.2215294839
#> 25  -2.381301709  0.449261106 -1.25216810 -0.25845245  1.18124533 -0.5727465585
#> 26   0.353957553 -0.050823362  0.85036201 -0.14776194 -0.47994665  0.2418995922
#> 27   0.375031818 -0.418843620 -0.20778331 -1.00367150  0.16581627  0.0933652564
#> 28   0.270804106 -0.045172406  0.85546945 -0.13742808 -0.48488499  0.3948561814
#> 29   0.563976594 -0.273063202 -0.49510828 -1.05483443  0.32174537  0.1189769259
#> 30  -0.446684505  0.257493172  1.01511244  0.84298907 -0.87571857  0.3679121853
#> 31  -0.108550733  0.469482481 -1.09938488  0.62866909  1.58390195 -0.4709979065
#> 32   1.235413991 -0.012338762  0.88416515 -0.07526738 -0.51406115 -0.2148005388
#> 33   0.168158571 -0.097494010  0.80624515 -0.22977103 -0.43978696 -0.4070168008
#> 34  -1.106591505 -0.113443075 -0.48660706  5.18583199 -2.09742668  0.3693411311
#> 35  -0.343624584  0.118083599 -0.98662303 -0.95649233  0.65337642 -0.3265813102
#> 36  -0.187604810 -0.470506753  0.01758158 -0.92315712  0.04702558 -0.2949826005
#> 37   1.049307715  0.456166840 -1.25036708 -0.20722708  1.20833707  0.2756353390
#> 38  -0.771617893  0.306595533 -0.62634903  1.99650796  2.07484681 -0.3300400336
#> 39   1.772366407  0.116745014  0.97938610  0.22722155 -0.64523593 -0.2842662642
#> 40  -1.753412761  0.220283593  0.48150883  2.90173837 -1.49593247 -0.5912420393
#> 41   0.334675077 -0.337671507  0.50455820 -0.60504338 -0.22541026 -0.1978575234
#> 42  -1.437684263  0.300468709  0.83722313  1.83118661 -1.19143701 -0.5614204304
#> 43   0.446516242  0.176598157  1.00950819  0.42580113 -0.72384774 -0.5625464732
#> 44  -0.543654748  0.311473940 -1.17732493 -0.72695845  0.88351036  0.3471377073
#> 45  -1.585808109 -0.007623760  0.88814876 -0.06601000 -0.51833192  0.0118566753
#> 46   1.287960970 -0.287602322 -0.47185329 -1.05293146  0.30868684  0.2237485008
#> 47   0.701860105 -0.289756818  0.57900306 -0.53183615 -0.27247468 -0.4419737686
#> 48  -1.755167376  0.294781258  0.96734687  1.24868261 -1.01115106 -0.4574644400
#> 49   0.275947926  0.294472464  0.96831288  1.24290522 -1.00928820  0.2967883485
#> 50  -0.442748259 -0.179515813 -0.63143283 -1.05702361  0.40093618 -0.5913833048
#> 51   0.363150960  0.480978197 -1.20316892  0.18033835  1.39494440  0.2177712125
#> 52   0.448273778 -0.134023104  0.76919040 -0.29057740 -0.40879271  0.3776380827
#> 53  -0.775736649 -0.355678242 -0.35124001 -1.03656394  0.24243759  0.3340677144
#> 54   0.975492561 -0.243339609 -0.54067080 -1.05731938  0.34766842  0.3776337257
#> 55   1.878247549  0.239544549 -1.11179213 -0.83929883  0.78608563 -0.4492617740
#> 56  -0.649627893 -0.435809615 -0.15754183 -0.98880034  0.13928878 -0.4324125106
#> 57  -0.350391071 -0.465994620 -0.02405206 -0.94075290  0.06900233 -0.4193090018
#> 58  -1.392924670  0.477976933 -1.22192253  0.06827286  1.34377978  0.1755683305
#> 59   0.538982292  0.244391766  1.01932158  0.75373669 -0.84447295 -0.1163720609
#> 60  -0.395404436 -0.418007742  0.34139603 -0.73828479 -0.12913564  0.4274689536
#> 61   0.588464378 -0.170255325  0.73003240 -0.34869636 -0.37806898  0.3249666349
#> 62  -0.159000093  0.310386059 -1.17640496 -0.72901316  0.88188088 -0.5890182968
#> 63  -0.138579738  0.385660381 -1.23192321 -0.54617610  1.01246683  0.2866395919
#> 64  -0.245924499  0.404052421 -1.24179212 -0.48303324  1.05247474 -0.5786141411
#> 65  -2.223613772  0.291803138  0.97574721  1.19684644 -0.99437561  0.3326838181
#> 66   0.917439916  0.078576949  0.95457696  0.12432455 -0.60238332  0.2309440251
#> 67  -0.311424578  0.439935717 -1.25254771 -0.31737001  1.14915987 -0.3418114230
#> 68   1.279506910 -0.281471776  0.59085898 -0.51929567 -0.28021466  0.1630178394
#> 69  -0.466881627 -0.458300821 -0.06949225 -0.95852087  0.09293424  0.4059422514
#> 70   0.243828499  0.131075118 -1.00062411 -0.94702018  0.66622756 -0.5913863045
#> 71   1.186722882  0.213176126  1.01960295  0.58239475 -0.78269337  0.4041455505
#> 72   2.268702774  0.411785462 -0.90362587  1.25411002  1.81961933 -0.4998318012
#> 73   0.107783259 -0.332961921 -0.39407319 -1.04359256  0.26571663  0.3963622807
#> 74   0.611878143 -0.387478067 -0.28486914 -1.02311642  0.20677741  0.4389552664
#> 75  -0.498364170 -0.125285998  0.77826552 -0.27625877 -0.41619209  0.3844847350
#> 76  -0.100119118 -0.471686108  0.03596604 -0.91497772  0.03729784  0.4041490642
#> 77   0.273822343  0.445329467 -1.00731733  0.94105734  1.70475802 -0.2433846810
#> 78  -0.279900459  0.012106261 -0.13802550  4.39278102 -1.89145023 -0.4933707345
#> 79   0.217826992  0.131311146  0.98785704  0.27071586 -0.66288543 -0.4249135315
#> 80   0.615463001  0.302832872  0.86643591  1.72026071 -1.15812614 -0.5777891705
#> 81   0.128380611  0.461893544 -1.24744304 -0.15759904  1.23393894 -0.5528417214
#> 82   0.910198384 -0.171048978 -0.64288557 -1.05645260  0.40784199 -0.2229457686
#> 83  -1.740963372 -0.087026634  0.81644546 -0.21184668 -0.44871791  0.3383443738
#> 84   1.112124276 -0.175995996  0.72359010 -0.35774254 -0.37318159 -0.0413253372
#> 85  -0.668126428 -0.277433152 -0.94135699  6.21932423 -2.36574077 -0.3672004455
#> 86   2.092744702  0.183539846  1.01204193  0.45289997 -0.73421228 -0.3678359003
#> 87   0.740744173  0.151407784 -1.02228300 -0.93092725  0.68689686  0.4073588999
#> 88  -0.915098316 -0.403270325  0.37728142 -0.71173390 -0.14967874  0.3529529973
#> 89   0.697268548 -0.004120457 -0.84975406 -1.02042300  0.54342467  0.2884080154
#> 90  -0.698037950 -0.446927564 -0.11864177 -0.97608990  0.11880001  0.2468243907
#> 91  -1.931290709 -0.058666024 -0.78541153 -1.03758725  0.49853701 -0.2534125099
#> 92  -0.574116411 -0.428315552 -0.18078739 -0.99589667  0.15155055  0.3132809516
#> 93  -1.431526648  0.366833994 -1.21993601 -0.60138950  0.97564629  0.3575789526
#> 94  -1.225191429 -0.470719746  0.02038833 -0.92192477  0.04554151  0.3849934678
#> 95   0.860299804  0.481073950 -1.20187134  0.18740804  1.39810949 -0.2559774277
#> 96  -0.205686846 -0.383756979 -0.29312319 -1.02495582  0.21118915  0.3233296350
#> 97   1.652482082 -0.308919276  0.55052653 -0.56092366 -0.25417524  0.0002855917
#> 98   0.148844620 -0.239181613  0.64773489 -0.45523113 -0.31848132  0.4526909869
#> 99  -0.002493923 -0.433863045  0.29712203 -0.76916143 -0.10417291 -0.2870052633
#> 100  0.141076987 -0.358855211 -0.34498408 -1.03542813  0.23905666  0.0570309550
#>          x5.pen2      x5.pen3    x5.null1
#> 1   -0.787622552 -0.882298933  0.47829439
#> 2    1.100624100 -0.718169382 -0.55694144
#> 3    1.183839211  1.034758776 -1.27718300
#> 4    1.328173121  0.047214462 -0.92891795
#> 5    0.097961783  4.138054285  2.14735761
#> 6   -0.810200201 -0.862590356  0.49327324
#> 7    1.245630859  0.775100653 -1.19276694
#> 8   -0.462314122  2.822411496  1.81022739
#> 9   -1.064697979  1.030330878  1.30224271
#> 10  -0.039341015 -1.177938463  0.07500249
#> 11   0.408695793 -1.144267936 -0.14418701
#> 12  -0.588249188 -1.018582616  0.35760095
#> 13   0.635306767 -1.070689096 -0.26133475
#> 14   0.411040588 -1.143719774 -0.14536274
#> 15   1.021140674 -0.810460909 -0.49571642
#> 16   0.168857366  3.631874508 -2.00240395
#> 17   0.060249070  3.865740937 -2.06254443
#> 18  -0.550171910 -1.038474928  0.33626731
#> 19   0.993818237  1.653689803 -1.46557729
#> 20   0.645277893 -1.066406798 -0.26670067
#> 21   1.284959433  0.566884751 -1.12200780
#> 22   0.327530501 -1.160733636 -0.10380903
#> 23   0.378117569 -1.151039527 -0.12890514
#> 24  -0.742387410 -0.918760950  0.44925193
#> 25  -0.127750538 -1.168291101  0.11826523
#> 26   1.300846835 -0.246175548 -0.80438412
#> 27  -0.009406024  3.893525884  2.08567861
#> 28   1.284901902  0.567235854 -1.12212975
#> 29   1.217896756 -0.520381012 -0.67076564
#> 30  -1.185624363  0.095348104  0.97856469
#> 31   0.500412102 -1.119654848 -0.19065634
#> 32  -0.749848663 -0.913006403  0.45396071
#> 33   0.617763038 -1.077989934 -0.25194696
#> 34  -1.185711083  0.101859503  0.98110059
#> 35  -0.616071851 -1.002963825  0.37347020
#> 36   0.785157805 -0.995161968 -0.34476387
#> 37  -1.156147471 -0.221167406  0.84763417
#> 38  -0.611543376 -1.005570922  0.37087001
#> 39  -0.900172630  5.895127206 -2.57902416
#> 40   0.020888398 -1.181403186  0.04570118
#> 41  -0.768350372 -0.898303594  0.46577242
#> 42   0.262137914 -1.170556487 -0.07164515
#> 43  -0.168816791 -1.161926562  0.13851294
#> 44  -1.182657305  0.008113249  0.94404313
#> 45  -0.969513565 -0.683931429  0.61252202
#> 46   1.290850915 -0.295592360 -0.78178999
#> 47  -0.445836867 -1.085049088  0.27971638
#> 48   0.527219155 -1.111186844 -0.20445195
#> 49   0.686383596  2.457999176 -1.69224346
#> 50   0.050102623 -1.182189990  0.03151612
#> 51   1.287344808 -0.311016163 -0.77462210
#> 52   1.310763598  0.380300993 -1.05576774
#> 53  -1.179424210 -0.040819474  0.92420308
#> 54   1.310767840  0.380262812 -1.05575387
#> 55   0.542874345 -1.105962642 -0.21256023
#> 56   0.573806300 -1.095018521 -0.22870221
#> 57   0.596852504 -1.086311828 -0.24084190
#> 58  -1.096147774 -0.442329693  0.74419628
#> 59   0.998679517 -0.832939686 -0.47971520
#> 60  -0.902621898  1.614793688  1.47834388
#> 61   0.778888652  2.229635850 -1.62945277
#> 62  -0.021440552 -1.179228881  0.06628338
#> 63   1.320279262 -0.103739068 -0.86669371
#> 64  -0.099009048 -1.172029410  0.10415778
#> 65  -0.601396824  2.469498948  1.71639502
#> 66  -1.131694501 -0.330147187  0.79821485
#> 67  -1.052401969  6.216122539 -2.66062280
#> 68   1.251182193 -0.436269232 -0.71410559
#> 69   1.242641517  0.789077824 -1.19741206
#> 70   0.049888337 -1.182186331  0.03162013
#> 71   1.126100339  1.242087078 -1.34206562
#> 72   0.438294226 -1.137041813 -0.15907212
#> 73   1.071702793  1.419474513 -1.39607888
#> 74  -0.954439491  1.444532731  1.42843963
#> 75   1.009198099  1.608906705 -1.45243528
#> 76   1.253584221  0.736919514 -1.18001417
#> 77   0.852128071 -0.952513311 -0.38449364
#> 78  -0.350167760 -1.118698503  0.22978308
#> 79   0.587095651 -1.090056606 -0.23569010
#> 80  -0.103386604 -1.171497752  0.10630344
#> 81   0.293143288 -1.166274398 -0.08686153
#> 82  -0.740808560 -0.919966144  0.44825939
#> 83   0.825208460  2.111698256 -1.59658833
#> 84   1.075452004 -0.749863282 -0.53664608
#> 85   0.681507220 -1.050009595 -0.28639515
#> 86   0.680532221 -1.050468509 -0.28586087
#> 87   1.230224494  0.845231609 -1.21595182
#> 88  -1.183782042  0.031251432  0.95330207
#> 89   1.320857697 -0.097357435 -0.86939771
#> 90   0.533111804  2.820309964 -1.78991150
#> 91  -0.706101843 -0.945402311  0.42676379
#> 92  -1.172588135 -0.111266120  0.89497842
#> 93  -1.184522321  0.050329875  0.96087973
#> 94   1.011628963  1.601769081 -1.45033456
#> 95   0.836243644 -0.963204863 -0.37489953
#> 96  -0.575845133  2.535665495  1.73414647
#> 97   0.205862955  4.382161945  2.20871267
#> 98  -1.074665865  0.987233794  1.28865583
#> 99   0.795851513 -0.988764825 -0.35098726
#> 100 -1.007085147 -0.626747392  0.64607527