Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>          x1.pen1     x1.pen2     x1.pen3     x1.null1      x2.pen1      x2.pen2
#> 1   -0.124531651  0.78668423 -0.81267927 -0.498608223 -0.059939142 -1.037873756
#> 2    0.039302699  0.96802718 -0.66910921 -0.628881592 -0.578545563 -0.417573469
#> 3   -0.350319721 -0.58229180 -0.87988812  0.225634166 -0.242788766 -0.703347465
#> 4   -0.489335753  1.45192419  7.91538067  3.220844962 -0.299600428 -0.832250137
#> 5    0.119233747  1.04885076 -0.57400683 -0.700061694 -0.640601463 -0.176769119
#> 6    0.396807174 -1.30789977  0.88538639  1.199441995  0.024252239  0.872919079
#> 7    0.393840296 -1.11791913  1.74030944  1.492022954  0.311894514 -0.766826367
#> 8    0.369431408 -1.34084313  0.55884439  1.074969700  0.301511620 -1.116653883
#> 9    0.460635519  1.25851413  0.52216915 -1.253181958 -0.170975555  0.715599379
#> 10  -0.256006264  0.61674157 -0.89855432 -0.394917462  0.370366089  1.032669310
#> 11   0.395760029 -1.31056785  0.86641941  1.192454350  0.057140058 -1.107812125
#> 12   0.335857732 -1.34415485  0.32898292  0.980968884 -0.589633804  0.156687679
#> 13   0.318723333  0.61853492  2.88085107 -2.039335327 -0.369270168 -0.754471724
#> 14  -0.459482751 -0.33012125 -0.96033755  0.093482008  0.125712193 -0.135494702
#> 15   0.399808828  0.84188172  2.25769151 -1.851728214 -0.539053936  0.264784593
#> 16   0.438454946  0.96422290  1.87978487 -1.733104166 -0.538848271 -0.503007967
#> 17  -0.498662592 -0.16237443 -0.99026196  0.009046916  0.136415710 -1.141110337
#> 18   0.477435383  1.22467951  0.80087857 -1.361549237  0.237074686 -1.154126453
#> 19  -0.464091874 -0.31531354 -0.96368855  0.085954730 -0.046588508  0.820045690
#> 20  -0.246803362  0.62968885 -0.89318342 -0.402427922 -0.593031934  0.147990197
#> 21   0.201984083 -1.26570812 -0.16599091  0.751078090  0.434380552  0.838573863
#> 22   0.384819348 -1.32905088  0.71145856  1.134300752 -0.008827116  0.848811498
#> 23   0.452118256  1.26477876  0.43630154 -1.218254921 -0.346590602  0.537516457
#> 24   0.477974300  1.14090844  1.23075573 -1.516998076  0.131993047  0.944077485
#> 25   0.394936486 -1.12437230  1.71763327  1.484716439  0.197131018 -0.347687882
#> 26   0.214991431 -1.27633449 -0.13171840  0.768663630  0.211931839 -1.155691240
#> 27   0.476982386  1.22665093  0.78797174 -1.356680626 -0.094023126  0.782012179
#> 28  -0.485864908  0.14189540 -1.00140469 -0.142754748 -0.512745892  0.311190360
#> 29   0.400574852 -1.16199650  1.58061083  1.440126605 -0.635114701 -0.220042059
#> 30  -0.491664840  0.11533874 -1.00263406 -0.129415084  0.228138918 -0.446316034
#> 31   0.136303805 -1.20674563 -0.31464734  0.670669172  0.164673031 -1.149129703
#> 32  -0.155468742  0.74919826 -0.83477049 -0.474642334  0.183000883 -0.304330683
#> 33   0.393478777 -1.11583341  1.74759241  1.494365393  0.327937236 -0.861386357
#> 34  -0.468132096  0.20696134 -0.99656288 -0.175620380 -0.405076482  0.466945978
#> 35   0.073207734 -0.08747881  4.43049688  2.273799523  0.025680043 -1.091160121
#> 36  -0.463403235  0.22181889 -0.99508710 -0.183168247  0.201972170  0.216774520
#> 37   0.293204166 -0.73015022  2.88066174  1.839152351 -0.361543260  0.520154781
#> 38   0.167707874 -1.23592773 -0.24812353  0.707549923 -0.126723268  0.754526322
#> 39   0.380725607 -1.05097988  1.96438359  1.563203703 -0.005095149  0.851582364
#> 40   0.319110674 -1.33995957  0.24122107  0.943326319 -0.499128091  0.333524184
#> 41   0.396397228 -1.30897284  0.87782971  1.196661265  0.114280080  0.933199467
#> 42   0.336074273 -1.34419002  0.33020544  0.981485773 -0.613711840 -0.316745197
#> 43   0.032228531  0.96066723 -0.67651511 -0.622942626 -0.571540552  0.199254176
#> 44  -0.020748379 -1.04218956 -0.56292687  0.515119451  0.011062019 -1.082857618
#> 45  -0.210161985 -0.80531342 -0.76346900  0.353516709 -0.515379808 -0.546180716
#> 46   0.341838181 -1.34491722  0.36370760  0.995574743  0.212929907  0.988848244
#> 47   0.292944555 -0.72931447  2.88284647  1.839787588 -0.194487498  0.734963008
#> 48  -0.490808083 -0.20860658 -0.98374445  0.032166961 -0.072870166 -1.028891784
#> 49  -0.225936819  0.65835490 -0.88067017 -0.419252008  0.307117370  1.025982435
#> 50  -0.480410350  0.16393526 -1.00005757 -0.153855913  0.201276943 -1.155145417
#> 51   0.099908524 -1.17107775 -0.38324482  0.630848653 -0.310130569  0.578157920
#> 52   0.100903272 -1.17207591 -0.38147708  0.631900181 -0.480698521 -0.603464257
#> 53  -0.289250651 -0.68707190 -0.83151622  0.283901106  0.201273994 -1.155145181
#> 54  -0.155043197  0.74972278 -0.83447513 -0.474972688  0.323880301 -0.833660229
#> 55   0.458730018  1.03981345  1.62345551 -1.649905887  0.196702455  0.980602953
#> 56   0.341487088  1.23643272 -0.12048599 -0.966242018 -0.299021422 -0.832855929
#> 57   0.016149708 -1.08314170 -0.51426175  0.548431292  0.300766962 -1.117613511
#> 58  -0.027559662 -1.03447604 -0.57145234  0.509101935 -0.257005350  0.633621968
#> 59   0.003991698  0.93093409 -0.70474951 -0.599666334  0.027160123  0.874989281
#> 60   0.453324804  1.26409018  0.44745309 -1.222837148 -0.620788078  0.062748961
#> 61   0.306464149 -0.77335115  2.76671000  1.805894209 -0.289966196  0.599698855
#> 62   0.247159875  1.16630792 -0.36126475 -0.835834162 -0.244655177 -0.887115882
#> 63  -0.386838074 -0.51090577 -0.90743605  0.187277780 -0.375083029 -0.747501089
#> 64  -0.158467509  0.74549467 -0.83684533 -0.472313547 -0.269535597  0.620905828
#> 65   0.230264574  1.15192578 -0.39511702 -0.815824743  0.149242974  0.954328975
#> 66  -0.335418046  0.49521853 -0.94125644 -0.326681303 -0.643723099 -0.123246225
#> 67   0.342786097 -1.34499619  0.36940413  0.997956013 -0.112763972  0.766389073
#> 68   0.007367930  0.09385598  4.84371948  2.386481687  0.129324493 -1.138728998
#> 69   0.185288647 -1.25150221 -0.20737497  0.729402273  0.281344071  1.018085734
#> 70   0.405637074 -1.21475203  1.37030280  1.370085668  0.222689101 -1.155558262
#> 71  -0.507497651 -0.02751826 -1.00198517 -0.058110851 -0.010462783 -0.228815535
#> 72   0.335054390  1.23237459 -0.14059213 -0.956017973  0.341219033  1.032458239
#> 73  -0.488301514  0.13120820 -1.00195098 -0.137381951 -0.220711794 -0.909468887
#> 74  -0.489353388  0.12639755 -1.00217419 -0.134965510 -0.326173539  0.560556337
#> 75   0.463903031  1.06196384  1.54347464 -1.623414451  0.228044381  0.996133524
#> 76  -0.023774344 -1.03876888 -0.56673059  0.512441774  0.316529843 -1.091112778
#> 77   0.460413026  1.04682877  1.59840743 -1.641638421 -0.498312489 -0.575208684
#> 78   0.386646191 -1.07939470  1.87152744  1.533923234  0.084054211  0.913862457
#> 79   0.082353923  1.01211105 -0.62068702 -0.666209444 -0.123169704  0.757564689
#> 80   0.079182263  1.00890515 -0.62446585 -0.663383237 -0.469934423  0.378365433
#> 81   0.337985067 -1.34447642  0.34110326  0.986084735 -0.633112082  0.007851653
#> 82   0.346013395  1.23920301 -0.10590489 -0.973592618 -0.289057711 -0.843184668
#> 83   0.064495338 -1.13478667 -0.44269736  0.594610147 -0.251192436  0.886811749
#> 84   0.174032853 -1.24159907 -0.23377705  0.715303899  0.336127731 -0.987085303
#> 85   0.451796861  1.26495288  0.43337840 -1.217051425 -0.291122407  0.598480506
#> 86  -0.320244351  0.52004297 -0.93359740 -0.340325314  0.442884104  0.936915869
#> 87  -0.095711268  0.82051702 -0.79090232 -0.520921103 -0.605081155 -0.345226894
#> 88   0.372375754 -1.33934446  0.58456058  1.085124037 -0.328899507  0.557522653
#> 89  -0.444871682  0.27399014 -0.98878884 -0.209825333 -0.353462802 -0.773025734
#> 90  -0.641688047  1.86843399  8.85732958  3.476693711  0.371158300  0.620576242
#> 91  -0.507325330 -0.06232743 -0.99999080 -0.040791508 -0.497632342 -0.576330004
#> 92  -0.056810362 -1.00077209 -0.60666140  0.483612971  0.225907553 -0.439026825
#> 93   0.278463643 -1.32094668  0.06883219  0.865927524  0.397765849  0.698701837
#> 94   0.153318559 -1.22275910 -0.27953650  0.690330856  0.323116430 -1.073940462
#> 95  -0.496178161 -0.17856761 -0.98812592  0.017135671 -0.187615550  0.700420031
#> 96  -0.197949034 -0.82228921 -0.75231581  0.363948047 -0.637036426 -0.015952233
#> 97  -0.284210571  0.57575962 -0.91446307 -0.371475788 -0.643825180 -0.116013286
#> 98   0.405999182 -1.25536935  1.18578256  1.306813026  0.119298276 -0.117124577
#> 99   0.349209031 -1.34519589  0.40954570  1.014621506  0.063397799  0.039858785
#> 100 -0.257971119  0.61395163 -0.89968945 -0.393305918  0.402932373  0.715179483
#>         x2.pen3     x2.null1       x3.pen1     x3.pen2     x3.pen3    x3.null1
#> 1   -0.55663657  0.566457357 -0.4593804346  0.04599488 -0.92707350 -0.03063678
#> 2   -1.17954162  0.147653830  0.3206108682 -1.03861220  0.87933914  0.93401763
#> 3    4.64294573 -2.178545298 -0.1882953149 -0.69144306 -0.43004857  0.38212636
#> 4   -0.87721360  0.395386890  0.3413621646 -0.77968948  2.11564837  1.31392303
#> 5   -1.25522101  0.021483417 -0.2094284009  0.61847637 -1.00388296 -0.32886714
#> 6   -0.82023161 -0.599748042  0.3413549924  0.72414393  2.00362233 -1.90507691
#> 7    1.53897602  1.262098762 -0.4608969081 -0.14405456 -0.85041583  0.06643225
#> 8    0.43536160  0.941573817 -0.4665414818 -0.03939461 -0.89598062  0.01270672
#> 9   -1.01354579 -0.474403536 -0.4220476774  0.20427026 -0.97118834 -0.11067867
#> 10  -0.11055015 -0.915917847 -0.2277964143 -0.64330201 -0.48474875  0.34980223
#> 11  -0.35660186  0.655150819  0.2238927031 -1.03294582  0.39032929  0.75829412
#> 12  -1.26862728 -0.149820497 -0.1747166584  0.66842585 -0.99871707 -0.35730392
#> 13  -0.95708730  0.343305969  0.1136800592  1.02176483 -0.86973393 -0.59606853
#> 14   2.84844324  1.595820502  0.0484835297 -0.92090616 -0.05484733  0.57312450
#> 15  -1.25013214 -0.206693377 -0.2781984870 -0.57562495 -0.55365326  0.30668713
#> 16  -1.13789808  0.194290135 -0.2261173883 -0.64542734 -0.48244151  0.35119733
#> 17  -0.18942439  0.722685457 -0.0727508355 -0.81424340 -0.26097971  0.47377970
#> 18   0.10632509  0.831613949  0.4390014913  1.29421500 -0.18565624 -1.07275168
#> 19  -0.89757594 -0.553489357 -0.2729492649  0.51959342 -1.00771128 -0.27444807
#> 20  -1.26961124 -0.145296803  0.0649803334 -0.93384567 -0.02291652  0.58757534
#> 21   0.93999808 -1.252361081 -0.4556063746  0.07014270 -0.93492144 -0.04285146
#> 22  -0.85754741 -0.577949716 -0.3708449201  0.33545819 -0.99483805 -0.17762570
#> 23  -1.14480587 -0.359633742 -0.0379982453  0.84646915 -0.95871390 -0.46647690
#> 24  -0.67999992 -0.674796963  0.4552058124  1.29222190 -0.07525645 -1.12758539
#> 25   2.43776322  1.494215093  0.4288136231  0.97396455  1.32159587 -1.68050411
#> 26   0.01783781  0.800208642 -0.3457388601 -0.46905623 -0.64697119  0.24291290
#> 27  -0.94445052 -0.523184688  0.4090149992  1.28617052 -0.33241597 -0.99471763
#> 28  -1.23848840 -0.231578836 -0.4625682935 -0.13033697 -0.85687650  0.05933825
#> 29  -1.24583654  0.043785515 -0.4667244545 -0.05077242 -0.89143444  0.01850845
#> 30   2.23959397  1.444375702  0.0473157323  0.94714748 -0.91600282 -0.53681036
#> 31  -0.11948432  0.749531296 -0.1377591061 -0.74808635 -0.35820184  0.42244565
#> 32   2.52321775  1.515528711  0.3063178700  0.63271464  2.23442051 -1.97776030
#> 33   1.30341681  1.197970286  0.0226555200  0.91866311 -0.93002027 -0.51600311
#> 34  -1.18126616 -0.318199265  0.3541994784  1.25341537 -0.51217719 -0.88812660
#> 35  -0.41466223  0.630407547  0.4816958213  1.24835213  0.30437970 -1.29834935
#> 36   2.59250093 -1.685762297 -0.3854135624  0.30216098 -0.98994484 -0.16052724
#> 37  -1.15446492 -0.349279028  0.4673594825  1.28439624  0.03859264 -1.18135833
#> 38  -0.97479314 -0.502441695 -0.1951544771 -0.68334819 -0.43962021  0.37657898
#> 39  -0.85344787 -0.580389385 -0.2003157152  0.63182940 -1.00272629 -0.33639838
#> 40  -1.23204823 -0.243682155  0.0284599339  0.92540858 -0.92685460 -0.52085626
#> 41  -0.70543071 -0.661878227  0.1080656647  1.01556583 -0.87419810 -0.59083243
#> 42  -1.21832666  0.094107842  0.3904707081  1.27695694 -0.40238862 -0.95492682
#> 43  -1.26274365 -0.172060700  0.3234811564  1.22959926 -0.58613693 -0.83930092
#> 44  -0.44042765  0.619183638 -0.4460326231 -0.22298470 -0.81022308  0.10765028
#> 45  -1.11349617  0.218439054  0.4388993743  1.29420869 -0.18625941 -1.07244383
#> 46  -0.54641357 -0.738682070  0.0461038901 -0.91900819 -0.05935299  0.57106735
#> 47   4.44249269  1.980111052 -0.1548325686 -0.72950840 -0.38276914  0.40890882
#> 48  -0.57643591  0.557092680  0.4821762138  1.24421640  0.32764457 -1.30811835
#> 49  -0.33064886 -0.831093901  0.1290240255 -0.98021709  0.11582464  0.64798233
#> 50  -0.01596641  0.787964699  0.4003383786  0.88742181  1.56917823 -1.76401872
#> 51  -1.12028613 -0.384345468 -0.2387068860 -0.62930451 -0.49971346  0.34068130
#> 52  -1.07729480  0.251231598  0.1246923305 -0.97729457  0.10553030  0.64362288
#> 53  -0.01597550  0.787961386  0.0275131962 -0.90391101 -0.09376414  0.55520353
#> 54   1.37463954  1.217546194  0.4819048647  1.20076638  0.53903957 -1.39406141
#> 55  -0.57598129 -0.725062072  0.4411331964  1.01438522  1.19987388 -1.63842206
#> 56  -0.87653074  0.395809619 -0.0326914654 -0.85176408 -0.19726700  0.50572311
#> 57   0.42994993  0.939839270 -0.2396551146  0.57279658 -1.00666137 -0.30345207
#> 58  -1.08206793 -0.419314131  0.2722080490 -0.52531477  2.85470488  1.51757113
#> 59  -0.81684144 -0.601684796 -0.1869476256 -0.69302131 -0.42816365  0.38321362
#> 60  -1.27541677 -0.101249367  0.4328079710  1.29351037 -0.22066122 -1.05472276
#> 61  -1.10613194 -0.397741300 -0.4136200645  0.22922990 -0.97658715 -0.12334248
#> 62  -0.81081776  0.435009987 -0.4513520713 -0.19961148 -0.82267331  0.09536692
#> 63  -0.96356300  0.338824663 -0.3787395802  0.31770662 -0.99232404 -0.16849889
#> 64  -1.09135318 -0.411155023 -0.4019902904  0.26095982 -0.98284204 -0.13948217
#> 65  -0.65411050 -0.687678485  0.1615102179 -0.20580448  3.65641637  1.72822196
#> 66  -1.26438485 -0.005985138  0.4760312562  1.27079586  0.15913925 -1.23573983
#> 67  -0.96202580 -0.511288458 -0.3842627190  0.30487964 -0.99037285 -0.16192001
#> 68  -0.20593781  0.716234448 -0.3689051110  0.33972349 -0.99540976 -0.17982273
#> 69  -0.39902322 -0.802981176  0.3853655690  1.27403092 -0.41974126 -0.94475082
#> 70   0.05405719  0.813172719 -0.3987387604 -0.36280065 -0.72525831  0.18291583
#> 71   3.59702922 -1.928750089 -0.2766247818  0.51350377 -1.00768765 -0.27116438
#> 72  -0.22380830 -0.873247113  0.4159419557  1.28892492 -0.30290862 -1.01095366
#> 73  -0.78078937  0.452049036 -0.1166720394 -0.77031027 -0.32736040  0.43910089
#> 74  -1.13124457 -0.373557345  0.3480775609 -0.81449130  1.99995319  1.28086092
#> 75  -0.51725507 -0.751856978 -0.2934064220  0.48507376 -1.00719845 -0.25592248
#> 76   0.56341195  0.982005842  0.3633662969  1.25991120 -0.48710882 -0.90391797
#> 77  -1.09571916  0.234940074  0.1881729621 -1.01607782  0.27398978  0.71283271
#> 78  -0.74644793 -0.640441111 -0.4496044780 -0.20766840 -0.81843486  0.09959310
#> 79  -0.97156861 -0.504693029  0.4827286853  1.23716215  0.36566953 -1.32394349
#> 80  -1.21740528 -0.268269064 -0.4128160165  0.23151619 -0.97706062 -0.12450377
#> 81  -1.27552390 -0.073077821  0.2583754879 -1.04404105  0.52520642  0.80900672
#> 82  -0.86472648  0.403063938  0.1962549117  1.10997468 -0.78954414 -0.67899894
#> 83   4.71774982  2.046137780  0.3379246366 -1.02300029  1.03918875  0.98734562
#> 84   0.94655760  1.097037295  0.1241816026 -0.10558625  3.89403787  1.78935780
#> 85  -1.10695507 -0.396977665  0.0007607776 -0.88135432 -0.14109844  0.53291916
#> 86   0.56801379 -1.142197754 -0.4601867490  0.04011659 -0.92510093 -0.02766141
#> 87  -1.20844861  0.109100772  0.2530679867  1.16655580 -0.71422172 -0.74426455
#> 88  -1.13308009 -0.371711541  0.2650321611 -1.04538559  0.55471123  0.81984529
#> 89  -0.93933983  0.355370170  0.3695181673  1.26408278 -0.46936448 -0.91489071
#> 90   1.59084326 -1.431202471  0.1303776101  1.04006408 -0.85578552 -0.61191461
#> 91  -1.09500944  0.235582131  0.4595579431  1.29030372 -0.03882377 -1.14507495
#> 92   2.25444063  1.448131149  0.3264482515 -1.03466172  0.92752189  0.95026995
#> 93   1.37306141 -1.372928133  0.1184990476  0.17253953  3.32481438 -2.30758176
#> 94   0.63781734  1.004997158 -0.4249881907  0.19507567 -0.96909335 -0.10601938
#> 95  -1.02747587 -0.463835078  0.2743071076 -0.53195466  2.83697453  1.51281205
#> 96  -1.27469419 -0.060889438  0.4826454613  1.23852989  0.35844530 -1.32095016
#> 97  -1.26541750 -0.009690294  0.4826168043  1.21090787  0.49372645 -1.37603725
#> 98   2.88322785  1.604340877 -0.0065848239  0.88426735 -0.94480966 -0.49189468
#> 99   3.17685421  1.675850868 -0.1024489444  0.76566154 -0.98151937 -0.41511516
#> 100  1.32538480 -1.359981375 -0.2696247234  0.52506175 -1.00770776 -0.27740269
#>          x4.pen1     x4.pen2     x4.pen3    x4.null1      x5.pen1     x5.pen2
#> 1    0.227768182  0.45935194  3.37915774 -1.85166830  0.285367241  1.17173353
#> 2   -0.069784448 -0.89975139 -0.74307729  0.50693864  0.439529660 -1.19630314
#> 3    0.201049585  1.17174089  0.10723554 -0.78261682 -0.429495282  0.30980253
#> 4    0.255208886 -1.21909279 -0.33539983  0.81609459 -0.472827960  0.02066847
#> 5   -0.422086807 -0.39199411 -0.91254534  0.21156045  0.284056093  1.22337085
#> 6   -0.227433332 -0.70701513 -0.83975051  0.38454125 -0.466485080  0.12259035
#> 7    0.438543475 -1.29036577  0.40414646  1.17689750  0.541877686 -1.19780560
#> 8    0.115314007  1.10812459 -0.09881188 -0.68405003 -0.031739040 -0.81011036
#> 9   -0.485156747  0.29175989 -0.83272470 -0.13353366  0.473404514 -1.20990485
#> 10  -0.413629092 -0.40931370 -0.91060332  0.22058378  0.197035119  1.30130001
#> 11  -0.127572717  0.87271632 -0.46748787 -0.47180634  0.267268357 -1.07946393
#> 12   0.198906595  1.17035066  0.10120964 -0.77986965 -0.240323294 -0.57949314
#> 13   0.444085272 -1.28521092  0.46100887  1.20046363 -0.408641833  0.37854024
#> 14   0.014011543 -0.99097194 -0.67431504  0.57399918 -0.151937034 -0.68457657
#> 15   0.370513710  0.96287824  1.98036045 -1.45193911  0.017542454  1.12608530
#> 16  -0.494793437 -0.20417367 -0.92052361  0.11569325 -0.391592870  0.42757915
#> 17  -0.443508757  0.40068806 -0.79091955 -0.19034767 -0.359070454 -0.40412948
#> 18  -0.352054618  0.57438886 -0.70345866 -0.28514257  0.338337293 -1.13374283
#> 19  -0.482305179 -0.24435640 -0.92076728  0.13596120 -0.467255219  0.11558347
#> 20  -0.198323320 -0.74526701 -0.82453411  0.40739562  0.519592548 -1.21367817
#> 21  -0.527044906  0.08837476 -0.88801136 -0.03050376 -0.290452486  0.65583862
#> 22  -0.407790178 -0.42094970 -0.90917822  0.22666853  0.064619837  1.18257534
#> 23  -0.313471635  0.63467784 -0.66602219 -0.31978102 -0.233848526  0.75909757
#> 24  -0.509931528 -0.14490218 -0.91828911  0.08595195  0.129859272 -0.96186467
#> 25   0.342794122  0.83816763  2.36997340 -1.56775282 -0.331547160 -0.45028437
#> 26  -0.465684776  0.34680994 -0.81273980 -0.16205229  0.230416708 -1.04927424
#> 27   0.273670014 -1.23334259 -0.29432165  0.84017574 -0.043253734  1.04712396
#> 28   0.263585276  1.20556407  0.31263194 -0.87234908 -0.398938510 -0.32648531
#> 29  -0.152388203  0.84456425 -0.49643090 -0.45203963 -0.472566807  0.03582631
#> 30  -0.037814497  0.96772945 -0.35323632 -0.54440979 -0.166046866  0.87002338
#> 31  -0.520314070  0.14546419 -0.87530865 -0.05915442  0.189911017  0.61654533
#> 32   0.349331490  0.86458075  2.29122348 -1.54469155 -0.282226785  0.67160563
#> 33   0.267091380 -1.22834769 -0.30935185  0.83145061 -0.279639323 -0.52724865
#> 34   0.170283274 -1.14600254 -0.48878299  0.71837113  0.270448585  1.28303717
#> 35  -0.377271823  0.53170525 -0.72759207 -0.26123570  0.192588150 -1.01716239
#> 36  -0.397624236  0.49485253 -0.74697623 -0.24094789 -0.117639624 -0.54806303
#> 37  -0.011097466 -0.96427497 -0.69653825  0.55348122  0.534366938 -1.20617114
#> 38   0.381447566  1.03465106  1.72775448 -1.37423496 -0.402719686 -0.31815509
#> 39   0.111449889  1.10493990 -0.10670758 -0.68005785 -0.471511193 -0.03666597
#> 40  -0.407288097  0.47645098 -0.75617778 -0.23092794 -0.435219654  0.28845132
#> 41   0.376881441 -1.29367219  0.02662878  1.00830010 -0.197160657 -0.63259302
#> 42   0.227802304  1.18795404  0.18746545 -0.81853163 -0.113881146 -0.72603413
#> 43   0.019689464  1.02359230 -0.26933171 -0.59328893 -0.352764951 -0.41512729
#> 44   0.413509417 -1.29951386  0.21402740  1.09492922  0.540754355 -1.19939407
#> 45   0.221112538 -1.19106799 -0.40321183  0.77457412  0.208506804  1.30630601
#> 46   0.454894946 -1.26743208  0.60838819  1.25980910 -0.347217501 -0.42458099
#> 47  -0.509441000  0.20311317 -0.86028946 -0.08826579 -0.462522161 -0.12061403
#> 48  -0.417334995  0.45660360 -0.76575917 -0.22019671 -0.135807149  0.91610545
#> 49  -0.048245626  0.95719429 -0.36742747 -0.53580462  0.174250122  1.28762493
#> 50   0.345603939  1.21107709  0.75220519 -1.04512643  0.040183573  1.15384552
#> 51   0.305324050  1.21724256  0.49745705 -0.94769188 -0.038657631 -0.25781227
#> 52  -0.096555744  0.90670193 -0.42985684 -0.49661025  0.541319639 -1.19861375
#> 53   0.178173680  1.15629020  0.04558450 -0.75415705  0.262423845 -1.07556403
#> 54   0.416384878 -1.29923220  0.23223388  1.10301216  0.515061393 -1.21466041
#> 55  -0.460699617 -0.30377905 -0.91921067  0.16614279  0.238054802  1.31040475
#> 56   0.298932977  0.68101621  2.81044150 -1.69402823 -0.298797716 -0.50008206
#> 57   0.411199986 -1.29963750  0.19988914  1.08861507  0.043685547  1.15804561
#> 58   0.363626703 -1.28861481 -0.02728994  0.98204930  0.073500689 -0.91040290
#> 59   0.077210618  1.07579221 -0.17260162 -0.64600067  0.067681744  1.18607494
#> 60  -0.495579442 -0.20143798 -0.92046948  0.11431683 -0.171444587 -0.66255288
#> 61   0.025498268  1.02903114 -0.26023177 -0.59840206 -0.429708642 -0.25053657
#> 62  -0.311454806 -0.58688067 -0.87760205  0.31590936  0.282946445  1.23428934
#> 63   0.165678041  0.28037075  3.81219362 -1.96876189  0.276982807  1.06469325
#> 64   0.292372659  1.21496923  0.43419034 -0.92239823 -0.345229893  0.54194254
#> 65  -0.417290398 -0.40188644 -0.91146223  0.21670945 -0.287962650  0.66064217
#> 66  -0.419183182 -0.39800523 -0.91189545  0.21468776 -0.446466194 -0.19615078
#> 67  -0.238263844  0.73945309 -0.59013158 -0.38305676  0.237513615  1.31048567
#> 68   0.231326054 -0.43648980  3.21914508  2.10850078 -0.352001440 -0.41644038
#> 69   0.451271792 -1.27505364  0.55114127  1.23703837 -0.218024393  0.78605931
#> 70  -0.506934425 -0.15799199 -0.91897352  0.09250714 -0.423616033  0.33044708
#> 71  -0.003020396  1.00197180 -0.30372541 -0.57364442 -0.435457862  0.28753211
#> 72   0.364331622 -1.28891235 -0.02455733  0.98339589 -0.453159634 -0.16902911
#> 73  -0.268232328  0.69939963 -0.62093109 -0.35833609 -0.467396896 -0.08512977
#> 74  -0.470341545  0.33444826 -0.81742618 -0.15561645  0.027551516  1.13848258
#> 75  -0.356053581 -0.51509919 -0.89394679  0.27666911  0.222025716  1.31005724
#> 76   0.178059212  1.15620968  0.04529006 -0.75401919  0.546391929 -1.09207793
#> 77   0.446222728 -1.28268826  0.48546367  1.21047869  0.363963705 -1.15164429
#> 78   0.135322319 -0.19566300  3.79263688  2.27204029  0.549917238 -1.11270797
#> 79   0.015384686 -0.13038686  4.76191200 -2.22031879  0.432172406 -1.19267340
#> 80  -0.068657705  0.93620776 -0.39443044 -0.51912977 -0.029736728  0.16141739
#> 81   0.281187247 -1.23892971 -0.27657851  0.85035430 -0.217549454  0.78685750
#> 82  -0.497379631  0.25118381 -0.84603042 -0.11272515  0.247069969 -1.06306243
#> 83  -0.207785934  0.77826003 -0.55788967 -0.40775726 -0.452220936  0.21410307
#> 84   0.391683649 -1.29774806  0.09426371  1.04033830 -0.300897105  0.63537967
#> 85   0.385558297  1.10525597  1.44168770 -1.28316330 -0.169797154  0.86417682
#> 86   0.384670349  1.12564985  1.34732511 -1.25228811  0.424556390 -1.18871479
#> 87   0.356547235  0.89535901  2.19718607 -1.51693390 -0.104522774 -0.73595674
#> 88   0.132797498 -1.11077294 -0.54247394  0.68037693  0.002120402 -0.10781727
#> 89  -0.342461939  0.58989367 -0.69421955 -0.29394707 -0.421486762  0.33764836
#> 90  -0.238770488  0.73879231 -0.59065934 -0.38264288 -0.021325846 -0.82034392
#> 91  -0.198283844  0.79000071 -0.54762568 -0.41539367  0.244004404  1.30898423
#> 92   0.196916522  1.16904853  0.09566242 -0.77733428  0.233263274  1.31087172
#> 93  -0.223764152  0.75814158 -0.57491681 -0.39485274 -0.446437530  0.24177351
#> 94  -0.288287763 -0.62169587 -0.86806416  0.33537299 -0.379223996  0.46037243
#> 95   0.246689521  0.51600363  3.23795533 -1.81299451 -0.323701718 -0.46265681
#> 96  -0.243148804 -0.68572819 -0.84751237  0.37205607 -0.334709505 -1.34577270
#> 97  -0.080603133  0.92369271 -0.40979701 -0.50945500 -0.369346976  0.48522653
#> 98  -0.487408491  0.28472099 -0.83511801 -0.12991202  0.523652019 -1.21236940
#> 99  -0.496702889 -0.19747658 -0.92038266  0.11232444  0.050013519  1.16556702
#> 100  0.225637159  1.18672697  0.18059952 -0.81550431  0.249979951  0.89511788
#>         x5.pen3    x5.null1
#> 1    1.34656019 -1.43757887
#> 2   -0.05347023  0.89807855
#> 3   -0.93909011 -0.18229554
#> 4   -0.98194013 -0.03875751
#> 5    1.12186607 -1.35948135
#> 6   -0.97248179 -0.08919722
#> 7    0.58152689  1.17539258
#> 8   -0.77920774  0.42647417
#> 9    0.08058837  0.96196516
#> 10   0.16712953 -0.98332351
#> 11  -0.45831131  0.67468836
#> 12  -0.90086000  0.27797674
#> 13  -0.92139085 -0.21699174
#> 14  -0.85482606  0.34214833
#> 15  -0.40203003 -0.68964286
#> 16  -0.90685672 -0.24202743
#> 17  -0.95229527  0.17906547
#> 18  -0.32657291  0.75391786
#> 19  -0.97332648 -0.08573134
#> 20   0.34917494  1.08052301
#> 21  -0.81550213 -0.36348598
#> 22  -0.29665477 -0.75154904
#> 23  -0.75863399 -0.42259131
#> 24  -0.63953311  0.54843821
#> 25  -0.94141205  0.20432783
#> 26  -0.51474911  0.63795716
#> 27  -0.51306827 -0.61832173
#> 28  -0.96682393  0.13753434
#> 29  -0.98091800 -0.04626816
#> 30  -0.68289970 -0.49088082
#> 31   2.96932438 -1.94451648
#> 32  -0.80753982 -0.37229462
#> 33  -0.91917067  0.24761154
#> 34   0.77469485 -1.23251244
#> 35  -0.56616728  0.60264056
#> 36   5.59650343 -2.68196985
#> 37   0.48570032  1.13699795
#> 38  -0.96811582  0.13314192
#> 39  -0.98458450 -0.01028200
#> 40  -0.94397502 -0.17159410
#> 41  -0.87925854  0.30982423
#> 42  -0.83270092  0.36889444
#> 43  -0.94985993  0.18504275
#> 44   0.56523574  1.16893209
#> 45   0.22931228 -1.01098094
#> 46  -0.94768848  0.19020141
#> 47  -0.98492026  0.03170999
#> 48  -0.64572103 -0.52130080
#> 49   0.05917673 -0.93374732
#> 50  -0.35381967 -0.71856561
#> 51   4.95862627 -2.50526123
#> 52   0.57332328  1.17214264
#> 53  -0.46613040  0.66971314
#> 54   0.31497110  1.06601305
#> 55   0.42613386 -1.09484314
#> 56  -0.92763554  0.23215194
#> 57  -0.34597241 -0.72317364
#> 58  -0.69511195  0.50352904
#> 59  -0.28907394 -0.75582587
#> 60  -0.86560149  0.32830155
#> 61  -0.97679016  0.09787873
#> 62   1.06832949 -1.34043971
#> 63   1.73196227 -1.56574840
#> 64  -0.86630793 -0.30166340
#> 65  -0.81310203 -0.36616226
#> 66  -0.98151579  0.06996108
#> 67   0.42187886 -1.09308326
#> 68  -0.94956266  0.18575812
#> 69  -0.74178212 -0.43865646
#> 70  -0.93409375 -0.19267424
#> 71  -0.94417894 -0.17113407
#> 72  -0.98315249  0.05616450
#> 73  -0.98529748  0.01390599
#> 74  -0.38123921 -0.70224918
#> 75   0.31154465 -1.04665733
#> 76   1.22398900  1.41284853
#> 77  -0.26936640  0.78604364
#> 78   1.12634170  1.37859134
#> 79  -0.07866244  0.88564134
#> 80   4.88073564  2.50237505
#> 81  -0.74126919 -0.43913672
#> 82  -0.49008977  0.65424630
#> 83  -0.95879332 -0.13453667
#> 84  -0.82547827 -0.35215651
#> 85  -0.68734845 -0.48712052
#> 86  -0.10357206  0.87319505
#> 87  -0.82702237  0.37544082
#> 88   4.62871030 -2.41382662
#> 89  -0.93228704 -0.19630247
#> 90  -0.77177387  0.43385346
#> 91   0.47512424 -1.11495254
#> 92   0.38950095 -1.07961974
#> 93  -0.95367157 -0.14829627
#> 94  -0.89621041 -0.25893051
#> 95  -0.93819216  0.21118264
#> 96   7.34961087 -3.16762636
#> 97  -0.88762793 -0.27183983
#> 98   0.38230858  1.09443534
#> 99  -0.33150911 -0.73159787
#> 100  2.24320201 -1.72712567