Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>           x1.pen1      x1.pen2      x1.pen3     x1.null1      x2.pen1
#> 1    0.0642369025 -0.566645547 -0.896276711  0.376561338 -0.157006411
#> 2   -0.5404341191 -1.073718554 -0.251511981  1.089623465 -0.406749112
#> 3   -0.5734481537 -1.049100960  0.110482948  1.284199912 -0.533197744
#> 4    0.0239741842  0.903042212  0.214281867 -0.572419639 -0.467263139
#> 5   -0.3124840432 -0.924785489 -0.753604303  0.698113264 -0.253781445
#> 6   -0.5435279592 -0.916919526  0.750008610  1.569030031  0.033311833
#> 7    0.2737997760  0.577660811 -0.327191812 -0.298111417 -0.544315294
#> 8    0.2855747124  0.556052901 -0.350613805 -0.283939783  0.246176755
#> 9    0.3346143561 -0.206224792 -0.836949306  0.155005365  0.029000431
#> 10   0.3746550291  0.341937455 -0.544818300 -0.153055200 -0.540546460
#> 11   0.4101256005  0.066455211 -0.722019636  0.002421170 -0.450004341
#> 12   0.0180581055 -0.614348200 -0.894030276  0.409909374  0.397687390
#> 13   0.1844792060 -0.430716947 -0.888025321  0.288166257 -0.011191060
#> 14   0.0009634855  0.923735208  0.278696641 -0.600439316 -0.048724763
#> 15  -0.4314510008 -1.018755128 -0.593285244  0.853959537  0.336830544
#> 16   0.3983640582 -0.027109921 -0.767809260  0.054337789  0.345276655
#> 17  -0.0471180509 -0.679145349 -0.885591924  0.458009635  0.416223021
#> 18   0.0866186421  0.839176292  0.059205213 -0.501780155  0.258181642
#> 19   0.3690545032  0.360143689 -0.530552770 -0.163687266 -0.121532613
#> 20  -0.1426425622  0.441754203  2.919405187 -1.471016163  0.050025860
#> 21   0.3509937975  0.412177421 -0.487711791 -0.194483036  0.077142863
#> 22   0.2934680726  0.541013960 -0.366425141 -0.274211250 -0.477451948
#> 23   0.3818172618 -0.091889944 -0.795607809  0.090441009  0.402501693
#> 24  -0.1878833577  0.718093261  2.181535163 -1.257331395  0.398974013
#> 25   0.3907320949  0.280523063 -0.590374220 -0.117642349  0.403365139
#> 26   0.4098995798  0.062512274 -0.724086192  0.004608591  0.274059975
#> 27   0.4041974411  0.006726806 -0.752026689  0.035550085 -0.077158108
#> 28   0.0803242133 -0.549564261 -0.896368270  0.364964594 -0.056939098
#> 29   0.3982237758 -0.027809333 -0.768126310  0.054726496  0.310497527
#> 30   0.0774774510  0.849147641  0.080438333 -0.511744255  0.270499542
#> 31   0.3297663309 -0.215630523 -0.839907517  0.160385426 -0.546525324
#> 32  -0.1812850864  0.945314334  1.308359104 -0.983630541 -0.428803011
#> 33  -0.3035500485 -0.917130638 -0.761897951  0.688195045  0.371971400
#> 34   0.0929153265  0.832179606  0.044798507 -0.494962028 -0.064517461
#> 35   0.2278442239  0.654292603 -0.236510045 -0.350567028  0.289838708
#> 36   0.3396324312 -0.196189737 -0.833718200  0.149278877 -0.077537950
#> 37   0.3420650316  0.435148537 -0.467762561 -0.208298183 -0.053999968
#> 38  -0.1953129065  0.844803908  1.762191545 -1.129557778 -0.330293793
#> 39   0.2169513676  0.670970564 -0.214959501 -0.362528984 -0.160703171
#> 40  -0.2567305361 -0.876083106 -0.799527604  0.638998210  0.375799008
#> 41  -0.4745425816 -1.046856398 -0.498038176  0.928128572  0.194189572
#> 42   0.1837865094 -0.431568188 -0.888136032  0.288696200 -0.546118766
#> 43   0.3042447992  0.519675031 -0.388208775 -0.260581884 -0.470417562
#> 44  -0.0754993617  0.977849168  0.545495572 -0.709858483 -0.193660519
#> 45  -0.5633323691 -1.068410717 -0.074131366  1.189455342  0.308715533
#> 46   0.3899191287  0.284089635 -0.587832565 -0.119681961  0.033290060
#> 47   0.0485262502  0.879271820  0.150563164 -0.543972728 -0.527635044
#> 48  -0.1568967516  0.981840164  1.026073781 -0.887437355 -0.433268681
#> 49  -0.5055194092 -1.063156440 -0.405137506  0.993090114 -0.238485801
#> 50  -0.5744187142 -1.031364085  0.227679897  1.340705912  0.049041800
#> 51  -0.1372527512  0.990915688  0.873290036 -0.833141526  0.429042228
#> 52  -0.1370148269  0.990966326  0.871656889 -0.832551463  0.368837971
#> 53   0.2329840107 -0.368233328 -0.878083053  0.249890515  0.344761235
#> 54   0.1926716142  0.706447456 -0.166539091 -0.388781817 -0.542112111
#> 55   0.3304169111  0.463106983 -0.442561641 -0.225321988  0.041101503
#> 56   0.2338767707  0.644835130 -0.248416404 -0.343880326 -0.292431341
#> 57   0.1169892610  0.804455274 -0.008824183 -0.469156024 -0.544687005
#> 58  -0.5745173415 -1.036468590  0.196330767  1.325830268  0.409763590
#> 59  -0.1345011787  0.991431845  0.854657542 -0.826396669 -0.465769438
#> 60   0.1951348235 -0.417490752 -0.886217724  0.279963438  0.256066900
#> 61   0.0198054605  0.906909049  0.225575582 -0.577384088  0.366518809
#> 62  -0.5235116690 -0.868356000  0.928170014  1.640579501 -0.364784889
#> 63  -0.1224544077 -0.751858195 -0.866721622  0.517314712 -0.548927571
#> 64  -0.3562261797 -0.961270154 -0.706768120  0.749643280  0.227587144
#> 65   0.2218378577  0.663550769 -0.224636696 -0.357179575  0.115951487
#> 66   0.0707381077  0.856358123  0.096351458 -0.519147437 -0.161980237
#> 67  -0.1296011964  0.991996754  0.822775704 -0.814789181  0.379712888
#> 68   0.3497520369  0.415463784 -0.484898138 -0.196450604  0.370076506
#> 69  -0.1470706301  0.464046477  2.864617441 -1.455520090  0.136188686
#> 70  -0.0785319262 -0.709682116 -0.879047198  0.482112587  0.428519254
#> 71   0.3316080895 -0.212088947 -0.838801648  0.158358174 -0.510746669
#> 72   0.0223706241  0.904535631  0.218608165 -0.574324051  0.003796691
#> 73   0.3389147419  0.442915902 -0.460864985 -0.213003686 -0.473617889
#> 74   0.3473934450  0.421620995 -0.479590631 -0.200144815 -0.548808005
#> 75   0.3437909695  0.430822325 -0.471570561 -0.205685010 -0.264901110
#> 76   0.3276163344  0.469539230 -0.436614499 -0.229273709 -0.068854000
#> 77  -0.2617155406 -0.880521597 -0.795947241  0.644033215 -0.113784883
#> 78   0.4109256481  0.086523836 -0.711310936 -0.008716786  0.312097109
#> 79  -0.5678995498 -0.991081629  0.438936337  1.437041196  0.393878616
#> 80   0.0336470693  0.893877718  0.188637177 -0.561061641 -0.517205033
#> 81  -0.5105987420 -0.839283180  1.028972843  1.680001632  0.271118592
#> 82   0.2326585065 -0.368674196 -0.878165405  0.250156599 -0.276629665
#> 83  -0.1632812026  0.550784119  2.645244489 -1.392946588 -0.470978490
#> 84  -0.2737278129 -0.891153691 -0.786923542  0.656354790  0.024006050
#> 85   0.2299249432 -0.372363569 -0.878847928  0.252385363 -0.399009030
#> 86  -0.0704401944  0.975197210  0.524438886 -0.701560704  0.064676595
#> 87   0.3308556915  0.462089725 -0.443496957 -0.224698274  0.391434510
#> 88  -0.5585616960 -0.958820464  0.582211991  1.499093675  0.430680512
#> 89   0.0996542426  0.824575205  0.029564668 -0.487700526  0.294083866
#> 90  -0.0248949357  0.944911029  0.358382699 -0.634166637  0.383037689
#> 91   0.3732615979  0.346583606 -0.541212073 -0.155762048  0.195037994
#> 92  -0.0713187483  0.975669836  0.528051149 -0.702987875  0.430470723
#> 93  -0.3159319507 -0.465290986  2.145952944  2.082390951 -0.541238263
#> 94  -0.4591756711 -1.037404748 -0.535571471  0.899978935 -0.011434256
#> 95   0.1809273029 -0.435071125 -0.888584392  0.290879579  0.399368566
#> 96  -0.5603913147 -1.070406918 -0.105146742  1.172709139  0.389684804
#> 97   0.3910559365  0.279083574 -0.591396529 -0.116819684  0.352482730
#> 98  -0.0242933475 -0.656718358 -0.889291000  0.440938239  0.202886637
#> 99   0.4111510164  0.102710774 -0.702439629 -0.017708086  0.191087418
#> 100 -0.1867348665  0.927254436  1.407725595 -1.016405190 -0.285746504
#>         x2.pen2     x2.pen3     x2.null1      x3.pen1      x3.pen2      x3.pen3
#> 1   -0.55937883  5.13281187 -2.281945843  0.469388893  0.827580684 -1.153943215
#> 2    0.51298346 -1.03231407 -0.216138560  0.012583851 -1.038772683  0.198619825
#> 3    0.20956148 -1.18070693 -0.047883330 -0.128911071 -0.904969674  0.460351183
#> 4   -0.32000177 -1.26390232  0.226392448 -0.422711876  0.193995660  1.113646863
#> 5   -0.70295090 -1.19535450  0.436647695  0.228389505 -1.119708128 -0.588839804
#> 6   -1.04970928 -0.98211652  0.674505141  0.246810509 -1.100584949 -0.748470657
#> 7    0.14163732 -1.20277941 -0.012096940  0.155133753 -1.124997323 -0.205377802
#> 8    0.37144903  3.02705764 -1.734677720  0.007120859 -1.034324342  0.210260175
#> 9   -1.04522568 -0.98662693  0.670717278 -0.493032394 -0.151977987  1.037976955
#> 10  -0.05636113 -1.24730257  0.090378079  0.046322311  0.006840512 -3.684224029
#> 11   0.43649187 -1.07829759 -0.171982630 -0.053256062 -0.981225170  0.329255755
#> 12  -1.28904751 -0.02680999  1.203461269 -0.103132548  0.636778925  1.060517015
#> 13  -1.00251227 -1.02615622  0.636126835  0.296343351  0.993607336  0.728148397
#> 14   0.91774044 -0.62909284 -0.495460102  0.374460729  1.046096659  0.575421496
#> 15   1.14042625  0.19499317 -0.868595147 -0.484173870  0.001223578  1.082791473
#> 16   0.63510026  2.36257307 -1.553883456  0.264822285 -0.869458536 -1.659143150
#> 17   0.87842305  1.65688533 -1.352001090 -0.407835275 -0.488920257  0.871093304
#> 18   0.40159603  2.95411219 -1.715154835 -0.410033076 -0.483923868  0.874358892
#> 19  -0.87617092 -1.11577297  0.545169849  0.209010911  0.925292028  0.846207161
#> 20   0.99537713 -0.49002155 -0.569818707  0.268242012 -1.036555071 -1.082605811
#> 21  -1.09422052 -0.93277550  0.714139365 -0.016207369 -1.014646565  0.258178537
#> 22   0.37922629 -1.10865477 -0.139824372  0.196481735 -0.526116864 -2.538557566
#> 23  -1.01640418  1.14695596  1.636979465  0.349634750  1.030634289  0.630624355
#> 24   1.12413215  0.54370036 -0.996345298  0.243465341 -1.105245003 -0.715264642
#> 25  -1.02031087  1.13515840  1.633067421  0.014414422 -1.040249285  0.194680581
#> 26   1.12576513 -0.02609007 -0.780869966  0.514600700  1.059654153 -0.139365065
#> 27  -0.92867389 -1.08264771  0.581319876  0.211480207 -1.127817370 -0.475858117
#> 28   0.91079793 -0.63980095 -0.489428469  0.421490156  1.070702871  0.445226399
#> 29  -1.28252343 -0.46887577  0.998498270  0.238048237 -0.708406281 -2.097950069
#> 30   1.12446004 -0.03663488 -0.776522668  0.122806678 -1.112173112 -0.090004949
#> 31  -0.00187549 -1.23787091  0.062350274  0.234476701 -1.114930520 -0.636113453
#> 32  -0.40728252 -1.25864858  0.272303050  0.314001303  1.006364798  0.698506490
#> 33   0.71589011  2.14240906 -1.492212549  0.261032256  0.374601907 -2.458115290
#> 34   0.90432454 -0.64958276 -0.483875508  0.352501595  1.032487493  0.624619587
#> 35   0.48296191  2.75388275 -1.661193020  0.370715669  1.043854050  0.584240753
#> 36   0.89304720 -0.66618021 -0.474356425 -0.380716738 -0.546245404  0.831590175
#> 37  -0.95514434 -1.06387670  0.600360913  0.252366047 -0.784846524 -1.897694044
#> 38  -0.58693867 -1.22937813  0.369985757  0.115004073 -1.108294913 -0.064939034
#> 39  -0.82770465 -1.14218417  0.513370583  0.476409674  0.846449765 -1.090280980
#> 40  -1.29777705 -0.18143443  1.135973641 -0.460636574  1.650621597 -6.999583662
#> 41   1.08870409 -0.22983489 -0.693682432  0.228105642  0.940852083  0.823802878
#> 42   0.12386292 -1.20795913 -0.002802873  0.066548748 -1.079098198  0.072563220
#> 43   1.38426314  6.55304740  3.196576687 -0.216765916  0.509688253  1.097176807
#> 44  -0.78520503 -1.16250715  0.486499165  0.403620052  1.062235608  0.499523710
#> 45   0.53309071  2.62769852 -1.626877260 -0.037706615 -0.995569088  0.300093700
#> 46   0.98296474 -0.51514768 -0.556883398  0.459783905  1.083247010  0.298247509
#> 47   0.23382535 -1.17192350 -0.060785731 -0.414616650 -0.473301834  0.881210710
#> 48  -0.39784875 -1.25948530  0.267303340  0.514555591  0.975356908 -0.600093100
#> 49  -0.72451725 -1.18750551  0.449476641  0.267913718 -0.897934118 -1.573381497
#> 50   0.99465569 -0.49151971 -0.569053032  0.429403446  1.074011638  0.418772163
#> 51   0.94988342  1.41350882 -1.279021566  0.262569997  0.377645096 -2.450130325
#> 52   0.70601380  2.16990891 -1.499973487  0.007374678 -1.034532392  0.209723104
#> 53   0.63360383  2.36655659 -1.554989957  0.386474911  1.053046859  0.545794333
#> 54  -0.04474101 -1.24546580  0.084406068 -0.311941979  0.385056858  1.114730708
#> 55  -0.11197783  4.14014521 -2.026276727 -0.470901908  0.061829750  1.095569976
#> 56   0.67251008 -0.91248724 -0.314299226 -0.414331411  0.211827781  1.115038439
#> 57   0.13831929 -1.20376480 -0.010359964  0.280582424  0.413454704 -2.355706424
#> 58   1.11311767  0.63873881 -1.029392642  0.257147157 -1.080517012 -0.871212829
#> 59  -0.32369128 -1.26379159  0.228319061 -0.304288799 -0.679364613  0.723340458
#> 60   1.11878034 -0.07773858 -0.759417317 -0.063403822  0.677528239  1.043726252
#> 61   0.69878203  2.18993366 -1.505614030 -0.447812450  0.133723098  1.107083396
#> 62   0.57702174 -0.98847506 -0.254402409 -0.455876422  0.111183713  1.103901683
#> 63   0.05789577 -1.22509661  0.031488067 -0.004017891  0.735758910  1.014358624
#> 64   1.10596561 -0.15207572 -0.727797030  0.014743518 -1.040513995  0.193970284
#> 65   1.04125642 -0.38265821 -0.622986500  0.518033681  1.047907225 -0.222929649
#> 66  -0.82608836 -1.14300326  0.512332491  0.520022440  1.028661859 -0.339369340
#> 67   0.74087053  2.07202385 -1.472268425  0.271536770 -0.949506114 -1.408223100
#> 68   1.13814327  0.35443257 -0.928397927  0.009968266 -1.036650652  0.204214345
#> 69   1.05432353 -0.34634389 -0.640273009  0.150570770 -0.351688672 -2.929613261
#> 70   1.07155541  0.89840831 -1.116598070  0.259508692  0.371588823 -2.466015437
#> 71   0.29270941 -1.14855762 -0.092406957  0.132136988 -1.116444307 -0.121279907
#> 72   0.96035354 -0.55774362 -0.534476860 -0.386176785 -0.535274307  0.839449316
#> 73  -0.30396140 -1.26427144  0.218028605  0.275247153  0.977836242  0.760665972
#> 74   0.04581014 -1.22788757  0.037741546 -0.487747898 -0.022183578  1.077106712
#> 75  -0.68696918 -1.20084419  0.427239336  0.319402108  1.010177448  0.688951932
#> 76  -0.93823637 -1.07604001  0.588128880 -0.482502703 -0.247438951  1.000770599
#> 77   0.86058529 -0.71110405 -0.447940864  0.165438462  0.188727094 -2.936426395
#> 78  -1.28328732 -0.46347778  1.001262250  0.350786014  1.031380465  0.628223736
#> 79  -1.29157560 -0.05796429  1.190157277 -0.368154814 -0.570573171  0.813631375
#> 80   0.27207600 -1.15708197 -0.081273250  0.230493808 -1.118192381 -0.604699457
#> 81  -1.26006398 -0.58638288  0.936104319  0.260273028 -1.071835756 -0.917611290
#> 82  -0.66981850 -1.20643650  0.417230399 -0.050681012 -0.983631895  0.324487961
#> 83   0.39356707 -1.10136044 -0.147813820 -0.394296437  0.251285798  1.117205599
#> 84  -1.04000818 -0.99178252  0.666350237 -0.180830038  0.551870540  1.087360878
#> 85   0.52542006 -1.02420215 -0.223468389 -0.152515140  0.583714229  1.078471221
#> 86  -1.08175931 -0.94750228  0.702633700  0.071162522 -1.082204770  0.060630613
#> 87  -1.29292533 -0.07680862  1.182041569  0.267804099  0.972147720  0.771459224
#> 88   0.96293851  1.36587546 -1.264487358 -0.191094453  0.540036456  1.090332719
#> 89  -0.64311921  2.13465295  1.946854892 -0.381626817  0.274405076  1.117881252
#> 90   1.13383381  0.43100114 -0.956252069  0.419651970  0.706571720 -1.535687422
#> 91  -1.20296408 -0.75569702  0.836757814 -0.027425619  0.713156515  1.026582890
#> 92   1.06216290  0.94678378 -1.132399860 -0.260178362 -0.744133522  0.660402238
#> 93  -0.05136708 -1.24652472  0.087811722  0.503973706  1.075362099  0.006359583
#> 94  -1.00224872 -1.02638253  0.635920952  0.497837514  0.909921341 -0.863814054
#> 95   0.80933233  1.87209144 -1.414943963  0.205591354  0.922474336  0.850049124
#> 96  -0.96230521  1.30569374  1.689031015 -0.348501185  0.329908253  1.117654679
#> 97   1.14060740  0.26476432 -0.895082855 -0.472435419  0.055984237  1.094459638
#> 98   1.09340313 -0.21044516 -0.702295436  0.465865105  0.818346368 -1.184603974
#> 99   0.23679395  3.34626119 -1.819371979  0.058004500  0.793730315  0.977313213
#> 100 -0.65626729 -1.21064236  0.409382857  0.167944340  0.890885415  0.889240223
#>        x3.null1      x4.pen1      x4.pen2      x4.pen3     x4.null1     x5.pen1
#> 1   -1.57076226 -0.450220610  0.286822987  0.894649687 -0.107203786  0.25776556
#> 2    0.60546672  0.384600345  1.010019521 -1.882312590 -1.428178367  0.35485417
#> 3    0.47395374  0.401451109 -0.915551759 -1.874888862  1.815900890  0.34821476
#> 4   -0.16649819  0.184030152  1.155482667  0.026593029 -0.723876067 -0.45539396
#> 5    0.92237769  0.051669046 -0.006371258 -4.570667199 -2.186981204 -0.52872878
#> 6    0.97839491  0.372527271  1.216636778 -0.846095442 -1.082523866 -0.16306328
#> 7    0.77841040 -0.222509150  0.728038133  0.652835414 -0.353050376  0.36029876
#> 8    0.60001919  0.337446500 -1.256863689  0.308748098  0.936568073  0.23353780
#> 9    0.01674921  0.276584688 -1.207299898  0.465138556  0.842330156 -0.52585618
#> 10   1.83376422  0.393693435  1.168025026 -1.202417442 -1.207827691  0.44728527
#> 11   0.54236435  0.299415820 -1.227022796  0.412271034  0.875440766  0.42620435
#> 12  -0.42646441 -0.299317481  0.614975758  0.733625535 -0.285284355 -0.51987362
#> 13  -0.75366303 -0.454718625 -0.148064294  0.988580495  0.112492132 -0.25542669
#> 14  -0.85097834  0.413956724 -0.955454180 -1.754504428  1.776329606  0.13016352
#> 15  -0.06393399 -0.023994512  0.963334215  0.412632306 -0.517714976 -0.15109675
#> 16   1.26973394  0.403465614 -1.293916573  0.053342796  1.071931092  0.38139872
#> 17   0.19984823 -0.456579317 -0.140810121  0.987977285  0.108839259 -0.35981037
#> 18   0.19700605  0.062248925 -0.985973537  0.775820414  0.603222000 -0.42170252
#> 19  -0.66539635  0.166511101 -0.286838794 -3.515919898  2.319509851  0.35631059
#> 20   1.09009595  0.060559703  1.047699437  0.282871181 -0.592755051 -0.51061462
#> 21   0.57724799 -0.034293259 -0.873652820  0.853414759  0.519319139 -0.52876453
#> 22   1.52327691  0.202238819  1.169171811 -0.021473548 -0.746594210 -0.48657166
#> 23  -0.81745391  0.073706329  1.060137402  0.260061517 -0.605242370  0.37282278
#> 24   0.96690015  0.470413994 -1.221869659 -0.768251474  1.426591607  0.36997492
#> 25   0.60730306  0.386017677  1.018138565 -1.853210373 -1.419135878  0.14264994
#> 26  -1.19731519  0.397320998 -1.291702988  0.083452671  1.056883162 -0.48303929
#> 27   0.88147528 -0.425175015  0.363393659  0.865086453 -0.147117547 -0.13149304
#> 28  -0.92446887 -0.481927793  0.104486971  0.948648311 -0.014316219  0.23302646
#> 29   1.39862113 -0.240853664  0.702585544  0.672487891 -0.337359440 -0.36860249
#> 30   0.73175971  0.085169014 -1.011724401  0.753288987  0.624659838 -0.32267920
#> 31   0.93917171 -0.130303833  0.845122664  0.548858578 -0.429591553  0.16515691
#> 32  -0.77375866 -0.483167344  0.057693841  0.959018821  0.009223906  0.44524400
#> 33  -1.97809334 -0.119805323  0.857498013  0.536327257 -0.438199640  0.26598900
#> 34  -0.82118109  0.374619754  0.959687045 -2.055394000 -1.481369801  0.14466318
#> 35  -0.84572948 -0.455536480 -0.144903566  0.988321615  0.110900283 -0.31339272
#> 36   0.23289625 -0.268559430  0.662422945  0.701701423 -0.313142847  0.05061713
#> 37   1.34048210 -0.251606004  0.687259246  0.683886622 -0.328043630 -0.38966092
#> 38   0.72137174  0.381808597 -1.284699710  0.152204769  1.021699936  0.36177821
#> 39  -1.54935421 -0.276884814  0.649901848  0.710387395 -0.305716895 -0.34754301
#> 40   2.70877987  0.202773360 -1.137047073  0.602746614  0.748156866 -0.38748189
#> 41  -0.68336721 -0.482731948  0.087249130  0.952629224 -0.005635735 -0.52391516
#> 42   0.66255442  0.330292163  1.231157993 -0.519139865 -0.958906461 -0.51372735
#> 43  -0.34566817  0.116879193 -1.046726564  0.718911897  0.655557611 -0.03777688
#> 44  -0.89466365  0.286814458  1.219740553 -0.305333931 -0.872062825 -0.21861731
#> 45   0.55684539  0.178653086 -1.112449423  0.639296631  0.720635429 -0.36240980
#> 46  -1.00027145  0.398703203 -0.907000648 -1.900228741  1.824166099 -0.39519327
#> 47   0.19098238  0.386704927  1.022207794 -1.838482562 -1.414548359 -0.25663152
#> 48  -1.37681485  0.123360749  1.105278194  0.165356208 -0.655222203 -0.40246387
#> 49   1.24384961 -0.382988929 -0.346131221  0.992424438  0.213491195 -0.32744941
#> 50  -0.93860641  0.255476160  1.204103406 -0.185717038 -0.820820300 -0.17193318
#> 51  -1.97572622  0.210920174 -1.145196503  0.589587490  0.757767013 -0.41049927
#> 52   0.60027124 -0.474534973 -0.051718253  0.977957415  0.064098731  0.40887126
#> 53  -0.86833687 -0.019238193  0.968288340  0.405953962 -0.521762866  0.39353983
#> 54  -0.27231169 -0.430269804 -0.228623839  0.993110803  0.153228168 -0.38898777
#> 55  -0.09593453  0.396828627  1.127201359 -1.411758312 -1.277921904  0.34940647
#> 56  -0.17614259  0.200976696  1.168247419 -0.018027939 -0.744982023 -0.10057975
#> 57  -1.94765148 -0.249121243  0.690829185  0.681259815 -0.330205048 -0.48099658
#> 58   1.02022633 -0.372912064  0.483911138  0.809269766 -0.211661852 -0.13411978
#> 59   0.31364007 -0.106380029 -0.784495877  0.897858328  0.460267349  0.39772831
#> 60  -0.45423491 -0.297429197  0.617987327  0.731677391 -0.287030676  0.18819420
#> 61  -0.13414438  0.141947380  1.121314990  0.125900296 -0.675249208 -0.52282902
#> 62  -0.12212736 -0.214802395  0.738486946  0.644494420 -0.359576502  0.44450169
#> 63  -0.49610771 -0.465888224  0.223336563  0.915980909 -0.074582577 -0.24298675
#> 64   0.60763379  0.388285386  1.189720361 -1.068188718 -1.161596592  0.14721209
#> 65  -1.23153734 -0.406443055 -0.291897544  0.993838851  0.185522436  0.25568538
#> 66  -1.27787024  0.460681987 -1.136584540 -1.140266267  1.564936143  0.35058376
#> 67   1.19324644  0.124074036 -0.182156040 -3.766758215  2.392863800  0.34796304
#> 68   0.60285254  0.192218314  1.161730253  0.005405598 -0.733951395 -0.32610833
#> 69   1.63098948 -0.077610706 -0.820744959  0.881251002  0.483669837  0.02909898
#> 70  -1.98043426 -0.471070712 -0.073123812  0.980805350  0.074833511  0.32450385
#> 71   0.74458954  0.190528569  1.160452901  0.009833468 -0.731853925 -0.47807226
#> 72   0.22650491 -0.197822231  0.761036288  0.625921540 -0.373841149  0.31498017
#> 73  -0.73078293 -0.283521134  0.639756559  0.717285325 -0.299740095 -0.27822735
#> 74  -0.05160503  0.081558349 -1.007691751  0.756961368  0.621235170 -0.05322327
#> 75  -0.78009534 -0.249404951 -0.585649645  0.961504776  0.342139443 -0.13397648
#> 76   0.06738469  0.429320552 -1.298172655 -0.099310682  1.145279672  0.11450953
#> 77  -2.11820248  0.433145289 -1.297828879 -0.126888345  1.158060502  0.10911644
#> 78  -0.81894647  0.299806458  1.224688053 -0.362072415 -0.895659608  0.26410640
#> 79   0.24719266 -0.387099860 -0.337060503  0.992794713  0.208791820  0.44243199
#> 80   0.92803199 -0.391248528 -0.327734605  0.993118983  0.203969690  0.33933856
#> 81   1.03578588  0.335548657  1.231293368 -0.550786377 -0.971314138  0.39128440
#> 82   0.54474862  0.049761237  1.037347519  0.300978944 -0.582704482  0.31554294
#> 83  -0.19762305 -0.375649436  0.478363641  0.812108335 -0.208634024  0.44748138
#> 84  -0.37166777 -0.194019234  0.766001703  0.621722701 -0.377017107  0.15852425
#> 85  -0.39179440 -0.451527414 -0.160015240  0.989504219  0.118514710  0.37643411
#> 86   0.66779118  0.299091004 -1.226750617  0.413066373  0.874953012  0.43507912
#> 87  -0.72297523  0.430521250 -1.298100411 -0.107785020  1.149221199  0.39123383
#> 88  -0.36430184  0.142419442 -1.074329043  0.688199650  0.681612029  0.36298322
#> 89  -0.21030896 -0.346520066  0.534494496  0.782087484 -0.239572832 -0.37643648
#> 90  -1.69543724  0.447136296 -1.292933937 -0.246436011  1.212026574 -0.50816953
#> 91  -0.47950874  0.132612627  1.113325917  0.146019239 -0.665091313 -0.06783106
#> 92   0.35573346 -0.483167188  0.065942405  0.957290713  0.005079387  0.34694804
#> 93  -1.13539092 -0.158880039  0.810534552  0.582168589 -0.406120571  0.02734553
#> 94  -1.47148217 -0.079044396 -0.818960907  0.882112429  0.482500137  0.38284911
#> 95  -0.66224441  0.338415052 -1.257563781  0.305804268  0.938243603  0.33022561
#> 96  -0.24111704  0.006949614  0.995106686  0.368020673 -0.544346340  0.42625884
#> 97  -0.09284141  0.228700989  0.475131475 -3.415661515 -1.873849602 -0.13668457
#> 98  -1.58100378  0.394046147  1.075859013 -1.633741168 -1.349936046 -0.35909821
#> 99  -0.54117310 -0.259113860 -0.570614455  0.964750431  0.333728490 -0.42392195
#> 100 -0.62875406 -0.370513197  0.488721316  0.806786200 -0.214292306 -0.22798558
#>         x5.pen2     x5.pen3    x5.null1
#> 1    1.14500817 -0.61274575 -0.86771228
#> 2   -0.74419979  2.11458117  1.37861456
#> 3    0.71124825  1.83492380 -1.85229108
#> 4   -0.42617087 -0.87731138  0.14765407
#> 5   -0.12085718 -1.02078787 -0.02078655
#> 6   -0.82727747 -0.50923329  0.41069655
#> 7   -0.76601520  2.05313576  1.36090747
#> 8   -1.10069610  0.21436933  0.75298709
#> 9   -0.04399855 -1.04427643 -0.06159900
#> 10   1.17488229  0.28731708 -1.31014473
#> 11   1.20868917  0.01994724 -1.19683307
#> 12  -0.21839565 -0.98408936  0.03166785
#> 13   0.59071587 -1.06690612 -0.40615723
#> 14  -1.05322319 -0.04201969  0.64512899
#> 15  -0.83882456 -0.49337294  0.41985857
#> 16  -0.86770423  1.74848362  1.27152619
#> 17   0.42773520 -1.09158842 -0.31325187
#> 18  -0.49371012 -0.83293173  0.18724195
#> 19  -0.74993225  2.09854474  1.37400280
#> 20  -0.26530621 -0.96354134  0.05726264
#> 21  -0.11858207 -1.02155024 -0.02200005
#> 22   0.14617339 -1.08319217 -0.16186187
#> 23  -0.82159118  1.89086303  1.31364477
#> 24  -1.07360883  0.90086960  1.00395632
#> 25  -1.06016970 -0.01569540  0.65671379
#> 26   0.15720518 -1.08463412 -0.16768652
#> 27   0.75292701 -1.01485849 -0.50656729
#> 28   1.12541862 -0.65956039 -0.83704498
#> 29   0.41228201 -1.09272074 -0.30471474
#> 30   0.48956477 -1.08504671 -0.34782312
#> 31   1.06654357 -0.76487126 -0.76177287
#> 32   1.18147860  0.24645998 -1.29343651
#> 33  -0.45794258  2.84661197  1.58310331
#> 34   1.04759296 -0.79138751 -0.74106862
#> 35   0.50427640 -1.08300053 -0.35615790
#> 36   0.95541037 -0.89019816 -0.65472297
#> 37  -0.54884530 -0.79254768  0.22052475
#> 38   1.20840841 -0.33050525 -1.02938357
#> 39  -0.61254375 -0.74063445  0.26030531
#> 40   0.37783293 -1.09454630 -0.28581252
#> 41  -0.19004589 -0.99557841  0.01632592
#> 42   0.04023685 -1.06481731 -0.10604268
#> 43   0.86143421 -0.95903781 -0.58135975
#> 44   0.64164015 -1.05390520 -0.43659849
#> 45  -0.59099101 -0.75887627  0.24666758
#> 46   0.36320448 -1.09503480 -0.27783547
#> 47   0.58900213 -1.06729614 -0.40514678
#> 48  -0.52764539 -0.80854594  0.20761282
#> 49  -0.64031348 -0.71603258  0.27817554
#> 50  -0.81857032 -0.52089179  0.40388706
#> 51   0.33305738 -1.09551276 -0.26147964
#> 52   1.21561489 -0.10587524 -1.13969630
#> 53   1.21617858 -0.19211227 -1.09876002
#> 54  -0.54993296 -0.79171044  0.22119130
#> 55  -0.72333063  2.17234167  1.39517096
#> 56   0.78984960 -0.99812857 -0.53112992
#> 57   0.16341036 -1.08540618 -0.17096462
#> 58   0.74973355 -1.01620840 -0.50447913
#> 59   0.86234524  1.42460805 -1.72051067
#> 60  -1.08299237  0.08983948  0.70191445
#> 61  -0.19853695 -0.99220938  0.02091186
#> 62   1.18350883  0.23326232 -1.28799676
#> 63   0.60822954 -1.06274651 -0.41653368
#> 64   1.04997512 -0.78820297 -0.74360025
#> 65  -1.10686243  0.28482058  0.78093670
#> 66  -1.09544764  0.73350108  0.94650675
#> 67   1.20258842 -0.37970914 -1.00341994
#> 68   0.48406249 -1.08576260 -0.34471733
#> 69   0.93320746 -0.90869236 -0.63640314
#> 70  -1.10834365  0.57232891  0.88918902
#> 71   0.17209061 -1.08643903 -0.17555270
#> 72  -0.60513705  2.48424599  1.48322563
#> 73  -0.70288340 -0.65558567  0.31986028
#> 74   0.35233353  4.63032082  2.05371484
#> 75   0.74990801 -1.01613503 -0.50459307
#> 76   1.01891242 -0.82681076 -0.71198171
#> 77  -0.04477617  3.78126589  1.83250072
#> 78  -1.10866239  0.31372097  0.79222319
#> 79   1.18842744  0.19980729 -1.27411069
#> 80   1.19840803 -0.40796142 -0.98815598
#> 81  -0.97030068  1.39252320  1.16311376
#> 82  -1.11037301  0.52626968  0.87240278
#> 83   1.06969102  0.76222536 -1.49225068
#> 84  -0.16966133  3.50723809  1.76030064
#> 85   1.21314385 -0.27182624 -1.05941095
#> 86   1.20039761  0.10574623 -1.23428013
#> 87  -0.97980449  1.35543182  1.15152601
#> 88   1.20885838 -0.32595375 -1.03174776
#> 89  -0.56976169 -0.77615700  0.23341649
#> 90   0.06653245 -1.07014622 -0.11989104
#> 91  -0.91331438 -0.37797227  0.48337323
#> 92  -1.09806838  0.70783942  0.93751986
#> 93  -0.98707478 -0.23106678  0.55766785
#> 94  -0.87678340  1.71943591  1.26285271
#> 95  -1.10651448  0.60380839  0.90055322
#> 96   1.20865131  0.02041522 -1.19704045
#> 97  -0.85243628 -0.47405725  0.43086446
#> 98  -0.59587226 -0.75480774  0.24973943
#> 99   0.30521224 -1.09533133 -0.24646292
#> 100 -0.76038740 -0.59290609  0.36028907