Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>          x1.pen1     x1.pen2     x1.pen3    x1.null1      x2.pen1     x2.pen2
#> 1    0.108653636  1.06734892 -0.33791300 -0.59750624  0.098373337  1.03327465
#> 2   -0.016194072  0.94240381 -0.52876259 -0.48755657  0.383084539 -0.93257318
#> 3   -0.182755634 -0.74683333 -1.06316096  0.42813572 -0.395351599  0.20431017
#> 4    0.040852666 -1.05931628 -0.92045630  0.63171934  0.442639006  1.37856865
#> 5    0.383432326  0.87653929  2.54517301 -1.64603355 -0.056160237  0.83564345
#> 6   -0.038183329  0.91901236 -0.55730915 -0.46968079 -0.270669406  0.50582654
#> 7    0.444602880 -1.29460818  0.92630727  1.61516288  0.029988935 -0.93408565
#> 8   -0.208094050 -0.70707222 -1.07310165  0.40561035  0.254509411  1.21575151
#> 9    0.104435665  1.06333796 -0.34540818 -0.59345351  0.286686727  1.25009110
#> 10  -0.389446445 -0.35767237 -1.11058365  0.22209978  0.472729977  1.38370175
#> 11  -0.471598005  0.13770051 -1.04011672 -0.02199705  0.494052765  1.24345349
#> 12  -0.128956458 -0.82739668 -1.03848996  0.47543745  0.499043269  1.27454928
#> 13  -0.473260349 -0.02629732 -1.07867809  0.05881154  0.098030625  1.03285302
#> 14   0.264441279  1.19985780  0.02172425 -0.77431780 -0.368835350  0.28858930
#> 15  -0.199762137 -0.72029285 -1.06994878  0.41304699  0.291557772 -1.17559317
#> 16  -0.209070252 -0.70551336 -1.07346365  0.40473682 -0.426282255 -0.06502796
#> 17  -0.468770734 -0.06667788 -1.08581818  0.07859015  0.350188064 -1.20032639
#> 18   0.113908167 -1.15216406 -0.84130081  0.70930699 -0.402263493 -0.21846731
#> 19  -0.060440487 -0.92415849 -0.99933681  0.53608635  0.475768913  1.38304996
#> 20   0.132670495  1.08985773 -0.29349144 -0.62115369  0.076911998  1.00673388
#> 21  -0.117338040 -0.84421204 -1.03249489  0.48563946  0.409326749 -1.08208889
#> 22   0.375556049 -1.41436460 -0.20105460  1.12759009 -0.408758492 -0.19066965
#> 23   0.445783740 -1.37737504  0.51538104  1.45418893 -0.378622929  0.25999318
#> 24  -0.469395327 -0.06189393 -1.08501972  0.07624771  0.400700808 -1.01666026
#> 25  -0.249833506 -0.71341535  6.65215421 -2.75160767 -0.212117647 -0.62733614
#> 26  -0.362518006 -0.42145339 -1.10961999  0.25425221  0.464806899  1.12191203
#> 27   0.281534968 -1.34121962 -0.53801909  0.93521286 -0.422426385  0.06573971
#> 28  -0.260854513 -0.61944378 -1.09042996  0.35748532  0.131889429  1.07413969
#> 29   0.243000257 -0.58407206  3.01229227  2.30798361 -0.415410279  0.11446196
#> 30   0.386588467  1.25739561  0.54926003 -0.99339095 -0.418402173  0.09600077
#> 31  -0.200782403 -0.71868192 -1.07034095  0.41213809 -0.425232118  0.03663280
#> 32   0.124437269 -1.16515843 -0.82792811  0.72126241 -0.386120521 -0.27557543
#> 33  -0.039164454 -0.95318685 -0.98513534  0.55534742  0.122120522  1.06230665
#> 34   0.388019924  0.89152782  2.49703795 -1.63200244  0.474852656  1.15885569
#> 35  -0.454822056 -0.14660945 -1.09728730  0.11773780  0.500181921  1.34952463
#> 36  -0.390291716 -0.35554402 -1.11057480  0.22103349  0.318673566 -1.19017899
#> 37   0.237705239 -1.29638400 -0.64063677  0.86696773  0.106689599 -1.01559154
#> 38   0.373044586 -1.41311093 -0.21360690  1.12105367  0.410635983  0.94990911
#> 39  -0.411229049  0.36854827 -0.95753228 -0.13911147 -0.427400003 -0.03682892
#> 40  -0.338609064 -0.47294358 -1.10706293  0.28053476  0.083508907 -0.99156061
#> 41   0.008401274  0.96802683 -0.49541579 -0.50790634 -0.010005985  0.89652173
#> 42  -0.127124815 -0.83006003 -1.03756119  0.47704498 -0.417519415  0.10171202
#> 43   0.444399710  1.16916291  1.41056271 -1.29743747  0.429182361  1.00554617
#> 44  -0.475999343  0.01607156 -1.07020252  0.03803003 -0.361106095  0.30970682
#> 45   0.230094410 -1.28817816 -0.65632378  0.85595639 -0.135626077  0.72476057
#> 46   0.296534722 -1.35542560 -0.49712807  0.96084003 -0.111157565 -0.76731519
#> 47   0.118687237  1.07682215 -0.31972383 -0.60726464 -0.005698205 -0.89423421
#> 48   0.439282314 -1.25669657  1.07854245  1.67169488 -0.064923402  0.82383218
#> 49   0.217289345 -1.27413337 -0.68150641  0.83790640  0.365941580 -1.20091669
#> 50  -0.382650476 -0.37447500 -1.11056140  0.23053190  0.399213046 -1.17732186
#> 51  -0.193074316  0.73728895 -0.73501816 -0.34642704  0.303485784 -0.65979895
#> 52  -0.224841611 -0.68002707 -1.07909694  0.39055047 -0.397945438  0.19458817
#> 53  -0.404514656 -0.31827664 -1.10999811  0.20242268  0.301217219 -0.65289577
#> 54  -0.080416164 -0.89650601 -1.01173510  0.51823989  0.391637096 -1.18830024
#> 55  -0.214031601  0.70967479 -0.75690955 -0.32941270  0.139733259 -1.04878208
#> 56   0.177522637 -1.22884089 -0.75121233  0.78519658 -0.317225509  0.41356474
#> 57  -0.085190188 -0.88983476 -1.01457230  0.51400123  0.406716077  0.93842829
#> 58   0.295596246 -0.73917200  2.62162644  2.18766356  0.081003179 -0.98892955
#> 59  -0.378373674 -0.38478491 -1.11046631  0.23571863  0.461011189  1.38373643
#> 60  -0.457475663  0.22061595 -1.01448651 -0.06344695  0.494211739  1.24432175
#> 61  -0.169994254 -0.76638450 -1.05775407  0.43939659  0.183167005 -1.09012192
#> 62   0.120066021 -1.15977697 -0.83354720  0.71627253  0.423763280  1.36916858
#> 63  -0.339219693 -0.47167675 -1.10714544  0.27988412  0.397533380 -1.18032414
#> 64   0.183020881  1.13497264 -0.18937782 -0.67432837 -0.143189136 -0.72531667
#> 65   0.425947095 -1.18749005  1.32916442  1.76185385  0.351605007  1.31372185
#> 66   0.444261595  1.15845601  1.46633668 -1.31565607  0.153574706 -1.06226416
#> 67   0.253826217 -1.31338417 -0.60547592  0.89105240  0.116143538  1.05503442
#> 68  -0.476189337  0.02121812 -1.06910369  0.03550261  0.135796156 -0.18977598
#> 69   0.443348298  1.19094664  1.28800862 -1.25689200 -0.427545071 -0.01916503
#> 70   0.405204038  1.25524747  0.68748631 -1.04572220  0.117774863  1.05702165
#> 71   0.307254326 -1.36514030 -0.46568542  0.98002816  0.405159823  1.35724017
#> 72   0.447422533 -1.36191365  0.60614592  1.49095209  0.490610295  1.37302074
#> 73   0.368642377 -0.97022260  1.99970562  1.98932898  0.496161721  1.36325893
#> 74  -0.458796542  0.21460030 -1.01649168 -0.06042056  0.400428873 -1.17487026
#> 75  -0.153970099  0.78653895 -0.69311088 -0.37769520  0.274394895 -0.57304599
#> 76   0.094576481 -1.12803230 -0.86446472  0.68790308 -0.266697229  0.51313689
#> 77   0.400180804  0.93279761  2.36129422 -1.59213984  0.169867071  1.11947010
#> 78   0.363426249 -1.40778883 -0.25903756  1.09706038  0.378086962  0.85681272
#> 79   0.356706376  0.79364913  2.80213526 -1.72008996  0.407822266  1.35907716
#> 80  -0.462593125  0.19599367 -1.02254749 -0.05108019 -0.208516289 -0.63276438
#> 81  -0.463354453 -0.10233349 -1.09136982  0.09604770 -0.426155080 -0.06714434
#> 82   0.295442742 -1.35441457 -0.50022240  0.95892814 -0.188007654  0.64574306
#> 83   0.282450036  1.21232532  0.07839559 -0.79970280  0.410320535 -1.09815367
#> 84   0.340216466  1.24500464  0.29891461 -0.89381909  0.259321430 -1.15377479
#> 85   0.180139932 -1.23189210 -0.74698074  0.78852555 -0.051086929 -0.84147152
#> 86  -0.001491344  0.95778642 -0.50902261 -0.49966923 -0.091254232 -0.79249933
#> 87  -0.134721928  0.80982637 -0.67188240 -0.39294800 -0.423799310  0.05285730
#> 88   0.393797060 -1.42136499 -0.09920807  1.17926716  0.345780753 -0.79476435
#> 89  -0.448701884  0.25635365 -1.00209009 -0.08149736  0.031133479 -0.93534241
#> 90  -0.291786590  0.59749959 -0.83576666 -0.26336783  0.174519471  1.12494399
#> 91  -0.182324941  0.75110669 -0.72364576 -0.35507419 -0.315619240 -0.44878881
#> 92   0.373764216 -0.98761785  1.95003915  1.97303771  0.288912096 -1.17396284
#> 93  -0.442602768  0.27808686 -0.99413895 -0.09253862  0.410656112 -1.11200527
#> 94   0.430260115 -1.41452605  0.21004740  1.32415448 -0.198635409 -0.64746206
#> 95  -0.200250379  0.72793711 -0.74255403 -0.34062626  0.038449958 -0.94334668
#> 96   0.150810238  1.10645837 -0.25782095 -0.63970866 -0.407877207  0.15309954
#> 97   0.440769751  1.20800395  1.18008466 -1.22056084 -0.132171778  0.72978113
#> 98   0.435186339 -1.40852546  0.27757861  1.35386331 -0.319775775 -0.44032724
#> 99   0.295659675 -1.35461572 -0.49960922  0.95930735  0.217780997  1.17489722
#> 100  0.403618345  0.94496021  2.32029177 -1.58000975 -0.330259893 -0.41830926
#>          x2.pen3     x2.null1      x3.pen1     x3.pen2     x3.pen3    x3.null1
#> 1   -0.690040394 -0.564407458 -0.172441306 -0.81448988  0.72791490  0.41649658
#> 2    2.338647856  1.598345738  0.168709911  1.42963892  0.60867118 -1.03346172
#> 3   -0.853228280 -0.072968487 -0.376045012  0.60146940  1.24900271 -0.37795728
#> 4    0.042060683 -1.086202563 -0.176156725 -0.81131479  0.73174857  0.41403658
#> 5   -0.792362424 -0.422849761 -0.516400418 -0.31704651  1.10840734  0.10053112
#> 6   -0.859744035 -0.230530724  0.405056727 -1.14880903 -0.16302805  0.85525397
#> 7   -0.231984678  0.583974203  0.116506316  1.37818168  0.74741682 -0.94610121
#> 8   -0.500056254 -0.742328267  0.197598317  1.45281089  0.51251151 -1.08910406
#> 9   -0.442972317 -0.786408800  0.300469966  1.45814068 -0.11197827 -1.39347304
#> 10   0.243560964 -1.188665176  0.510762691 -1.15355915 -0.56743510  1.00990918
#> 11   1.176266439 -1.587869089 -0.145149792  1.02662879  1.11608416 -0.62840326
#> 12   1.043645007 -1.536212006  0.082061502  1.33936611  0.82119154 -0.89515225
#> 13  -0.690341553 -0.564064828  0.132750222 -1.02754529  0.37097279  0.61700608
#> 14  -0.860557208 -0.116172329 -0.526756903  0.12317628  1.24147005 -0.13043462
#> 15   0.407215811  0.907477926  0.300748354  1.09133744 -1.45415671 -1.89912863
#> 16  -0.802282630  0.063955863 -0.195454712  0.94565520  1.15561156 -0.57584667
#> 17   0.703765859  1.031835561  0.232660310 -1.08112793  0.21699829  0.69121995
#> 18  -0.753749027  0.142942575 -0.079954942 -0.88817723  0.62993200  0.47670050
#> 19   0.270335029 -1.201669180 -0.525263480 -0.27940144  1.12527685  0.07990262
#> 20  -0.708145449 -0.543263037 -0.129131649  1.05143364  1.10211058 -0.64525136
#> 21   1.749490917  1.409819091  0.150305549  1.41270458  0.66199635 -1.00100181
#> 22  -0.763660592  0.128504228 -0.083834159  1.11923623  1.05850980 -0.69359741
#> 23  -0.858532911 -0.101465472 -0.354493496 -0.62970309  0.91254459  0.28606163
#> 24   2.028048561  1.500506781  0.375716909 -1.14063543 -0.08275192  0.82241524
#> 25  -0.538084678  0.369530081  0.229123279  1.47167658  0.38411048 -1.15866175
#> 26   1.620288952 -1.752840055  0.018116318 -0.13977953 -4.18386715 -2.73706708
#> 27  -0.832283877 -0.002521925  0.123215611 -1.02204693  0.38417198  0.61037076
#> 28  -0.658561983 -0.598753260  0.564243211 -1.11243101 -0.97506303  1.15221204
#> 29  -0.840906346 -0.027249537  0.242606139  1.47689860  0.31879228 -1.19232841
#> 30  -0.837800310 -0.017877959  0.098914075  1.35876020  0.78656927 -0.91952497
#> 31  -0.826477169  0.012248942  0.443447164 -1.15612583 -0.28382262  0.90319368
#> 32  -0.731747829  0.172847079 -0.158962092  1.00487264  1.12757754 -0.61393772
#> 33  -0.668164386 -0.588565800 -0.178687884 -0.80914111  0.73435672  0.41235805
#> 34   1.494149299 -1.707080250 -0.188908969  0.95647072  1.15082350 -0.58267589
#> 35   0.640727759 -1.370347849  0.189925836  1.44712130  0.53974862 -1.07368651
#> 36   0.527957298  0.959441176  0.251974048  0.82704583 -2.11465422 -2.11507175
#> 37  -0.100195095  0.660020143 -0.544821514 -0.07736903  1.19720026 -0.02734846
#> 38   2.151090927 -1.937827375 -0.075863397  1.13082374  1.05015368 -0.70225544
#> 39  -0.809611342  0.049583648 -0.450752397  0.41343180  1.26265436 -0.27897886
#> 40  -0.143128021  0.635990471 -0.422596735 -0.53430946  0.98458893  0.22603382
#> 41  -0.767776390 -0.463145144 -0.394425096 -0.57673063  0.95412381  0.25226140
#> 42  -0.838782167 -0.020776785  0.565712954 -1.10967717 -0.99383745  1.15851189
#> 43   1.987210248 -1.881873391  0.095846380  1.35528469  0.79307192 -0.91501293
#> 44  -0.861744140 -0.127070998 -0.120898154 -0.85674928  0.67398687  0.45023534
#> 45  -0.825870584 -0.354278198  0.229327556  1.47177055  0.38317513 -1.15915139
#> 46  -0.422424823  0.458880780  0.417567820 -1.15166801 -0.20018090  0.87017885
#> 47  -0.284906002  0.551302580  0.381147387 -1.14228869 -0.09695527  0.82828670
#> 48  -0.796578220 -0.415277613 -0.493764595  0.27399731  1.25870086 -0.20735014
#> 49   0.815020649  1.075929812 -0.522408773  0.14783154  1.24520863 -0.14300923
#> 50   1.179521086  1.212872061  0.322067370  1.37602090 -0.54727224 -1.57193094
#> 51   3.183265122  1.851401767  0.083212603 -0.99825610  0.43746037  0.58308628
#> 52  -0.852118663 -0.068008640 -0.495304717 -0.38534249  1.07463411  0.13866764
#> 53   3.202795682  1.857077806  0.152096771  1.41441648  0.65703941 -1.00407326
#> 54   1.067293813  1.171802723 -0.340226247  0.67880151  1.23651641 -0.41999906
#> 55  -0.033441150  0.696152659 -0.534351734 -0.22893805  1.14608106  0.05262087
#> 56  -0.863689064 -0.181269341  0.183779418  1.44230758  0.56063351 -1.06168755
#> 57   2.184199557 -1.949021592  0.518984355 -0.77762675 -2.35390190  1.57567203
#> 58  -0.147594590  0.633451927 -0.254037666  0.84465284  1.19403535 -0.51434855
#> 59   0.153922525 -1.144145306 -0.083277473  1.12004877  1.05793330 -0.69420032
#> 60   1.172713191 -1.586501152  0.319720585  1.24556973 -1.00929880 -1.74435322
#> 61   0.066227354  0.747680203 -0.427299294  0.47762972  1.26052391 -0.31236717
#> 62  -0.050409782 -1.035956841  0.463874817 -1.15794390 -0.35777622  0.93175623
#> 63   1.151676064  1.202764269  0.329986452 -1.12460468  0.02702170  0.77607393
#> 64  -0.460337376  0.431009166  0.577309047 -0.99143589 -1.59209045  1.35023843
#> 65  -0.296786829 -0.888166489  0.038104637 -0.96999966  0.49411732  0.55313725
#> 66  -0.003284388  0.712030337 -0.468549588  0.36002467  1.26251122 -0.25142886
#> 67  -0.673862018 -0.582406109 -0.275840460 -0.71815186  0.83293821  0.34564646
#> 68   4.417058377  2.199972507  0.129557966 -1.02571174  0.37541521  0.61477810
#> 69  -0.813958764  0.040594848  0.319690227  1.24520417 -1.01044464 -1.74476433
#> 70  -0.672320054 -0.584081878 -0.473300245  0.34489004  1.26215933 -0.24365136
#> 71  -0.126921631 -0.992482597  0.579070215 -1.00821705 -1.52093272  1.32823787
#> 72   0.438677032 -1.280716897  0.221333596  1.46778986  0.41866963 -1.14040788
#> 73   0.534641901 -1.323927341 -0.034438139  1.18945548  1.00289951 -0.74832226
#> 74   1.201128321  1.220679613  0.194060521 -1.06132197  0.28039018  0.66133810
#> 75   3.424486734  1.921066809  0.270795919  1.47933235  0.15122075 -1.27439482
#> 76  -0.859190830 -0.234490696 -0.340985974  0.67722248  1.23681467 -0.41913032
#> 77  -0.617543619 -0.639882910  0.317971443  1.40852185 -0.40286153 -1.51479081
#> 78   2.413739807 -2.025705951  0.237589018 -0.17998468 -4.00568088  2.02480307
#> 79  -0.116684474 -0.998409693  0.045902691  1.29551297  0.88728655 -0.84594279
#> 80  -0.534121524  0.372826971 -0.421980521  0.49141111  1.25973224 -0.31958161
#> 81  -0.801713203  0.065035790 -0.152428745 -0.83128298  0.70715055  0.42967437
#> 82  -0.842486444 -0.308198978 -0.470827909 -0.44526320  1.04132672  0.17302754
#> 83   1.672937636  1.384331612 -0.049709755  1.16814820  1.02111296 -0.73110277
#> 84   0.287023774  0.853600443  0.003470879  1.24080315  0.95297562 -0.79266742
#> 85  -0.346960697  0.511089901  0.271429196 -1.09969451  0.14714378  0.72318414
#> 86  -0.398120674  0.476138233  0.574408681 -1.08673138 -1.13594332  1.20557443
#> 87  -0.829774646  0.004014966 -0.165243752 -0.82058878  0.72046989  0.42124954
#> 88   2.786481660  1.734531463 -0.535066832 -0.22408083  1.14797840  0.05001519
#> 89  -0.230214532  0.585043682 -0.223577012 -0.76905946  0.78020033  0.38216695
#> 90  -0.612083592 -0.645101349 -0.479920307  0.32305631  1.26141140 -0.23245048
#> 91  -0.650265006  0.266304907 -0.362064671 -0.62017913  0.92032921  0.27989121
#> 92   0.396515473  0.902773337 -0.032463420 -0.92255673  0.57709399  0.50729512
#> 93   1.603147366  1.360861850  0.575351448 -0.97739381 -1.64968091  1.36790802
#> 94  -0.523201867  0.381811999  0.288405714  1.47142138  0.01207545 -1.33864047
#> 95  -0.218779233  0.591918104 -0.302503185  0.75439043  1.21999556 -0.46217529
#> 96  -0.846773076 -0.046881986  0.462109701 -1.15785667 -0.35106018  0.92918539
#> 97  -0.824625767 -0.357274620  0.203111532  1.45666319  0.49207703 -1.10050755
#> 98  -0.654809420  0.261616672 -0.146520036  1.02448589  1.11724695 -0.62696601
#> 99  -0.556112945 -0.695775973 -0.177405296  0.97526928  1.14215758 -0.59467658
#> 100 -0.666344971  0.249485752 -0.321073819  0.71783967  1.22854232 -0.44163076
#>          x4.pen1     x4.pen2      x4.pen3     x4.null1     x5.pen1     x5.pen2
#> 1   -0.393210533 -0.40983992 -1.205993166  0.275474484  0.18454233  1.23921715
#> 2    0.377459713 -1.30363233 -0.139589513  1.255694872  0.44251415 -1.03774788
#> 3   -0.178156812 -0.77318111 -1.177406796  0.481917564 -0.47074490 -0.43531199
#> 4    0.404183644  0.97432969  1.062660167 -1.035255939 -0.23048594  0.82230801
#> 5    0.395614872 -1.16532460  0.544196441  1.539355095  0.03332678 -1.04725164
#> 6   -0.131336812 -0.83862246 -1.158524841  0.523336345  0.41609003  1.28692076
#> 7    0.114426118  0.97345087 -0.045821876 -0.634347001  0.27861787 -0.50213866
#> 8    0.408762585  0.79398086  1.799286929 -1.258983195  0.07514745  1.14794461
#> 9    0.353431327 -1.31539469 -0.327690216  1.165615029  0.19995409  1.25062391
#> 10  -0.117108471 -0.85788015 -1.151911928  0.535932888  0.42370074  1.12497677
#> 11  -0.015236811  0.89888827 -0.266942242 -0.536197990 -0.51844368  0.33480222
#> 12  -0.108508462  0.83295428 -0.404834003 -0.469510514 -0.30820642 -0.68921756
#> 13  -0.247803701 -0.66895299 -1.197582063  0.419436705 -0.35595845  0.64873995
#> 14  -0.305699481 -0.57424308 -1.206971483  0.365360771 -0.16722326 -0.85513262
#> 15   0.393432370  0.99361890  0.944533102 -0.997107056 -0.30448403  0.72411854
#> 16  -0.113458287 -0.86277635 -1.150147568  0.539168895 -0.04683049 -0.97547893
#> 17   0.281430786 -0.69156447  1.867095889  1.993631843  0.14362714  1.20697864
#> 18   0.020084253  0.92116838 -0.210930585 -0.561981144 -0.14755592 -0.87593558
#> 19   0.043922509  0.93538717 -0.171589216 -0.579693880 -0.28936983  0.74504745
#> 20  -0.143545370 -0.82187311 -1.163869000  0.512541336  0.33434602 -1.24446877
#> 21   0.387369512 -1.28876610 -0.020096403  1.309529724 -0.57323273  0.13651714
#> 22  -0.532866940  0.13016638 -1.065670384 -0.009578467  0.45145856 -1.12378211
#> 23   0.017670577 -1.02802461 -1.065730942  0.659269189 -0.50547377  0.36789978
#> 24  -0.455849600 -0.26339025 -1.188482310  0.197895256  0.25524177  0.52456706
#> 25   0.031170953  0.92786362 -0.192801665 -0.570182512  0.42958076  1.21721976
#> 26  -0.074860702 -0.91349918 -1.129727590  0.573582444 -0.14410569  0.92556600
#> 27   0.303705706 -1.30278642 -0.562309600  1.040753004  0.41808371  1.08824650
#> 28  -0.416542989  0.50506435 -0.819962506 -0.225879962  0.37413025 -0.77346426
#> 29  -0.527561698  0.02272831 -1.110704506  0.047792805  0.29245874 -1.22874606
#> 30  -0.109691950  0.83204836 -0.406509905 -0.468668576  0.29654994 -1.23057914
#> 31  -0.082751436 -0.90328112 -1.134171297  0.566510164 -0.39447075 -0.56707581
#> 32   0.097177834 -1.11916846 -0.987440139  0.740806096  0.13161139  0.18562860
#> 33   0.008124187 -1.01663790 -1.073475492  0.650099909 -0.06760524 -0.95584236
#> 34   0.141257779  0.98625078  0.006862003 -0.656434084  0.45151228 -1.12639032
#> 35  -0.432261420  0.47905329 -0.842607879 -0.209576702  0.03615431  1.11204399
#> 36  -0.082567201  0.85236045 -0.367691466 -0.487964599 -0.57284406 -0.14481872
#> 37  -0.031905959 -0.96784017 -1.102915710  0.612596636 -0.51827027 -0.33105546
#> 38  -0.304849606  0.65235682 -0.669199739 -0.324409337 -0.31126440 -0.68521893
#> 39  -0.383888198 -0.42903649 -1.207156983  0.285779346  0.42994944 -0.97382465
#> 40  -0.327677997  0.62609593 -0.699292361 -0.305920031  0.45154773 -1.13687849
#> 41   0.200316256 -1.22473606 -0.833184125  0.867491206 -0.57213610 -0.14865824
#> 42   0.075400859  0.95313669 -0.117378764 -0.603603466  0.38399249 -0.80480944
#> 43  -0.347522759 -0.49964779 -1.209094872  0.324073954  0.25117617 -0.43023203
#> 44  -0.371017650  0.57150691 -0.757096196 -0.268872124 -0.14742784  0.92178424
#> 45  -0.478220771  0.38780596 -0.914433274 -0.154309116 -0.43502555  1.21310030
#> 46   0.408002792  0.96448579  1.116550926 -1.052413648  0.42925819  1.22463062
#> 47   0.373880063 -1.30694406 -0.174138627  1.239677341  0.45033455 -1.10108265
#> 48  -0.221632162  0.73730323 -0.559471293 -0.387987572  0.27330975 -1.21947096
#> 49   0.390662214  0.99734304  0.919064185 -0.988779401 -0.46776030  0.45238873
#> 50   0.090895212 -1.11221992 -0.994668788  0.733951982  0.45156815 -1.13396486
#> 51   0.286338478 -1.29299872 -0.621718906  1.006112834 -0.57310210  0.13727899
#> 52  -0.063356839 -0.92826205 -1.122990129  0.583938114 -0.02929202 -0.99171765
#> 53   0.403027686  0.97690321  1.047967459 -1.030552070 -0.30999664  0.71635969
#> 54   0.387169218 -1.28918941 -0.023025695  1.308236989  0.02175478 -1.03727195
#> 55   0.229338826 -1.25054421 -0.773463180  0.909633229  0.35514431 -0.71560512
#> 56  -0.262331483 -0.64597264 -1.200590815  0.406114270 -0.58479698 -0.04170841
#> 57  -0.447771417 -0.28453970 -1.191940405  0.209013540  0.44690922 -1.18381275
#> 58   0.398276556  0.74248973  1.964386244 -1.306516350  0.03669505  0.09683864
#> 59  -0.024847043  0.89257573 -0.281767036 -0.529256247 -0.20206999  0.85746929
#> 60   0.253343961 -1.26999328 -0.716125835  0.947650477  0.05883617 -1.06880121
#> 61  -0.023134381  0.89370859 -0.279137084 -0.530491453 -0.15119040 -0.87212762
#> 62   0.394835785  0.99157753  0.958052758 -1.001512415  0.33496532 -1.24464279
#> 63  -0.426769647 -0.33588323 -1.199037675  0.236105232 -0.58093084 -0.08990892
#> 64  -0.169297391  0.78395334 -0.489141537 -0.426081645  0.35213306  1.32432370
#> 65  -0.529630697  0.04332940 -1.102785376  0.036872710  0.36426082 -0.74302869
#> 66   0.166819139  0.99746542  0.060384714 -0.678432151 -0.88053783  2.26684324
#> 67  -0.073836533 -0.91481990 -1.129140331  0.574502127  0.12272724 -1.12000676
#> 68   0.150360542 -0.29043233  2.788432540  2.277385222  0.05674882  0.04900568
#> 69  -0.149094768 -0.81418860 -1.166199009  0.507635780 -0.20799580  0.85023990
#> 70  -0.153448228  0.79723162 -0.467478188 -0.437458902  0.38428691  1.31967513
#> 71  -0.512857172  0.28565450 -0.982594629 -0.095296516 -0.50464754 -0.36369350
#> 72   0.216200041  1.01590844  0.176181080 -0.724637348  0.38938296  0.95536661
#> 73   0.064871816 -1.08294717 -1.022555679  0.706379372  0.08219188 -1.08798555
#> 74  -0.241651288 -0.67854344 -1.196180880  0.425039659  0.44953662 -1.16815438
#> 75  -0.498083890  0.33576785 -0.950642673 -0.123920201 -0.58194876 -0.08028315
#> 76  -0.494072907 -0.14644849 -1.163741968  0.136651604  0.25775749  1.28877372
#> 77  -0.484699826  0.37210186 -0.925701124 -0.145060792  0.41031748  1.29735366
#> 78   0.411193869  0.95398149  1.170666886 -1.069497012  0.06849954  1.14193370
#> 79  -0.505102592  0.31374798 -0.965021645 -0.111270791  0.34483080  0.79648510
#> 80   0.277159749 -1.28711379 -0.650137123  0.988991382  0.06552760 -1.07435118
#> 81   0.318154612 -1.30927912 -0.505976182  1.072321584 -0.28147165 -0.88375925
#> 82  -0.284022567 -0.61072141 -1.204281447  0.385939516 -0.07222109 -0.95141909
#> 83  -0.008010870 -0.99717518 -1.085907564  0.634815588  0.01927285 -1.03511493
#> 84  -0.397059493 -0.40176185 -1.205423583  0.271149839  0.39077967 -1.24919077
#> 85  -0.317801557  0.63764339 -0.686256234 -0.313991950 -0.45639860 -0.46261752
#> 86   0.008570722 -1.01717263 -1.073119870  0.650526594 -0.25576712 -0.75482643
#> 87   0.292721879  1.03223776  0.404446929 -0.810996593 -0.37857180 -0.59125152
#> 88   0.170330641  0.99892504  0.068034562 -0.681541673  0.41849750 -1.23667586
#> 89   0.212261793  1.01461830  0.166226396 -0.720734904 -0.57143722  0.14667714
#> 90  -0.345207588 -0.50394353 -1.209091301  0.326426146  0.18543413  1.23988914
#> 91  -0.519063465 -0.03648075 -1.131658719  0.079022446 -0.55436562 -0.22484934
#> 92   0.393360595  0.99372036  0.943853414 -0.996885304 -0.56784879 -0.17009233
#> 93  -0.243662248  0.71623781 -0.588665849 -0.371604969 -0.44273594 -0.48735577
#> 94  -0.132639998 -0.83684465 -1.159109709  0.522183652 -0.14111502  0.51793171
#> 95   0.251511746  0.31314463  3.113561926 -1.620641511  0.21974756 -1.18859492
#> 96   0.267729844 -1.28063111 -0.677505245  0.972128908 -0.39241174 -0.57025741
#> 97  -0.456218575  0.43501374 -0.878682337 -0.182555221 -0.58496009  0.02765160
#> 98  -0.391421469 -0.41356366 -1.206239844  0.277470294  0.40150111  1.30818149
#> 99  -0.356754736  0.59024668 -0.737929758 -0.281396011  0.45139222 -1.14399107
#> 100 -0.530495014  0.17837419 -1.042320824 -0.035741110 -0.22142819  0.83365357
#>          x5.pen3     x5.null1
#> 1   -0.484834813 -0.772292667
#> 2    1.439833860  1.427336767
#> 3   -1.105638566  0.225346033
#> 4   -0.993313223 -0.424980891
#> 5   -0.635907714  0.617723201
#> 6    0.523589334 -1.201186651
#> 7    2.946676527  1.855201097
#> 8   -0.672701952 -0.665805618
#> 9   -0.452272662 -0.789325500
#> 10   1.243197790 -1.444001690
#> 11  -1.174276192 -0.159867613
#> 12  -0.982381529  0.366213850
#> 13  -1.083466668 -0.324406297
#> 14  -0.858355587  0.470349659
#> 15  -1.048588932 -0.366863850
#> 16  -0.733478856  0.557444472
#> 17  -0.563069936 -0.729809608
#> 18  -0.839354839  0.484511094
#> 19  -1.037804083 -0.378955688
#> 20   0.006539635  0.928843674
#> 21  -1.194503694 -0.061309648
#> 22   1.111429196  1.324807898
#> 23  -1.168292373 -0.176583073
#> 24   2.884495244 -1.923263355
#> 25   0.886633571 -1.327559760
#> 26  -0.920175743 -0.490875329
#> 27   1.368079095 -1.483356311
#> 28   2.245390088  1.662738422
#> 29  -0.133439564  0.869304341
#> 30  -0.121056473  0.874705080
#> 31  -1.049863392  0.296462620
#> 32   3.662081556 -2.132737664
#> 33  -0.756634590  0.542253687
#> 34   1.100392934  1.321276792
#> 35  -0.725935314 -0.632373893
#> 36  -1.178968447  0.077354523
#> 37  -1.139215988  0.171177594
#> 38  -0.984868712  0.363855295
#> 39   1.654429124  1.492005175
#> 40   1.055116712  1.306726363
#> 41  -1.178394039  0.079262604
#> 42   2.158550494  1.638224996
#> 43   3.121953172  1.902067323
#> 44  -0.923181196 -0.488351188
#> 45   6.834246030  2.858289088
#> 46   0.853837115 -1.316503109
#> 47   1.204216787  1.354262714
#> 48  -0.188385759  0.844995202
#> 49  -1.149419818 -0.219780280
#> 50   1.067844897  1.310827353
#> 51  -1.194475707 -0.061684968
#> 52  -0.713310105  0.570376781
#> 53  -1.052460197 -0.362417122
#> 54  -0.650918543  0.608802292
#> 55   2.401801765  1.706467268
#> 56  -1.190695523  0.026363287
#> 57   0.828672268  1.232284925
#> 58   4.340462373  2.219661249
#> 59  -0.970351621 -0.446773951
#> 60  -0.601500200  0.637744137
#> 61  -0.842902564  0.481895460
#> 62   0.008862476  0.929805415
#> 63  -1.186095995  0.050147253
#> 64   0.032702830 -1.011970046
#> 65   2.328252004  1.685969969
#> 66   9.187309190  3.460737963
#> 67  -0.505506932  0.690810886
#> 68   4.232806491  2.191984356
#> 69  -0.975224423 -0.442244061
#> 70   0.220063988 -1.087296737
#> 71  -1.129631658  0.187975013
#> 72   1.776121040 -1.607793368
#> 73  -0.568209566  0.656587555
#> 74   0.909584839  1.259220307
#> 75  -1.187137641  0.045391265
#> 76  -0.310696939 -0.859688648
#> 77   0.449656383 -1.174237317
#> 78  -0.682180603 -0.659978749
#> 79   2.210578625 -1.734671600
#> 80  -0.592148599  0.643087898
#> 81   5.995307717 -2.744387567
#> 82  -0.761677465  0.538894123
#> 83  -0.654092847  0.606900606
#> 84   0.266465861  1.032004057
#> 85  -1.095368989  0.239806830
#> 86  -0.938497719  0.405803913
#> 87  -1.037817460  0.309928389
#> 88   0.458417164  1.103349299
#> 89  -1.194099408 -0.066316110
#> 90  -0.483000697 -0.773261584
#> 91  -1.164902723  0.117319697
#> 92  -1.175003018  0.089930115
#> 93  -1.085503963  0.253022864
#> 94   5.281898159  2.460845347
#> 95  -0.320701907  0.783874639
#> 96  -1.048312480  0.298226315
#> 97  -1.194631976 -0.007765073
#> 98   0.358176417 -1.140220334
#> 99   1.023520382  1.296509776
#> 100 -0.986104102 -0.431948580