Title

Title

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>         x1.pen1     x1.pen2     x1.pen3    x1.null1     x2.pen1     x2.pen2
#> 1   -0.56010989  0.16692389 -0.98580944 -0.06553189  0.04826682 -1.01461826
#> 2    0.02983060 -1.09260340 -0.26802030  0.69240900  0.26534744 -1.23672634
#> 3   -0.42177922  0.50924021 -0.92536633 -0.24787593  0.39506188  0.99110169
#> 4   -0.39811662  0.54395422 -0.91324513 -0.26765851  0.20360953 -1.18401781
#> 5    0.10624595 -1.15379778 -0.08192714  0.78336408 -0.12216022 -0.79811255
#> 6    0.04421017 -1.10526634 -0.23715466  0.70820256 -0.31666583 -0.50326311
#> 7    0.21858631 -0.60011848  2.52075294  1.65856480 -0.47787112  0.41856954
#> 8   -0.46762660  0.43315679 -0.94765152 -0.20557181  0.37162335 -1.06293258
#> 9   -0.23605154 -0.79283326 -0.66752267  0.44602967 -0.25573329  0.75753514
#> 10  -0.52005027 -0.24949664 -0.94090954  0.14424544 -0.44231780  0.49624985
#> 11   0.27213782 -1.14299337  0.84189148  1.14631041 -0.07840301 -0.85636399
#> 12  -0.28356056 -0.72562662 -0.72021775  0.40385516  0.25160725 -0.62159138
#> 13  -0.53309711  0.28306360 -0.97621384 -0.12541953 -0.17294509  0.84114566
#> 14   0.28337350 -0.98221642  1.48360351  1.35526654 -0.15203152  0.86045874
#> 15   0.16485893 -0.38025379  3.02982516  1.79833560 -0.06352613 -0.87569927
#> 16   0.23848731 -1.18875541  0.49683374  1.02249449 -0.34026516  0.65658586
#> 17   0.49319025  1.17239086  0.41934035 -1.06063941  0.17427039  1.08947524
#> 18  -0.32783482  0.63520432 -0.87485061 -0.32142748  0.31532474 -1.26578177
#> 19  -0.43352890 -0.46792480 -0.86658442  0.25771951 -0.32150752 -0.49482306
#> 20  -0.34250020  0.61735390 -0.88315780 -0.31068359 -0.52881501  0.19245240
#> 21  -0.25263567  0.71942174 -0.82972131 -0.37391847  0.03579747 -0.07787263
#> 22   0.28425385  0.39777098  3.00219293 -1.86061588 -0.50079345 -0.06927606
#> 23   0.19367343 -1.19495026  0.23541904  0.92067432 -0.33428370 -0.47217293
#> 24  -0.52245498  0.31382607 -0.97190780 -0.14155411  0.35493668 -1.26843826
#> 25  -0.44831619  0.46690641 -0.93846441 -0.22417219  0.35266131  1.11164411
#> 26  -0.04519282 -1.02007483 -0.40701539  0.61684651  0.28964693  1.12554929
#> 27  -0.49167990  0.38630711 -0.95865675 -0.18013847  0.18539798  0.28080022
#> 28  -0.54703871 -0.14473988 -0.96364004  0.09134099  0.16745510 -0.36805145
#> 29   0.49315217  1.17241731  0.41906305 -1.06053333 -0.51275915 -0.01530233
#> 30  -0.10104062 -0.95989958 -0.49367862  0.56520244 -0.34659351  0.64816598
#> 31   0.11579061  0.05681173  3.81906628 -2.07945922  0.35147808 -1.26963278
#> 32   0.28183143 -0.96245534  1.54688729  1.37477345  0.36472893 -1.26225017
#> 33  -0.07738188 -0.98600679 -0.45832319  0.58676427  0.16042007 -1.14088203
#> 34   0.46657460  1.18346155  0.25859657 -0.99763279  0.32501996  0.66036498
#> 35  -0.55081215 -0.12586378 -0.96690652  0.08185847 -0.08402318  0.91903586
#> 36   0.30218927  1.14187761 -0.24348655 -0.77371949  0.28876407 -0.74098140
#> 37  -0.13480750  0.83343812 -0.74958746 -0.45123492 -0.23036686 -0.69375618
#> 38  -0.21323171 -0.82339820 -0.64086165  0.46606743  0.09711513 -0.16544796
#> 39  -0.55753083 -0.08666894 -0.97289460  0.06219526  0.35068019 -0.96552514
#> 40  -0.56434086  0.13276002 -0.98671433 -0.04815632 -0.03027138  0.96119521
#> 41  -0.50723881 -0.28926959 -0.93018307  0.16451531 -0.12231811 -0.79789837
#> 42   0.08749151  1.01006861 -0.55412778 -0.59739550  0.38042848 -1.23284784
#> 43   0.06854961 -1.12559428 -0.18090858  0.73620933 -0.27255909 -0.79124678
#> 44  -0.54782287  0.23065104 -0.98182036 -0.09821210 -0.22725090 -0.68655636
#> 45  -0.55973088 -0.07139557 -0.97493906  0.05453902  0.31008315 -1.26365116
#> 46  -0.11789384  0.84843709 -0.73700581 -0.46212035 -0.40618100 -0.33150055
#> 47  -0.28677066 -0.72089533 -0.72365199  0.40097055 -0.08130890 -0.85256078
#> 48  -0.04298936  0.91150462 -0.67741152 -0.51035857 -0.52911150  0.18640869
#> 49   0.28499914 -1.06344190  1.19834904  1.26512839  0.35848986  0.77372540
#> 50   0.44180115  1.18550485  0.14584531 -0.95140086  0.29952174  1.12596197
#> 51  -0.18949254 -0.85412394 -0.61208917  0.48686355  0.33472303  1.12067340
#> 52  -0.35951656 -0.60588774 -0.79769628  0.33354259  0.34677110 -0.94948772
#> 53  -0.54600195  0.23806252 -0.98115733 -0.10203956  0.34808431 -0.95482708
#> 54  -0.05487433 -1.01000558 -0.42287279  0.60769212 -0.52802625  0.20512748
#> 55   0.02633173  0.96552296 -0.61557749 -0.55579183 -0.33341319 -0.47373437
#> 56   0.43183778  0.73009400  2.13651777 -1.61837299  0.08850691 -1.06178349
#> 57  -0.43833914 -0.45789882 -0.87086892  0.25236626 -0.48272720 -0.13423924
#> 58   0.27859590 -0.92989998  1.64743731  1.40544156  0.06843365 -1.03847599
#> 59   0.25988491 -1.16824321  0.68517562  1.09133407  0.39503682  0.99823272
#> 60   0.49417884  1.17168968  0.42660498 -1.06341542  0.34891949  0.73912424
#> 61  -0.55737579 -0.08768671 -0.97275262  0.06270553  0.38936294  0.91948595
#> 62   0.28369810 -1.08367500  1.11819453  1.23907073  0.08071738  1.03753291
#> 63   0.52200682  1.02016234  1.21923284 -1.33959379 -0.51370371 -0.01038925
#> 64  -0.52268177 -0.24075450 -0.94310978  0.13980632  0.02398636  1.00031220
#> 65  -0.51670137  0.32891920 -0.96951299 -0.14951987  0.09403638  0.05917991
#> 66   0.28172852 -1.10119737  1.04428372  1.21472337  0.31031348 -1.26375059
#> 67  -0.12863827 -0.92830231 -0.53276280  0.54047298  0.20626595  0.33323856
#> 68  -0.08530937 -0.97735925 -0.47037524  0.57949517 -0.48662696 -0.12131354
#> 69  -0.53507331  0.27683423 -0.97699326 -0.12216796 -0.08622483  0.91723413
#> 70   0.13748881 -1.17330740  0.01409691  0.82689924 -0.50185294  0.34937175
#> 71  -0.49066720 -0.33547084 -0.91621519  0.18823958  0.20630450  0.33333629
#> 72   0.20727601 -1.19622785  0.30334020  0.94794233  0.29290317  0.56459116
#> 73   0.11709116 -1.16102404 -0.05023742  0.79795330  0.37917408  0.86072265
#> 74  -0.34080825 -0.63703635 -0.77930811  0.35134783 -0.02412062  0.96580092
#> 75  -0.47524215 -0.37461955 -0.90306767  0.20852338 -0.33367756 -0.47326045
#> 76   0.25626736 -1.17321172  0.64790365  1.07796706  0.38532805 -1.19015863
#> 77   0.46353941  0.81312724  1.89858194 -1.54883381  0.38248385 -1.13591877
#> 78  -0.55832914 -0.08131495 -0.97362968  0.05951116 -0.39458093 -0.35608432
#> 79  -0.41115860 -0.51258618 -0.84636781  0.28179632  0.36207625  0.78736316
#> 80   0.36641124  1.16935773 -0.09746730 -0.84431666  0.38562917  0.89526768
#> 81  -0.32511666 -0.66226467 -0.76354873  0.36599986 -0.07938940  0.92280822
#> 82   0.21656287  1.09475891 -0.39140447 -0.69534073  0.36312182  0.79142250
#> 83   0.50992106  0.96172815  1.42944845 -1.40634452  0.38524724 -1.17614066
#> 84   0.52166409  1.01805216  1.22721620 -1.34216757 -0.24209370 -0.62506924
#> 85   0.11561179  1.02961528 -0.52312168 -0.61728829  0.13881696  1.07139810
#> 86  -0.55955019  0.17066155 -0.98565883 -0.06743836  0.08050524  1.03740100
#> 87  -0.10836184  0.85676019 -0.72978179 -0.46824880 -0.30111545 -0.52987680
#> 88   0.51748842  1.14125325  0.65994186 -1.14976232  0.36673646  1.09901662
#> 89  -0.33296341  0.62902321 -0.87777357 -0.31769357  0.33316684 -1.27056181
#> 90  -0.04953121 -1.01558173 -0.41416924  0.61273185  0.22071102 -1.19986692
#> 91   0.24430113 -1.18496083  0.54157067  1.03914943 -0.13876678 -0.77541400
#> 92  -0.54797550 -0.14021049 -0.96444663  0.08906466 -0.50178969  0.34958701
#> 93   0.01744054 -1.08134288 -0.29335044  0.67920331  0.06630521 -1.03597815
#> 94  -0.11713766  0.84910073 -0.73643629 -0.46260663 -0.28247265  0.72762703
#> 95  -0.42451166  0.50506265 -0.92673949 -0.24551704  0.38602448  1.06298162
#> 96   0.01255940 -1.07682316 -0.30302854  0.67409643 -0.14860220 -0.76180320
#> 97   0.28116430 -0.95502084  1.57023203  1.38192856  0.31484162 -0.82988244
#> 98   0.38199327  1.17451685 -0.05530780 -0.86371200 -0.52846125  0.19859419
#> 99  -0.51612818 -0.26213328 -0.93762941  0.15067196 -0.51563983  0.29447518
#> 100  0.08392564 -1.13762720 -0.14244890  0.75483402  0.36951600  1.09562367
#>         x2.pen3    x2.null1     x3.pen1      x3.pen2     x3.pen3     x3.null1
#> 1   -0.98317450  0.65380960  0.31337522  1.209329955 -0.90241085 -0.872502936
#> 2   -0.59413077  0.94734495  0.45269085  1.242470314 -0.17987187 -1.299431482
#> 3    1.45603896 -1.17783261 -0.46041097 -0.161334070 -0.90101775  0.052284736
#> 4   -0.75274355  0.84507395  0.23467846  1.133063995 -1.02810163 -0.757132151
#> 5   -1.10408075  0.49424641  0.43462111  1.269840107 -0.42902865 -1.174831450
#> 6   -1.15930179  0.32275893 -0.33014858 -0.499595879 -0.60082997  0.248395123
#> 7   -0.90927959 -0.15415036 -0.45219653 -0.215519955 -0.86243044  0.081732916
#> 8    0.89208904  1.59860034 -0.20307776 -0.658629306 -0.38885810  0.358575622
#> 9   -0.61842266 -0.36193298 -0.32281480  0.383560491 -1.15199888 -0.232062014
#> 10  -0.85678076 -0.19804132  0.25972174  1.158673533 -0.99467999 -0.790917253
#> 11  -1.08104962  0.53274881 -0.07006930  0.770809247 -1.19314339 -0.452350763
#> 12   2.13792877  2.00856288 -0.03594012 -0.808285904 -0.09574123  0.490612740
#> 13  -0.51237702 -0.42329093 -0.02040690  0.834724583 -1.18564980 -0.494237837
#> 14  -0.48473782 -0.43848077 -0.14267902  0.672443115 -1.19531182 -0.392126888
#> 15  -1.07216793  0.54607471  0.17722261 -0.927552328  0.43657485  0.694621599
#> 16  -0.72407238 -0.29477907  0.26671844 -0.923508662  0.87830201  0.842810294
#> 17   0.07060629 -0.69821552 -0.43042026 -0.297627033 -0.79790808  0.127460622
#> 18  -0.40330217  1.05458816  0.33269833  1.225682642 -0.85989845 -0.906019484
#> 19  -1.15971571  0.31821259 -0.28095072  0.459361146 -1.16964144 -0.272480130
#> 20  -1.02838662 -0.03336764 -0.34945768  0.330613215 -1.13722581 -0.204228733
#> 21   4.30703579 -1.95235760  0.38296250  0.907373886  1.02187010 -1.782003782
#> 22  -1.11587177  0.09957479  0.28453166 -0.620020083  2.21806425  1.233374423
#> 23  -1.16055550  0.30607718 -0.15668746 -0.705349688 -0.31119768  0.395458605
#> 24  -0.15861701  1.17618749  0.25977138  1.158723233 -0.99460851 -0.790986552
#> 25   0.76773451 -0.95712252 -0.36363953  0.300400808 -1.12790667 -0.188457380
#> 26   0.43119322 -0.83816717  0.35927807  0.838130769  1.21125430 -1.848471191
#> 27   3.47898962 -1.73922082 -0.13598307  0.681807453 -1.19553598 -0.397683210
#> 28   2.75503006  2.19556646 -0.37478360  0.275411486 -1.11971220 -0.175460768
#> 29  -1.10172380  0.07240816 -0.08832669  0.746685642 -1.19464609 -0.437165656
#> 30  -0.73200528 -0.28943636 -0.42208904 -0.321301438 -0.77782824  0.140940968
#> 31  -0.18652308  1.16300785 -0.45635349  0.002316533 -1.00075869 -0.034213491
#> 32  -0.06673790  1.21854362  0.12870710  1.015659729 -1.12517016 -0.633923588
#> 33  -0.83497960  0.78478008  0.43236177  1.071995936  0.52771731 -1.599214713
#> 34   2.52001474 -1.48360118 -0.42877163 -0.302524529 -0.79381118  0.130237363
#> 35  -0.39135788 -0.48774139 -0.01525358  0.841230447 -1.18457209 -0.498657534
#> 36   1.83132484  1.91262511  0.05510621 -0.870354004  0.09271819  0.567050127
#> 37   5.68782161 -2.30340685  0.19533463 -0.318631004  3.04096644  1.450012354
#> 38   3.22809114  2.33512998  0.19926952 -0.932506559  0.52078284  0.724029094
#> 39   1.20301495  1.70677320  0.29460416 -0.891856522  1.14163456  0.925238509
#> 40  -0.31255146 -0.52715743 -0.39657851  0.222415866 -1.10087499 -0.148007185
#> 41  -1.10415602  0.49410886  0.44482600  1.125819202  0.34272750 -1.526236219
#> 42   0.16991405  1.32144843  0.27256468  1.171376553 -0.97542821 -0.809183915
#> 43   5.90612591 -2.35887921 -0.33073927 -0.498701609 -0.60185111  0.247822968
#> 44   5.67169953 -2.29931013 -0.07999206 -0.773550938 -0.17747039  0.455622760
#> 45  -0.42752035  1.04169378 -0.29359700  0.437248999 -1.16492778 -0.260607743
#> 46  -1.15739012  0.23244565  0.29858757 -0.882868059  1.19772577  0.942341869
#> 47  -1.08271911  0.53016252 -0.46336459 -0.076425377 -0.95573009  0.007037994
#> 48  -1.03096910 -0.03024253 -0.26312103  0.489567372 -1.17549235 -0.288825201
#> 49   2.19720051 -1.39402042 -0.37634238 -0.422308520 -0.68368899  0.200333327
#> 50   0.47316594 -0.85352939  0.29628820 -0.675769008  2.04332777  1.185705064
#> 51   0.65086741 -0.91684402  0.24744070 -0.481153289  2.61625045  1.339610674
#> 52   1.25121752  1.72311252  0.28865150  1.186813420 -0.94898379 -0.833102063
#> 53   1.23524556  1.71771013 -0.38089112 -0.413687636 -0.69230553  0.195128601
#> 54  -1.02287724 -0.03993495 -0.41261887  0.178566549 -1.08377986 -0.125363592
#> 55  -1.16051015  0.30691081 -0.06704862  0.774765557 -1.19283156 -0.454870865
#> 56  -0.93854397  0.69745566 -0.16874842 -0.693638376 -0.33155050  0.385942974
#> 57  -1.13035023  0.13226023  0.43472942  1.269786569 -0.42807766 -1.175336716
#> 58  -0.96181759  0.67528766  0.14486710  1.034313728 -1.11405178 -0.651071492
#> 59   1.42630649 -1.16877038  0.27697021  1.175659133 -0.96845635 -0.815613794
#> 60   2.29859015 -1.42240193  0.43601240  1.269103081 -0.41657079 -1.181429619
#> 61   1.72831089 -1.25927105  0.29248439  1.190405951 -0.94226342 -0.838987753
#> 62  -0.12766589 -0.61338621  0.29046518 -0.899311299  1.09114761  0.909715466
#> 63  -1.10033921  0.06993259 -0.17314426  0.628946518 -1.19319017 -0.366720499
#> 64  -0.22676802 -0.56815565 -0.22986065  0.543339343 -1.18418168 -0.318341434
#> 65   3.99538693 -1.87264512  0.29209356 -0.896564554  1.11024058  0.915600570
#> 66  -0.42648279  1.04224995 -0.36982130 -0.434288062 -0.67151880  0.207613682
#> 67   3.35348758 -1.70644534 -0.28739615  0.448167419 -1.16730035 -0.266460781
#> 68  -1.12768961  0.12575356 -0.38699347  0.246493305 -1.10967953 -0.160464752
#> 69  -0.39448002 -0.48614158  0.08062748  0.958932205 -1.15222368 -0.585588512
#> 70  -0.95069508 -0.11624009  0.44059611  1.106140815  0.41244066 -1.554090699
#> 71   3.35325219 -1.70638371  0.13589726  1.023987416 -1.12035993 -0.641492407
#> 72   2.77576226 -1.55310858  0.16730995  1.059825980 -1.09667561 -0.675748116
#> 73   1.92733141 -1.31722673  0.30752166 -0.848015381  1.38429644  0.998229006
#> 74  -0.30317842 -0.53172878  0.24536090 -0.932158645  0.74110339  0.798278774
#> 75  -1.16052411  0.30665775  0.31040968 -0.812660394  1.54423259  1.045030024
#> 76   0.39746903  1.41376777 -0.14566539 -0.715802864 -0.29247042  0.404125076
#> 77   0.62778282  1.50213272 -0.39003358 -0.395620619 -0.70999114  0.184310245
#> 78  -1.15894626  0.24513377  0.04910338  0.078523149  3.98908665  1.689523380
#> 79   2.15641709 -1.38253539  0.28568493 -0.906260934  1.03970453  0.893767739
#> 80   1.81239265 -1.28390509  0.23629592 -0.933782908  0.69133740  0.781819842
#> 81  -0.38476200 -0.49111115  0.10356857 -0.064124548  3.65777125  1.606609044
#> 82   2.14418072 -1.37908154  0.38381588  1.261028182 -0.70859032 -1.011984359
#> 83   0.46123489  1.43869241 -0.20538771 -0.656178595 -0.39268906  0.356714394
#> 84  -1.14681371  0.39014585 -0.45246472 -0.214152257 -0.86344185  0.080983148
#> 85  -0.01115191 -0.66405673 -0.35176125 -0.465454456 -0.63873593  0.226832701
#> 86  -0.12805937 -0.61321072  0.31049648 -0.806999074  1.56805888  1.051922399
#> 87  -1.15762953  0.33718753 -0.40714156  0.194125315 -1.09000308 -0.133393321
#> 88   0.88245738 -0.99573921 -0.14895919  0.663598787 -1.19502288 -0.386908571
#> 89  -0.30952533  1.10292167 -0.42086278  0.153716441 -1.07347814 -0.112544389
#> 90  -0.71453477  0.87115982  0.16645900 -0.924309686  0.39887090  0.681247292
#> 91  -1.11170677  0.47981525  0.37893691  1.258337486 -0.72638407 -1.000377613
#> 92  -0.95057349 -0.11635652  0.09810185 -0.049616978  3.69181102  1.615157747
#> 93  -0.96416403  0.67298464 -0.07233463  0.767835900 -1.19336539 -0.450462476
#> 94  -0.65192776 -0.34137581  0.11282877  0.997118891 -1.13505426 -0.617537962
#> 95   1.11747485 -1.07235549 -0.22688343  0.548009020 -1.18482920 -0.320933154
#> 96  -1.11595159  0.47129539  0.13687811 -0.153765538  3.44521595  1.553036284
#> 97   1.59270600  1.83611344 -0.01092158 -0.092378116  3.42634686 -2.555568576
#> 98  -1.02573286 -0.03654760 -0.20734585  0.578136095 -1.18858268 -0.337776502
#> 99  -0.98030487 -0.08681848 -0.02732009 -0.814725734 -0.07917439  0.497558929
#> 100  0.90884828 -1.00450293  0.25348549 -0.929785463  0.78950093  0.814124645
#>          x4.pen1     x4.pen2      x4.pen3    x4.null1      x5.pen1      x5.pen2
#> 1    0.316184095  1.21653130 -0.216971352 -0.83261820 -0.448831603  0.289840873
#> 2   -0.129150200  0.82386143 -0.816144418 -0.43710236  0.385299653 -1.212180515
#> 3   -0.474612580  0.28707920 -1.014989421 -0.13074043  0.027611124  0.960018995
#> 4   -0.471647571  0.29514974 -1.013881060 -0.13492069  0.189502112 -0.502538482
#> 5    0.284871902  1.19675505 -0.292464592 -0.79351690  0.377534573  1.197430127
#> 6   -0.276828871  0.64368583 -0.916276798 -0.32514962 -0.310658022  0.588480219
#> 7   -0.389604714 -0.49005258 -0.914401532  0.26752081  0.318160214 -1.441983135
#> 8    0.009072131 -1.02097978 -0.508443116  0.61126327  0.395004926  1.202172900
#> 9   -0.504902440 -0.18561765 -1.003163439  0.10874161  0.200235837  1.097271744
#> 10   0.055782035  1.01009496 -0.648656489 -0.57597309  0.084697323  0.053799746
#> 11   0.393412092 -1.12923568  1.181094940  1.32975689  0.090765407 -1.177925207
#> 12  -0.309874484  0.59728568 -0.935608274 -0.29831040  0.398463077 -1.441163439
#> 13   0.443181855  1.25640610  0.278393855 -1.05738422  0.474740512  1.178377862
#> 14   0.301395618  1.20746454 -0.253908374 -0.81370245  0.443217746  0.894386709
#> 15   0.225364747  0.35482076  3.336568382 -2.03889242  0.415628634  0.813110661
#> 16  -0.218638037  0.71921604 -0.879611784 -0.37036534  0.211390815 -1.334494602
#> 17   0.266189279  1.18399895 -0.333022181 -0.77175611 -0.470329516  0.147815322
#> 18   0.098848004 -0.08176542  3.975938831  2.14542994  0.479632185  1.168623732
#> 19  -0.506811719  0.17915279 -1.025427489 -0.07538140 -0.010287257  0.926552514
#> 20   0.382157231 -1.20397291  0.801399408  1.19840812 -0.816217087 -1.824114005
#> 21   0.070270156  1.02336114 -0.632613398 -0.58750480 -0.267366066  0.649653541
#> 22   0.200722102  0.29708611  3.474941437 -2.07730868 -0.435324828 -0.179535986
#> 23  -0.515838579  0.13485322 -1.027409661 -0.05288882 -0.467855117  0.007599598
#> 24  -0.489649772  0.24239171 -1.020291488 -0.10770510  0.200991442  1.097813729
#> 25   0.136348725  1.08175252 -0.551131586 -0.64282651  0.371106872  0.695620832
#> 26   0.214807223 -1.19856334 -0.110542783  0.82674064 -0.359358923  0.509196182
#> 27  -0.518896070 -0.10959270 -1.015031458  0.07029204  0.304553451  1.164558613
#> 28  -0.482603226 -0.26827000 -0.985812276  0.15090284 -0.514296746  1.646902415
#> 29   0.234129931 -1.21085343 -0.054699076  0.85313456  0.483094279  1.052256299
#> 30  -0.339198834 -0.57994503 -0.873704395  0.31711952 -0.392399185  0.445311160
#> 31   0.402140487  1.25482083  0.065305967 -0.96620625  0.229535897 -1.355642093
#> 32   0.007431274 -1.01933154 -0.510725225  0.60983214  0.263966975  1.140432349
#> 33  -0.249031000 -0.71702968 -0.795398979  0.39695956  0.343900058  1.184404663
#> 34  -0.458229910  0.32940927 -1.008652730 -0.15274236 -0.377465620 -0.362302584
#> 35  -0.310546004 -0.62620847 -0.849656644  0.34340928  0.478735144  1.170704053
#> 36   0.487557950  1.19890642  0.786515624 -1.25312177 -0.432457778  0.345995137
#> 37  -0.408331733 -0.45296366 -0.929070315  0.24752965  0.350092464  0.643526620
#> 38  -0.328721074 -0.59720689 -0.864993352  0.32686083 -0.470773204  0.049591044
#> 39   0.252688286 -1.22143165  0.004131617  0.88022020 -0.195025164 -0.740221808
#> 40   0.101121659  1.05105181 -0.596427678 -0.61270126 -0.446447645 -0.131697240
#> 41   0.150984129  0.18216385  3.746689589 -2.15220933  0.445334507  0.901080971
#> 42  -0.503333676 -0.19244463 -1.001907934  0.11220743  0.096838061 -1.186309111
#> 43  -0.265051201 -0.69433847 -0.809902112  0.38330663  0.225518979 -1.351033673
#> 44   0.281805573 -1.23496737  0.109410826  0.92705130  0.029480005 -0.005771730
#> 45  -0.113474961 -0.88887007 -0.659660852  0.50865926  0.249059181  0.408921289
#> 46  -0.525633670  0.05338281 -1.027644334 -0.01174456  0.398103473 -1.442033636
#> 47   0.273318920 -0.59192805  2.787939008  1.81844219  0.405848990 -1.355982623
#> 48   0.391125017 -1.07970251  1.378296906  1.39471825  0.401573900  1.203490857
#> 49   0.467348264  1.24339699  0.469258867 -1.13394795 -0.222398601  0.706908716
#> 50   0.035452808  0.99118961 -0.670264979 -0.56005610  0.069436122  0.995484555
#> 51  -0.526903110 -0.01466638 -1.024491549  0.02250276  0.192583300  1.091746981
#> 52  -0.525413677 -0.04694695 -1.021936869  0.03874138  0.146999502 -1.253781932
#> 53   0.127337923  1.07399520 -0.563165529 -0.63496428  0.456833415  0.939202480
#> 54   0.353070616 -1.23882203  0.491181162  1.08312732  0.361125941 -1.463750278
#> 55   0.462578228  0.99500872  1.608600352 -1.53057526  0.481231033  1.041134887
#> 56  -0.303475304 -0.63719378 -0.843609478  0.34974105  0.113133403  1.031049106
#> 57   0.105386209 -1.11216323 -0.355543192  0.70107251  0.247082465 -0.689648042
#> 58   0.173200540  1.11271956 -0.498359059 -0.67622522 -0.209709383  0.722148498
#> 59  -0.263314816 -0.69682788 -0.808343469  0.38479481 -0.243042973  0.681296738
#> 60   0.490247633  1.16192816  0.974400446 -1.31993732  0.073784429 -0.141346468
#> 61   0.347135135  1.23361009 -0.130767202 -0.87531996  0.281877822  0.483001448
#> 62   0.486200815  1.10929740  1.199500920 -1.39704087  0.041470639 -0.037117327
#> 63   0.085845354 -1.09457887 -0.390162402  0.68168148 -0.313387522  0.584372833
#> 64  -0.475455177 -0.29035598 -0.980367382  0.16225694  0.379887274  1.198162629
#> 65  -0.421151762  0.41071576 -0.992744914 -0.19561330  0.364287836  1.192833323
#> 66  -0.123194356  0.83042666 -0.811584061 -0.44148876 -0.374523941  0.481206207
#> 67   0.420081666  1.25774310  0.148001502 -1.00240705 -0.346338122  0.531795238
#> 68  -0.341881788  0.54924259 -0.953323287 -0.27114488  0.189636061  1.089602222
#> 69  -0.090723174  0.86546593 -0.785901188 -0.46538442  0.452408011  1.199279967
#> 70  -0.180837465  0.76489307 -0.853909614 -0.39882150  0.333860136  1.179743508
#> 71   0.374445907  1.24595973 -0.041942586 -0.91746281 -0.127332680 -0.389216494
#> 72   0.374678012  0.72590081  2.398445362 -1.77124575 -0.996364190 -2.199347684
#> 73  -0.508733560 -0.16792084 -1.006271306  0.09976899  0.091364637  1.013512951
#> 74  -0.502357492  0.19743709 -1.024223518 -0.08469837  0.304378049  1.164461459
#> 75   0.233523517 -1.21048631 -0.056530420  0.85227992  0.401430412 -1.309767979
#> 76   0.292760877  1.20193967 -0.274392154 -0.80303676 -0.446433395  0.299165939
#> 77   0.216897503 -1.19994894 -0.104739532  0.82951643 -0.004387253 -0.133035814
#> 78   0.305555070  1.21006693 -0.243764402 -0.81893703 -0.182223123 -0.762290091
#> 79  -0.322522952  0.57870356 -0.942727564 -0.28773316 -0.093046414 -0.907652592
#> 80  -0.188357510  0.75599141 -0.859146678 -0.39320095  0.453898876  0.929150131
#> 81  -0.318008279  0.58539324 -0.940204213 -0.29153044  0.035103947 -1.099272617
#> 82   0.390563128  1.25165906  0.017927724 -0.94493962 -0.466029737 -0.010827976
#> 83   0.406882053  1.25584971  0.085957722 -0.97535259  0.089810421 -1.176602919
#> 84   0.393038441 -1.14441054  1.114679131  1.30742036 -0.294591008 -0.553246043
#> 85  -0.359502087 -0.54522005 -0.890327922  0.29774763 -0.284277778  0.626575448
#> 86  -0.175016513 -0.81481351 -0.724465287  0.45842010  0.060029369 -1.134874072
#> 87   0.472600748  1.03590223  1.470657624 -1.48638323 -0.451102188 -0.108900487
#> 88  -0.380418144 -0.50741535 -0.907122980  0.27696691  0.371159332 -1.464396905
#> 89   0.370668111 -0.94903748  1.817925777  1.53338778 -0.471681835  0.075698804
#> 90   0.180274433 -1.17397602 -0.199225499  0.78328485  0.479905190  1.033866280
#> 91  -0.073370670 -0.93419193 -0.613912031  0.54156526  0.088013007 -1.174111407
#> 92   0.392202292 -1.15694397  1.056877081  1.28777612 -0.469848549  0.155272670
#> 93   0.486497234  1.20437994  0.755144735 -1.24171774  0.389469199 -1.455781230
#> 94   0.372219214 -1.22231686  0.668119588  1.14990445 -0.420184184  0.380349809
#> 95  -0.480209672 -0.27583610 -0.983986088  0.15478768  0.247713879  0.405916735
#> 96   0.336746802 -1.24393383  0.378321134  1.03888756 -0.388372373 -0.332697369
#> 97   0.304636624 -0.69421046  2.529300307  1.74458406  0.447433141  1.201413479
#> 98  -0.429764148 -0.40711911 -0.945614708  0.22313429 -0.357136270 -0.413938271
#> 99   0.462833941  1.24732195  0.426594040 -1.11719318 -0.088890767 -0.914131322
#> 100  0.393329665 -1.13570555  1.153207732  1.32040787  0.488450655  1.119215897
#>         x5.pen3     x5.null1
#> 1   -1.30565338 -0.080925155
#> 2    0.85585963  1.862373324
#> 3   -0.71742423 -0.511264132
#> 4    2.72410677  2.489479842
#> 5    0.08892585 -0.880211974
#> 6   -1.12418078 -0.248372563
#> 7   -0.76901381  1.137304917
#> 8    0.16678969 -0.910749811
#> 9   -0.42506009 -0.659376319
#> 10   4.43350698 -2.190344896
#> 11  -1.28117465  0.752715411
#> 12  -0.13772080  1.460031745
#> 13   0.76342288 -1.127930638
#> 14   2.08527060 -1.544426563
#> 15   2.35405726 -1.622538167
#> 16  -1.08847394  0.923951894
#> 17  -1.36601663 -0.006964019
#> 18   0.84406215 -1.155455954
#> 19  -0.76924166 -0.482366592
#> 20   9.08998773 -3.412163824
#> 21  -1.07522468 -0.285816909
#> 22  -1.45668307  0.158008433
#> 23  -1.41257494  0.064206261
#> 24  -0.42348087 -0.660120336
#> 25   2.71523863 -1.725186407
#> 26  -1.18087516 -0.201761752
#> 27  -0.16701169 -0.775001875
#> 28   7.51765088  3.946100515
#> 29   1.48230479 -1.362361145
#> 30  -1.22170535 -0.165496078
#> 31  -1.04797381  0.954487696
#> 32  -0.27830841 -0.726501068
#> 33  -0.04030332 -0.828087686
#> 34  -1.48091601  0.250109132
#> 35   0.82772626 -1.149909056
#> 36  -1.27766333 -0.110951777
#> 37   2.86787116 -1.767881394
#> 38  -1.39989861  0.043023260
#> 39  -1.46812273  0.451571435
#> 40  -1.44728640  0.134059999
#> 41   2.06221434 -1.537648621
#> 42  -1.27387585  0.760339090
#> 43  -1.05728997  0.947582372
#> 44   3.85255726  2.836665930
#> 45   3.51752804 -1.946006005
#> 46  -0.14469691  1.456864802
#> 47   0.32784609  1.658814653
#> 48   0.19853773 -0.923029038
#> 49  -1.02462600 -0.322298122
#> 50  -0.65646748 -0.544092658
#> 51  -0.44086432 -0.651902274
#> 52  -1.20506286  0.826849224
#> 53   1.92764151 -1.497821784
#> 54  -0.53756880  1.265857058
#> 55   1.53036114 -1.377280231
#> 56  -0.58721947 -0.580022351
#> 57   2.27803347  2.347943412
#> 58  -1.01028549 -0.332279559
#> 59  -1.04787679 -0.305789401
#> 60   3.54952536  2.744441425
#> 61   3.31779110 -1.891787106
#> 62   4.66124597 -2.250361027
#> 63  -1.12729656 -0.245908123
#> 64   0.09891363 -0.884163844
#> 65   0.03519894 -0.858770229
#> 66  -1.19926361 -0.185744637
#> 67  -1.16543096 -0.214851721
#> 68  -0.44686081 -0.649052876
#> 69   0.52072617 -1.042724684
#> 70  -0.07477402 -0.813852164
#> 71   5.53498158 -2.479696936
#> 72  10.01964161 -3.656009101
#> 73  -0.62253913 -0.561867672
#> 74  -0.16753043 -0.774780093
#> 75   0.51461160  1.732972577
#> 76  -1.30117682 -0.085875745
#> 77   4.90001914 -2.313118234
#> 78  -1.46430741  0.464226972
#> 79  -1.42809336  0.551933322
#> 80   1.96370321 -1.508539861
#> 81  -1.33942550  0.686426564
#> 82  -1.41780777  0.073476800
#> 83  -1.28230410  0.751524129
#> 84  -1.48591037  0.349064241
#> 85  -1.09427344 -0.271520540
#> 86  -1.31516252  0.715354912
#> 87  -1.44235984  0.122649875
#> 88  -0.46227888  1.304691650
#> 89  -1.39148061  0.029803883
#> 90   1.56113176 -1.386791710
#> 91  -1.28441654  0.749287510
#> 92  -1.36319285 -0.010791007
#> 93  -0.27797886  1.395024160
#> 94  -1.25927697 -0.129594729
#> 95   3.52554502 -1.948173515
#> 96  -1.47825067  0.235076529
#> 97   0.48005587 -1.028063293
#> 98  -1.48437742  0.276489315
#> 99  -1.42596970  0.556058369
#> 100  1.15839355 -1.259565544