Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth

The smooth object from construct_smooth_data

dat

The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)
#>          x1.pen1      x1.pen2     x1.pen3    x1.null1       x2.pen1     x2.pen2
#> 1   -0.114663925 -0.827228667  0.41346625  0.37467788  0.2856839078  1.02873658
#> 2   -0.458628416 -0.221554148  0.92937888 -0.03887655  0.2823341803  1.06058429
#> 3    0.393613864  0.943395543 -1.33639650 -2.14977141 -0.3737647991  0.52051487
#> 4   -0.462982231 -0.077904550  0.99371394 -0.12278832 -0.4549475658 -0.10606125
#> 5    0.363707002  1.284526642  0.50990316 -1.25518608 -0.0454219848  0.98035033
#> 6    0.133268342 -0.130753786 -4.11514786  2.01680477  0.2689971261  0.81843327
#> 7    0.417070965 -1.012805497 -1.56103124  1.22519940 -0.4242849011  0.39936730
#> 8    0.242763516 -1.098195610 -0.26841893  0.72602911  0.3771631252 -0.57468487
#> 9   -0.356828308  1.005115929 -6.75858878  2.77522879  0.4062216317 -1.06537355
#> 10  -0.256973931 -0.665717176  0.60639033  0.24750433 -0.0002491553  1.02355729
#> 11   0.145852686 -0.160864346 -4.04296208  1.99584645  0.0206508191  1.04221922
#> 12   0.380841195  0.900883812 -1.45572981 -2.19630399 -0.4166956816  0.42006528
#> 13   0.330348595 -1.125256206 -0.59869768  0.86691625  0.0895050165  0.09986760
#> 14   0.450556065  1.193264072 -0.51329658 -1.80489781 -0.4538405622 -0.10987539
#> 15   0.439091207  1.121308908 -0.78343807 -1.92363007 -0.4140495484 -1.56166477
#> 16   0.081066819 -0.007294165 -4.40800044  2.10147542  0.1655738130 -0.16028605
#> 17   0.317637882 -1.123733628 -0.53979533  0.84269201  0.1805384621  1.14459648
#> 18   0.400887699  1.301917959  0.34422369 -1.36194189  0.3701653519 -0.56015058
#> 19  -0.422897877 -0.368519940  0.84608229  0.04998349 -0.4779648774  0.16744688
#> 20   0.153334843  1.078703449  0.92915387 -0.89578139 -0.3138418063  0.63316604
#> 21   0.240178891  0.488924472 -2.49876218 -2.58074441 -0.3501298290  0.56777548
#> 22   0.409159622 -0.926518834 -1.91373098  1.34420794 -0.4704611212  0.22382604
#> 23   0.395781490  1.300419241  0.37126910 -1.34525444  0.2617689860  0.77655144
#> 24   0.433698010  1.297655692  0.10342718 -1.50080547 -0.4581363545  0.28515193
#> 25   0.233431643  1.168245887  0.82344761 -1.00716588  0.2885555465 -0.98977727
#> 26  -0.249962396  0.524098134  1.11137949 -0.47328814  0.5401668672 -1.04756374
#> 27  -0.154168843 -0.786543370  0.46825290  0.34047907 -0.0098990465  1.01465174
#> 28  -0.382104424 -0.465794009  0.77959989  0.11133297  0.0785399220  1.08898669
#> 29  -0.162966612 -0.777108152  0.48027341  0.33279318 -0.9678036677  1.93568296
#> 30  -0.449922613  0.026010774  1.03099181 -0.18250512  0.3450918454 -1.02901790
#> 31   0.154391488 -1.047643837 -0.04124245  0.62040382 -0.1641478470  0.84833120
#> 32  -0.038993754 -0.898459222  0.30339880  0.43975272  0.0130694639  1.03555041
#> 33  -0.452444696 -0.259657314  0.90959244 -0.01619469 -0.0719191850 -0.67856609
#> 34  -0.257794669  0.510576621  1.11141228 -0.46499260 -0.3234623100  0.61654135
#> 35   0.391125123 -1.105187251 -1.00512186  1.02571641  0.4695205743 -1.09080745
#> 36   0.402801663 -0.891881388 -2.03882722  1.38531717 -0.0557508174  0.96993541
#> 37  -0.121128112 -0.820750027  0.42253299  0.36910838 -0.1644216607 -0.74069125
#> 38   0.415197106 -1.044206166 -1.40805710  1.17193823 -0.2513079016  0.73158544
#> 39   0.390689221 -0.838260178 -2.22072572  1.44420400  0.1744604635  1.14278899
#> 40  -0.462986458 -0.178188653  0.95046100 -0.06444431 -0.4569850673  0.28999195
#> 41   0.364588813  1.285078393  0.50664563 -1.25740081  0.2451198294 -0.95714225
#> 42   0.201039754 -1.076254497 -0.15237164  0.67313282  0.2490816276  1.13487560
#> 43  -0.425498303 -0.360984186  0.85083099  0.04532692  0.2061469121  1.14874785
#> 44   0.408495083 -0.922512905 -1.92856543  1.34911056 -0.4788161208  0.15838483
#> 45  -0.340669615 -0.542421596  0.71972214  0.16161714  0.0653025518 -0.80739588
#> 46   0.128349077 -1.030216495  0.01416039  0.59325245 -0.0559472017  0.96973547
#> 47   0.452166979  1.211960175 -0.43392319 -1.76866252 -0.3800317163 -0.28477745
#> 48   0.410489596 -1.068251096 -1.27396371  1.12430078  0.2819366508  1.06304645
#> 49  -0.455018846 -0.245383487  0.91714622 -0.02471714 -0.4805261362  0.04289411
#> 50  -0.457383847 -0.021247252  1.01495455 -0.15540991  0.0450489437  1.06286644
#> 51  -0.153044664  0.678252802  1.10064355 -0.57130916 -0.2232675415  0.77134868
#> 52  -0.422438943  0.140057982  1.06362856 -0.24774136 -0.0089238654 -0.73962116
#> 53   0.365316225  1.285529549  0.50393924 -1.25923683  0.2606745459  1.12209369
#> 54   0.416254390 -0.986718001 -1.67582580  1.26447859  0.1921418654  0.46894485
#> 55   0.052642909  0.956521580  1.01565650 -0.77765433  0.2849095455  0.95068033
#> 56   0.409845179 -0.930786478 -1.89781134  1.33893812  0.1739669941 -0.17611310
#> 57   0.016940438 -0.101203700 -3.84696306 -3.04586925  0.2738516064  0.84984986
#> 58   0.369221324  1.287887955  0.48910313 -1.26923838 -0.3393860344  0.58793523
#> 59   0.416887911 -1.022356379 -1.51650674  1.20980957 -0.4196556318  0.41213318
#> 60  -0.024409152 -0.911290882  0.28115993  0.45238305  0.0162732176 -0.76308644
#> 61   0.303902627  1.237812439  0.68599381 -1.12614960 -0.4043909877 -0.23722108
#> 62   0.352625739 -1.124616307 -0.71702626  0.91457053  0.2820115620  1.06259348
#> 63   0.121405025 -1.025418544  0.02825018  0.58624740 -0.0506467537 -0.69959428
#> 64   0.400057656  0.965567611 -1.27269594 -2.12464483  0.2017713369  0.50668798
#> 65   0.337458463  0.765854713 -1.81588243 -2.33303042  0.5304614049 -0.95664321
#> 66   0.435793442  1.105416907 -0.83796217 -1.94683933  0.3117178155 -1.00636019
#> 67  -0.431970409 -0.341092185  0.86310532  0.03309349 -0.4773715119  0.17321074
#> 68   0.270517412  1.206261426  0.75789274 -1.06658048 -0.4424202535  0.34399084
#> 69  -0.463968956 -0.160610125  0.95858641 -0.07474277  0.4821467323 -1.09312041
#> 70   0.425930483  1.301896098  0.17602090 -1.46057208  0.2234742788 -0.94024425
#> 71  -0.334399048  0.365407757  1.10353484 -0.37804904 -0.3478318851 -0.34024181
#> 72   0.407671438 -0.917704611 -1.94623643  1.35494071  0.2865612587  0.99738278
#> 73  -0.017898047  0.222796011 -4.94509557  2.25576229  0.3868650839 -1.05474850
#> 74   0.242018165  1.177281820  0.80939905 -1.02038902  0.2464854211  0.69867283
#> 75  -0.106785567  0.745350012  1.08923874 -0.61659253 -0.4419328486  0.34561861
#> 76  -0.098296010  0.757314534  1.08670851 -0.62488632  0.4702290586 -1.09097196
#> 77  -0.461852357 -0.062685676  0.99963927 -0.13156959 -0.3936485361 -0.25893631
#> 78  -0.016117180 -0.918465914  0.26833663  0.45959388 -0.4049609889  0.44998011
#> 79  -0.002620605  0.886128640  1.04803453 -0.71977251 -0.0647445193  0.26752521
#> 80   0.050676063  0.954055780  1.01696422 -0.77552725  0.2488196731  0.70991845
#> 81   0.359187893  0.832046320 -1.64244325 -2.26782978 -0.4383772031 -0.15685222
#> 82   0.075646935 -0.992322539  0.11527850  0.54200983  0.2670799349 -0.97387120
#> 83  -0.342236013  0.348807985  1.10171675 -0.36830514  0.2733488823  1.09710750
#> 84   0.265783463  1.201564491  0.76701179 -1.05864172 -0.4446055351 -0.13912556
#> 85   0.254688770  1.190368239  0.78747524 -1.04046490  0.2812806058  0.90850945
#> 86  -0.430135563 -0.346913040  0.85955232  0.03666469  0.2863344592  0.98137167
#> 87  -0.289617293 -0.621205763  0.65002578  0.21565924  0.1213756188  0.21004438
#> 88   0.406560576 -1.079977012 -1.20010921  1.09764936  0.2761118241  0.86588108
#> 89   0.304050642 -0.562532515 -3.02990073  1.69600879 -0.4819025966  0.08204223
#> 90  -0.163091456 -0.776973217  0.48044358  0.33268387  0.2721650559  1.10025007
#> 91   0.350016763 -1.124958072 -0.70195088  0.90857028 -0.1938286521 -0.54779694
#> 92  -0.229739099 -0.700246145  0.57010124  0.27299582 -0.4684577038 -0.05052523
#> 93   0.116693303  1.035251873  0.96550448 -0.85049496 -0.4648901899  0.25417889
#> 94   0.104405932 -1.013419222  0.06168146  0.56945577 -0.4629492264  0.26359869
#> 95  -0.339538004  0.354570537  1.10236855 -0.37168402 -0.0731986648 -0.44067804
#> 96   0.369778777 -1.119981909 -0.82703389  0.95777104 -0.4635500135  0.26073405
#> 97   0.317025069 -1.123632238 -0.53708319  0.84156808  0.5407635106 -1.04166170
#> 98   0.416176856 -1.034689424 -1.45655320  1.18894271  0.0728258474  1.08472983
#> 99   0.179283288 -1.063371138 -0.09844072  0.64781683 -0.0651572982  0.96027837
#> 100  0.322753717  1.254168187  0.63835472 -1.16310960 -0.4817853882  0.07383537
#>           x2.pen3    x2.null1     x3.pen1       x3.pen2       x3.pen3
#> 1    1.2899222928 -1.15940334 -0.08065030 -0.7014752309 -1.176618e+00
#> 2    1.1579109177 -1.11592193 -0.19649265 -0.5431665358 -1.188389e+00
#> 3   -0.7421021461 -0.22577312 -0.42011402 -1.2318033245  6.862538e+00
#> 4   -0.9712815701  0.11273197 -0.05930416 -0.7289331502 -1.171710e+00
#> 5   -0.2603928314 -0.54534009 -0.32142930 -0.3414161657 -1.171532e+00
#> 6    1.9604580157 -1.36841996  0.42218231 -1.2101467383 -3.637378e-01
#> 7   -0.8124154167 -0.15746712  0.33392348 -0.7463156042  1.516868e+00
#> 8    2.1251857248  1.82579062  0.17834790  1.0622286209  2.999515e-01
#> 9   -0.3557403145  0.88007128 -0.46665566  0.1190752487 -1.026284e+00
#> 10  -0.1756536988 -0.58906515  0.18599664  1.0650046653  3.296998e-01
#> 11  -0.1332705141 -0.61012949 -0.42301587  0.3804810279 -8.732755e-01
#> 12  -0.8014345782 -0.16894113  0.12585316  1.0366008055  1.239304e-01
#> 13   3.6165787196 -1.83607112  0.25634459  1.0694330752  6.936421e-01
#> 14  -0.9717369614  0.11477439  0.21883812 -1.0582620688 -9.815910e-01
#> 15   6.9871025734 -2.73925187  0.01790255 -0.0487643863  4.182443e+00
#> 16   3.2088430348  2.14799003  0.30387664  0.9836968043  1.319364e+00
#> 17   0.3277332951 -0.81411434 -0.00962425 -0.7914623092 -1.156524e+00
#> 18   2.1654824459  1.83816959 -0.26806088 -0.4333861464 -1.183087e+00
#> 19  -0.9100345064 -0.03234759 -0.34453805 -1.0278245499  6.400728e+00
#> 20  -0.6619968917 -0.29255099 -0.38105302 -0.2205763771 -1.147330e+00
#> 21  -0.7104033270 -0.25332377 -0.35936900  0.5217979964 -7.623132e-01
#> 22  -0.8903553271 -0.06234748  0.08038369  1.0077229628  5.953658e-05
#> 23   2.0737499579 -1.40218570 -0.29177671  0.6313016026 -6.578284e-01
#> 24  -0.8660808040 -0.09521696 -0.44126992  0.3224785479 -9.125130e-01
#> 25  -0.5856448400  0.73743415 -0.46077368  0.2279658713 -9.697124e-01
#> 26   0.4643767080  1.25829286 -0.42053789  0.3873719342 -8.683887e-01
#> 27  -0.1944837153 -0.57954167  0.22090809 -1.0604568563 -9.788043e-01
#> 28  -0.0008865208 -0.67296477 -0.45253611 -0.0008851124 -1.078161e+00
#> 29   8.3480996791  3.61666011 -0.36294293  0.5151839428 -7.680539e-01
#> 30  -0.4908684323  0.79990427  0.16006451  0.3526033170  3.238101e+00
#> 31  -0.4529926376 -0.43590610 -0.46486433  0.0867528958 -1.041296e+00
#> 32  -0.1489138910 -0.60241299  0.08926752  1.0137418371  2.264883e-02
#> 33  -0.8929806898  0.45013072  0.25384980 -1.0945284223 -9.298320e-01
#> 34  -0.6748384625 -0.28244113  0.30478249  0.9289849610  1.558497e+00
#> 35  -0.1501633093  0.98846140 -0.42231376 -0.1139484843 -1.117813e+00
#> 36  -0.2786742140 -0.53559743  0.14636759 -0.9782260048 -1.061896e+00
#> 37   5.3259487846 -2.29463194 -0.02489015  0.9259213992 -2.260012e-01
#> 38  -0.5774266554 -0.35476600  0.30520304  0.9402206543  1.512690e+00
#> 39   0.3027289802 -0.80396578  0.24308455  0.6186021332  2.551323e+00
#> 40  -0.8640341855 -0.09782466 -0.11871457 -0.6513905503 -1.183189e+00
#> 41  -0.6450877791  0.69460848 -0.43062314 -0.0877823941 -1.109408e+00
#> 42   0.7180828908 -0.96329037  0.34784418 -1.1787029374 -7.196579e-01
#> 43   0.4455542114 -0.86086949  0.13267846  1.0404723519  1.445072e-01
#> 44  -0.9129675239 -0.02753954  0.30520136  0.9623431554  1.417982e+00
#> 45  -0.8136605401  0.54713157 -0.28714866 -0.4017341121 -1.179807e+00
#> 46  -0.2790183340 -0.53541291 -0.35929766 -1.0676605698  6.490917e+00
#> 47  -0.9803556011  0.21002025 -0.30892991  0.6058150325 -6.838662e-01
#> 48   1.1469867184 -1.11228148  0.30127569  0.8869201004  1.719475e+00
#> 49  -0.9448904044  0.03358118  0.21946505  1.0729844575  4.777072e-01
#> 50  -0.0805549326 -0.63565885 -0.15698324  0.7985869130 -4.517529e-01
#> 51  -0.5384548756 -0.38131812 -0.30873624 -0.3643363424 -1.174983e+00
#> 52  -0.8601636618  0.49408508  0.05135339 -0.8662742534 -1.129280e+00
#> 53   0.8241975868 -1.00136940 -0.46228424  0.2162084505 -9.762836e-01
#> 54   2.8200918496 -1.61689312  0.28766340  1.0418722163  9.820869e-01
#> 55   1.5668934865 -1.24788948  0.12468271 -0.9532699090 -1.080485e+00
#> 56   3.1691423654  2.13647910 -0.23271476  0.7107019826 -5.686365e-01
#> 57   1.8725326806 -1.34194915  0.19103360  1.0666640664  3.500123e-01
#> 58  -0.6960688572 -0.26526885  0.43226755 -1.2056996529 -2.683995e-01
#> 59  -0.8056905572 -0.16453566 -0.42084096 -0.1183542950 -1.119183e+00
#> 60  -0.8454178866  0.51186362  0.37077638 -1.1942106030 -6.412248e-01
#> 61  -0.9804433407  0.18375171  0.26100191 -1.1016897708 -9.179246e-01
#> 62   1.1490056980 -1.11295480 -0.17767198  0.7758473786 -4.842231e-01
#> 63  -0.8825086224  0.46493248  0.36493360 -1.1905585426 -6.628233e-01
#> 64   2.7341478614 -1.59273508 -0.25487954  0.6822363022 -6.021411e-01
#> 65   0.9007031255  1.42251915 -0.46113868  0.2252672109 -9.712309e-01
#> 66  -0.5495390744  0.76197486 -0.21833903 -0.5109158594 -1.187855e+00
#> 67  -0.9081360906 -0.03540742  0.14023095  0.2941362488  3.380426e+00
#> 68  -0.8398596189 -0.12708393 -0.31940036  0.5895881530 -6.998668e-01
#> 69  -0.0943169614  1.01581827 -0.46106712  0.0507062089 -1.057127e+00
#> 70  -0.6712972759  0.67458624  0.21278825 -0.4379696903  2.325321e+00
#> 71  -0.9778286650  0.24112537  0.43943569 -1.0842717067  4.380278e-01
#> 72   1.4073378171 -1.19733862  0.25711248  1.0691380019  6.990848e-01
#> 73  -0.4032960972  0.85284364 -0.46490126  0.1899996981 -9.905193e-01
#> 74   2.2748599450 -1.46126313 -0.39447786 -0.1889442641 -1.139365e+00
#> 75  -0.8390919579 -0.12797113 -0.03354520  0.9184282607 -2.421817e-01
#> 76  -0.1472163370  0.98992429  0.30387346  0.9837287531  1.319210e+00
#> 77  -0.9806359005  0.19570493 -0.40150630  0.4353154765 -8.329808e-01
#> 78  -0.7848375255 -0.18565459 -0.45500465  0.2632742332 -9.492712e-01
#> 79   4.2633239112  2.45035305  0.30119327  0.8861997489  1.722110e+00
#> 80   2.2464798440 -1.45298865  0.01465123  0.9587147335 -1.483310e-01
#> 81  -0.9764279943  0.14002768  0.23983260 -1.0802374169 -9.517881e-01
#> 82  -0.6162682523  0.71578337  0.14479444  0.3074822566  3.348135e+00
#> 83   0.9798516236 -1.05570379 -0.06053440  0.8943322158 -2.910862e-01
#> 84  -0.9748580417  0.13047588 -0.36197663  0.5169828368 -7.664984e-01
#> 85   1.6998857048 -1.28923305 -0.33580541 -0.3144786145 -1.167004e+00
#> 86   1.4637985665 -1.21535235  0.04324098  0.9809230398 -8.738954e-02
#> 87   3.3851963880 -1.77312093 -0.05431522 -0.7352937080 -1.170430e+00
#> 88   1.8265416114 -1.32800503 -0.20809287 -0.5261619338 -1.188218e+00
#> 89  -0.9351946706  0.01287758 -0.42952701 -0.0913755505 -1.110590e+00
#> 90   0.9624015892 -1.04969578 -0.46152102  0.2223498619 -9.728656e-01
#> 91  -0.9421901596  0.36407932  0.02152185  0.9641693932 -1.341042e-01
#> 92  -0.9633955872  0.08310380 -0.40483507  0.4274668642 -8.389504e-01
#> 93  -0.8787213997 -0.07857890  0.40520492 -1.2094854842 -4.802318e-01
#> 94  -0.8749600044 -0.08363037  0.29547892 -1.1347667024 -8.527233e-01
#> 95   4.7189035274 -2.13215186 -0.14639291  0.8098993976 -4.349203e-01
#> 96  -0.8761115032 -0.08209342  0.41552243 -1.2107622926 -4.141610e-01
#> 97   0.4993342497  1.27207828  0.21889437 -1.0583217870 -9.815157e-01
#> 98  -0.0151868738 -0.66637428  0.27761061  1.0553679971  8.695994e-01
#> 99  -0.2950176627 -0.52678665  0.03649416 -0.8482275471 -1.136909e+00
#> 100 -0.9373223345  0.01721757  0.29719114  1.0193280949  1.130449e+00
#>        x3.null1      x4.pen1      x4.pen2      x4.pen3     x4.null1
#> 1    0.47583086  0.411141962  1.231600245  0.372661566 -1.163607895
#> 2    0.37589543 -0.455986567  0.511329150 -1.314617372 -0.207412233
#> 3   -2.65634906 -0.543989528 -0.284088884 -1.271912397  0.195493006
#> 4    0.49447207  0.161964254 -0.165751328  3.141988359  1.852073521
#> 5    0.26000736 -0.374584111 -0.580118542 -1.119588221  0.358811444
#> 6    1.22359935  0.188444298 -1.078272282 -0.382332601  0.782156268
#> 7    2.01010071 -0.416690271  1.194193902  6.091006409  2.586722521
#> 8   -0.75323369 -0.493773129 -0.393113699 -1.226620494  0.253415994
#> 9    0.01156707  0.293753474  1.288491070 -0.406558540 -0.861179047
#> 10  -0.76497468 -0.006639357  1.073753826 -0.970818057 -0.558959787
#> 11  -0.13385902  0.203878437 -1.085905674 -0.346828244  0.797466342
#> 12  -0.68158696  0.395216210 -0.851107383  1.406475518  1.388762637
#> 13  -0.90163381  0.334735606 -1.118654237  0.081631992  0.966236146
#> 14   0.79030841  0.386855571 -1.089709241  0.419658981  1.084630743
#> 15  -1.93732487 -0.463920082  0.496642676 -1.318252768 -0.199687374
#> 16  -1.11452504 -0.552011569  0.295723558 -1.350287153 -0.096633947
#> 17   0.53899246 -0.172374957  0.905031727 -1.131328112 -0.435112531
#> 18   0.31173238 -0.082119693  1.000773730 -1.050208159 -0.501756311
#> 19  -2.53248378 -0.562910313 -0.231672872 -1.289839832  0.168271094
#> 20   0.19386661 -0.502718844  0.418651690 -1.334509394 -0.159170884
#> 21  -0.21811752 -0.148447219  0.931378617 -1.111083586 -0.452726785
#> 22  -0.62860430 -0.482420303 -0.414172720 -1.216537402  0.264851016
#> 23  -0.28821850  0.406739735 -1.042629133  0.692268983  1.173720164
#> 24  -0.10068552 -0.590380878 -0.125518481 -1.319089719  0.114050026
#> 25  -0.04787970 -0.277282614  0.779896521 -1.210797242 -0.356871424
#> 26  -0.13784656  0.314352397 -1.119367325 -0.008284257  0.932861245
#> 27   0.79313044  0.411108673 -1.009000000  0.844982929  1.221673373
#> 28   0.07603402  0.362080325 -0.718649726  1.792919584  1.497354044
#> 29  -0.21403929  0.109933547 -1.032906332 -0.537005039  0.712191668
#> 30  -1.68031493 -0.603032178  0.005546332 -1.342929786  0.048229472
#> 31   0.02901129 -0.389771373 -0.559476271 -1.133510940  0.346678888
#> 32  -0.63844422 -0.563753920  0.259051018 -1.352769180 -0.078197971
#> 33   0.84013839  0.346408755  0.824435044  1.689073041 -1.573742280
#> 34  -1.19069881 -0.500546021  0.423342810 -1.333672858 -0.161586322
#> 35   0.13650426 -0.085019924  0.997845549 -1.053028612 -0.499597216
#> 36   0.69933446 -0.252096285  0.811521005 -1.193017944 -0.375929635
#> 37  -0.52495098  0.332612931 -1.118899361  0.071587168  0.962552294
#> 38  -1.17628590  0.389482278 -1.086058026  0.445434115  1.093269607
#> 39  -1.48728116  0.256135779 -0.405888549  2.587052505  1.709582020
#> 40   0.44297428  0.151884871 -0.140974770  3.197693780  1.866196306
#> 41   0.12249989 -0.436643098 -0.490839978 -1.175818572  0.307320465
#> 42   1.00621585 -0.536881957 -0.301651285 -1.265364612  0.204695601
#> 43  -0.69016944 -0.054499003 -0.912500366 -0.779313619  0.588436011
#> 44  -1.14622481  0.344648043 -0.660532712  1.950114447  1.540384986
#> 45   0.29375855 -0.596156974  0.110695148 -1.352718141 -0.004152383
#> 46  -2.55667404 -0.464446569  0.495655003 -1.318490560 -0.199169056
#> 47  -0.27141694 -0.578187527 -0.179788873 -1.305285995  0.141637368
#> 48  -1.24073972  0.277601407  0.611348927  2.213544692 -1.720661916
#> 49  -0.82201145  0.411269058  1.231017624  0.375579773 -1.164620756
#> 50  -0.40939875 -0.081069174 -0.890241671 -0.812446983  0.569705255
#> 51   0.27276201 -0.574609504  0.220284887 -1.354310269 -0.058784505
#> 52   0.59725101  0.193019801 -1.080587406 -0.372015202  0.786631399
#> 53  -0.04139920  0.040680485  1.116396074 -0.914049539 -0.596375632
#> 54  -1.00259971  0.289490756 -1.116283391 -0.102196172  0.896972097
#> 55   0.67458036  0.295587896 -1.117368857 -0.080476995  0.905372495
#> 56  -0.34299045  0.234209321 -1.099289640 -0.270610013  0.829524195
#> 57  -0.77293576 -0.600574931  0.069146825 -1.349823048  0.016528598
#> 58   1.27443117  0.297020443  1.289584154 -0.395528970 -0.866074919
#> 59   0.13886405  0.382508126  0.956685039  1.332464705 -1.469751874
#> 60   1.05905095 -0.441567435 -0.483119628 -1.180217497  0.302978261
#> 61   0.85092267 -0.526947830 -0.324766020 -1.256319986  0.216876939
#> 62  -0.39148201  0.277460757  1.282272019 -0.458507345 -0.837772780
#> 63   1.04485691 -0.248398686 -0.730432018 -0.998002357  0.452662614
#> 64  -0.32288411  0.343565062  0.814900682  1.713646705 -1.580771753
#> 65  -0.04639071 -0.566333940  0.250341552 -1.353211908 -0.073830378
#> 66   0.35671011 -0.537484606 -0.300197950 -1.265916995  0.203932382
#> 67  -1.71952832  0.408041856 -1.035865482  0.724792317  1.184039983
#> 68  -0.26088646  0.194524019 -1.081339119 -0.368585866  0.788114068
#> 69   0.04839118 -0.601030772 -0.045146953 -1.335270313  0.073583294
#> 70   2.28558544 -0.318534890  0.725530349 -1.238242268 -0.325006233
#> 71   1.60022312 -0.492276755 -0.395946580 -1.225290471  0.254949026
#> 72  -0.90359214  0.380952706  1.285590538  0.015550385 -1.034052717
#> 73  -0.02701263  0.282070060  1.284153692 -0.444297560 -0.844233701
#> 74   0.17678182  0.019050676  1.097201859 -0.940773963 -0.579074177
#> 75  -0.51712113  0.277874311  1.282444552 -0.457247179 -0.838347588
#> 76  -1.11447519 -0.587857264  0.161710039 -1.354549014 -0.029554257
#> 77  -0.16589980  0.396881262 -0.859421401  1.380447484  1.381285144
#> 78  -0.06745012 -0.564873037  0.255306677 -1.352966390 -0.076319806
#> 79  -1.24155130 -0.407216632  0.594800496 -1.290281669 -0.252019635
#> 80  -0.56171659 -0.040439538  0.316088329  4.198034027  2.116540994
#> 81   0.81962888 -0.053572545  1.029099737 -1.021590591 -0.523133485
#> 82  -1.71065062 -0.448635676 -0.471839827 -1.186520504  0.296662123
#> 83  -0.49307014 -0.364011765  0.661229190 -1.266126266 -0.288548048
#> 84  -0.21514697  0.145444754  0.250069948  3.029802997 -1.939714263
#> 85   0.24511481  0.340514651  0.804771418  1.739615780 -1.588183059
#> 86  -0.58968416  0.380585415 -1.096971013  0.363927417  1.065781133
#> 87   0.49886219  0.285751768  1.285589526 -0.432679990 -0.849482754
#> 88   0.36574074 -0.524396806  0.369062349 -1.342282272 -0.133781046
#> 89   0.12442210  0.075697670 -1.010389850 -0.594346321  0.684684167
#> 90  -0.04478210 -0.344409531  0.689570329 -1.254411994 -0.304469453
#> 91  -0.56831170 -0.420985231 -0.514691938 -1.161781187  0.320838828
#> 92  -0.16126823 -0.341007244  0.694389182 -1.252331661 -0.307198927
#> 93   1.15798445  0.364610349 -1.109206854  0.246910699  1.025392558
#> 94   0.90644688  0.415591409  1.198776421  0.524304305 -1.215414862
#> 95  -0.41854084  0.389029374 -0.822433579  1.494344247  1.413842815
#> 96   1.19571700 -0.600386789 -0.053058999 -1.333898811  0.077550479
#> 97   0.79038490 -0.584106839  0.180146034 -1.354744048 -0.038744098
#> 98  -0.96387455  0.411517301 -1.000491778  0.880643518  1.232694526
#> 99   0.58260542  0.298513995  0.673082792  2.066203931 -1.679993826
#> 100 -1.05255389 -0.599026282  0.086164433 -1.351161644  0.008057060
#>          x5.pen1      x5.pen2    x5.pen3     x5.null1
#> 1   -0.275973719  0.621790015 -0.3513436  0.283326158
#> 2    0.337993924  0.785720823  1.8966682  1.179636247
#> 3   -0.291552650  0.602810115 -0.3815385  0.266900357
#> 4    0.322489600 -1.269369354 -1.0007933 -1.316521558
#> 5   -0.408175677  0.001033699 -0.9919799 -0.149999101
#> 6    0.335305098  0.837719780  1.6652400  1.102912320
#> 7    0.332663459  0.851956875  1.5937449  1.078814187
#> 8   -0.425696608  0.319718218 -0.7313811  0.054076899
#> 9   -0.311213124  0.577123317 -0.4206272  0.245264082
#> 10   0.363609544 -1.295228948 -0.8143861 -1.481538903
#> 11  -0.231716143 -0.443659135 -1.2227642 -0.425043849
#> 12  -0.350754914 -0.183209316 -1.1036876 -0.262880511
#> 13   0.377732032 -1.102592599  0.2471268 -2.127395114
#> 14   0.334982525 -0.934735382  0.7511333 -2.368003421
#> 15   0.306714293 -1.254709496 -1.0488490 -1.267069757
#> 16   0.395403299 -1.268330817 -0.4381430 -1.746230603
#> 17  -0.349250355  0.520081866 -0.5010349  0.199315397
#> 18  -0.287155436 -0.332319379 -1.1772324 -0.354793641
#> 19  -0.183805982  0.716464633 -0.1798480  0.372425762
#> 20   0.326926431  0.690174914  2.2535974  1.294641328
#> 21   0.347096172 -1.287791989 -0.9038377 -1.406650388
#> 22  -0.190857693 -0.519174615 -1.2489962 -0.474057216
#> 23  -0.163481826 -0.563525495  5.4640285  2.243634958
#> 24   0.116865140  0.907696434  0.4371008  0.651129576
#> 25  -0.277297730 -0.353064972 -1.1863242 -0.367742663
#> 26   0.392039474 -1.280385154 -0.5235311 -1.691302823
#> 27   0.208706284 -1.138724141 -1.2241181 -1.032660330
#> 28  -0.037368008 -0.775144448 -1.3059686 -0.655692334
#> 29  -0.337629782  0.538760169 -0.4756011  0.214071267
#> 30  -0.365801195 -0.142092504 -1.0808269 -0.237728806
#> 31   0.226878660 -1.162399093 -1.2023578 -1.069781553
#> 32   0.290130348  0.919606606  1.1260891  0.915666392
#> 33  -0.373370688 -0.120059262 -1.0681021 -0.224257116
#> 34  -0.007408442  0.845987018  0.1548012  0.530181490
#> 35   0.383660691 -1.292446419 -0.6472986 -1.607008516
#> 36  -0.322567312  0.561246460 -0.4438472  0.232201701
#> 37   0.319989103 -1.267166949 -1.0090108 -1.308328495
#> 38  -0.434514958  0.227313185 -0.8175914 -0.007295481
#> 39  -0.123982530 -0.634817907 -1.2813750 -0.552360359
#> 40  -0.040575087 -0.770075039 -1.3054100 -0.651735845
#> 41  -0.252260879 -0.403799712 -1.2073888 -0.399663594
#> 42  -0.419783394  0.349082204 -0.7017116  0.074167169
#> 43  -0.412745224  0.376375816 -0.6730260  0.093149031
#> 44  -0.253665073 -0.401021078 -1.2062783 -0.397904946
#> 45   0.202423434 -1.130378505 -1.2308477 -1.020294984
#> 46  -0.379085629 -0.102687023 -1.0578314 -0.213633291
#> 47   0.160256741 -0.385499125  2.1338154 -2.951365648
#> 48   0.077029642 -0.014770421  4.1691287  1.870201485
#> 49  -0.137368975 -0.612301887 -1.2758541 -0.536732886
#> 50  -0.369266576 -0.132132960 -1.0751164 -0.231639307
#> 51   0.332208434 -1.277416932 -0.9663163 -1.349864855
#> 52  -0.094371495 -0.683723860 -1.2919561 -0.587093662
#> 53   0.194033034 -1.119115807 -1.2392528 -1.004125451
#> 54   0.341228460 -1.284030384 -0.9300836 -1.383311895
#> 55  -0.179608062 -0.539235922 -1.2553091 -0.487323830
#> 56  -0.417560433  0.358374002 -0.6920699  0.080594273
#> 57  -0.221051011 -0.463814087 -1.2301404 -0.437996386
#> 58  -0.037620136 -0.774746167 -1.3059258 -0.655380745
#> 59   0.148444167  0.155360996  3.7542201  1.749312022
#> 60   0.193904595  0.930625476  0.6688609  0.744213614
#> 61  -0.432223200  0.149951783 -0.8825081 -0.057012080
#> 62  -0.051032931 -0.753492247 -1.3033924 -0.638932494
#> 63   0.331977480  0.854889365  1.5784133  1.073620397
#> 64   0.193622307  0.930574153  0.6678734  0.743827044
#> 65   0.263363153 -0.697909102  1.3757860 -2.641318834
#> 66   0.238632957 -1.177305320 -1.1862669 -1.094983101
#> 67   0.329354081  0.864418845  1.5269092  1.056102952
#> 68   0.258599581 -1.201769678 -1.1546950 -1.140295808
#> 69   0.378022514 -1.103931110  0.2427224 -2.125181510
#> 70   0.125422887  0.910970716  0.4597719  0.660458722
#> 71  -0.376154460 -0.111683305 -1.0631765 -0.219135312
#> 72  -0.409676242  0.007472768 -0.9876257 -0.146028427
#> 73   0.067248101 -0.936811125 -1.3062809 -0.795219909
#> 74   0.279150312  0.925435378  1.0545790  0.889740826
#> 75   0.335732376 -0.937401646  0.7436690 -2.364595946
#> 76  -0.171590916  0.727284643 -0.1574715  0.383581106
#> 77  -0.412135455  0.018401486 -0.9801612 -0.139281528
#> 78  -0.399671452  0.416124420 -0.6291790  0.121390216
#> 79   0.394115941 -1.274058047 -0.4763356 -1.721952483
#> 80   0.346678669 -1.287540355 -0.9057845 -1.404942975
#> 81   0.320261222  0.657235582  2.3644879  1.329665759
#> 82   0.188902071 -1.112165384 -1.2440811 -0.994421449
#> 83   0.179998946  0.233580341  3.5580061  1.691657571
#> 84  -0.255630924 -0.397118546 -1.2047102 -0.395437141
#> 85  -0.193711835 -0.514039209 -1.2473351 -0.470679321
#> 86  -0.336674162  0.540240929 -0.4735491  0.215252539
#> 87  -0.401438452  0.411300325 -0.6346381  0.117922367
#> 88   0.243396666 -1.183245339 -1.1792453 -1.105492166
#> 89   0.007799224 -0.845859532 -1.3106354 -0.713257105
#> 90   0.375993854 -1.094687162  0.2729675 -2.140337028
#> 91  -0.266703814  0.632584621 -0.3336317  0.292850164
#> 92   0.330781258  0.859521987  1.5537129  1.065232883
#> 93   0.331474840 -1.276840352 -0.9690714 -1.347257808
#> 94  -0.060474016 -0.738446655 -1.3013182 -0.627489671
#> 95   0.335139827 -1.279665644 -0.9550333 -1.360444028
#> 96  -0.331146758 -0.232694215 -1.1297029 -0.293218661
#> 97  -0.371378326 -0.125959862 -1.0715427 -0.227864966
#> 98   0.062299035  0.883641021  0.3044347  0.595447731
#> 99  -0.186270299 -0.527389312 -1.2516152 -0.479475720
#> 100  0.391861025 -1.280823891 -0.5270673 -1.688975841