Title

Title

make_predict_dat(Smooth, dat)

Arguments

Smooth
The smooth object from construct_smooth_data

dat
The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)

#>          x1.base1      x1.base2    x1.base3    x1.base4     x2.base1
#> 1    0.3172625383 -0.7403854043 -1.79016226  1.61523248 -0.137796113
#> 2    0.0528571505 -1.0702195396  1.05377955  0.56548039 -0.068953811
#> 3    0.0541943078 -0.0409679279 -3.43700090  2.05107800 -0.396906463
#> 4    0.2616079363 -1.2376531108  0.67219029  0.77796333 -0.039897257
#> 5   -0.5521731425  0.2256060573  1.37507951 -0.14115273  0.464680350
#> 6    0.3055338679 -1.2598494914  0.53512566  0.84144553 -0.349091086
#> 7    0.1163992054 -0.1973742529 -3.08669766  1.96095040 -0.162100909
#> 8   -0.2338813521  0.7709699273  1.06158161 -0.43854784  0.004995578
#> 9   -0.6059025828  1.5865124645 -7.01381362  2.96102487  0.474185236
#> 10   0.4034101777 -1.0511606031 -0.89481575  1.35476381  0.076209479
#> 11  -0.1113621416 -0.4356805894 -4.50709439 -2.15909140  0.198028118
#> 12   0.2107038215 -1.2039666567  0.79633727  0.71586836  0.085812311
#> 13   0.0002481843  0.9910076734  0.77887337 -0.60319146  0.031364990
#> 14   0.4168139705 -1.1687422694 -0.45393977  1.21451304 -0.344907122
#> 15   0.3594461877 -0.8749974913 -1.42825350  1.51299126  0.087532442
#> 16  -0.5086911205 -0.2913580067  1.43287884  0.10816944 -0.248650138
#> 17  -0.3348493494 -0.6180898211  1.36727111  0.27354812  0.152763552
#> 18   0.4050364346 -1.2384858179 -0.08744995  1.08845191 -0.608915916
#> 19  -0.5753963699  0.0698648208  1.41275008 -0.06542432 -0.297180784
#> 20   0.1564645034  1.1011405280  0.51122549 -0.72965053  0.307256823
#> 21   0.3834802406  1.1173201510 -0.45381704 -1.08571191  0.233742727
#> 22  -0.4152735177  0.5339935119  1.24030186 -0.29964680 -0.130193753
#> 23   0.2180082272  0.3745843887 -2.71554436 -1.71690607  0.187942281
#> 24  -0.5754806457  0.0677070838  1.41314474 -0.06438535  0.340890203
#> 25   0.2208384531 -1.2111920293  0.77387986  0.72748389  0.392643498
#> 26   0.0616070311 -1.0784122059  1.04285414  0.57271652  0.307310154
#> 27   0.1591260955 -1.1641053236  0.89717013  0.66120460  0.068934164
#> 28   0.1670639917  1.1073497741  0.48829111 -0.73962921  0.380974825
#> 29   0.3796584997 -0.9477644678 -1.21882998  1.45209277 -0.089007889
#> 30   0.3979727397  1.0105973045 -0.95254751 -1.23975390 -0.297699761
#> 31   0.2277247482 -1.2159679085  0.75803835  0.73556891 -0.161456890
#> 32   0.1043771573  1.0679518125  0.61307978 -0.68387223  0.153572143
#> 33  -0.2908847388  0.7052297039  1.12006733 -0.39758241  0.153286719
#> 34   0.4166853378 -1.1592806648 -0.49447158  1.22786314 -0.385206503
#> 35   0.2928772209 -0.6683773053 -1.97407296  1.66596933  0.194652030
#> 36   0.4166198523 -1.1799498275 -0.40408475  1.19794647  0.439695356
#> 37  -0.2180821091 -0.7768099167  1.29848583  0.36224433 -0.381359364
#> 38  -0.1823864690 -0.6057476024 -4.87453140 -2.24876918 -0.295911208
#> 39  -0.0662236513  0.9353964117  0.86843081 -0.55554982  0.038844299
#> 40  -0.5732626801  0.1030170671  1.40625714 -0.08141499 -0.325649994
#> 41  -0.0768354699 -0.9399622576  1.18930075  0.46542945 -0.395938011
#> 42   0.3839806693  0.9042962424 -1.32024382 -1.34613156  0.184011949
#> 43   0.0106186270 -1.0296236884  1.10285678  0.53164617 -0.395380429
#> 44   0.3984170537  1.0426283554 -0.82462728 -1.20145628 -0.234227854
#> 45  -0.1007641130 -0.9140371116  1.21012297  0.44783523  0.322071898
#> 46  -0.4595109997  0.4580889798  1.28193976 -0.25904734  0.328046477
#> 47  -0.0555650967  0.9446505019  0.85470058 -0.56307507 -0.127136678
#> 48   0.3983061777  1.0184530798 -0.92219354 -1.23073278 -0.201854617
#> 49   0.3752115317 -1.2676965566  0.19851283  0.98155996 -0.014623336
#> 50  -0.4468395601  0.4810457158  1.26999971 -0.27118558  0.407261859
#> 51   0.3433615861 -1.2704390774  0.37995206  0.90839538 -0.315483562
#> 52   0.4003338229 -1.0356233381 -0.94612759  1.37046026 -0.382970316
#> 53   0.2194942339  1.1345848858  0.36073413 -0.79322794  0.152445653
#> 54  -0.2841001456  0.7133844194  1.11324632 -0.40253870  0.035178986
#> 55   0.3984315706  1.0420896040 -0.82687869 -1.20213683 -0.198330810
#> 56   0.3951480202 -1.2539169617  0.03239279  1.04473425  0.101849153
#> 57  -0.0864041933  0.9175146647  0.89385188 -0.54138395 -0.181801707
#> 58   0.3955139812  1.0763046852 -0.67483529 -1.15561281 -0.266319001
#> 59  -0.4172270483 -0.4844675092  1.40486492  0.20396900  0.151549023
#> 60   0.1808997077 -0.3631665318 -2.70772390  1.86229077 -0.124313220
#> 61   0.0831174599 -1.0982321311  1.01470224  0.59092004 -0.300781420
#> 62   0.3401957474  1.1561068855 -0.09995795 -0.96679774 -0.258688379
#> 63  -0.5199694441 -0.2607213256  1.43474596  0.09329622 -0.327823358
#> 64  -0.0058447938 -0.1827560237 -3.96013884 -2.02553938 -0.254316268
#> 65   0.3524862490 -0.8515379941 -1.49340907  1.53166035 -0.175824054
#> 66  -0.1644179276 -0.8418276668  1.26068717  0.40144202  0.243226513
#> 67  -0.5748571489  0.0002094386  1.42379206 -0.03196911 -0.390125402
#> 68   0.2029197860  0.3348936396 -2.80805100 -1.74029747 -0.313325247
#> 69  -0.2860341124  0.7110696727  1.11519422 -0.40112852 -0.316573969
#> 70   0.1645201898 -1.1684870011  0.88755229  0.66661474  0.102350212
#> 71  -0.1138494277 -0.8995888033  1.22106940  0.43826499 -0.321271717
#> 72  -0.5604573315  0.1896221474  1.38542575 -0.12349809  0.285388232
#> 73  -0.4880750822  0.4015531346  1.30908801 -0.22960698 -0.393549419
#> 74   0.1638157650 -1.1679173406  0.88881941  0.66590455 -0.068227225
#> 75  -0.5752704121  0.0728140534  1.41220515 -0.06684473 -0.126479356
#> 76   0.3985508726  1.0312916376 -0.87122347 -1.21549273 -0.382184017
#> 77   0.1377205815 -1.1462962465  0.93350217  0.64034209  0.084470060
#> 78   0.3039057795  1.1595162598  0.07863308 -0.90257249 -0.198863733
#> 79   0.3299223139 -1.2679099585  0.44045309  0.88283340  0.348949093
#> 80   0.4165654406 -1.1813653620 -0.39762889  1.19578911 -0.082170616
#> 81   0.3487592275 -0.8392260572 -1.52720825  1.54129667  0.238585921
#> 82   0.0150433278 -1.0339575377  1.09797503  0.53511676 -0.314130293
#> 83   0.3605557244  0.8028066171 -1.62559570 -1.43103181  0.349622537
#> 84   0.3289980488 -1.2676793840  0.44436723  0.88115701  0.381263855
#> 85   0.2099226696  1.1301038602  0.38601234 -0.78284477  0.387638391
#> 86  -0.2734291584  0.7260198068  1.10244599 -0.41028410 -0.285090354
#> 87   0.1903734410  1.1202309202  0.43471737 -0.76251622  0.129017438
#> 88   0.3966926864  0.9919366056 -1.02243509 -1.26037500 -0.071391277
#> 89   0.0403662503  1.0221619752  0.71947649 -0.63309832  0.469183652
#> 90   0.3372684499  0.7201948331 -1.85499260 -1.49315537  0.365499902
#> 91   0.1547513494  1.1001178801  0.51485492 -0.72806096 -0.272077411
#> 92  -0.6432594717 -1.7093032127 -7.25881024 -2.83068340 -0.146294757
#> 93  -0.2366276078 -0.7533294211  1.31061052  0.34856660  0.143014999
#> 94   0.0246350338 -1.0432869626  1.08719525  0.54269429  0.243118226
#> 95  -0.0605555965  0.9403338298  0.86115645 -0.55954743 -0.292900043
#> 96   0.3735084189  1.1333242026 -0.34362166 -1.04970304 -0.275255963
#> 97   0.3884102279  0.9286865675 -1.24141249 -1.32374169 -0.162777106
#> 98   0.3752662641 -0.9312142924 -1.26755164  1.46638877 -0.070636568
#> 99   0.1087356138  1.0708813131  0.60517141 -0.68751861  0.353914996
#> 100  0.1730712638  1.1107750236  0.47491232 -0.74539900  0.161548323
#>        x2.base2    x2.base3    x2.base4     x3.base1     x3.base2     x3.base3
#> 1    0.53868098  4.21354206  1.76152119  0.072759655  1.015046390 -0.188106111
#> 2    0.35073781  3.82965125  1.64808196  0.192356077 -1.244226249 -0.980924295
#> 3   -0.20956913 -0.80513733 -0.03669531  0.058035983 -0.128707341  4.008196962
#> 4    0.27126571  3.66712607  1.60004117  0.421544537  0.736356723  1.917089356
#> 5    1.03040329 -0.91531270 -1.45474726 -0.685466920 -1.598755679  7.177789353
#> 6   -0.42518045 -0.51135429  0.13413463 -0.093248615  0.933353017 -0.450616780
#> 7   -0.64660828 -0.02051080  0.36907098  0.126224397  0.009928139  3.703394789
#> 8    0.55556604 -1.34169713 -0.58111261 -0.417333383  0.659727571 -0.905858106
#> 9    0.98378828 -0.54889022 -1.70929798  0.407372261  1.023604517  0.811006792
#> 10  -0.05773092  2.97898011  1.39547669 -0.559695688  0.100106111 -1.315297488
#> 11  -0.55106904  1.75539353  1.01761307 -0.348191925 -0.504747991 -1.444984572
#> 12   0.65502910 -1.36770593 -0.66802223  0.430757468  0.777413515  1.793301948
#> 13   0.58830213 -1.35173112 -0.60851758  0.247438382 -1.282211642 -0.804956203
#> 14   0.01063294 -1.02124836 -0.19151600 -0.257491343  0.819971599 -0.682025587
#> 15   0.65711917 -1.36809677 -0.66997734  0.330575468 -1.070544641  0.518332821
#> 16   0.20072509 -1.16394478 -0.32103307 -0.566341306  0.275366510 -1.226267054
#> 17  -0.72847687  0.98882234  0.76058454  0.331840129 -1.248631495 -0.174429305
#> 18   1.82485331  6.84065123  2.53782748 -0.217959662  0.850790728 -0.627527620
#> 19  -0.51059473 -0.35747810  0.21290730  0.398513818  0.653649746  2.152803015
#> 20   0.90988066 -1.32792592 -0.97929861 -0.502487987 -0.147281305 -1.399595045
#> 21   0.82936047 -1.36701337 -0.85884686 -0.532051792  0.454173567 -1.103740388
#> 22  -0.66923160  0.05716316  0.40263089 -0.506008377 -0.136279958 -1.396780361
#> 23   0.77675408 -1.37589674 -0.79403314  0.439840536  0.831275310  1.621739452
#> 24   0.94436105 -1.29509839 -1.04428947  0.332807128 -1.244933725 -0.154178541
#> 25   0.99249618 -1.21339783 -1.16491851  0.436517335  0.808762872  1.694891407
#> 26   0.90993676 -1.32788291 -0.97939568 -0.445605972 -0.299926496 -1.430008057
#> 27  -0.74870925  0.61183215  0.62381079 -0.324215746 -0.549113233 -1.444065735
#> 28   0.72885194  0.37010514 -2.19709039 -0.103058742 -0.897575060 -1.370255708
#> 29   0.40548755  3.94148389  1.68112846 -0.462081601  0.597173636 -0.974298794
#> 30   0.11288064 -1.10233506 -0.26113718  0.128329031 -0.251239002  2.545313142
#> 31  -0.64709313 -0.01894984  0.36975316  0.444278730  0.908135225  1.350904746
#> 32   0.73638212 -1.37689525 -0.74914306 -0.215008582 -0.732495166 -1.422019281
#> 33   0.73604407 -1.37688591 -0.74878193 -0.557915953  0.350388071 -1.179295187
#> 34  -0.32527149 -0.66080452  0.05131188 -0.106388229 -0.892931556 -1.372253811
#> 35  -0.64759863  1.41621259  0.90666703 -0.317586820 -0.561071364 -1.443545368
#> 36   1.02565612 -1.07193219 -1.31902072  0.147712026  1.041696175 -0.050084553
#> 37  -0.10123543 -0.91945574 -0.11421580 -0.225501134 -0.715973409 -1.425342527
#> 38   0.11632961 -1.10488970 -0.26348223 -0.040857476 -0.981638446 -1.326875154
#> 39  -0.74445513  0.51261950  0.58623831  0.433591599  0.975794054  1.067942850
#> 40   0.05547437 -1.05811370 -0.22209900 -0.381675525  0.702166753 -0.854261355
#> 41  -0.19159507 -0.82534219 -0.04981807 -0.014403015  0.975917446 -0.331348689
#> 42  -0.68521146  1.24986387  0.85075442 -0.447246303  0.619353773 -0.950980852
#> 43  -0.18468107 -0.83297397 -0.05483418  0.402239919  0.665889581  2.118889709
#> 44  -0.58384440 -0.19600332  0.29012637  0.136464517 -1.191885320 -1.105226505
#> 45   0.92530162 -1.31494097 -1.00695072  0.115593148 -1.170026365 -1.142547652
#> 46   0.93142133 -1.30907581 -1.01851790  0.004392999 -0.235988446  4.241388694
#> 47   0.38332424 -1.26993238 -0.44909044 -0.558364205  0.090084862 -1.319603216
#> 48   0.27532401 -1.21079700 -0.37251051  0.235072000 -1.275362251 -0.851062567
#> 49   0.20169794  3.52426404  1.55776799  0.348726533  1.059945283  0.497524472
#> 50   0.78930131  0.18347269 -2.10749619  0.405144551  0.675669663  2.091569914
#> 51   0.07720712 -1.07523755 -0.23688441  0.389553472  1.040085348  0.695634152
#> 52  -0.33425319 -0.64843098  0.05845272  0.228084193  0.225457610  3.216868602
#> 53   0.73504769 -1.37685679 -0.74771885  0.189260646  1.053230750  0.035392817
#> 54   0.59301094 -1.35306375 -0.61254631  0.333270107 -1.243006480 -0.143866249
#> 55   0.28068288 -1.21397366 -0.37624351  0.203635525  1.056585936  0.067001063
#> 56   0.67446820 -1.37105398 -0.68644539 -0.515819696  0.496545362 -1.068794267
#> 57   0.30542076 -1.22831102 -0.39355289  0.311778565 -0.957257677  0.852755623
#> 58  -0.54929857 -0.27648945  0.25225229  0.182161556 -1.235487536 -1.006774526
#> 59  -0.72941710  0.98118635  0.75789673  0.284105415  1.067280791  0.273121596
#> 60   0.50187225  4.13835729  1.73930422  0.011553264 -0.221732801  4.210499939
#> 61  -0.50563516 -0.36727034  0.20805887  0.173530644  1.049169381  0.002086406
#> 62  -0.55797439 -0.25708820  0.26148836  0.410346147  1.020053462  0.833223776
#> 63   0.05067619 -1.05426878 -0.21883229 -0.160914131 -0.814665344 -1.400833839
#> 64  -0.56280796 -0.24605576  0.26670952 -0.081709829 -0.926996182 -1.356699535
#> 65   0.31421569 -1.23327877 -0.39973981 -0.087094614 -0.919632459 -1.360242756
#> 66   0.84005236 -1.36389898 -0.87311080  0.225228080  1.060918274  0.116934873
#> 67  -0.14384849 -0.87654203 -0.08413655  0.340209722  0.488207154  2.586964557
#> 68   0.08168376 -1.07870626 -0.23992841  0.321879504  1.065484284  0.395077281
#> 69   0.07492825 -1.07346414 -0.23533467 -0.032727108 -0.992259278 -1.320269707
#> 70  -0.74823545  0.73706964  0.67023558 -0.562904997  0.314557828 -1.202462626
#> 71  -0.47514236 -0.42495299  0.17906667  0.043371260 -0.158157271  4.072396777
#> 72   0.52401917  0.95384095 -2.46011415 -0.123416806 -0.868944490 -1.382003505
#> 73  -0.27992607 -0.72053826  0.01600321 -0.541435772  0.424151900 -1.126983476
#> 74   0.46225947 -1.30647990 -0.50757662 -0.549673587  0.036297745 -1.341374286
#> 75   0.38423021 -1.27038454 -0.44974737  0.175486090 -1.229546483 -1.022840939
#> 76  -0.10468234 -0.91607582 -0.11179907  0.336670757 -1.223736086 -0.048010908
#> 77  -0.08265025  2.92490077  1.37925634 -0.475112945  0.576170700 -0.995492480
#> 78   0.27987444 -1.21349604 -0.37567999  0.298207243  1.067278069  0.316178686
#> 79   0.95231431 -1.28527713 -1.06115509  0.152606061  1.043185835 -0.040412049
#> 80   0.38682157  3.90335715  1.66986208 -0.077355039  0.942491890 -0.427172737
#> 81   0.83483016 -1.36548335 -0.86609122  0.356626172  0.531342763  2.477474848
#> 82  -0.48624843 -0.40442410  0.18947158  0.438624567  0.822437745  1.650726706
#> 83   0.95297265 -1.28441665 -1.06259077  0.371059027  1.051333888  0.597558561
#> 84   0.98260557 -1.23575682 -1.13548819 -0.490281855  0.549465686 -1.021270342
#> 85   0.98820630 -1.22361417 -1.15172143  0.337109350 -1.220097614 -0.031114155
#> 86   0.13673425 -1.11977258 -0.27736167 -0.535945299  0.442355665 -1.113033800
#> 87  -0.22473706  2.60718569  1.28329710  0.374639104  1.049501148  0.615268209
#> 88  -0.70392208  0.20412381  0.46414372  0.444177879  0.910956170  1.340159359
#> 89   1.02778904 -0.86509513 -1.49375184  0.230661489  1.061861746  0.130030047
#> 90   0.96818383 -1.26216432 -1.09773852 -0.253795360 -0.670371088 -1.433046436
#> 91  -0.54253772 -0.29126858  0.24516914  0.305999947  0.403341792  2.796539258
#> 92  -0.65816666  0.01786510  0.38574876  0.076379472 -1.126413405 -1.202665550
#> 93   0.72385018 -1.37637002 -0.73590477  0.444264458  0.908577097  1.349226251
#> 94   0.43733228  1.18730512 -2.56011073 -0.286492749 -0.615574256 -1.439642946
#> 95  -0.51635665 -0.34594591  0.21859108  0.369083605  0.565515964  2.389060933
#> 96   0.15465913 -1.13252483 -0.28956730 -0.530119223  0.459757097 -1.099282813
#> 97  -0.64609792 -0.02214958  0.36835443 -0.498436510 -0.159637906 -1.402656178
#> 98  -0.70431050  0.20606009  0.46493824 -0.128758406  0.911869146 -0.502116700
#> 99   0.95714490 -1.27878027 -1.07184199 -0.188090142 -0.773989889 -1.412333793
#> 100 -0.72054523  1.04788255  0.78127163  0.104839871 -1.158374152 -1.160242213
#>         x3.base4     x4.base1    x4.base2    x4.base3     x4.base4     x5.base1
#> 1   -0.580376073 -0.415926163  0.20488955 -0.83096591  0.019406160 -0.452375745
#> 2    1.107515311  0.429644964 -0.77384978  0.40202839  2.039956105 -0.310593006
#> 3   -1.876198671  0.483659259 -1.03411958 -0.87520923  1.318615667  0.280770024
#> 4   -1.300375479  0.453392636 -1.02173330 -1.00812724  1.205309001 -0.428691046
#> 5   -2.711448220  0.061454439  0.81889690  0.54731830 -0.643080139  0.326034984
#> 6   -0.461375459 -0.222557090  0.63524527 -0.21096514 -0.325698417  0.386947551
#> 7   -1.795231641 -0.363165714  0.42899681 -0.56717539 -0.143465547 -0.435186556
#> 8   -0.212599614 -0.134688066 -0.40614069 -1.24278509  0.426985869  0.177223699
#> 9   -0.955524981 -0.414145960  0.15386859 -0.88010660  0.054102170  0.355086063
#> 10   0.137140538 -0.204996329 -0.31255899 -1.20190200  0.362124409  0.146398069
#> 11   0.475113663 -0.375981284  0.39886064 -0.60813773 -0.120185453 -0.283015601
#> 12  -1.263888010 -0.303009905 -0.16061784 -1.11922840  0.260941321 -0.386085923
#> 13   1.220886337 -0.225443721  0.63216317 -0.21771139 -0.322514997 -0.332257070
#> 14  -0.343773219 -0.414998427  0.22840875 -0.80713555  0.003193359  0.191585716
#> 15   1.829572633  0.130529506 -0.71407671 -1.30985254  0.677291637  0.387582152
#> 16   0.036732835  0.301235316 -0.89312469 -1.25449434  0.892941039 -0.427846850
#> 17   1.543906003 -0.231724447  0.62534887 -0.23240643 -0.315550946  0.297844400
#> 18  -0.372817487 -0.404126396  0.08805864 -0.93870462  0.098076916 -0.477872620
#> 19  -1.368783780 -0.389935295  0.03402993 -0.98313872  0.133711449  0.265727377
#> 20   0.274319927  0.042728252 -0.61637048 -1.30262581  0.588144662  0.226910047
#> 21  -0.072104065  0.024072426 -0.59527142 -1.29900129  0.570461499 -0.242047661
#> 22   0.268254324 -0.311014558 -1.09675169  5.54114801 -2.173156350  0.048932480
#> 23  -1.212599471  0.077322740  0.82105720  0.60782303 -0.665844933  0.332451428
#> 24   1.552996086 -0.185326088  0.67241045 -0.12405067 -0.365980702  0.382270940
#> 25  -1.234575446 -0.037596128  0.78106221  0.24117552 -0.522991626  0.335487421
#> 26   0.358814071 -0.334759336  0.48447482 -0.48593232 -0.187973935  0.381472476
#> 27   0.501119314 -0.345288804  0.46531314 -0.51491711 -0.172332969  0.168804829
#> 28   0.727062251 -0.341463169 -0.08755613 -1.07229024  0.213169168  0.097513242
#> 29  -0.167423693  0.405428852 -0.98786360 -1.12820756  1.080974421  0.114000197
#> 30   2.519171360 -0.407285977  0.10378956 -0.92516138  0.087631614 -0.472253665
#> 31  -1.129702254 -0.338950523 -0.09273267 -1.07577415  0.216546881 -0.406760951
#> 32   0.613829444 -0.153680779  0.70060496 -0.04963522 -0.399470577 -0.400324261
#> 33  -0.007896224  0.335222264 -0.57911393  0.97993835  2.291771805 -0.389802603
#> 34   0.723643576  0.053253872  0.03668345  3.20214470 -1.495838315  0.389753323
#> 35   0.508196636  0.170983872 -0.75806509 -1.30663466  0.722358931  0.273873531
#> 36  -0.638416147  0.490874753 -1.03454577 -0.82745524  1.355425588  0.003163096
#> 37   0.603235721  0.476453555 -0.88382591  0.03008342  1.863009586  0.354004292
#> 38   0.792369193 -0.021093175  0.78961345  0.28668826 -0.541416706 -0.021938570
#> 39  -1.040103928  0.177660885  0.74850910  1.28663670 -0.905114389 -0.476441914
#> 40  -0.244849117  0.508473611 -0.99330821 -0.44215000  1.609146998  0.193216563
#> 41  -0.517018075  0.044859307 -0.61877284 -1.30299759  0.590188641  0.086358090
#> 42  -0.183152908  0.503252279 -1.02864717 -0.70338075  1.444278040  0.346323596
#> 43  -1.359022010 -0.062528177 -0.32897951  3.96417138 -1.717102179  0.384153649
#> 44   1.014922566  0.118406711  0.81743730  0.79606453 -0.734944412  0.386115619
#> 45   0.984109289  0.071890068  0.82049123  0.58652300 -0.657863641  0.351026414
#> 46  -1.937855564  0.121150473  0.81653781  0.81089792 -0.740287136  0.155804616
#> 47   0.142766477 -0.334072984  0.48567700 -0.48407894 -0.188965561  0.387827245
#> 48   1.192703168  0.121396803  0.81645177  0.81224781 -0.740772646 -0.350348629
#> 49  -0.847132264 -0.365351982 -0.03412680 -1.03487634  0.178312074 -0.144736709
#> 50  -1.351139222  0.200949006 -0.79012893 -1.30095098  0.757753446  0.366426318
#> 51  -0.916367019  0.375303731 -0.96249912 -1.17761009  1.018748510  0.054060097
#> 52  -1.664629708  0.032310025 -0.60460397 -1.30068269  0.578225367  0.362434311
#> 53  -0.673081053  0.441690126 -1.01453163 -1.04417876  1.170854559 -0.419093740
#> 54   1.557603338 -0.314555565 -0.13988723 -1.10639397  0.247357284 -0.353255837
#> 55  -0.685674101  0.080066761 -0.65825325 -1.30781048  0.624767864  0.230069874
#> 56  -0.099383649 -0.411413668  0.26483837 -0.76865483 -0.022237531 -0.194634643
#> 57   1.954079153  0.324489607 -0.91571787 -1.23552318  0.929288075 -0.334448279
#> 58   1.089311922 -0.411437569  0.12951944 -0.90239048  0.070461939 -0.347873221
#> 59  -0.765124175  0.137085097  0.80888105  0.90541037 -0.774008137  0.286066593
#> 60  -1.929699110  0.177165970  0.75015679  1.27885293 -0.902506621  0.030647241
#> 61  -0.659682140  0.010664539 -0.10071432  3.49239440 -1.580429054 -0.317209895
#> 62  -0.962976291  0.178287051  0.74633874  1.29677945 -0.908508590 -0.264305903
#> 63   0.668287369  0.186003272  0.70407796  1.47512287 -0.967507412  0.251242964
#> 64   0.749138656 -0.112401689 -0.43417918 -1.25347402  0.447009379  0.385719925
#> 65   0.743542807  0.472265389 -1.03091666 -0.93452938  1.270300249  0.310987177
#> 66  -0.705334189  0.075408892  0.11082103  3.04199856 -1.448886645 -0.462700093
#> 67  -1.491686796 -0.409970211  0.27487583 -0.75770134 -0.029319324  0.043908003
#> 68  -0.810210737  0.155085696  0.79308125  1.03676520 -0.820014985 -0.277134011
#> 69   0.801172394  0.406007596 -0.98832744 -1.12710939  1.082260776 -0.471525488
#> 70   0.013576116  0.354656503 -0.61774515  0.87020987  2.245564013 -0.281483554
#> 71  -1.893193602  0.059357273  0.05687859  3.15881324 -1.483156941  0.251339339
#> 72   0.706243100 -0.341872080 -0.08670704 -1.07171644  0.212615182  0.370714676
#> 73  -0.053182282  0.002773271 -0.57099978 -1.29407673  0.550662470 -0.425144417
#> 74   0.172817875  0.485867658 -1.03441901 -0.86166807  1.329226526  0.373177709
#> 75   1.077748190 -0.412562963  0.13837615 -0.89436933  0.064524650  0.388901280
#> 76   1.599763876 -0.389118786  0.36229940 -0.65521561 -0.092646063 -0.228612464
#> 77  -0.152794419 -0.223286395  0.63446949 -0.21266919 -0.324895113  0.200065022
#> 78  -0.781191542 -0.066398903 -0.34094137  3.98874353 -1.724208565  0.217451545
#> 79  -0.642385201  0.498765805 -1.03237441 -0.75845924  1.405909275  0.252099628
#> 80  -0.472544443  0.097021133 -0.67711550 -1.30917567  0.642000560 -0.413127169
#> 81  -1.461026444 -0.037726926  0.78099140  0.24082068 -0.522847112  0.026966071
#> 82  -1.221327368 -0.402306274  0.31472741 -0.71262204 -0.057794892 -0.444467623
#> 83  -0.882426838 -0.220718240  0.63719263 -0.20666963 -0.327720952 -0.299938337
#> 84  -0.134537827 -0.155775702  0.69882719 -0.05459394 -0.397265712  0.355003724
#> 85   1.607077180 -0.287798125 -0.18672319 -1.13485415  0.278098029 -0.085460298
#> 86  -0.064617924  0.508277565 -0.99165124 -0.43298474  1.614527326  0.257575596
#> 87  -0.888602380  0.108287293  0.22644719  2.78506529 -1.373006339  0.022506482
#> 88  -1.126359280  0.016012031 -0.58611168 -1.29723457  0.562925324  0.281452380
#> 89  -0.710444112  0.117567388  0.26085451  2.70645359 -1.349628359 -0.472117913
#> 90   0.574511688 -0.137783511 -0.40219326 -1.24122087  0.424191864  0.322254093
#> 91  -1.549830255  0.238172176 -0.82922991 -1.28925352  0.804548907 -0.094729231
#> 92   0.930354379  0.181232834  0.56317015  1.93926541 -1.115902886  0.329310757
#> 93  -1.129180350  0.011352621  0.80406578  0.38129580 -0.579043830  0.078412353
#> 94   0.540869452 -0.413522135  0.24671325 -0.78804531 -0.009533501 -0.154033735
#> 95  -1.436112079 -0.042220247  0.77853044  0.22868331 -0.517895650  0.061089418
#> 96  -0.075658745 -0.278057409  0.56987292 -0.34230438 -0.262088748  0.245061694
#> 97   0.281133896 -0.374222329 -0.01166078 -1.01833253  0.163638720 -0.236616319
#> 98  -0.436406753  0.331733039 -0.92262255 -1.22876928  0.941124808 -0.257193391
#> 99   0.640933706 -0.355958623  0.44430375 -0.54555533 -0.155516770 -0.014733994
#> 100  0.968868361  0.166462497 -0.75318664 -1.30723532  0.717175015 -0.461775495
#>        x5.base2     x5.base3     x5.base4
#> 1    0.31463046 -0.906306289 -0.105972384
#> 2    0.58635194 -0.648787100 -0.270131541
#> 3   -1.34758976 -0.916590014  1.032140939
#> 4   -0.17947638 -1.183576678  0.153338908
#> 5   -1.38259339 -0.754275500  1.146568081
#> 6   -1.24688241  0.309333628  1.683138744
#> 7    0.36640340 -0.864878807 -0.135307190
#> 8   -1.23245247 -1.125785856  0.843763125
#> 9    0.74214395  2.251785091 -1.312463882
#> 10  -1.19316585 -1.164899412  0.798168757
#> 11   0.62130469 -0.606772583 -0.293684926
#> 12   0.47148404 -0.770703924 -0.197310547
#> 13   0.55674057 -0.682434476 -0.250740213
#> 14  -0.51938525  2.273689285  2.395588241
#> 15  -1.25170523  0.291581299  1.675676989
#> 16   0.38508325 -0.849171459 -0.146071265
#> 17  -1.36253957 -0.863435966  1.071649777
#> 18   0.13231559 -1.030813839 -0.007217722
#> 19  -0.75800885  1.701420620  2.203619257
#> 20   0.36358814  3.370853371 -1.613269926
#> 21  -0.60496936 -1.292376323  0.376140398
#> 22   0.91610696 -0.071026661 -0.549715499
#> 23   1.01346481  0.882620832 -0.901044907
#> 24  -1.37774996 -0.329178107  1.389874262
#> 25  -1.38750493 -0.709293503  1.175481885
#> 26  -1.37911128 -0.341189491  1.383714704
#> 27  -1.22188846 -1.137234142  0.830964037
#> 28  -1.12790218 -1.213238573  0.732059523
#> 29   0.95630442  0.065516619 -0.605964994
#> 30   0.02727838 -1.089854703  0.047542494
#> 31  -0.24696922 -1.208267049  0.187772032
#> 32   0.44485713 -0.795962847 -0.181244317
#> 33   0.46474909 -0.777188971 -0.193221637
#> 34  -1.35536942 -0.172127798  1.467735359
#> 35  -1.34109893 -0.935967583  1.017129867
#> 36  -0.99439011 -1.271630954  0.620265588
#> 37  -1.07899608  0.849593625  1.897891969
#> 38  -0.21762293  4.831185116 -1.986083000
#> 39   0.07601043 -1.063498073  0.022271823
#> 40  -1.25213855 -1.102273029  0.868875872
#> 41   0.93987513  0.004729268 -0.581258890
#> 42   0.70837141  2.364858836 -1.343870876
#> 43  -1.37403477 -0.298520059  1.405455506
#> 44  -1.24084055  0.331308707  1.692333469
#> 45   0.99992241  1.037361630 -0.951415221
#> 46  -1.20532622 -1.153758599  0.811722245
#> 47  -1.25362577  0.284457416  1.672673983
#> 48   0.53022760 -0.711181952 -0.233768038
#> 49  -0.76960485 -1.301843936  0.470737904
#> 50   0.97676854  1.220908053 -1.009558541
#> 51  -1.06746960 -1.245016470  0.678361249
#> 52  -1.11545220  0.741418214  1.856551404
#> 53  -0.21030765 -1.195202891  0.169057338
#> 54   0.52579175 -0.715871682 -0.230961871
#> 55  -1.29526878 -1.037764924  0.931418959
#> 56   0.71837390 -0.474231022 -0.363852891
#> 57  -0.42411645 -1.259880203  0.279182187
#> 58  -0.39455455 -1.252595081  0.263758497
#> 59  -1.35240785 -0.900936132  1.044018916
#> 60  -0.03017022  3.382932370  2.752100133
#> 61  -0.46056391 -1.268118374  0.298331025
#> 62   0.64356126 -0.578591704 -0.309103597
#> 63  -1.31830599 -0.992649388  0.970997264
#> 64  -1.37025889 -0.269847463  1.419851713
#> 65  -0.91392997  1.304676437  2.064974124
#> 66  -0.03825140 -1.122632811  0.081250493
#> 67   0.91278871 -0.080730362 -0.545611004
#> 68   0.62841798 -0.597892639 -0.298575096
#> 69   0.02098429 -1.093133322  0.050792255
#> 70  -0.53172364 -1.281784726  0.336230798
#> 71   1.01644682  0.467076759 -0.758228878
#> 72   0.96574122  1.293238072 -1.032040627
#> 73  -0.19114376 -1.188045473  0.159285826
#> 74  -1.38836672 -0.443377718  1.329952890
#> 75  -1.35947360 -0.197203615  1.455613484
#> 76   0.68331287 -0.525147331 -0.337578552
#> 77  -1.26040336 -1.091440854  0.879975264
#> 78   1.00615786  0.346908807 -0.714406976
#> 79   1.01662687  0.470008077 -0.759281806
#> 80  -0.22840808 -1.201755684  0.178292799
#> 81  -1.02884402 -1.260464131  0.646855762
#> 82   0.34019788 -0.886224360 -0.120371294
#> 83  -0.49562787 -1.275257375  0.316912892
#> 84  -1.39307514 -0.595289368  1.244744860
#> 85  -0.86230259 -1.296758635  0.528481534
#> 86  -1.32489401 -0.977758964  0.983467008
#> 87  -1.02242434 -1.262719968  0.641799884
#> 88   1.02138121  0.593357579 -0.802950578
#> 89   0.02607718 -1.090482593  0.048162924
#> 90   0.62697005  2.622790321 -1.414526296
#> 91   0.34292866  4.210010450  3.013059234
#> 92  -1.38442328 -0.739277593  1.156320841
#> 93  -1.10158628 -1.228357151  0.707934648
#> 94  -0.75463213 -1.301893211  0.461771362
#> 95   0.92401220 -0.047124070 -0.559762528
#> 96   0.41004131  3.245531151 -1.580501371
#> 97  -0.61470256 -1.293506575  0.381526394
#> 98   0.65175049 -0.567920654 -0.314867092
#> 99  -0.96819322 -1.278679345  0.600882938
#> 100 -0.04337401 -1.125070980  0.083875354