Title `

Title `

make_predict_dat(Smooth, dat)

Arguments

Smooth
The smooth object from construct_smooth_data

dat
The testing data to construct the new design matrix.

Value

a data frame containing the trasnformed desgin matrix for the testing data

Examples

raw_dat <- sim_Bai(100, 5)$dat %>% data.frame
test_dat <- sim_Bai(100, 5)$dat %>% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_smooth <- construct_smooth_data(sm_df, raw_dat)$Smooth

make_predict_dat(dsn_smooth, test_dat)

#>          x1.pen1     x1.pen2     x1.pen3     x1.null1       x2.pen1
#> 1    0.524508906  0.96890173 -0.58055032 -1.285277014  0.3172625383
#> 2    0.028961816 -1.02175617  0.04014289  0.655521023  0.0528571505
#> 3   -0.448552433  1.75509207 -7.13245877  2.660637461  0.0541943078
#> 4    0.197496584 -1.06329350 -0.68394739  0.927582853  0.2616079363
#> 5   -0.374053895 -0.76427223 -5.34371818 -2.688392701 -0.5521731425
#> 6    0.200404334 -0.63076732 -2.19928834  1.384548769  0.3055338679
#> 7   -0.271996804 -0.69341430  0.60667939  0.372097784  0.1163992054
#> 8   -0.282584948 -0.67748401  0.62306338  0.361976694 -0.2338813521
#> 9    0.392410445  1.03893507  0.57551286 -0.773984698 -0.6059025828
#> 10   0.527999225  1.04960215 -0.13934210 -1.116426791  0.4034101777
#> 11  -0.067697240 -0.93882467  0.25685010  0.558571309 -0.1113621416
#> 12  -0.368273850 -0.52862782  0.75515653  0.273039468  0.2107038215
#> 13   0.475859868  0.82363040 -1.13237240 -1.474867399  0.0002481843
#> 14   0.530591654  1.00498045 -0.40917544 -1.221996705  0.4168139705
#> 15   0.232756204 -0.85746665 -1.57956045  1.208440792  0.3594461877
#> 16  -0.250849880 -0.72402372  0.57370874  0.391974647 -0.5086911205
#> 17   0.166113962  0.86880691  0.89810750 -0.538946445 -0.3348493494
#> 18  -0.075859229 -0.93086251  0.27288615  0.550954430  0.4050364346
#> 19  -0.410695446  0.20031774  1.06803598 -0.109379802 -0.5753963699
#> 20   0.506336843  0.90513627 -0.84097395 -1.377085242  0.1564645034
#> 21  -0.160483782 -0.83993720  0.42653720  0.474059060  0.3834802406
#> 22  -0.103807107 -0.90253449  0.32596150  0.525230881 -0.4152735177
#> 23  -0.068699315  0.40878828 -4.41909696  1.966119767  0.2180082272
#> 24   0.314551217  0.48407974 -2.14982603 -1.789520502 -0.5754806457
#> 25  -0.335923451  0.33414598  1.07844390 -0.180938234  0.2208384531
#> 26   0.333608704  1.00076833  0.69385399 -0.699568719  0.0616070311
#> 27   0.033403826  0.75129447  0.98684873 -0.439822949  0.1591260955
#> 28  -0.053026854 -0.95278022  0.22733732  0.572412519  0.1670639917
#> 29  -0.479381818 -0.10096161  0.99252400  0.045794170  0.3796584997
#> 30   0.150471805 -1.07897939 -0.38729598  0.823198127  0.3979727397
#> 31  -0.066811264  0.65569363  1.03035287 -0.372175886  0.2277247482
#> 32  -0.152097918  0.56715718  1.05618797 -0.315357485  0.1043771573
#> 33  -0.114034991  0.60766525  1.04582474 -0.340805471 -0.2908847388
#> 34  -0.035000168  0.68687089  1.01821191 -0.393372976  0.4166853378
#> 35   0.086854103 -0.15176568 -3.27593276  1.672072928  0.2928772209
#> 36   0.086995926  0.06245096 -3.25728087 -2.106663802  0.4166198523
#> 37  -0.083561019 -0.92322034  0.28778508  0.543814700 -0.2180821091
#> 38   0.040285491  0.75760591  0.98324885 -0.444604867 -0.1823864690
#> 39  -0.419091749 -0.41179606  0.83779552  0.208193080 -0.0662236513
#> 40  -0.072108493  0.65041539  1.03223694 -0.368656555 -0.5732626801
#> 41  -0.427561809  0.16159432  1.06228294 -0.089118357 -0.0768354699
#> 42   0.218495751 -1.03170425 -0.90169870  0.999860861  0.3839806693
#> 43   0.153196016  0.85775270  0.90885009 -0.528495324  0.0106186270
#> 44   0.098703536 -1.06447096 -0.17094846  0.741456188  0.3984170537
#> 45   0.372530334  1.02678334  0.61981967 -0.747164505 -0.1007641130
#> 46  -0.001965291 -0.99765372  0.11613227  0.622685946 -0.4595109997
#> 47  -0.373011697 -0.51899203  0.76259088  0.267559177 -0.0555650967
#> 48   0.321073683  0.99188074  0.71478712 -0.685365079  0.3983061777
#> 49  -0.004090593  0.71638651  1.00498339 -0.414161543  0.3752115317
#> 50   0.528868424  1.04661043 -0.16141232 -1.125381984 -0.4468395601
#> 51  -0.207931105 -0.78181238  0.50534728  0.431329587  0.3433615861
#> 52  -0.474836621 -0.02019581  1.01964257  0.004509547  0.4003338229
#> 53  -0.339668605 -0.58306908  0.71080438  0.304555691  0.2194942339
#> 54   0.214615872  0.29542457 -2.65612311 -1.936506830 -0.2841001456
#> 55   0.167658211 -1.07800938 -0.47919345  0.856376197  0.3984315706
#> 56   0.512942452  1.07056514  0.06660732 -1.029474298  0.3951480202
#> 57   0.194306071  0.89259108  0.87271697 -0.562507169 -0.0864041933
#> 58  -0.468722841  0.02146455  1.03166476 -0.016807977  0.3955139812
#> 59   0.524578434  0.96921373 -0.57916517 -1.284775478 -0.4172270483
#> 60   0.376124792  0.60566884 -1.80729278 -1.687133850  0.1808997077
#> 61   0.201644713 -1.05914360 -0.71971103  0.939678510  0.0831174599
#> 62   0.123063526 -1.07385260 -0.26389115  0.777253748  0.3401957474
#> 63   0.378080395  1.03026418  0.60795518 -0.754456889 -0.5199694441
#> 64  -0.288314118  0.40291042  1.07768424 -0.218980280 -0.0058447938
#> 65   0.120334100 -1.07299665 -0.25280018  0.773039007  0.3524862490
#> 66  -0.312469517 -0.63003772  0.66903204  0.332620627 -0.1644179276
#> 67  -0.419307877 -0.41121555  0.83816790  0.207878472 -0.5748571489
#> 68   0.457862068  0.78077020 -1.27432574 -1.520965097  0.2029197860
#> 69   0.306338800  0.98116659  0.73786920 -0.669258143 -0.2860341124
#> 70   0.430894988  0.72032381 -1.46545837 -1.581700168  0.1645201898
#> 71   0.053652204 -1.03900114 -0.02722526  0.683728263 -0.1138494277
#> 72   0.105163807 -0.22136548 -3.12882165  1.633683907 -0.5604573315
#> 73   0.464609811  1.07227247  0.35291475 -0.895234306 -0.4880750822
#> 74  -0.070283172 -0.93631730  0.26195914  0.556152096  0.1638157650
#> 75  -0.171232342 -0.82725172  0.44475504  0.464406293 -0.5752704121
#> 76  -0.436793178 -0.36043936  0.86940648  0.180592057  0.3985508726
#> 77   0.487771325  1.07625301  0.24313640 -0.948962907  0.1377205815
#> 78   0.235120256 -0.91954942 -1.37638351  1.148055654  0.3039057795
#> 79  -0.474929449 -0.18238626  0.95999576  0.087555366  0.3299223139
#> 80   0.231940253 -0.97624312 -1.16334120  1.082920243  0.4165654406
#> 81  -0.182521345  0.53340438  1.06314547 -0.294746977  0.3487592275
#> 82  -0.296983016 -0.65510960  0.64524736  0.347994181  0.0150433278
#> 83   0.523005705  1.06045002 -0.04877773 -1.078983789  0.3605557244
#> 84   0.531546700  1.02404456 -0.30569423 -1.182461000  0.3289980488
#> 85  -0.022589651  0.24529339 -4.08948569  1.881740929  0.2099226696
#> 86   0.093546157  0.80559714  0.95210193 -0.482692094 -0.2734291584
#> 87  -0.467972173  0.02558019  1.03278013 -0.018916511  0.1903734410
#> 88   0.234598920  0.92566796  0.83119331 -0.598204314  0.3966926864
#> 89  -0.300828021  0.38563448  1.07829107 -0.209321644  0.0403662503
#> 90   0.141062390 -0.36324030 -2.82160629  1.552764054  0.3372684499
#> 91   0.428131428  1.05809785  0.48063360 -0.828077675  0.1547513494
#> 92   0.019125391  0.73810798  0.99404139 -0.429978773 -0.6432594717
#> 93  -0.153830015 -0.84765552  0.41513824  0.480034948 -0.2366276078
#> 94   0.409340780  0.67418966 -1.60577226 -1.625430102  0.0246350338
#> 95   0.176009557  0.87721022  0.88950917 -0.547095059 -0.0605555965
#> 96  -0.402537602  0.21742092  1.07019484 -0.118382038  0.3735084189
#> 97   0.109954169 -1.06927070 -0.21225452  0.757501899  0.3884102279
#> 98  -0.155865505  0.56304804  1.05711347 -0.312821203  0.3752662641
#> 99   0.514234932  1.06970181  0.05427829 -1.034870352  0.1087356138
#> 100  0.401220915  0.65718632 -1.65648098 -1.641072836  0.1730712638
#>           x2.pen2     x2.pen3    x2.null1      x3.pen1     x3.pen2     x3.pen3
#> 1   -0.7403854043 -1.79016226  1.61523248 -0.137796113  0.53868098  4.21354206
#> 2   -1.0702195396  1.05377955  0.56548039 -0.068953811  0.35073781  3.82965125
#> 3   -0.0409679279 -3.43700090  2.05107800 -0.396906463 -0.20956913 -0.80513733
#> 4   -1.2376531108  0.67219029  0.77796333 -0.039897257  0.27126571  3.66712607
#> 5    0.2256060573  1.37507951 -0.14115273  0.464680350  1.03040329 -0.91531270
#> 6   -1.2598494914  0.53512566  0.84144553 -0.349091086 -0.42518045 -0.51135429
#> 7   -0.1973742529 -3.08669766  1.96095040 -0.162100909 -0.64660828 -0.02051080
#> 8    0.7709699273  1.06158161 -0.43854784  0.004995578  0.55556604 -1.34169713
#> 9    1.5865124645 -7.01381362  2.96102487  0.474185236  0.98378828 -0.54889022
#> 10  -1.0511606031 -0.89481575  1.35476381  0.076209479 -0.05773092  2.97898011
#> 11  -0.4356805894 -4.50709439 -2.15909140  0.198028118 -0.55106904  1.75539353
#> 12  -1.2039666567  0.79633727  0.71586836  0.085812311  0.65502910 -1.36770593
#> 13   0.9910076734  0.77887337 -0.60319146  0.031364990  0.58830213 -1.35173112
#> 14  -1.1687422694 -0.45393977  1.21451304 -0.344907122  0.01063294 -1.02124836
#> 15  -0.8749974913 -1.42825350  1.51299126  0.087532442  0.65711917 -1.36809677
#> 16  -0.2913580067  1.43287884  0.10816944 -0.248650138  0.20072509 -1.16394478
#> 17  -0.6180898211  1.36727111  0.27354812  0.152763552 -0.72847687  0.98882234
#> 18  -1.2384858179 -0.08744995  1.08845191 -0.608915916  1.82485331  6.84065123
#> 19   0.0698648208  1.41275008 -0.06542432 -0.297180784 -0.51059473 -0.35747810
#> 20   1.1011405280  0.51122549 -0.72965053  0.307256823  0.90988066 -1.32792592
#> 21   1.1173201510 -0.45381704 -1.08571191  0.233742727  0.82936047 -1.36701337
#> 22   0.5339935119  1.24030186 -0.29964680 -0.130193753 -0.66923160  0.05716316
#> 23   0.3745843887 -2.71554436 -1.71690607  0.187942281  0.77675408 -1.37589674
#> 24   0.0677070838  1.41314474 -0.06438535  0.340890203  0.94436105 -1.29509839
#> 25  -1.2111920293  0.77387986  0.72748389  0.392643498  0.99249618 -1.21339783
#> 26  -1.0784122059  1.04285414  0.57271652  0.307310154  0.90993676 -1.32788291
#> 27  -1.1641053236  0.89717013  0.66120460  0.068934164 -0.74870925  0.61183215
#> 28   1.1073497741  0.48829111 -0.73962921  0.380974825  0.72885194  0.37010514
#> 29  -0.9477644678 -1.21882998  1.45209277 -0.089007889  0.40548755  3.94148389
#> 30   1.0105973045 -0.95254751 -1.23975390 -0.297699761  0.11288064 -1.10233506
#> 31  -1.2159679085  0.75803835  0.73556891 -0.161456890 -0.64709313 -0.01894984
#> 32   1.0679518125  0.61307978 -0.68387223  0.153572143  0.73638212 -1.37689525
#> 33   0.7052297039  1.12006733 -0.39758241  0.153286719  0.73604407 -1.37688591
#> 34  -1.1592806648 -0.49447158  1.22786314 -0.385206503 -0.32527149 -0.66080452
#> 35  -0.6683773053 -1.97407296  1.66596933  0.194652030 -0.64759863  1.41621259
#> 36  -1.1799498275 -0.40408475  1.19794647  0.439695356  1.02565612 -1.07193219
#> 37  -0.7768099167  1.29848583  0.36224433 -0.381359364 -0.10123543 -0.91945574
#> 38  -0.6057476024 -4.87453140 -2.24876918 -0.295911208  0.11632961 -1.10488970
#> 39   0.9353964117  0.86843081 -0.55554982  0.038844299 -0.74445513  0.51261950
#> 40   0.1030170671  1.40625714 -0.08141499 -0.325649994  0.05547437 -1.05811370
#> 41  -0.9399622576  1.18930075  0.46542945 -0.395938011 -0.19159507 -0.82534219
#> 42   0.9042962424 -1.32024382 -1.34613156  0.184011949 -0.68521146  1.24986387
#> 43  -1.0296236884  1.10285678  0.53164617 -0.395380429 -0.18468107 -0.83297397
#> 44   1.0426283554 -0.82462728 -1.20145628 -0.234227854 -0.58384440 -0.19600332
#> 45  -0.9140371116  1.21012297  0.44783523  0.322071898  0.92530162 -1.31494097
#> 46   0.4580889798  1.28193976 -0.25904734  0.328046477  0.93142133 -1.30907581
#> 47   0.9446505019  0.85470058 -0.56307507 -0.127136678  0.38332424 -1.26993238
#> 48   1.0184530798 -0.92219354 -1.23073278 -0.201854617  0.27532401 -1.21079700
#> 49  -1.2676965566  0.19851283  0.98155996 -0.014623336  0.20169794  3.52426404
#> 50   0.4810457158  1.26999971 -0.27118558  0.407261859  0.78930131  0.18347269
#> 51  -1.2704390774  0.37995206  0.90839538 -0.315483562  0.07720712 -1.07523755
#> 52  -1.0356233381 -0.94612759  1.37046026 -0.382970316 -0.33425319 -0.64843098
#> 53   1.1345848858  0.36073413 -0.79322794  0.152445653  0.73504769 -1.37685679
#> 54   0.7133844194  1.11324632 -0.40253870  0.035178986  0.59301094 -1.35306375
#> 55   1.0420896040 -0.82687869 -1.20213683 -0.198330810  0.28068288 -1.21397366
#> 56  -1.2539169617  0.03239279  1.04473425  0.101849153  0.67446820 -1.37105398
#> 57   0.9175146647  0.89385188 -0.54138395 -0.181801707  0.30542076 -1.22831102
#> 58   1.0763046852 -0.67483529 -1.15561281 -0.266319001 -0.54929857 -0.27648945
#> 59  -0.4844675092  1.40486492  0.20396900  0.151549023 -0.72941710  0.98118635
#> 60  -0.3631665318 -2.70772390  1.86229077 -0.124313220  0.50187225  4.13835729
#> 61  -1.0982321311  1.01470224  0.59092004 -0.300781420 -0.50563516 -0.36727034
#> 62   1.1561068855 -0.09995795 -0.96679774 -0.258688379 -0.55797439 -0.25708820
#> 63  -0.2607213256  1.43474596  0.09329622 -0.327823358  0.05067619 -1.05426878
#> 64  -0.1827560237 -3.96013884 -2.02553938 -0.254316268 -0.56280796 -0.24605576
#> 65  -0.8515379941 -1.49340907  1.53166035 -0.175824054  0.31421569 -1.23327877
#> 66  -0.8418276668  1.26068717  0.40144202  0.243226513  0.84005236 -1.36389898
#> 67   0.0002094386  1.42379206 -0.03196911 -0.390125402 -0.14384849 -0.87654203
#> 68   0.3348936396 -2.80805100 -1.74029747 -0.313325247  0.08168376 -1.07870626
#> 69   0.7110696727  1.11519422 -0.40112852 -0.316573969  0.07492825 -1.07346414
#> 70  -1.1684870011  0.88755229  0.66661474  0.102350212 -0.74823545  0.73706964
#> 71  -0.8995888033  1.22106940  0.43826499 -0.321271717 -0.47514236 -0.42495299
#> 72   0.1896221474  1.38542575 -0.12349809  0.285388232  0.52401917  0.95384095
#> 73   0.4015531346  1.30908801 -0.22960698 -0.393549419 -0.27992607 -0.72053826
#> 74  -1.1679173406  0.88881941  0.66590455 -0.068227225  0.46225947 -1.30647990
#> 75   0.0728140534  1.41220515 -0.06684473 -0.126479356  0.38423021 -1.27038454
#> 76   1.0312916376 -0.87122347 -1.21549273 -0.382184017 -0.10468234 -0.91607582
#> 77  -1.1462962465  0.93350217  0.64034209  0.084470060 -0.08265025  2.92490077
#> 78   1.1595162598  0.07863308 -0.90257249 -0.198863733  0.27987444 -1.21349604
#> 79  -1.2679099585  0.44045309  0.88283340  0.348949093  0.95231431 -1.28527713
#> 80  -1.1813653620 -0.39762889  1.19578911 -0.082170616  0.38682157  3.90335715
#> 81  -0.8392260572 -1.52720825  1.54129667  0.238585921  0.83483016 -1.36548335
#> 82  -1.0339575377  1.09797503  0.53511676 -0.314130293 -0.48624843 -0.40442410
#> 83   0.8028066171 -1.62559570 -1.43103181  0.349622537  0.95297265 -1.28441665
#> 84  -1.2676793840  0.44436723  0.88115701  0.381263855  0.98260557 -1.23575682
#> 85   1.1301038602  0.38601234 -0.78284477  0.387638391  0.98820630 -1.22361417
#> 86   0.7260198068  1.10244599 -0.41028410 -0.285090354  0.13673425 -1.11977258
#> 87   1.1202309202  0.43471737 -0.76251622  0.129017438 -0.22473706  2.60718569
#> 88   0.9919366056 -1.02243509 -1.26037500 -0.071391277 -0.70392208  0.20412381
#> 89   1.0221619752  0.71947649 -0.63309832  0.469183652  1.02778904 -0.86509513
#> 90   0.7201948331 -1.85499260 -1.49315537  0.365499902  0.96818383 -1.26216432
#> 91   1.1001178801  0.51485492 -0.72806096 -0.272077411 -0.54253772 -0.29126858
#> 92  -1.7093032127 -7.25881024 -2.83068340 -0.146294757 -0.65816666  0.01786510
#> 93  -0.7533294211  1.31061052  0.34856660  0.143014999  0.72385018 -1.37637002
#> 94  -1.0432869626  1.08719525  0.54269429  0.243118226  0.43733228  1.18730512
#> 95   0.9403338298  0.86115645 -0.55954743 -0.292900043 -0.51635665 -0.34594591
#> 96   1.1333242026 -0.34362166 -1.04970304 -0.275255963  0.15465913 -1.13252483
#> 97   0.9286865675 -1.24141249 -1.32374169 -0.162777106 -0.64609792 -0.02214958
#> 98  -0.9312142924 -1.26755164  1.46638877 -0.070636568 -0.70431050  0.20606009
#> 99   1.0708813131  0.60517141 -0.68751861  0.353914996  0.95714490 -1.27878027
#> 100  1.1107750236  0.47491232 -0.74539900  0.161548323 -0.72054523  1.04788255
#>        x3.null1      x4.pen1      x4.pen2      x4.pen3     x4.null1
#> 1    1.76152119  0.072759655  1.015046390 -0.188106111 -0.580376073
#> 2    1.64808196  0.192356077 -1.244226249 -0.980924295  1.107515311
#> 3   -0.03669531  0.058035983 -0.128707341  4.008196962 -1.876198671
#> 4    1.60004117  0.421544537  0.736356723  1.917089356 -1.300375479
#> 5   -1.45474726 -0.685466920 -1.598755679  7.177789353 -2.711448220
#> 6    0.13413463 -0.093248615  0.933353017 -0.450616780 -0.461375459
#> 7    0.36907098  0.126224397  0.009928139  3.703394789 -1.795231641
#> 8   -0.58111261 -0.417333383  0.659727571 -0.905858106 -0.212599614
#> 9   -1.70929798  0.407372261  1.023604517  0.811006792 -0.955524981
#> 10   1.39547669 -0.559695688  0.100106111 -1.315297488  0.137140538
#> 11   1.01761307 -0.348191925 -0.504747991 -1.444984572  0.475113663
#> 12  -0.66802223  0.430757468  0.777413515  1.793301948 -1.263888010
#> 13  -0.60851758  0.247438382 -1.282211642 -0.804956203  1.220886337
#> 14  -0.19151600 -0.257491343  0.819971599 -0.682025587 -0.343773219
#> 15  -0.66997734  0.330575468 -1.070544641  0.518332821  1.829572633
#> 16  -0.32103307 -0.566341306  0.275366510 -1.226267054  0.036732835
#> 17   0.76058454  0.331840129 -1.248631495 -0.174429305  1.543906003
#> 18   2.53782748 -0.217959662  0.850790728 -0.627527620 -0.372817487
#> 19   0.21290730  0.398513818  0.653649746  2.152803015 -1.368783780
#> 20  -0.97929861 -0.502487987 -0.147281305 -1.399595045  0.274319927
#> 21  -0.85884686 -0.532051792  0.454173567 -1.103740388 -0.072104065
#> 22   0.40263089 -0.506008377 -0.136279958 -1.396780361  0.268254324
#> 23  -0.79403314  0.439840536  0.831275310  1.621739452 -1.212599471
#> 24  -1.04428947  0.332807128 -1.244933725 -0.154178541  1.552996086
#> 25  -1.16491851  0.436517335  0.808762872  1.694891407 -1.234575446
#> 26  -0.97939568 -0.445605972 -0.299926496 -1.430008057  0.358814071
#> 27   0.62381079 -0.324215746 -0.549113233 -1.444065735  0.501119314
#> 28  -2.19709039 -0.103058742 -0.897575060 -1.370255708  0.727062251
#> 29   1.68112846 -0.462081601  0.597173636 -0.974298794 -0.167423693
#> 30  -0.26113718  0.128329031 -0.251239002  2.545313142  2.519171360
#> 31   0.36975316  0.444278730  0.908135225  1.350904746 -1.129702254
#> 32  -0.74914306 -0.215008582 -0.732495166 -1.422019281  0.613829444
#> 33  -0.74878193 -0.557915953  0.350388071 -1.179295187 -0.007896224
#> 34   0.05131188 -0.106388229 -0.892931556 -1.372253811  0.723643576
#> 35   0.90666703 -0.317586820 -0.561071364 -1.443545368  0.508196636
#> 36  -1.31902072  0.147712026  1.041696175 -0.050084553 -0.638416147
#> 37  -0.11421580 -0.225501134 -0.715973409 -1.425342527  0.603235721
#> 38  -0.26348223 -0.040857476 -0.981638446 -1.326875154  0.792369193
#> 39   0.58623831  0.433591599  0.975794054  1.067942850 -1.040103928
#> 40  -0.22209900 -0.381675525  0.702166753 -0.854261355 -0.244849117
#> 41  -0.04981807 -0.014403015  0.975917446 -0.331348689 -0.517018075
#> 42   0.85075442 -0.447246303  0.619353773 -0.950980852 -0.183152908
#> 43  -0.05483418  0.402239919  0.665889581  2.118889709 -1.359022010
#> 44   0.29012637  0.136464517 -1.191885320 -1.105226505  1.014922566
#> 45  -1.00695072  0.115593148 -1.170026365 -1.142547652  0.984109289
#> 46  -1.01851790  0.004392999 -0.235988446  4.241388694 -1.937855564
#> 47  -0.44909044 -0.558364205  0.090084862 -1.319603216  0.142766477
#> 48  -0.37251051  0.235072000 -1.275362251 -0.851062567  1.192703168
#> 49   1.55776799  0.348726533  1.059945283  0.497524472 -0.847132264
#> 50  -2.10749619  0.405144551  0.675669663  2.091569914 -1.351139222
#> 51  -0.23688441  0.389553472  1.040085348  0.695634152 -0.916367019
#> 52   0.05845272  0.228084193  0.225457610  3.216868602 -1.664629708
#> 53  -0.74771885  0.189260646  1.053230750  0.035392817 -0.673081053
#> 54  -0.61254631  0.333270107 -1.243006480 -0.143866249  1.557603338
#> 55  -0.37624351  0.203635525  1.056585936  0.067001063 -0.685674101
#> 56  -0.68644539 -0.515819696  0.496545362 -1.068794267 -0.099383649
#> 57  -0.39355289  0.311778565 -0.957257677  0.852755623  1.954079153
#> 58   0.25225229  0.182161556 -1.235487536 -1.006774526  1.089311922
#> 59   0.75789673  0.284105415  1.067280791  0.273121596 -0.765124175
#> 60   1.73930422  0.011553264 -0.221732801  4.210499939 -1.929699110
#> 61   0.20805887  0.173530644  1.049169381  0.002086406 -0.659682140
#> 62   0.26148836  0.410346147  1.020053462  0.833223776 -0.962976291
#> 63  -0.21883229 -0.160914131 -0.814665344 -1.400833839  0.668287369
#> 64   0.26670952 -0.081709829 -0.926996182 -1.356699535  0.749138656
#> 65  -0.39973981 -0.087094614 -0.919632459 -1.360242756  0.743542807
#> 66  -0.87311080  0.225228080  1.060918274  0.116934873 -0.705334189
#> 67  -0.08413655  0.340209722  0.488207154  2.586964557 -1.491686796
#> 68  -0.23992841  0.321879504  1.065484284  0.395077281 -0.810210737
#> 69  -0.23533467 -0.032727108 -0.992259278 -1.320269707  0.801172394
#> 70   0.67023558 -0.562904997  0.314557828 -1.202462626  0.013576116
#> 71   0.17906667  0.043371260 -0.158157271  4.072396777 -1.893193602
#> 72  -2.46011415 -0.123416806 -0.868944490 -1.382003505  0.706243100
#> 73   0.01600321 -0.541435772  0.424151900 -1.126983476 -0.053182282
#> 74  -0.50757662 -0.549673587  0.036297745 -1.341374286  0.172817875
#> 75  -0.44974737  0.175486090 -1.229546483 -1.022840939  1.077748190
#> 76  -0.11179907  0.336670757 -1.223736086 -0.048010908  1.599763876
#> 77   1.37925634 -0.475112945  0.576170700 -0.995492480 -0.152794419
#> 78  -0.37567999  0.298207243  1.067278069  0.316178686 -0.781191542
#> 79  -1.06115509  0.152606061  1.043185835 -0.040412049 -0.642385201
#> 80   1.66986208 -0.077355039  0.942491890 -0.427172737 -0.472544443
#> 81  -0.86609122  0.356626172  0.531342763  2.477474848 -1.461026444
#> 82   0.18947158  0.438624567  0.822437745  1.650726706 -1.221327368
#> 83  -1.06259077  0.371059027  1.051333888  0.597558561 -0.882426838
#> 84  -1.13548819 -0.490281855  0.549465686 -1.021270342 -0.134537827
#> 85  -1.15172143  0.337109350 -1.220097614 -0.031114155  1.607077180
#> 86  -0.27736167 -0.535945299  0.442355665 -1.113033800 -0.064617924
#> 87   1.28329710  0.374639104  1.049501148  0.615268209 -0.888602380
#> 88   0.46414372  0.444177879  0.910956170  1.340159359 -1.126359280
#> 89  -1.49375184  0.230661489  1.061861746  0.130030047 -0.710444112
#> 90  -1.09773852 -0.253795360 -0.670371088 -1.433046436  0.574511688
#> 91   0.24516914  0.305999947  0.403341792  2.796539258 -1.549830255
#> 92   0.38574876  0.076379472 -1.126413405 -1.202665550  0.930354379
#> 93  -0.73590477  0.444264458  0.908577097  1.349226251 -1.129180350
#> 94  -2.56011073 -0.286492749 -0.615574256 -1.439642946  0.540869452
#> 95   0.21859108  0.369083605  0.565515964  2.389060933 -1.436112079
#> 96  -0.28956730 -0.530119223  0.459757097 -1.099282813 -0.075658745
#> 97   0.36835443 -0.498436510 -0.159637906 -1.402656178  0.281133896
#> 98   0.46493824 -0.128758406  0.911869146 -0.502116700 -0.436406753
#> 99  -1.07184199 -0.188090142 -0.773989889 -1.412333793  0.640933706
#> 100  0.78127163  0.104839871 -1.158374152 -1.160242213  0.968868361
#>          x5.pen1     x5.pen2     x5.pen3     x5.null1
#> 1   -0.415926163  0.20488955 -0.83096591  0.019406160
#> 2    0.429644964 -0.77384978  0.40202839  2.039956105
#> 3    0.483659259 -1.03411958 -0.87520923  1.318615667
#> 4    0.453392636 -1.02173330 -1.00812724  1.205309001
#> 5    0.061454439  0.81889690  0.54731830 -0.643080139
#> 6   -0.222557090  0.63524527 -0.21096514 -0.325698417
#> 7   -0.363165714  0.42899681 -0.56717539 -0.143465547
#> 8   -0.134688066 -0.40614069 -1.24278509  0.426985869
#> 9   -0.414145960  0.15386859 -0.88010660  0.054102170
#> 10  -0.204996329 -0.31255899 -1.20190200  0.362124409
#> 11  -0.375981284  0.39886064 -0.60813773 -0.120185453
#> 12  -0.303009905 -0.16061784 -1.11922840  0.260941321
#> 13  -0.225443721  0.63216317 -0.21771139 -0.322514997
#> 14  -0.414998427  0.22840875 -0.80713555  0.003193359
#> 15   0.130529506 -0.71407671 -1.30985254  0.677291637
#> 16   0.301235316 -0.89312469 -1.25449434  0.892941039
#> 17  -0.231724447  0.62534887 -0.23240643 -0.315550946
#> 18  -0.404126396  0.08805864 -0.93870462  0.098076916
#> 19  -0.389935295  0.03402993 -0.98313872  0.133711449
#> 20   0.042728252 -0.61637048 -1.30262581  0.588144662
#> 21   0.024072426 -0.59527142 -1.29900129  0.570461499
#> 22  -0.311014558 -1.09675169  5.54114801 -2.173156350
#> 23   0.077322740  0.82105720  0.60782303 -0.665844933
#> 24  -0.185326088  0.67241045 -0.12405067 -0.365980702
#> 25  -0.037596128  0.78106221  0.24117552 -0.522991626
#> 26  -0.334759336  0.48447482 -0.48593232 -0.187973935
#> 27  -0.345288804  0.46531314 -0.51491711 -0.172332969
#> 28  -0.341463169 -0.08755613 -1.07229024  0.213169168
#> 29   0.405428852 -0.98786360 -1.12820756  1.080974421
#> 30  -0.407285977  0.10378956 -0.92516138  0.087631614
#> 31  -0.338950523 -0.09273267 -1.07577415  0.216546881
#> 32  -0.153680779  0.70060496 -0.04963522 -0.399470577
#> 33   0.335222264 -0.57911393  0.97993835  2.291771805
#> 34   0.053253872  0.03668345  3.20214470 -1.495838315
#> 35   0.170983872 -0.75806509 -1.30663466  0.722358931
#> 36   0.490874753 -1.03454577 -0.82745524  1.355425588
#> 37   0.476453555 -0.88382591  0.03008342  1.863009586
#> 38  -0.021093175  0.78961345  0.28668826 -0.541416706
#> 39   0.177660885  0.74850910  1.28663670 -0.905114389
#> 40   0.508473611 -0.99330821 -0.44215000  1.609146998
#> 41   0.044859307 -0.61877284 -1.30299759  0.590188641
#> 42   0.503252279 -1.02864717 -0.70338075  1.444278040
#> 43  -0.062528177 -0.32897951  3.96417138 -1.717102179
#> 44   0.118406711  0.81743730  0.79606453 -0.734944412
#> 45   0.071890068  0.82049123  0.58652300 -0.657863641
#> 46   0.121150473  0.81653781  0.81089792 -0.740287136
#> 47  -0.334072984  0.48567700 -0.48407894 -0.188965561
#> 48   0.121396803  0.81645177  0.81224781 -0.740772646
#> 49  -0.365351982 -0.03412680 -1.03487634  0.178312074
#> 50   0.200949006 -0.79012893 -1.30095098  0.757753446
#> 51   0.375303731 -0.96249912 -1.17761009  1.018748510
#> 52   0.032310025 -0.60460397 -1.30068269  0.578225367
#> 53   0.441690126 -1.01453163 -1.04417876  1.170854559
#> 54  -0.314555565 -0.13988723 -1.10639397  0.247357284
#> 55   0.080066761 -0.65825325 -1.30781048  0.624767864
#> 56  -0.411413668  0.26483837 -0.76865483 -0.022237531
#> 57   0.324489607 -0.91571787 -1.23552318  0.929288075
#> 58  -0.411437569  0.12951944 -0.90239048  0.070461939
#> 59   0.137085097  0.80888105  0.90541037 -0.774008137
#> 60   0.177165970  0.75015679  1.27885293 -0.902506621
#> 61   0.010664539 -0.10071432  3.49239440 -1.580429054
#> 62   0.178287051  0.74633874  1.29677945 -0.908508590
#> 63   0.186003272  0.70407796  1.47512287 -0.967507412
#> 64  -0.112401689 -0.43417918 -1.25347402  0.447009379
#> 65   0.472265389 -1.03091666 -0.93452938  1.270300249
#> 66   0.075408892  0.11082103  3.04199856 -1.448886645
#> 67  -0.409970211  0.27487583 -0.75770134 -0.029319324
#> 68   0.155085696  0.79308125  1.03676520 -0.820014985
#> 69   0.406007596 -0.98832744 -1.12710939  1.082260776
#> 70   0.354656503 -0.61774515  0.87020987  2.245564013
#> 71   0.059357273  0.05687859  3.15881324 -1.483156941
#> 72  -0.341872080 -0.08670704 -1.07171644  0.212615182
#> 73   0.002773271 -0.57099978 -1.29407673  0.550662470
#> 74   0.485867658 -1.03441901 -0.86166807  1.329226526
#> 75  -0.412562963  0.13837615 -0.89436933  0.064524650
#> 76  -0.389118786  0.36229940 -0.65521561 -0.092646063
#> 77  -0.223286395  0.63446949 -0.21266919 -0.324895113
#> 78  -0.066398903 -0.34094137  3.98874353 -1.724208565
#> 79   0.498765805 -1.03237441 -0.75845924  1.405909275
#> 80   0.097021133 -0.67711550 -1.30917567  0.642000560
#> 81  -0.037726926  0.78099140  0.24082068 -0.522847112
#> 82  -0.402306274  0.31472741 -0.71262204 -0.057794892
#> 83  -0.220718240  0.63719263 -0.20666963 -0.327720952
#> 84  -0.155775702  0.69882719 -0.05459394 -0.397265712
#> 85  -0.287798125 -0.18672319 -1.13485415  0.278098029
#> 86   0.508277565 -0.99165124 -0.43298474  1.614527326
#> 87   0.108287293  0.22644719  2.78506529 -1.373006339
#> 88   0.016012031 -0.58611168 -1.29723457  0.562925324
#> 89   0.117567388  0.26085451  2.70645359 -1.349628359
#> 90  -0.137783511 -0.40219326 -1.24122087  0.424191864
#> 91   0.238172176 -0.82922991 -1.28925352  0.804548907
#> 92   0.181232834  0.56317015  1.93926541 -1.115902886
#> 93   0.011352621  0.80406578  0.38129580 -0.579043830
#> 94  -0.413522135  0.24671325 -0.78804531 -0.009533501
#> 95  -0.042220247  0.77853044  0.22868331 -0.517895650
#> 96  -0.278057409  0.56987292 -0.34230438 -0.262088748
#> 97  -0.374222329 -0.01166078 -1.01833253  0.163638720
#> 98   0.331733039 -0.92262255 -1.22876928  0.941124808
#> 99  -0.355958623  0.44430375 -0.54555533 -0.155516770
#> 100  0.166462497 -0.75318664 -1.30723532  0.717175015