Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                 x1           x2           x3          x4 y
#>   [1,] -1.29017232 -0.966865579  1.270547451 -0.64559369 0
#>   [2,] -0.12488604  1.664079212  0.093559956 -0.73423935 0
#>   [3,] -1.45357655 -0.815469131  1.767194256  0.10379193 1
#>   [4,] -0.33266976  0.300766397  1.575587897 -0.62068087 1
#>   [5,]  1.23813125  0.248165447  0.316924855  0.02265194 1
#>   [6,] -0.84273188  0.339317527 -1.083367759  1.72778213 0
#>   [7,]  0.28008701 -0.214155229 -0.375316135  0.13581867 1
#>   [8,] -0.19777759 -0.847771883  0.126918138 -1.00135772 0
#>   [9,]  0.51138824  0.272054165 -3.075281455 -1.27908498 0
#>  [10,]  0.01319499  0.249730095  1.418228332 -0.29651229 1
#>  [11,]  1.01268197 -1.171692016  0.757535273  0.91860440 0
#>  [12,]  1.12008102 -0.959741439  1.885191788 -2.00615991 0
#>  [13,]  1.45392774 -0.474759983  0.427798364  0.49102187 1
#>  [14,]  0.88899003 -2.251279941 -1.106936108 -0.34332840 0
#>  [15,]  0.31363377  1.004821146  0.774716185 -2.44884867 0
#>  [16,]  0.58426196  0.049536244 -0.308982106 -0.60443055 0
#>  [17,] -0.55440086 -1.786159842 -0.292495213  0.69691353 0
#>  [18,] -1.78432621  0.764196519 -0.645607895 -1.73068447 0
#>  [19,]  1.42134082  1.293393165 -0.461681878 -1.24401257 0
#>  [20,]  1.85561240 -2.218947280 -0.611256587 -0.21816720 0
#>  [21,] -0.34192460  1.464851113  1.540004353 -0.55696752 1
#>  [22,]  0.10396343  0.744239852  0.329111960 -1.78840989 0
#>  [23,] -0.59602361  0.056232767  0.183351803 -0.14181869 0
#>  [24,] -0.11742708  2.065774776 -0.918785123 -0.88062765 0
#>  [25,] -0.23515852 -1.239452479 -1.282556331 -0.86102903 0
#>  [26,]  0.84776396  0.303420998  0.753569188 -0.31895848 0
#>  [27,] -0.96168181  1.247598513 -0.209527232 -1.22084892 0
#>  [28,] -1.46864216  1.171950429 -0.233178547  1.45665361 0
#>  [29,] -0.64545807 -0.575333623  0.989189133 -0.31954071 1
#>  [30,]  0.35547544  1.451645982  0.353492879  0.27093159 1
#>  [31,]  0.73401733  0.110112230  0.002868081 -1.18671137 0
#>  [32,]  0.71818517  0.800539366 -0.029231585  0.31831590 0
#>  [33,]  1.94722998  1.304931183  0.189869526  0.64108332 0
#>  [34,] -0.02518654 -0.603315918 -0.777641866  1.09355584 0
#>  [35,]  0.19799781 -0.399876045 -1.792139306  0.95740812 0
#>  [36,] -1.31775213 -1.930524955 -0.245281991  0.29721312 0
#>  [37,]  1.02334613  0.588774615  0.699987454 -0.68362524 1
#>  [38,] -0.56717157 -1.262826014 -0.022804525 -0.25223795 1
#>  [39,] -0.55065751 -0.219097318 -0.729212813  1.84493507 0
#>  [40,]  0.07211109 -1.275845418 -0.701893009  0.31386396 0
#>  [41,]  0.52002277  0.662241878  0.710894849 -1.19308051 0
#>  [42,] -0.30212063  0.252370097 -0.315717556  1.04091725 0
#>  [43,]  0.47334357 -0.829789719 -0.021276594  1.57023606 0
#>  [44,]  0.21694070 -0.103449697  1.299416073 -1.00689892 1
#>  [45,] -0.89296832  0.190511521 -0.046185479 -0.44582637 0
#>  [46,] -0.23209893  1.214149159 -0.220072342  0.28550354 0
#>  [47,]  0.24055587 -0.187895638  0.213651294  0.66269152 1
#>  [48,]  0.22252403 -0.364567911 -0.466256951 -0.27892894 1
#>  [49,]  2.42392042 -0.185840147  0.960936971  0.52202831 1
#>  [50,]  1.04491331  1.739548087 -0.676359945 -1.16781978 0
#>  [51,]  0.50695975  1.343908565  0.077909533  1.32008397 0
#>  [52,]  0.20397807 -0.477751678  0.079969811 -0.96086174 1
#>  [53,] -1.23892805  0.585607595 -2.053475981 -0.57589800 0
#>  [54,] -0.31405265 -1.531843636 -0.778148683  0.87758764 0
#>  [55,] -0.51478997  0.223876761  0.034774149 -0.32538025 0
#>  [56,]  0.96014242  0.507432280  0.018424615 -1.77251129 0
#>  [57,] -0.41286379  0.176526168  0.511165562 -0.45197680 0
#>  [58,] -2.42618736  0.338162946  0.027474550 -0.71032054 0
#>  [59,] -1.89155799  0.254968212  1.822415498  0.79783017 1
#>  [60,] -0.55277616  0.296452797  1.060529309 -0.05654354 1
#>  [61,] -0.09933234 -1.676943747  0.032687225 -0.82754814 0
#>  [62,]  0.61258487  1.033628336 -0.065621420 -0.22178464 0
#>  [63,] -0.53715767 -1.103622088 -0.415196210  0.21492128 0
#>  [64,]  0.82664706 -0.785799948 -0.513338179 -1.75202739 0
#>  [65,]  2.31275814 -0.039829060 -0.970276811  0.56736394 0
#>  [66,] -0.26492992 -1.641402529  0.993124466 -0.87354710 0
#>  [67,] -0.90341210 -0.492866776  1.001564497  0.13856260 1
#>  [68,]  0.72816250  0.710538206 -0.654626706  0.79313567 0
#>  [69,] -0.69453432  0.874351295 -0.021002932 -0.22540037 1
#>  [70,]  2.36755040 -0.175849952  0.237277001  1.47496813 0
#>  [71,] -0.90748154 -0.891780566 -0.994936757 -0.59307842 0
#>  [72,] -0.03750922 -0.106623160 -0.462590171  0.90760087 0
#>  [73,] -1.44571654 -2.195685730  1.129289450 -0.34861839 1
#>  [74,]  0.55726181 -0.943028256  0.527314365 -0.32043498 0
#>  [75,] -1.20659020  0.757669339 -0.299337785  0.21536844 0
#>  [76,] -0.34942767  0.490916643 -0.430258563  2.22133863 0
#>  [77,]  1.38511295  0.670269777 -0.013458885  1.49938141 0
#>  [78,] -0.11170008  0.016710731 -0.689358749 -0.89862557 0
#>  [79,] -1.05845544 -0.637041496 -0.031770989  0.50933307 0
#>  [80,] -0.20856121  1.793309046  0.426070485 -0.95473410 0
#>  [81,] -2.13666045 -0.711191543  1.260761731 -0.16923968 1
#>  [82,] -0.16490492  0.863639197 -1.151345471 -0.72968650 0
#>  [83,] -0.29286235  0.153558681  1.146471571 -0.31404961 0
#>  [84,]  1.34155335  0.744119957  0.311525588  1.19050359 0
#>  [85,]  1.80108764  0.612435559 -0.549321295  0.80032879 0
#>  [86,] -1.88400663 -1.635150575 -1.470915941 -0.34257092 0
#>  [87,]  0.85205803 -0.351479568 -0.559761083 -0.07115129 0
#>  [88,]  1.08723076  0.876575815  1.510051904  1.01962160 1
#>  [89,] -0.55475089  0.694800947  0.279654807  0.32023109 1
#>  [90,] -1.26676182  0.653717000 -0.550811356  0.44583805 0
#>  [91,]  2.26321039  0.696941509 -0.052150985  0.68106993 1
#>  [92,]  0.30645699 -0.178293817  0.826592835 -0.79174673 1
#>  [93,] -1.01116781  0.004827452  0.329772846  1.49843134 0
#>  [94,] -0.91304431 -0.437286379  0.649474651  1.15861007 1
#>  [95,] -0.64696050 -0.287221784  1.904428476 -0.61299165 1
#>  [96,] -0.68577245  1.786021599  1.308771203 -0.96474476 1
#>  [97,] -0.76685324 -0.840705399  1.154732892  0.04231792 1
#>  [98,]  0.71939212 -0.666751778 -0.519871583  1.18441894 0
#>  [99,]  0.48792332 -0.663844795  2.299685922  0.20599319 1
#> [100,]  0.37165490 -1.307642024 -0.321557747 -1.19481826 0
#> 
#> $eta
#>   [1]  -4.6331366  -3.7178530   4.4501215   0.9666380   3.4270501 -18.5066877
#>   [7]   2.1521363 -14.0757651 -29.9038948   5.0601780  -0.7440588 -10.7662008
#>  [13]   5.0576809  -9.9966217 -25.7971543 -11.6347512  -6.6665647 -13.8684594
#>  [19] -10.5377537  -8.1437252   5.0050159 -10.4351741  -1.6205440 -18.2385708
#>  [25] -16.1994430  -2.2276607 -10.9877677 -17.8320256  10.1343635   6.9713937
#>  [31] -17.4256290  -6.3566834  -4.6602884 -11.2960916  -8.1338040 -11.9590657
#>  [37]   5.0873434   2.2872801 -20.1251375  -1.5551761  -1.5067352 -15.4113393
#>  [43] -16.4911024   4.4858169  -2.7232374 -10.8497143   3.0155727   3.9953210
#>  [49]   5.5900548  -8.8409041  -9.6045355   2.8992468 -16.4699251 -11.5610401
#>  [55]  -3.8222642 -14.7415246  -5.3377094  -5.8789104   7.5938761   5.6484662
#>  [61]  -7.8138377  -9.2233891  -4.6953839 -27.0116875  -7.2409464  -3.6093751
#>  [67]  10.8576511 -10.8055521   1.7088793  -7.1205237 -10.4181333 -11.1225994
#>  [73]   3.6278407  -4.5695984  -6.5967246 -29.4136790  -6.1496229 -16.5938632
#>  [79]  -3.9503280  -7.4289840   1.1230443 -16.2220180  -3.5096546  -0.4759952
#>  [85]  -8.8281150  -6.7031239  -5.2600839   3.8033427   4.3482404  -7.1972301
#>  [91]   3.8114219   5.2116910 -14.6652787   2.2604254  14.7149987   7.3783870
#>  [97]   6.8827906 -16.4490448  12.6302552  -3.9212022
#> 
#> $mu
#>   [1] 9.630560e-03 2.371023e-02 9.884576e-01 7.244489e-01 9.685393e-01
#>   [6] 9.175878e-09 8.958682e-01 7.708548e-07 1.030157e-13 9.936956e-01
#>  [11] 3.221172e-01 2.110033e-05 9.936799e-01 4.555149e-05 6.258039e-12
#>  [16] 8.852948e-06 1.271146e-03 9.484276e-07 2.651552e-05 2.904684e-04
#>  [21] 9.933404e-01 2.937979e-05 1.651299e-01 1.199746e-08 9.218733e-08
#>  [26] 9.729390e-02 1.690697e-05 1.801564e-08 9.999603e-01 9.990625e-01
#>  [31] 2.704864e-08 1.732106e-03 9.375010e-03 1.242122e-05 2.933637e-04
#>  [36] 6.400899e-06 9.938635e-01 9.078181e-01 1.818712e-09 1.743399e-01
#>  [41] 1.814231e-01 2.027404e-07 6.886606e-08 9.888579e-01 6.161602e-02
#>  [46] 1.940978e-05 9.532727e-01 9.819310e-01 9.962791e-01 1.446709e-04
#>  [51] 6.741770e-05 9.478092e-01 7.034000e-08 9.530154e-06 2.140980e-02
#>  [56] 3.961294e-07 4.783873e-03 2.790026e-03 9.994967e-01 9.964895e-01
#>  [61] 4.039410e-04 9.869386e-05 9.054624e-03 1.857690e-12 7.161200e-04
#>  [66] 2.635535e-02 9.999807e-01 2.028615e-05 8.466909e-01 8.076904e-04
#>  [71] 2.988472e-05 1.477441e-05 9.741144e-01 1.025585e-02 1.362971e-03
#>  [76] 1.681905e-13 2.129741e-03 6.214078e-08 1.888488e-02 5.934381e-04
#>  [81] 7.545530e-01 9.012952e-08 2.903877e-02 3.831982e-01 1.465328e-04
#>  [86] 1.225569e-03 5.168022e-03 9.781902e-01 9.872355e-01 7.480966e-04
#>  [91] 9.783619e-01 9.945771e-01 4.275139e-07 9.055460e-01 9.999996e-01
#>  [96] 9.993758e-01 9.989758e-01 7.182416e-08 9.999967e-01 1.943216e-02
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                 x1           x2           x3            x4   y
#>   [1,] -0.20082203 -2.761269884  2.318847850  0.7200622397   9
#>   [2,] -0.83097306 -0.985698163 -0.816583768 -0.4404615688   2
#>   [3,]  0.99533022 -1.615409414 -0.547853676 -0.7097845483   1
#>   [4,] -2.65615456 -1.031949254  1.277638300  1.1834083393   1
#>   [5,] -0.29966223 -1.043687441 -0.732116644 -1.4640447538   0
#>   [6,] -0.19159398  0.054815557  1.598874177  0.9032486516   1
#>   [7,]  1.51431806  0.246598992 -1.082500982 -1.5208563687   0
#>   [8,]  1.21792864 -1.331834847 -0.012572291 -0.5637639498  19
#>   [9,]  0.74638877  0.656247168  2.485103510 -0.0434109218 296
#>  [10,]  1.64343555 -0.518399693 -1.388371309 -2.2628197332   0
#>  [11,] -0.52552685  1.454109991  1.228675012 -0.2460105427  29
#>  [12,] -0.10993976  0.704570071  0.802359507 -0.3552202872   3
#>  [13,]  0.20124262  1.721302903  0.342122916 -1.8688884722   0
#>  [14,]  1.50021421 -0.613608060  0.244679160 -0.4757138837   6
#>  [15,]  0.04441777  0.143237133  1.570542529 -0.1547914024  11
#>  [16,]  2.64812543  1.113068089  0.211877677  0.5077025079   3
#>  [17,] -2.24066910 -1.313819376 -0.345043054 -0.5085492585   1
#>  [18,] -0.88931727  1.150332822  1.122576422 -0.8218940377   2
#>  [19,] -1.93933169  0.629056037  0.757899887  0.0239487232  16
#>  [20,]  0.80130229  1.449832385  1.868671898  1.8342822185   0
#>  [21,]  0.58030360 -0.194718023  0.683395339  2.8438688449   0
#>  [22,] -0.02249502  0.572704433  0.824042936 -1.4710187124   0
#>  [23,]  0.65192627 -0.081933614 -0.769888679  0.5588104280   1
#>  [24,]  0.61806034  1.408678354 -0.785876808  1.6200471399   0
#>  [25,]  0.46755047  0.521424686 -0.190847129  1.3006018855   0
#>  [26,]  0.82120396  0.172199465 -0.848495846 -0.0318132119   0
#>  [27,]  0.80784806  0.010535541  1.852860713 -0.2628785866   9
#>  [28,] -0.68782729 -0.207714292 -0.165124534 -0.5023129722   6
#>  [29,]  2.01373137 -0.018562370 -0.132614867 -0.8056382008   2
#>  [30,] -1.60052113 -0.867848897  0.822632296 -0.4725510344   4
#>  [31,] -0.13901278  0.211185061 -2.129964831  1.3072829850   0
#>  [32,] -0.31322532 -0.414671408 -1.773926828  0.0622227534   2
#>  [33,]  0.39528999 -0.389572394 -1.092031236 -0.8024554678   2
#>  [34,]  0.14756599  0.443910519 -0.560324018 -0.0753597634  10
#>  [35,]  0.61018608  0.218097080  0.368549075  2.0954175391   0
#>  [36,] -1.09896081  1.355816272  1.703533603  0.8855182110   3
#>  [37,] -0.42794626 -1.313000222  0.826811486  1.2042916854   0
#>  [38,]  0.18966891 -2.343400644 -0.140554830  2.5800497313   0
#>  [39,]  0.76713288  0.220830566  0.546479766  0.9766162036   0
#>  [40,] -0.56111480  1.283887050 -0.324292323 -0.8030797002   0
#>  [41,] -0.12015588 -1.191582482  1.523974324  0.1675033624   6
#>  [42,]  0.39478850  0.691304353 -0.035502768  0.2331218375  22
#>  [43,] -0.07890483 -0.233723614  1.474592060 -1.1015131976   0
#>  [44,]  0.16420204 -1.461952961  0.316533462 -0.0522589416  33
#>  [45,]  0.92168048  0.215193716 -0.917386524 -0.2603870756   0
#>  [46,]  0.19157019 -0.605765007  0.320375544  1.7629957749   0
#>  [47,] -0.32327081 -0.730534279 -0.899324806  0.7642487035   0
#>  [48,]  1.17727468  0.083392547  1.508294792  0.0520082096  15
#>  [49,]  1.20080514  1.002631036  0.892207182 -0.9489325812   4
#>  [50,]  0.70796057  0.097862105  1.001586725  1.0154873526   0
#>  [51,] -1.60363510  0.901693653  0.291987235 -1.2018556153   0
#>  [52,]  2.68640201 -0.947000000 -0.007864076 -0.3922315317   1
#>  [53,] -1.74165698 -0.263574257 -0.211985988  1.1328787586   4
#>  [54,] -0.36962521 -0.550819469  0.040951548  1.3909874893   0
#>  [55,] -0.60227914 -0.707366629 -0.298690323  0.8402443836   0
#>  [56,]  0.78538094 -0.163858150  0.168012272  0.6729909593   0
#>  [57,] -1.88431734 -0.593178826  0.514835210 -1.1818795922   2
#>  [58,]  0.17340709 -1.371740648 -0.073195638  0.9214017870   9
#>  [59,] -0.31689998 -0.195320327  0.693218275 -1.1588614882   0
#>  [60,] -1.22335385  0.579604193  1.243738118 -0.3611299996   7
#>  [61,] -1.95759259 -1.742625369 -0.980051931 -1.0677612286   0
#>  [62,]  2.86281257 -0.267638563 -0.636346289 -0.5864609760   2
#>  [63,] -2.17490236  1.295794405  0.193160732  0.7261439003   0
#>  [64,]  0.78978941  0.972261412 -0.583726059 -1.1788188741   0
#>  [65,]  0.08597671  0.763117269  1.156227398  1.2390045128   3
#>  [66,]  0.26786743 -2.014667869 -0.564581647  1.5636290710   0
#>  [67,]  0.24996706 -1.418696905 -1.220342657  0.9946893428   3
#>  [68,]  0.31071721 -0.079067099 -0.125956566  1.3139243833   0
#>  [69,] -0.41346239  0.638889054  1.215398262  1.2567543156   1
#>  [70,] -0.44623782  1.775948942 -0.754224121  0.9580946826   0
#>  [71,] -0.12652619  1.479249081 -0.155228270 -2.5617303569   0
#>  [72,] -1.93918179 -0.277884177  0.410973173  0.4262693742   5
#>  [73,] -1.79749669 -1.119729302 -0.178235622  0.7659695801   2
#>  [74,]  0.98215413 -0.566449372  1.015328935 -0.0002855235  13
#>  [75,]  0.07038413  0.696260916  0.002727530  0.4726787602   9
#>  [76,]  1.97706799 -0.427687210 -0.113113277  0.5324372766   5
#>  [77,] -1.07764383  0.205441136 -2.000405757 -0.6996684231   0
#>  [78,]  0.56302212  0.005456475 -1.828573859 -0.1298609862   1
#>  [79,]  0.17311259 -0.276450512  0.387719200 -1.1544635938   5
#>  [80,] -0.82517010 -0.694667784 -1.959491015  0.0166435506   8
#>  [81,] -2.45015603  1.306595255 -0.004934676  1.6152862835   1
#>  [82,]  0.74970461  0.779263229  0.876587860  0.3421527669   2
#>  [83,]  1.14114805  0.876079075  1.248182390  1.3464632409   1
#>  [84,] -0.19947004  0.734353620  0.010539654 -0.1867621344   5
#>  [85,]  1.77984890  1.199800549 -0.330930566 -0.5826520913   0
#>  [86,]  1.96930641  0.058111277  1.395185941 -0.7018639409   2
#>  [87,] -0.95868627  0.866585318 -2.250645465 -1.6220274674   0
#>  [88,]  0.20834674  1.423370645  1.153232054  0.9364587424   9
#>  [89,]  0.32782654 -1.272929928  0.031256537 -0.0218179073  22
#>  [90,] -0.84101241 -0.415597345 -1.026177467 -0.5323303248  14
#>  [91,] -1.87888471 -0.680088324  0.398164678  0.0327809020   9
#>  [92,] -0.07964825  0.651102762 -0.188181823 -0.3865493959   3
#>  [93,]  1.24979918 -0.059372360 -0.887738344  1.0026585101   0
#>  [94,] -1.22616543 -0.131925579  0.416601520  0.8370724446   0
#>  [95,] -1.02544665 -0.787272045 -1.316041267  1.8107587268   0
#>  [96,]  0.76820932  0.073624171  0.049606759  0.9744087748   0
#>  [97,]  1.43590721  2.004203670 -0.341212510  1.6101319056   0
#>  [98,]  0.06355516  1.863956011 -1.581325685 -0.4057041919   2
#>  [99,]  1.38767126 -1.714781335  0.698065731 -1.3534303602   0
#> [100,]  1.44732974 -0.489278707 -1.111742090 -1.0624740378   1
#> 
#> $eta
#>   [1]   2.535649256   0.745042231   0.533166125  -0.298875328  -4.760981661
#>   [6]  -0.574070438  -4.078418848   2.588157861   5.683628172  -9.108583882
#>  [11]   3.110367297   1.709963792  -3.839841311   1.406373592   2.230191496
#>  [16]  -0.801104580   0.551690642   0.616011957   3.176458431  -3.400090890
#>  [21] -14.714307100  -1.557715509  -0.783170288  -4.265634394  -0.916163842
#>  [26]  -0.567767425   1.570962888   1.991316077  -0.677446626   1.108480489
#>  [31]  -3.814134762   1.137344087   1.439082819   2.675147629  -8.486182798
#>  [36]   1.210671813  -0.642913585 -10.151586820  -1.842242656   0.090815337
#>  [41]   2.227358222   2.654864768   0.232655030   3.367808900  -0.106654860
#>  [46]  -3.336715073  -1.270078133   2.699888768   0.242477857  -1.932114643
#>  [51]  -1.184943402  -0.818131261   0.307828979  -2.500823476   0.253956668
#>  [56]  -0.449278566   0.006695591   1.513806410  -1.643011055   2.140574754
#>  [61]  -1.305061294   0.338231620  -0.670741517  -3.306165336   0.120769499
#>  [66]  -3.498464059   1.139179031  -1.715023101  -0.396181312  -0.648533083
#>  [71] -11.688489318   2.237857003   0.803921897   2.600010612   2.125278255
#>  [76]   1.645611266  -1.096947187  -0.248683239   0.427431993   1.794589930
#>  [81]  -4.272253264   1.037833846  -0.745272291   0.947701808  -1.102138038
#>  [86]   0.446485673  -4.193574731   2.035200176   2.895009129   2.417041397
#>  [91]   2.222430403   1.453434022  -0.488508981  -1.023000406  -6.279605978
#>  [96]  -2.471364322  -4.398714078   0.927558289  -1.486093602   0.106684979
#> 
#> $mu
#>   [1] 1.262462e+01 2.106530e+00 1.704320e+00 7.416519e-01 8.557205e-03
#>   [6] 5.632282e-01 1.693422e-02 1.330524e+01 2.940142e+02 1.107114e-04
#>  [11] 2.242928e+01 5.528761e+00 2.149701e-02 4.081129e+00 9.301647e+00
#>  [16] 4.488329e-01 1.736186e+00 1.851529e+00 2.396174e+01 3.337024e-02
#>  [21] 4.070592e-07 2.106167e-01 4.569550e-01 1.404296e-02 4.000508e-01
#>  [26] 5.667894e-01 4.811279e+00 7.325168e+00 5.079122e-01 3.029751e+00
#>  [31] 2.205679e-02 3.118475e+00 4.216826e+00 1.451449e+01 2.062992e-04
#>  [36] 3.355738e+00 5.257583e-01 3.901412e-05 1.584617e-01 1.095067e+00
#>  [41] 9.275330e+00 1.422306e+01 1.261946e+00 2.901488e+01 8.988358e-01
#>  [46] 3.555356e-02 2.808097e-01 1.487808e+01 1.274403e+00 1.448416e-01
#>  [51] 3.057635e-01 4.412555e-01 1.360468e+00 8.201743e-02 1.289116e+00
#>  [56] 6.380883e-01 1.006718e+00 4.543994e+00 1.933968e-01 8.504324e+00
#>  [61] 2.711559e-01 1.402465e+00 5.113293e-01 3.665647e-02 1.128365e+00
#>  [66] 3.024380e-02 3.124202e+00 1.799596e-01 6.728847e-01 5.228121e-01
#>  [71] 8.389838e-06 9.373223e+00 2.234286e+00 1.346388e+01 8.375228e+00
#>  [76] 5.184178e+00 3.338888e-01 7.798270e-01 1.533315e+00 6.017007e+00
#>  [81] 1.395031e-02 2.823095e+00 4.746051e-01 2.579774e+00 3.321602e-01
#>  [86] 1.562810e+00 1.509224e-02 7.653784e+00 1.808367e+01 1.121264e+01
#>  [91] 9.229736e+00 4.277779e+00 6.135405e-01 3.595146e-01 1.874139e-03
#>  [96] 8.446954e-02 1.229314e-02 2.528328e+00 2.262548e-01 1.112584e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                  x1          x2          x3            x4            y
#>   [1,]  1.152057136  1.04921251  0.25167267  0.0103275091  -0.58360068
#>   [2,] -1.945122794  1.03431448 -0.06964815  1.2324512455 -10.81161404
#>   [3,] -0.177569614  0.44073944 -1.53133168  1.0935012044 -21.09409536
#>   [4,] -0.582461894  1.09385711  0.63636438  0.0057244462   1.59480232
#>   [5,]  0.132763926  0.23336901  0.88531827 -0.1026758025   5.92429086
#>   [6,]  1.072274401  0.75827780  1.63414434 -1.1130552945   5.55457884
#>   [7,] -0.304096678  0.67887116 -0.24645815 -0.4792634550  -5.34410616
#>   [8,]  0.890515660 -0.17255254 -1.39207514 -0.2099913054 -13.55527687
#>   [9,]  0.202414177  1.70115550  0.89156547 -0.0003391099  11.88402337
#>  [10,] -1.615914853  1.33415459 -0.95049534 -0.5577290496  -5.73852321
#>  [11,] -0.064264610  1.92167737 -0.10733687  1.4606867026 -17.55136763
#>  [12,]  1.537415776  0.59455268  0.66091054 -3.0782913456 -41.33296527
#>  [13,]  1.250254250  0.76367517  1.16061076 -1.3106802187   0.64154679
#>  [14,]  0.785308105  0.90635405  0.31477045 -0.3899678634  -6.80155006
#>  [15,]  0.241186736 -1.35651758 -0.47084657  1.0940656693   0.06247738
#>  [16,]  0.324726241 -0.09910106  0.68970053 -0.7859488274   1.96979070
#>  [17,]  1.802275233  0.47456945  0.54951244  0.2976051217   0.75393193
#>  [18,]  0.112663870  0.49205657 -1.12206835 -0.7737918424  -3.80215796
#>  [19,]  1.088480300  1.60491076 -1.03736455 -0.0346662410   0.42900277
#>  [20,]  2.132185614  1.08957750  0.71586380  0.1666794628   2.92326609
#>  [21,]  0.814634854 -0.65026059  0.86436024 -0.0357089245  -1.36650038
#>  [22,] -0.994850793 -0.07945432  1.22851020 -0.1629232308   4.57326609
#>  [23,]  0.081969956 -0.23362658  0.57748618  1.1934335578  -4.33526514
#>  [24,] -0.020868562 -0.61513412  0.29102230  1.2430916490  -6.50994175
#>  [25,] -2.533530794  0.99077655  0.36511139  0.1273564786  -1.21348150
#>  [26,]  0.628507463  0.36619499 -0.46668024 -0.2126650842  -5.56120256
#>  [27,] -1.220287391  0.19799821 -0.24024934 -0.6559508064 -12.57246277
#>  [28,]  0.194411717 -1.34417607 -0.06954448 -0.0997313412   5.76553030
#>  [29,]  0.134701500 -0.69848602 -1.03453241 -0.8432880555  -7.10971234
#>  [30,]  0.790451705 -0.31036182  0.57316719  0.3254012976   0.61750021
#>  [31,]  0.542146934  1.22630721 -0.37234695 -0.6904494728  -7.29180332
#>  [32,]  2.180961359 -0.52170961 -1.63829448  1.3381989725 -14.39936670
#>  [33,]  0.124303695 -1.97135452  0.11178449  0.7396013118  -3.92652227
#>  [34,]  0.922953016 -0.01101469 -0.34037237 -0.0575862842  -9.46761306
#>  [35,]  0.411438678 -1.72844306 -0.69106347 -0.3809381117  -4.21258948
#>  [36,]  0.441159202 -1.75907035 -0.70232759 -0.1137848529  -6.41290696
#>  [37,]  1.769955122 -0.66356921  2.07060877  0.7889096744   3.42053687
#>  [38,]  1.593151372  1.43814339 -0.72512934 -2.2103687163 -29.16480441
#>  [39,]  0.961705970  0.07657237 -1.08552543  0.7305793067 -14.54910180
#>  [40,]  0.180400948 -0.32481270 -0.46513746 -0.3403460448   2.77899934
#>  [41,]  0.363191777  0.27197456 -0.38800920 -0.0277619923  -0.83174520
#>  [42,]  1.165594661  1.29898335  0.46387066  0.1051418526   6.00795637
#>  [43,] -0.458429599 -0.80108520  0.47665591  0.3487466199  -2.78622440
#>  [44,]  1.128614251 -0.11381784  1.02379428  0.0776963699   7.69867584
#>  [45,] -0.305940863 -1.39873250  1.00987360 -0.9257308272   1.29883009
#>  [46,]  0.293594267  1.77053235  0.22154911  0.8562040341   3.31357141
#>  [47,]  0.240840263  1.14920881 -0.99599579  0.7389283890  -8.35671599
#>  [48,]  0.638123998 -0.43875968  0.63204800 -0.6882970302  -0.30633284
#>  [49,] -0.148086891  1.60070819 -0.09941043 -2.6031842774 -37.97836983
#>  [50,]  1.119872588  0.32916871  0.50365653 -1.3355114173  -3.71911003
#>  [51,]  0.129062930  0.22266307 -0.85325417  0.3999427842  -5.71745229
#>  [52,]  0.241909518  0.32806962  1.16667488 -0.4804625850   8.22819255
#>  [53,]  0.037102428  0.02046152 -1.03447999 -0.4379052463 -12.23315499
#>  [54,] -1.019205965  0.33603627 -1.23713336 -0.3292122308 -11.18785796
#>  [55,]  0.055175255  0.68459720  0.22608906  1.9063957498 -14.57848003
#>  [56,]  0.825910916  0.40797970 -1.56756612 -0.8921658947 -17.13413673
#>  [57,]  0.485572987  0.97199660 -0.55042599 -0.7082417292  -9.13880771
#>  [58,] -0.218150146 -0.57651611 -1.00183866  2.3192155056 -36.91427110
#>  [59,]  1.090220008  0.65927683 -1.06192459 -0.4272692840  -4.12575828
#>  [60,] -0.142989410 -0.94491159 -1.37854886 -0.1267121493 -19.23155137
#>  [61,]  0.570082216 -2.19749128  0.38307714 -0.4149160830  -1.16919369
#>  [62,]  0.551329745 -0.30609299  0.58405100 -0.0373739068   5.35448124
#>  [63,] -0.574336332 -0.44213398  0.32506930 -1.0372959303   2.56990911
#>  [64,]  0.141716210  0.20544928 -1.92711124  0.6508636894 -14.07861363
#>  [65,] -0.791341177  0.09633152 -1.51629375  0.0810789667 -12.22389123
#>  [66,]  2.377765007  0.78759516 -1.94638827  0.1170149918  -8.27726192
#>  [67,] -1.365967691  0.89251316 -1.22204960  0.7381889430 -16.54365118
#>  [68,] -1.027109074  0.15333812  1.63188192  0.7831594326   2.06965034
#>  [69,] -0.078565921  0.76574391 -0.01683359  1.4145087140 -12.32766957
#>  [70,] -0.660985653 -0.54467081  1.56373589  0.0849358041  14.32610568
#>  [71,]  0.330495409 -1.23472456  0.05802387 -1.0909599426  -1.77570983
#>  [72,]  0.537616211 -0.63046397  0.19528439 -0.6394376848  -1.84815882
#>  [73,]  0.609159767 -1.43284886  0.15584416  0.0697436469   0.50730505
#>  [74,] -0.138402416 -1.06403903 -0.23332527  0.6204817783 -11.69845569
#>  [75,]  0.387166801 -0.97010493 -1.21167441  1.9075198057 -30.22767615
#>  [76,] -1.625817350  1.64786257  0.78083886  1.1882392437   3.14039012
#>  [77,] -1.368882163  0.82403958 -0.42934968  1.0191518171 -12.68161241
#>  [78,] -2.046985137  0.40742688  1.34735075 -0.1885578323   7.37973718
#>  [79,]  1.337139234  0.41433411  0.06793294 -0.2536060896   3.86047720
#>  [80,] -0.429878942 -0.53096148 -0.76770186  0.8227192495  -9.71273661
#>  [81,] -1.545190199 -0.61352294 -0.28192528 -0.5807919022  -2.05070770
#>  [82,]  3.818730328  0.73330809  0.79220141  1.0735330606  -3.44597007
#>  [83,]  0.558794889  0.84130694 -0.55955235 -0.3315423684  -7.54108817
#>  [84,] -0.460217593 -0.80789734  0.64760466  1.8599792260 -20.03218609
#>  [85,] -0.386565928 -1.37697605  0.28390579  0.2213390932   1.32071365
#>  [86,]  1.308232518  0.50987976 -0.12669072  0.1654462773   5.76351770
#>  [87,] -0.720580590 -1.07695097 -2.91893827 -0.0907427077 -16.40490838
#>  [88,] -0.650113578  0.42250404  0.06901337 -0.2006804415   5.84301665
#>  [89,]  0.128845180  0.12956254 -0.41454578  0.8264856148  -8.17166898
#>  [90,]  0.142995612  0.28227833  0.20838652 -1.3776182279  -8.06253809
#>  [91,]  0.563492545 -2.12321721  0.38362568  0.3828735886  -1.17965404
#>  [92,] -2.004978102 -0.18018793  0.14004550  1.2399168744  -6.42485488
#>  [93,] -0.378868885 -1.43344499  0.74679740  0.6253031764   1.33204046
#>  [94,]  0.356385649  0.12575348  0.78169388  1.7109792681 -10.98562474
#>  [95,]  0.136895049 -0.31649331 -0.41394530 -0.7551825721  -0.73848679
#>  [96,] -0.006956257 -0.04250093 -1.05231862  0.3331220468 -11.93359116
#>  [97,] -0.188411118  0.54101536  1.13211323  1.2472692764  -4.42246726
#>  [98,] -0.636642239  2.18311499 -1.03777985 -0.8778701103 -10.77149242
#>  [99,] -0.131607620 -0.90486090 -1.55494752 -1.3494862069 -27.41383508
#> [100,]  0.776980132  0.27627703 -0.59583288  0.7655282388 -13.82098497
#> 
#> $eta
#>   [1]   0.16481554 -11.66183528 -18.58568466   0.81036496   5.18954018
#>   [6]   6.03635028  -5.59468700 -13.19402225  11.51614629  -5.31519736
#>  [11] -16.96424467 -42.09627218   3.48362317  -7.10484695  -1.63578030
#>  [16]   2.27544653  -0.21798227  -4.56473989  -1.14205980   2.35566680
#>  [21]  -0.40816061   3.73626718  -1.14661940  -6.77220909  -1.73816803
#>  [26]  -7.22659810 -12.94887430   6.28080761  -8.22018417  -0.34871382
#>  [31]  -9.84348100 -12.52978855  -3.84123407  -9.25519091  -5.15721593
#>  [36]  -6.08652282   3.02319969 -30.51979823 -15.87263833   2.95916538
#>  [41]  -1.46034775   4.77695237  -3.46660952   6.83465787  -0.45704383
#>  [46]   2.42762832  -8.83757250   0.08227201 -36.02397402  -3.98647116
#>  [51]  -6.28413603   9.30147309 -11.23783447  -9.90433064 -13.45600447
#>  [56] -17.43543540  -9.97842054 -37.07950131  -2.59051499 -17.71606169
#>  [61]  -1.51651028   3.42412188   1.28469553 -14.10696744  -9.77622334
#>  [66]  -8.73027484 -16.82813823   0.79899463 -12.41272248  14.92625011
#>  [71]  -1.15633316  -2.54972231   0.23182161 -11.12433020 -27.51469777
#>  [76]   3.31656367 -12.79908879   6.83107568   4.88602792  -8.54948186
#>  [81]  -2.07768174  -4.77548426  -7.50681214 -19.19002554   0.54086624
#>  [86]   5.89456606 -16.35643846   4.96494214  -7.58709684  -6.99830789
#>  [91]  -3.04247533  -8.25778511   1.56382889 -11.35596612   0.13152690
#>  [96] -11.46709994  -3.23283801 -10.97849945 -27.59537454 -14.24817881
#> 
#> $mu
#>   [1]   0.16481554 -11.66183528 -18.58568466   0.81036496   5.18954018
#>   [6]   6.03635028  -5.59468700 -13.19402225  11.51614629  -5.31519736
#>  [11] -16.96424467 -42.09627218   3.48362317  -7.10484695  -1.63578030
#>  [16]   2.27544653  -0.21798227  -4.56473989  -1.14205980   2.35566680
#>  [21]  -0.40816061   3.73626718  -1.14661940  -6.77220909  -1.73816803
#>  [26]  -7.22659810 -12.94887430   6.28080761  -8.22018417  -0.34871382
#>  [31]  -9.84348100 -12.52978855  -3.84123407  -9.25519091  -5.15721593
#>  [36]  -6.08652282   3.02319969 -30.51979823 -15.87263833   2.95916538
#>  [41]  -1.46034775   4.77695237  -3.46660952   6.83465787  -0.45704383
#>  [46]   2.42762832  -8.83757250   0.08227201 -36.02397402  -3.98647116
#>  [51]  -6.28413603   9.30147309 -11.23783447  -9.90433064 -13.45600447
#>  [56] -17.43543540  -9.97842054 -37.07950131  -2.59051499 -17.71606169
#>  [61]  -1.51651028   3.42412188   1.28469553 -14.10696744  -9.77622334
#>  [66]  -8.73027484 -16.82813823   0.79899463 -12.41272248  14.92625011
#>  [71]  -1.15633316  -2.54972231   0.23182161 -11.12433020 -27.51469777
#>  [76]   3.31656367 -12.79908879   6.83107568   4.88602792  -8.54948186
#>  [81]  -2.07768174  -4.77548426  -7.50681214 -19.19002554   0.54086624
#>  [86]   5.89456606 -16.35643846   4.96494214  -7.58709684  -6.99830789
#>  [91]  -3.04247533  -8.25778511   1.56382889 -11.35596612   0.13152690
#>  [96] -11.46709994  -3.23283801 -10.97849945 -27.59537454 -14.24817881
#>