Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                 x1           x2          x3           x4 y
#>   [1,]  1.71527740  0.900485695 -0.12859619  2.954596222 0
#>   [2,]  0.81863768 -0.485637127  1.90236135  0.527579006 1
#>   [3,]  0.14137333 -0.139258394 -1.14347299  0.411368824 0
#>   [4,]  1.13587630 -1.024610954 -0.50104605  0.726587202 0
#>   [5,]  0.45832427  0.378110657 -1.67989462  0.053776797 0
#>   [6,]  0.53544240  0.415417657 -1.29964628  0.191509046 0
#>   [7,]  1.61171128  0.242903308 -0.27135450 -1.205431916 0
#>   [8,] -0.21863405  2.473661034  0.19157329  0.112096478 0
#>   [9,] -1.38782367 -1.434504475 -0.17065217  0.473153051 0
#>  [10,]  1.34686451 -1.129912499 -1.29336775  0.826352343 0
#>  [11,]  0.29029245 -1.388782881 -0.11949522 -0.825720016 1
#>  [12,]  0.55881140 -1.033127343  0.56483052  1.303509461 0
#>  [13,] -0.33733093 -1.525530688  1.91252004 -0.042764885 1
#>  [14,]  0.44276823 -0.576574237 -0.94746352  1.207860420 0
#>  [15,] -0.56443128  0.023948288 -0.80454830  0.051544857 0
#>  [16,]  0.96875367 -1.009784592 -0.02012031 -1.737584836 0
#>  [17,] -0.38902736 -1.970542551  0.39223840  0.409432781 0
#>  [18,]  1.33240569  0.553287406 -1.35526399  0.836503863 0
#>  [19,]  0.20116801  1.410024307  2.22129418 -0.280757640 1
#>  [20,] -0.23845068 -1.233356566  0.67046235  0.364585047 1
#>  [21,] -1.17004276  0.460352871 -0.10123860  1.151072447 0
#>  [22,]  1.41928480 -0.409649828 -2.42037818  0.786507073 0
#>  [23,] -0.91014616  0.992542311  0.28277724 -0.159704042 0
#>  [24,]  1.44201516  0.073468933 -1.67237644 -1.370446449 0
#>  [25,]  0.96631678 -0.584588367  0.32329570  1.261602994 0
#>  [26,]  0.71437158 -1.792564325 -1.31478832  0.002463102 0
#>  [27,] -0.59546914  1.337589052 -0.41893951  0.142409762 1
#>  [28,]  1.16412963 -0.727779051 -0.85712123 -0.560252426 0
#>  [29,]  0.29311478  1.055093984 -0.01549308  1.715164576 0
#>  [30,]  1.49891630  0.703011471  1.37810908  1.137134563 1
#>  [31,] -0.70591323  0.590532267 -0.28388922 -1.208850212 0
#>  [32,]  0.67940442  1.352060851 -0.11676639  0.840337992 0
#>  [33,] -2.72399290  1.358429419  0.14539765  0.640495624 1
#>  [34,] -0.28910156 -2.119439507  0.92612115 -0.576763578 0
#>  [35,]  0.93345971  0.714906306 -0.25479844 -0.042891318 0
#>  [36,]  1.28891406 -1.461375447 -1.10669373  0.959738589 0
#>  [37,]  1.46465291  0.702710039 -0.14236760  0.486316296 1
#>  [38,]  1.31232474 -0.871134916 -0.09228526  1.484777411 0
#>  [39,]  0.43560165  0.214870904  1.49807229 -0.794826006 1
#>  [40,] -0.11714825 -1.102159317 -0.31310134 -0.667378460 0
#>  [41,]  1.24218011 -0.349846691  1.35528586  0.827307131 1
#>  [42,]  1.26230763 -0.497298213 -0.07805580  0.695453684 1
#>  [43,] -0.13956093 -2.464628403 -0.21218869  0.634464850 0
#>  [44,] -0.73903133 -2.245568213 -0.24847821 -2.349633198 0
#>  [45,] -1.66274647  0.097371421 -1.64278943 -1.265363093 0
#>  [46,] -0.63345417  1.355629908 -0.86506865 -0.744144346 0
#>  [47,]  2.40600437 -0.806035006  0.20761947 -1.250594183 0
#>  [48,]  0.21220243  0.213944341  1.40676562  0.320592462 1
#>  [49,] -1.27115345  1.323912599  0.27805386 -0.055280901 0
#>  [50,] -1.40898906  0.326702804  1.16812375 -0.237471357 1
#>  [51,] -1.46560187  1.450341466 -1.36044300  0.152253857 0
#>  [52,]  0.07814297  2.367816002  0.57862783 -1.140729887 0
#>  [53,] -0.60079132  1.360530931 -0.68500878 -0.255576713 1
#>  [54,]  1.29198378 -1.163638728 -0.26715999 -0.286638567 1
#>  [55,] -1.26334282 -1.894184016 -0.23444947  0.944110165 0
#>  [56,] -0.38006404 -0.939754723  0.72690098 -0.889442296 0
#>  [57,]  0.54864260  0.620099705 -0.31171347  0.611393518 0
#>  [58,]  0.10686500 -0.324183141  0.14448165  1.059672791 1
#>  [59,] -0.05140532  0.753734490 -0.71272501  1.435658767 0
#>  [60,] -0.98142017  0.280445690 -0.25781196  0.306983280 0
#>  [61,]  0.75372419  0.960892061  0.89291679 -0.991995959 0
#>  [62,]  0.49469559  0.940101333 -0.32666674 -0.438770245 0
#>  [63,]  0.08648104 -0.586498743  0.53870298 -3.111583521 0
#>  [64,] -1.67374981 -0.652063702  1.00965521  1.175125914 1
#>  [65,]  0.93545350  0.341082166  1.52613399  0.129932538 1
#>  [66,]  1.18929581 -0.262831753  0.47180182  1.108523435 0
#>  [67,]  0.54130801 -1.486496548 -0.83540264  2.214191535 0
#>  [68,]  0.43244102 -0.461054207  0.18194249  0.116278427 1
#>  [69,] -0.10425155  0.169057269 -0.05171976 -0.902530835 0
#>  [70,] -0.07559918  0.212178983 -0.44028911  0.108653251 0
#>  [71,] -2.42564218 -0.186992181 -0.21345120  1.822167330 0
#>  [72,] -1.35379805 -0.047717583  1.35671150  0.578811606 0
#>  [73,] -0.34497645  0.671983611  1.32390986 -0.518886605 1
#>  [74,]  1.55131580  0.471570243 -1.29509656  1.024788028 0
#>  [75,] -1.16489953 -0.165783095  1.05484115  0.389574405 0
#>  [76,]  0.89902018  0.254383005  0.46752239  1.007144681 0
#>  [77,]  0.95102129 -0.052643607  0.81699963  1.501468074 0
#>  [78,] -0.18353654  0.689757322  0.02829276  0.819361449 0
#>  [79,] -1.73865030  1.775526777  1.16508918  0.673271321 1
#>  [80,]  1.28072951  0.899697827 -0.15047576 -0.112416899 0
#>  [81,]  1.19440826 -0.395239551 -0.67015261  0.481348739 1
#>  [82,] -1.03691098 -1.028606372  0.69915446  1.788866577 0
#>  [83,]  1.28144385  0.062870335 -0.78397758  0.639763086 0
#>  [84,]  0.27063568  1.099425607  1.62455255 -0.365961226 1
#>  [85,] -0.74739893  1.290576722  0.39026274 -0.852150223 1
#>  [86,]  0.31680658  1.085028321  0.82577904 -0.139408700 1
#>  [87,]  0.27631382 -1.372408460 -0.69023794  0.782633534 1
#>  [88,]  0.04911997  0.788742951 -0.05469698 -0.756828423 0
#>  [89,]  0.46876143 -0.405306765 -0.27048849 -0.081945463 1
#>  [90,] -0.27107016 -0.882969203 -0.32378698  0.212201663 0
#>  [91,] -0.29561819  1.619570118 -0.85292692  0.135492350 0
#>  [92,] -1.15065376  0.667813469 -0.12139410  0.557885392 0
#>  [93,]  2.38694847 -0.009996428  0.14243055  0.245682938 0
#>  [94,] -2.27497659  0.594800173 -0.99698107  0.565760519 0
#>  [95,]  1.17521254 -1.269405384  0.37724568  0.046293473 1
#>  [96,]  0.97649960 -0.182311395  0.09786360 -1.949921332 0
#>  [97,]  0.24563972 -0.490967553  0.11591142  0.005194811 1
#>  [98,] -0.40533724  1.422050433  0.44147999  0.631648079 0
#>  [99,]  0.71188508 -0.730796033  0.73535174  0.884587116 0
#> [100,] -0.12241376 -0.214453427  0.94236705 -1.478995194 0
#> 
#> $eta
#>   [1] -52.52499179   7.64915122  -6.10528038  -6.55029733  -9.09850211
#>   [6]  -8.53121905 -15.69408562  -2.17015187  -2.89843060  -9.57947196
#>  [11]   3.13580263 -11.45083270   8.85895559 -10.49559142  -8.12938190
#>  [16] -21.07794127  -6.99864292  -4.84010109  18.13698252  -1.58787201
#>  [21] -10.18527071 -12.69596203  -0.67225792 -23.14031565  -6.51456134
#>  [26] -17.04117847  -1.09117716  -5.33294002 -15.43701466   3.19344468
#>  [31]  -1.71000760  -9.68438368   2.97576949  -3.71595323  -2.42900204
#>  [36]  -4.02605868   0.42605052 -12.26475014   3.65835927 -10.62597981
#>  [41]  11.81661811   4.14082072  -4.78292426 -23.78739608 -19.07024019
#>  [46]  -5.08905373  -8.30054998   8.62919235  -4.94520205   2.44863968
#>  [51]  -8.09866578  -1.22525514  -2.20459763   1.14282895 -16.29162062
#>  [56]  -8.48592818  -2.87271566   0.15369106 -15.83412961  -4.15098485
#>  [61]  -8.99980205  -6.81581180 -42.07955943   2.55077620   5.88862846
#>  [66]   2.02156690 -28.64863011   5.45366592 -12.31334963  -9.17763325
#>  [71] -21.21844552  -1.79599901   4.30594792 -12.99214240  -0.34848195
#>  [76]  -8.55065664 -12.08180372  -6.17815977   9.54383505   0.73519421
#>  [81]   1.96870687 -17.56546368  -7.32559599   8.56677282   1.23969348
#>  [86]   3.92578368   0.04716594  -2.15829992   1.80056626 -12.51387068
#>  [91]  -7.63098777  -4.44080890  -1.06985897 -10.03915069   7.54123043
#>  [96] -20.36086944   7.80691196  -1.45214124  -7.33888710  -9.16799457
#> 
#> $mu
#>   [1] 2.220446e-16 9.995238e-01 2.226089e-03 1.427649e-03 1.118207e-04
#>   [6] 1.971755e-04 1.528077e-07 1.024631e-01 5.223120e-02 6.912866e-05
#>  [11] 9.583456e-01 1.064050e-05 9.998579e-01 2.765735e-05 2.946635e-04
#>  [16] 7.014011e-10 9.122873e-04 7.844236e-03 1.000000e+00 1.696835e-01
#>  [21] 3.772044e-05 3.063461e-06 3.379914e-01 8.918434e-11 1.479514e-03
#>  [26] 3.972924e-08 2.513967e-01 4.806634e-03 1.976012e-07 9.605868e-01
#>  [31] 1.531627e-01 6.224415e-05 9.514674e-01 2.375424e-02 8.098771e-02
#>  [36] 1.753168e-02 6.049302e-01 4.715032e-06 9.748729e-01 2.427644e-05
#>  [41] 9.999926e-01 9.843394e-01 8.301983e-03 4.669444e-11 5.222758e-09
#>  [46] 6.126090e-03 2.483185e-04 9.998212e-01 7.067176e-03 9.204619e-01
#>  [51] 3.038520e-04 2.270130e-01 9.933838e-02 7.581987e-01 8.406961e-08
#>  [56] 2.063092e-04 5.351892e-02 5.383473e-01 1.328388e-07 1.550472e-02
#>  [61] 1.234190e-04 1.095102e-03 2.220446e-16 9.276256e-01 9.972369e-01
#>  [66] 8.830429e-01 3.614582e-13 9.957377e-01 4.491364e-06 1.033141e-04
#>  [71] 6.094614e-10 1.423388e-01 9.866914e-01 2.278155e-06 4.137506e-01
#>  [76] 1.933806e-04 5.661570e-06 2.069948e-03 9.999284e-01 6.759441e-01
#>  [81] 8.774721e-01 2.351885e-08 6.580338e-04 9.998097e-01 7.755107e-01
#>  [86] 9.806549e-01 5.117893e-01 1.035582e-01 8.582179e-01 3.675305e-06
#>  [91] 4.849461e-04 1.164910e-02 2.554299e-01 4.365493e-05 9.994695e-01
#>  [96] 1.436768e-09 9.995933e-01 1.896722e-01 6.493513e-04 1.043146e-04
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                 x1           x2          x3           x4   y
#>   [1,] -0.84282302 -0.437194696 -1.11369571  0.746735148  12
#>   [2,]  0.10585479  0.884812787 -0.34670341 -1.181249101   2
#>   [3,]  1.12225966  1.585657366 -0.42215727  0.074995770  29
#>   [4,]  1.04063552 -1.970524918 -0.01107440  0.039407172   1
#>   [5,] -0.23070887 -0.002770688  0.27996199 -0.690859677   0
#>   [6,] -1.05445089  1.148710256 -0.19921323  1.977357867   0
#>   [7,] -0.79500049  1.748994785 -1.63272945 -0.519641698   6
#>   [8,]  0.20778218  0.109427264  1.12978750  0.922785767   1
#>   [9,]  0.84382305 -0.278161454 -0.31100351  1.564402515   0
#>  [10,]  0.01600106 -0.599156527  0.23798665  2.434873725   0
#>  [11,] -0.09215289  0.856867042 -0.10838282 -1.744884636   0
#>  [12,] -0.30125021  0.211136848 -0.03655741  0.773345673   0
#>  [13,]  0.76452203 -0.548191320 -0.60726001  0.265419095   3
#>  [14,] -0.80126256  1.745411301  0.70898230 -0.701573065  12
#>  [15,]  1.03917389 -1.010373661  0.64882771  0.480150765   1
#>  [16,]  1.22153932 -0.899305905  0.07668139 -0.270618430   5
#>  [17,] -1.44187877  1.111119498 -0.75154056 -1.011705332   0
#>  [18,] -1.49268773 -0.151223512  0.16038807 -3.207633813   0
#>  [19,]  0.64263583 -0.229298897  1.82285015 -0.439502411  11
#>  [20,]  2.64486196  0.592107046  2.97517710 -1.105699954 734
#>  [21,] -0.73343464 -2.021636783 -0.89540190 -1.267644614   1
#>  [22,]  1.33848736  0.866787709  0.22462060  0.229490570   3
#>  [23,]  1.44593860 -1.544834042  1.72430269  0.667858428  29
#>  [24,]  0.32732953  1.260873068  0.06440721 -0.745217145   3
#>  [25,] -0.30309133  1.037890269  0.79245140 -0.274163986   2
#>  [26,] -0.99970690  0.291532108 -0.25091034  0.461470433   2
#>  [27,]  0.50020877  0.172004807  1.18400487  0.132721388   3
#>  [28,]  1.69713530 -0.083436898 -0.11601851  0.215905568   1
#>  [29,]  0.81861119 -2.048813656  1.20089335  0.082224343   1
#>  [30,] -1.31688092  1.694468355  0.22136097 -1.143497578   0
#>  [31,] -1.41987513 -0.664835188  0.97791849 -1.150915053   0
#>  [32,]  0.48804148 -0.556526987 -0.59725059 -0.080153720   4
#>  [33,] -0.17602896 -0.812976434 -0.95443768 -0.097320041   1
#>  [34,]  1.23542760 -1.363936529  2.01617342 -0.876987393  88
#>  [35,]  1.51143001  0.960943269  0.41135664  0.108225851   2
#>  [36,]  0.51954697  0.766602252  1.58042540  1.301850726   1
#>  [37,] -1.86649411 -0.222895543  1.07454127  0.562110825  16
#>  [38,]  0.07254210  0.081754721 -0.58383171 -0.032253371   2
#>  [39,] -0.97651982 -1.517386308 -0.75605695  0.004864110   6
#>  [40,] -3.21540530  1.895085464  0.46250847  0.137365902   3
#>  [41,] -1.07079755 -0.139879616 -0.66865695 -0.001635881   2
#>  [42,]  1.67831460  0.221602585 -0.28354296  1.911903772   0
#>  [43,] -0.19030293  0.714266909 -2.62005628  0.714630226   1
#>  [44,] -2.41208098  0.402086133 -0.14043111  1.568640997   0
#>  [45,] -0.15033614  1.227241357 -1.04142333  1.237658203   0
#>  [46,] -0.51230866  0.849462415  0.13734710 -0.038613232   7
#>  [47,] -0.92747406  0.909431463  1.19901374  0.238877638  12
#>  [48,] -0.37583867  0.612857421 -0.33636414 -0.838979050   0
#>  [49,] -1.31590591  0.174347286 -0.47729771 -1.024458600   0
#>  [50,] -0.23911109 -0.668529248 -0.86158710 -0.995506171   1
#>  [51,] -0.38878563 -0.567166292  1.89223657  0.303254818  21
#>  [52,] -1.28886210 -0.176520647 -1.24856483 -0.352899598   1
#>  [53,] -0.39644258 -0.918543086 -1.08749701  0.398679311   1
#>  [54,]  0.41459172  0.310514905  1.01764378 -1.365786923   0
#>  [55,] -0.53124010  0.305828892  0.46729740  2.896025663   0
#>  [56,]  0.62407466 -2.105607527  1.40285905 -0.773022370   1
#>  [57,] -1.65094127  1.164619970 -2.15861437 -1.119813026   0
#>  [58,] -0.31785546 -0.613230706  0.19378267 -0.564873028   0
#>  [59,]  2.36279003  2.455386751  0.83194949  2.490255764   0
#>  [60,] -1.09738155 -0.802017584  0.88376238 -1.127671082   0
#>  [61,] -0.52128458 -2.261533573  0.75374598  0.410226932  13
#>  [62,]  1.46677740  1.094410952  0.38044109 -0.745422665   2
#>  [63,]  1.89527814 -0.448783783 -0.59044725  1.262638969   0
#>  [64,]  0.69612604  0.978430036 -0.20316638  0.184241859   1
#>  [65,] -1.09872244  1.664715654  1.10407523  0.178670845  10
#>  [66,]  0.77698319 -0.156077774  0.23702679 -0.510208854   2
#>  [67,] -1.63220741  0.689423412  0.24793647 -0.468333787  14
#>  [68,] -0.97652813  0.797626688  1.04364281  0.523481045   8
#>  [69,]  0.38522707 -0.610608758 -1.17206001  0.952155723   1
#>  [70,]  0.04545882 -1.609244216  0.24118045 -0.143379986   8
#>  [71,] -1.43767830 -0.418566621  0.08020736 -0.295628685   8
#>  [72,] -1.16153472  0.953611970  0.34324016  0.444952540   0
#>  [73,] -0.25934898 -0.331525049  1.47300449  0.645099093  10
#>  [74,] -1.41507183  0.916866160 -2.15729711 -1.027902767   0
#>  [75,] -0.93690947 -0.023143307  0.65871577 -1.214766672   0
#>  [76,] -0.74508655  0.512468939  1.53710968 -1.459172080   1
#>  [77,]  0.27826116 -1.091453811 -0.09550575 -0.232824424   5
#>  [78,] -2.87099635 -0.395029437 -0.74475634 -1.101235187   1
#>  [79,]  1.11875641  0.100385814 -0.71583569 -0.159210085   4
#>  [80,]  0.92646839  0.059403962 -0.84001542  1.354488565   0
#>  [81,]  2.53379874 -0.071132659 -0.31395262 -0.272051843   1
#>  [82,] -0.73408663 -0.805977539  0.53259154 -1.699226243   0
#>  [83,] -0.61965304  0.389986653  0.47607667 -0.605849403   6
#>  [84,]  0.67158770 -0.665612842 -1.07711443 -0.679157540   0
#>  [85,]  0.33658369 -1.684728240 -1.69012374  1.096655309   0
#>  [86,] -0.62849121 -0.240101567  0.82666881  0.125148297  20
#>  [87,]  1.68957762 -0.187568061  0.85795200 -0.605765430   3
#>  [88,]  0.75895026  0.355184732  0.04809181  1.147038142   0
#>  [89,] -0.50106683 -0.540478231  1.98648158  0.286491715  86
#>  [90,] -0.20233144 -0.348244255 -1.27011818 -1.427476586   0
#>  [91,] -0.96826609 -0.025349118 -0.04795043 -0.653163314   1
#>  [92,]  1.68409644  1.871251323  0.11114121  0.782537435   0
#>  [93,]  0.63989076 -0.647165624  0.82692378 -0.046934110   7
#>  [94,]  1.63104843 -0.360545924  0.30633253 -0.259924537   1
#>  [95,]  1.77768484 -0.418610047  0.06262036 -1.060134329   0
#>  [96,]  0.29926678  2.024902754 -1.41547604  1.572709530   0
#>  [97,]  0.12905551  0.396947674 -0.51410892 -0.200374934   8
#>  [98,] -1.88424097  1.019586944  0.67556683 -0.878831635   2
#>  [99,]  0.55633347  0.277503501  0.32260789  1.729824342   0
#> [100,] -0.21526931  1.830310350  0.95677354 -0.026675566   3
#> 
#> $eta
#>   [1]   1.84525427  -1.23087958   2.94510163   0.69418682  -1.28710927
#>   [6]  -7.73476157   1.98335476   0.51749244  -3.80642692  -9.81641168
#>  [11]  -5.44815630  -1.43645778   0.77590097   2.28145953   0.70339571
#>  [16]   1.27825318  -2.19598872 -19.27450959   2.67005248   6.59937202
#>  [21]  -1.86038919   2.01311489   3.39427357   0.78178023  -0.17757776
#>  [26]   0.65039263   1.43437389  -0.20212778   0.71532303  -1.27849670
#>  [31]  -1.03298655   1.93843730  -0.53949606   4.39908813   0.77354835
#>  [36]  -0.18389036   2.60714937   0.66129256   2.15483642   0.22093796
#>  [41]   0.58391401  -7.53734381  -0.25997437  -3.65842926  -3.33226793
#>  [46]   1.40855307   2.17174857   0.15281992  -2.59421958  -1.79883358
#>  [51]   3.43067770  -0.06633993  -0.78871706  -1.23232022 -15.08929321
#>  [56]   0.33253256  -1.50989541   0.28855084  -9.01371563  -1.88547777
#>  [61]   2.18845031  -0.31456680  -1.59424329  -0.48944530   2.42175441
#>  [66]  -0.08930731   2.58340080   1.93410307   0.33796758   2.12474102
#>  [71]   1.87689588  -0.38442234   1.76236741  -2.09693224  -1.59113465
#>  [76]   0.51402091   1.76336128   0.47647000   0.83218268  -3.93432521
#>  [81]   0.34925621  -3.85081046   1.96326166  -0.63713252  -0.56563727
#>  [86]   3.02935275   0.38815091  -2.04990350   4.48721299  -3.01962544
#>  [91]  -0.06433835  -0.95205639   1.28563670   1.60485143  -0.80689438
#>  [96]  -3.83880760   2.29457743  -0.02354816  -5.10028779   1.05900095
#> 
#> $mu
#>   [1] 6.329709e+00 2.920356e-01 1.901259e+01 2.002080e+00 2.760677e-01
#>   [6] 4.373566e-04 7.267081e+00 1.677815e+00 2.222746e-02 5.454897e-05
#>  [11] 4.304233e-03 2.377685e-01 2.172549e+00 9.790960e+00 2.020602e+00
#>  [16] 3.590363e+00 1.112485e-01 4.257816e-09 1.444073e+01 7.346337e+02
#>  [21] 1.556121e-01 7.486601e+00 2.979300e+01 2.185359e+00 8.372959e-01
#>  [26] 1.916293e+00 4.197016e+00 8.169905e-01 2.044847e+00 2.784556e-01
#>  [31] 3.559423e-01 6.947885e+00 5.830420e-01 8.137663e+01 2.167443e+00
#>  [36] 8.320270e-01 1.356034e+01 1.937295e+00 8.626479e+00 1.247246e+00
#>  [41] 1.793043e+00 5.328110e-04 7.710714e-01 2.577296e-02 3.571202e-02
#>  [46] 4.090033e+00 8.773612e+00 1.165115e+00 7.470415e-02 1.654918e-01
#>  [51] 3.089757e+01 9.358127e-01 4.544274e-01 2.916152e-01 2.797713e-07
#>  [56] 1.394495e+00 2.209331e-01 1.334492e+00 1.217287e-04 1.517565e-01
#>  [61] 8.921377e+00 7.301051e-01 2.030621e-01 6.129663e-01 1.126561e+01
#>  [66] 9.145645e-01 1.324210e+01 6.917836e+00 1.402095e+00 8.370729e+00
#>  [71] 6.533194e+00 6.808438e-01 5.826214e+00 1.228327e-01 2.036944e-01
#>  [76] 1.672001e+00 5.832007e+00 1.610380e+00 2.298330e+00 1.955889e-02
#>  [81] 1.418012e+00 2.126250e-02 7.122520e+00 5.288066e-01 5.679981e-01
#>  [86] 2.068384e+01 1.474252e+00 1.287473e-01 8.887341e+01 4.881950e-02
#>  [91] 9.376877e-01 3.859465e-01 3.616970e+00 4.977120e+00 4.462418e-01
#>  [96] 2.151925e-02 9.920243e+00 9.767269e-01 6.094992e-03 2.883489e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                 x1          x2           x3           x4            y
#>   [1,] -0.91030410  0.54137110 -0.752659672 -0.921266480  -2.62506559
#>   [2,] -1.00178631  0.95411468  0.350952215  0.217690700  -3.92665737
#>   [3,] -0.58159232  0.91735473 -0.420553225  0.233818352  -3.08186539
#>   [4,] -0.73236865 -0.48579867  0.789461150  0.540133712  10.58421501
#>   [5,]  2.30006136 -1.84883075  0.864943572 -1.295124683  -2.05217781
#>   [6,] -1.39561303 -0.88227241 -1.499859285  0.321070237 -19.05380196
#>   [7,] -0.46070929 -0.35072691  1.028657347  0.903755702   3.60265556
#>   [8,] -0.64249899 -1.15441684  0.276032612  2.111384444 -17.49079685
#>   [9,]  0.81600704 -1.36413450  0.468745052  0.317372926   0.27534489
#>  [10,]  0.08338789 -0.64777033  0.445980433  1.050051152  -1.46850624
#>  [11,]  0.65900185 -1.22969308  0.583898949 -0.443324134  -2.96051322
#>  [12,]  0.90244873 -1.83643839  1.533704413  1.207458269  -6.81298899
#>  [13,] -0.94217156  0.21041430 -0.798724646 -1.906692901 -24.85817169
#>  [14,]  1.21336732  0.19309342  0.005024061 -1.087494533  -5.89033698
#>  [15,] -1.94000886 -0.74621188 -0.555810860  1.535231373 -18.24715866
#>  [16,] -0.56623190  1.44320646  0.580358051  0.046119021   6.54321095
#>  [17,] -0.87957906 -0.55124127 -0.446643733 -0.254171972   3.53327640
#>  [18,]  0.46859623 -0.43766034 -0.671351493  0.314968321  -0.91342724
#>  [19,] -0.46306102  0.44262624  0.990073721  0.280840661   6.70054368
#>  [20,]  0.28706261 -1.59299387  0.076120547 -0.929178210   0.51368468
#>  [21,] -0.84854444  0.53538242 -0.049954970  0.093116451   6.24266393
#>  [22,]  1.41129715 -0.42106989  1.750294271  2.055668986  -5.51209329
#>  [23,] -0.56377384  1.13886408  2.053102282 -0.664392558   6.89492072
#>  [24,]  1.06679207 -1.22704308 -1.215949537 -0.527162079  -5.95832438
#>  [25,]  0.08954381 -1.98292976  1.138772586  0.337832604   4.60500963
#>  [26,]  0.10506111 -0.61452326  0.869324619  0.199175378   8.27391239
#>  [27,] -0.18205756 -0.65604287 -0.502456645  1.093367048 -15.40719286
#>  [28,] -0.17886466  0.84959292 -0.581065760  1.040958721 -14.44382797
#>  [29,]  1.90892261 -0.21901333  0.512816883  0.672590757  -3.18002303
#>  [30,] -0.01775922 -0.84188255  1.521161354  0.171265396   5.61576798
#>  [31,] -0.52790463 -0.69599607  0.345670776  1.284065861  -7.08312751
#>  [32,] -0.21765554  0.30566398 -0.891863685 -0.455721551 -13.96095376
#>  [33,] -0.81549777  0.34937886 -0.976834025 -0.377655907  -3.19937821
#>  [34,] -0.05727960  1.64646997  0.140909581 -0.754529470  -1.47493038
#>  [35,]  0.68561065  0.38349328  0.128446933 -0.001185693  -2.77559698
#>  [36,] -1.63617968 -0.12978607  0.193412755  0.795891367   0.29138844
#>  [37,]  1.05152006 -0.40320458 -0.106727679  0.923708635  -3.14232476
#>  [38,]  0.12365904 -0.12407591  1.470975028 -0.543286168   7.89184372
#>  [39,]  0.64319988  0.40083831  0.101736239 -0.176652393  -4.31956295
#>  [40,] -0.29822732 -1.83745414  0.100938342 -2.345959526 -36.64077377
#>  [41,]  0.31648532 -1.52104413  0.477026624  0.714352080   7.63754080
#>  [42,]  1.16979738 -0.90985936  0.716569227  0.075454182   2.78205589
#>  [43,] -0.21661729  1.85191193 -0.468422243  0.115332964 -10.22432219
#>  [44,]  0.57649559 -0.39338632  1.170258338  0.117965241   6.43341717
#>  [45,]  0.24526753  0.68867501  0.799482259 -0.323172153   9.69470297
#>  [46,] -1.24022140  0.73778643 -0.294313821 -0.746284656 -10.94628093
#>  [47,]  0.61470447 -1.65804408 -0.431041165 -0.508255335  -8.89850891
#>  [48,]  1.00486430  1.75224317 -0.962499200  0.445487074  -7.77373126
#>  [49,] -1.47432563 -0.29868145  0.182458371  0.238204581   1.00305884
#>  [50,] -0.37280436  0.36056390  2.048045316 -0.212232670   7.49116471
#>  [51,]  1.83997697  0.98748236  1.801184638 -1.734602866 -11.44271701
#>  [52,]  0.28041996  1.86294579  1.203171853 -1.603863236  -2.90881137
#>  [53,] -0.48641548 -0.50332188 -1.186897310 -1.785483956 -20.69810549
#>  [54,]  0.75665334 -0.13012920  0.608978847  0.807657686  -5.85159380
#>  [55,]  0.55936722 -0.43506979 -0.546904579 -0.381235139  -2.50810719
#>  [56,]  0.40591187 -1.13778206 -0.291201695 -0.917032188  -6.02133321
#>  [57,]  0.53011103  0.60951319  1.697137000 -1.145945024   5.38634652
#>  [58,] -0.02425783 -0.89880186 -1.661861758  1.185924158 -22.36191155
#>  [59,]  0.95832923 -0.09973147  0.419991355  0.856002079  -5.96339359
#>  [60,] -0.06058868 -0.06820262  0.022873312  1.444207920 -15.95312569
#>  [61,]  0.42978248  1.47562825  0.604017181  3.005395572 -34.79876166
#>  [62,] -0.40018817  0.82694697 -1.390645294 -0.097129719 -11.67122218
#>  [63,]  0.13569313 -0.67342828  1.494994993 -0.257307991  12.33677349
#>  [64,] -1.89210184  1.36563900 -1.610783017  1.057854771 -11.04700062
#>  [65,]  0.19257023  1.55994904 -1.158849367 -0.839471169  -2.69065801
#>  [66,] -0.91500577  2.00398281 -0.700339746  2.097678374 -29.34681096
#>  [67,] -0.29195066  1.11866506  0.830825964 -0.626075506  -6.87079014
#>  [68,]  0.17434285  1.28093577  0.274286125 -0.161515356   3.19562060
#>  [69,]  0.13210037  0.73653116 -0.688532989 -0.651475727  -2.31145345
#>  [70,]  0.85339160  1.01169271  0.537155599  0.661078874  -7.45962545
#>  [71,]  0.81696510 -2.24114881 -0.664462113 -0.165886136  -7.22152837
#>  [72,]  0.13218674 -1.64692912  0.406325998 -0.028142485   4.49626813
#>  [73,]  0.25937277 -0.76781001  1.124736704  1.637309953  -4.90353494
#>  [74,]  0.47347789  0.92998976  1.010281068  0.972407682  -2.97498610
#>  [75,]  1.00182537  1.05922571 -2.005480068 -0.797810817 -20.98215339
#>  [76,]  0.06910061  1.26040028 -0.515968276  2.793434721 -40.92052998
#>  [77,] -1.44527738 -0.51764910  0.175378647 -0.242444544  -0.06112831
#>  [78,]  1.89923502  0.74490331 -0.781743975  0.182745882  -6.56674290
#>  [79,]  0.27946056 -0.57843051  0.451809783  0.267168827   6.72758714
#>  [80,]  0.57608479 -0.46653079 -0.825378134 -1.903044437 -21.25826600
#>  [81,] -0.13947940  0.60305401  0.497674122 -2.225657669 -21.65560498
#>  [82,] -0.92462688 -1.91348027  1.049360650 -1.287914886  -4.32962569
#>  [83,] -1.46831005 -1.23667381 -1.184461011 -1.845656140 -24.21306762
#>  [84,]  1.07109093 -0.28548983  0.804359678  0.392610016   8.23966600
#>  [85,] -1.86008330 -0.38431514 -0.184805446  0.598965567   3.94395469
#>  [86,] -0.41899621  1.84820533 -0.177895108 -0.569540347  -4.54209775
#>  [87,]  0.02023352  0.37140375  0.282408129  0.402735857   1.74216191
#>  [88,]  0.53694109  0.26678119  2.216337907 -0.465822634   7.19256725
#>  [89,] -0.28474145  0.49855974 -0.001546135  0.098303502  -2.20140432
#>  [90,] -1.56091183 -0.89889714  0.455615020 -1.157050865  -7.26256723
#>  [91,]  0.69806307 -0.19598273  0.286425625  0.102091842  -3.99322613
#>  [92,] -0.52769502 -0.69172271 -0.619791503 -0.002851872  -3.96362558
#>  [93,]  1.15419849 -0.09660396 -0.515500853  0.051344621  -0.98877614
#>  [94,]  2.23440052  1.65541563  0.246304189  3.271766563 -45.46539773
#>  [95,] -0.18162044  1.43547679  0.972136863  0.292795019   2.59069502
#>  [96,]  0.02524121 -1.00035209 -0.288816962  0.335985445  -7.29559237
#>  [97,]  0.47935420  2.70717115 -0.546557921 -0.041552957  -1.34804707
#>  [98,]  1.18164719 -0.99948154 -0.094112713 -0.398001564  -2.47731696
#>  [99,]  0.18963373 -0.67829357  2.788099682  0.123666285  19.73658407
#> [100,]  0.46688942  0.39056138  0.843804328  0.467147280   3.52396223
#> 
#> $eta
#>   [1]  -3.70325785  -2.50754097  -3.94012346  10.41464882  -3.54448565
#>   [6] -17.94951423   2.98877672 -18.63298260  -0.01483995  -1.26827875
#>  [11]  -1.72860492  -5.91978917 -25.41553330  -5.53991850 -16.61279394
#>  [16]   6.61201435   1.65263623  -1.06738346   5.50426745   1.37503607
#>  [21]   6.07548007  -5.48250440   6.83618555  -6.75471314   3.73355500
#>  [26]   7.95984047 -14.68236756 -15.41492167  -2.19030466   3.40078713
#>  [31]  -7.43790656 -13.38544813  -2.99520376  -1.61014188  -3.99640103
#>  [36]  -0.74202309  -0.83950804   8.20724641  -3.22973864 -36.65287543
#>  [41]   6.60756735   3.15821879  -9.82694286   7.08267969  10.86894308
#>  [46]  -8.85997545  -8.36666301  -6.24645839   1.39369892   7.75295079
#>  [51] -13.30796043  -3.06641506 -21.51785856  -7.10521172  -3.34718114
#>  [56]  -5.47987896   5.10599078 -23.22553416  -5.65678546 -16.04529596
#>  [61] -37.72609464 -12.75907568  10.82831465 -12.72542635  -3.32903743
#>  [66] -28.71547631  -7.18218094   3.25734636  -1.85605252  -8.08535849
#>  [71]  -8.46865393   5.21383278  -5.46187383  -1.69634771 -20.62538746
#>  [76] -43.19328407   0.84943252  -7.28666651   7.95395585 -23.02711355
#>  [81] -23.16727407  -5.21326472 -25.01059885   7.36181799   3.34576755
#>  [86]  -7.09270726   1.05656224   8.02531710  -2.94611135  -7.55638845
#>  [91]  -3.93004700  -4.88684448  -2.27444515 -45.01404159   1.82804910
#>  [96]  -6.51375625  -1.35923156  -1.82733536  19.59902529   4.99986718
#> 
#> $mu
#>   [1]  -3.70325785  -2.50754097  -3.94012346  10.41464882  -3.54448565
#>   [6] -17.94951423   2.98877672 -18.63298260  -0.01483995  -1.26827875
#>  [11]  -1.72860492  -5.91978917 -25.41553330  -5.53991850 -16.61279394
#>  [16]   6.61201435   1.65263623  -1.06738346   5.50426745   1.37503607
#>  [21]   6.07548007  -5.48250440   6.83618555  -6.75471314   3.73355500
#>  [26]   7.95984047 -14.68236756 -15.41492167  -2.19030466   3.40078713
#>  [31]  -7.43790656 -13.38544813  -2.99520376  -1.61014188  -3.99640103
#>  [36]  -0.74202309  -0.83950804   8.20724641  -3.22973864 -36.65287543
#>  [41]   6.60756735   3.15821879  -9.82694286   7.08267969  10.86894308
#>  [46]  -8.85997545  -8.36666301  -6.24645839   1.39369892   7.75295079
#>  [51] -13.30796043  -3.06641506 -21.51785856  -7.10521172  -3.34718114
#>  [56]  -5.47987896   5.10599078 -23.22553416  -5.65678546 -16.04529596
#>  [61] -37.72609464 -12.75907568  10.82831465 -12.72542635  -3.32903743
#>  [66] -28.71547631  -7.18218094   3.25734636  -1.85605252  -8.08535849
#>  [71]  -8.46865393   5.21383278  -5.46187383  -1.69634771 -20.62538746
#>  [76] -43.19328407   0.84943252  -7.28666651   7.95395585 -23.02711355
#>  [81] -23.16727407  -5.21326472 -25.01059885   7.36181799   3.34576755
#>  [86]  -7.09270726   1.05656224   8.02531710  -2.94611135  -7.55638845
#>  [91]  -3.93004700  -4.88684448  -2.27444515 -45.01404159   1.82804910
#>  [96]  -6.51375625  -1.35923156  -1.82733536  19.59902529   4.99986718
#>