Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                  x1           x2          x3           x4 y
#>   [1,]  0.948368390 -1.690915180  2.43444399 -0.353911491 1
#>   [2,] -0.482840657 -0.742055916 -0.51468521 -2.074460653 0
#>   [3,] -0.090571212  0.567464546  1.69196516 -0.360921504 1
#>   [4,]  0.767238731  0.077994845  0.55514133  1.603542591 0
#>   [5,]  1.894105116 -0.586378918 -1.26721148  0.045956652 0
#>   [6,]  1.601498334 -1.661975342  0.05147039 -0.572141591 0
#>   [7,] -0.160162259  0.747843135  0.44186017 -0.340015825 0
#>   [8,] -0.391634321  1.414932884 -0.71718500  0.010579541 0
#>   [9,] -0.573377225 -1.016522698  2.24465969 -1.544763268 0
#>  [10,] -0.898637048 -1.984301728  0.33956565 -1.117422980 0
#>  [11,]  0.870490570  0.290135668  0.21992828 -1.444159397 0
#>  [12,] -0.379180046 -1.613179540 -0.12844223  1.132715509 0
#>  [13,] -0.229200830  0.609863409  0.34476668 -0.111341911 1
#>  [14,]  0.827002041 -0.470677483 -0.09710019  0.035531805 0
#>  [15,] -0.525399722 -2.368390904  0.43143604 -0.869786576 1
#>  [16,]  0.970734221 -0.155783587 -0.19403961 -0.237825353 0
#>  [17,] -0.139779033 -0.510662356  1.02083908 -2.002044435 0
#>  [18,]  0.636241327 -1.909883823  1.05868095  0.428611260 0
#>  [19,]  0.124833482  0.311264981 -0.17156355  0.264361929 1
#>  [20,]  0.656825240 -0.995151490 -0.39559556  0.188918842 0
#>  [21,] -2.352895100  0.444878542  0.51059788  0.656038131 0
#>  [22,]  0.298804036  0.618623791  0.62301949  1.580076973 0
#>  [23,]  1.037817653  1.321704676  0.13713840  1.718446333 0
#>  [24,]  0.402376944  0.573287458 -3.26060038  0.780582805 0
#>  [25,] -1.833439860 -1.145730415  0.54394526  0.130850363 1
#>  [26,]  0.295926715 -0.407260527  2.17654631  1.504298060 1
#>  [27,] -0.386517555 -0.323302470  0.29133443 -0.249371943 1
#>  [28,]  0.818177676  0.624351886  1.10278233 -1.751177805 0
#>  [29,]  1.059141472  1.259992019  0.60001856 -0.394722432 1
#>  [30,] -1.648313748 -0.356203334  0.77790931 -0.153474305 1
#>  [31,] -0.372256238 -0.657355753  0.18250402 -1.289806283 0
#>  [32,] -0.277531825 -0.994080271  1.41726592 -0.662170498 0
#>  [33,] -0.434227979  0.295420629  0.43462776 -2.439308116 0
#>  [34,]  0.004461456  0.737454385 -2.05541553  1.106444998 0
#>  [35,] -1.409898883  1.097627537 -1.01327815  0.092442995 0
#>  [36,] -1.004310382  1.826661762 -1.61732770  0.955495330 0
#>  [37,] -1.569766800  0.610712645  0.68317135  0.569080898 1
#>  [38,]  0.093900631  0.090771267  1.10806364 -0.720950888 0
#>  [39,] -1.267360968  0.433535186  1.94507560  0.133853325 1
#>  [40,] -0.441074209 -2.620331399  1.21346260  1.464803318 0
#>  [41,]  0.889302130  0.052249728 -0.18986074  0.642150132 0
#>  [42,] -0.878100417  0.408979276 -0.71421981  0.003670697 1
#>  [43,]  0.046230707  0.544609766  0.31695794  0.229491689 1
#>  [44,]  1.421608100  0.438780145 -0.48270555 -0.750129469 0
#>  [45,] -0.378337404  1.214762598  1.23678366  1.125890400 0
#>  [46,] -0.322248669 -1.251391420 -1.25359885 -0.265324633 0
#>  [47,]  0.982987234 -0.036336568  0.41432645  0.005306469 0
#>  [48,]  0.481543244 -0.777583952 -1.67476279 -0.446993237 0
#>  [49,]  0.827256383  0.864548021 -0.23112787 -0.476304085 0
#>  [50,] -1.053318779  0.366768370 -0.63480211  0.327917524 0
#>  [51,] -0.474223434  1.268748056 -0.43648491  0.869112563 0
#>  [52,] -0.887667500  0.183969295  0.57667680  0.380502189 1
#>  [53,]  0.273670128 -0.788178969  0.83561633  1.972623901 0
#>  [54,]  1.503826199  0.887808546  1.91686918 -0.797907993 1
#>  [55,] -0.090380862 -1.770930615 -0.06487549  2.076875230 0
#>  [56,]  1.791660236  1.105109183  2.24344121 -0.174079067 1
#>  [57,]  0.394933017  1.274961086  0.68081801  0.322303769 1
#>  [58,] -0.399161290  1.393595719  1.80042713  0.476198908 1
#>  [59,] -1.415580859 -0.789750653  0.93526867  1.233659933 0
#>  [60,]  1.494518364 -0.408584832 -1.88485103  0.231581677 0
#>  [61,]  1.401986875  0.406498484 -1.12938218  0.399530338 0
#>  [62,]  1.330557866 -0.619660448  1.88715959 -0.317383023 1
#>  [63,] -3.354048171  0.193095347 -1.26266729  0.699635668 0
#>  [64,]  1.925669552  0.483341678 -0.64748665 -0.602629816 0
#>  [65,]  0.670370307  0.474465933 -1.31929216  0.874642343 0
#>  [66,]  0.451307550 -0.438846129  0.90938444 -1.183755930 1
#>  [67,]  0.945971770  1.963501197 -0.94519540 -0.138756661 0
#>  [68,]  0.375328721 -1.588498227 -0.76922611  0.241950744 1
#>  [69,] -0.484013842  0.350542247 -0.02325507  1.247553894 0
#>  [70,]  0.459790399 -0.676735647 -1.32356893  1.459222794 0
#>  [71,] -0.034025890  0.083862595  0.41170300  0.796190371 0
#>  [72,]  1.362571375  1.903182160  0.22713449  0.467092849 0
#>  [73,] -0.951803858 -0.246841031  0.24358319 -0.509824650 1
#>  [74,]  0.858883146  1.071171842  0.22548596 -0.573615668 0
#>  [75,] -0.947873935  0.008056517 -0.69984743 -0.924725733 0
#>  [76,] -1.363212349  0.361634843  1.75875945  1.935065168 0
#>  [77,]  0.638815201  0.836168429  0.36981764 -1.919140832 0
#>  [78,]  0.124161211 -1.606996550 -0.44619603  1.183694897 0
#>  [79,] -0.362291444 -0.187233052 -0.30442124 -0.747415138 0
#>  [80,] -0.680623496 -1.020704268 -0.42970445  0.376028635 0
#>  [81,]  0.079313054 -0.266331017  0.11104410 -0.573016882 1
#>  [82,] -1.726699350  2.311273179 -0.96581063  0.886548062 0
#>  [83,] -1.171889935 -1.891035723 -0.15096378 -1.591450448 0
#>  [84,]  1.096164218 -0.659825873  0.21106302 -0.937471059 0
#>  [85,] -0.159273019 -0.298399437  0.33817007 -0.302576971 0
#>  [86,]  0.396534764  0.180012225 -1.24780292  1.240849224 0
#>  [87,]  1.321621174 -0.159737047 -1.04621760  0.676305420 0
#>  [88,]  0.256930438  0.839224028  0.60103346  0.565895686 1
#>  [89,] -0.725096879  1.226230449 -0.93385256 -0.234089054 0
#>  [90,] -0.802374967  0.717180933  0.22656175 -0.979580561 1
#>  [91,] -0.559761530 -0.226184538  2.08635599  1.331129316 1
#>  [92,] -1.298148785  1.257030198 -0.20358856 -2.539324215 0
#>  [93,] -0.945083766 -0.558207751  0.33094115 -0.471153486 1
#>  [94,] -0.549089080  0.127751735 -1.31917622 -1.373929232 0
#>  [95,]  0.020766983  1.515863626 -0.55454306  0.401332713 0
#>  [96,]  1.693200812 -1.902472630  0.35408382  0.395736544 0
#>  [97,] -1.210804905  1.539285371 -1.25606505 -0.139846043 0
#>  [98,] -0.506845715  0.746317094  0.01610225  0.359050708 1
#>  [99,] -0.443486712  2.016917925 -2.03976071 -0.083582289 0
#> [100,]  0.602930200 -0.022444004 -0.46197478  1.997678294 0
#> 
#> $eta
#>   [1]  10.37302996 -28.38325048   9.29462180 -19.99639722 -10.18509017
#>   [6]  -5.59258351  -1.68616507  -6.90721213  -1.03180033  -6.41443234
#>  [11] -15.22385313 -10.71948838  -0.52295322  -2.71279933   1.88868374
#>  [16]  -4.22867935 -15.88993456  -3.83391678   0.19378716 -11.78649951
#>  [21]  -1.92838253  -1.59811675 -14.46331472 -17.23508501   5.38842130
#>  [26]   9.02430866  -0.05549272 -10.92188591   1.44291678  10.74217239
#>  [31] -11.99089864  -3.69334229 -31.50320919 -17.62528582 -14.27851603
#>  [36] -15.83094525   7.02573108   1.34120486   7.54146514  -5.51842583
#>  [41] -11.84610987  -0.40556626   5.47430158  -2.31110549  -6.51982317
#>  [46] -11.91886713  -2.53283963 -14.46337823  -9.31534999  -6.56413785
#>  [51] -10.19381159   1.30942297 -13.69453591   5.26666508 -28.50699516
#>  [56]   3.03080739   3.78583074   6.77445045  -8.75330993  -8.91613914
#>  [61]  -4.32736863  14.94663045 -18.51743347  -6.16115880 -14.46897398
#>  [66]   2.42911166 -11.91486095  -0.91057983  -9.39748012 -18.99686750
#>  [71]  -7.25875819   0.78785471   2.00063848  -9.66151804 -11.96907199
#>  [76] -12.65703437 -21.69093711  -6.11378395  -9.50424673  -3.48467469
#>  [81]   2.18153857  -6.73721310 -23.41119043  -1.52116363  -1.25726821
#>  [86] -16.88868975  -5.83038874   5.25797540  -4.85685739   2.38899222
#>  [91]   5.36521999 -41.49612449   3.65926933 -21.15377686  -1.29829768
#>  [96]  -8.81807839 -10.11081687   0.16543647 -19.20581912 -30.44428467
#> 
#> $mu
#>   [1] 9.999687e-01 4.713146e-13 9.999081e-01 2.068593e-09 3.772725e-05
#>   [6] 3.711564e-03 1.562808e-01 9.995432e-04 2.627352e-01 1.635071e-03
#>  [11] 2.445482e-07 2.210934e-05 3.721619e-01 6.222231e-02 8.686054e-01
#>  [16] 1.436234e-02 1.256288e-07 2.116702e-02 5.482957e-01 7.606502e-06
#>  [21] 1.269297e-01 1.682450e-01 5.231931e-07 3.272636e-08 9.954516e-01
#>  [26] 9.998796e-01 4.861304e-01 1.805832e-05 8.089059e-01 9.999784e-01
#>  [31] 6.200350e-06 2.428427e-02 2.220446e-16 2.215316e-08 6.293888e-07
#>  [36] 1.332624e-07 9.991121e-01 7.926880e-01 9.994697e-01 3.996126e-03
#>  [41] 7.166329e-06 3.999757e-01 9.958244e-01 9.020738e-02 1.471760e-03
#>  [46] 6.663446e-06 7.358783e-02 5.231599e-07 9.002348e-05 1.408054e-03
#>  [51] 3.739966e-05 7.874166e-01 1.128595e-06 9.948657e-01 4.164562e-13
#>  [56] 9.539467e-01 9.778134e-01 9.988587e-01 1.579128e-04 1.341874e-04
#>  [61] 1.303021e-02 9.999997e-01 9.077804e-09 2.105365e-03 5.202406e-07
#>  [66] 9.190204e-01 6.690194e-06 2.868812e-01 8.292591e-05 5.620375e-09
#>  [71] 7.034865e-04 6.873705e-01 8.808641e-01 6.368371e-05 6.337170e-06
#>  [76] 3.185066e-06 3.799668e-10 2.207281e-03 7.452907e-05 2.975144e-02
#>  [81] 8.985794e-01 1.184543e-03 6.802197e-11 1.792902e-01 2.214445e-01
#>  [86] 4.627381e-08 2.928335e-03 9.948211e-01 7.714897e-03 9.159840e-01
#>  [91] 9.953453e-01 2.220446e-16 9.748952e-01 6.501768e-10 2.144517e-01
#>  [96] 1.480106e-04 4.063595e-05 5.412650e-01 4.560566e-09 2.220446e-16
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                  x1          x2           x3          x4  y
#>   [1,] -0.502218514  0.13208840 -1.315093163 -0.03021453  0
#>   [2,]  0.547526312 -1.30864777 -0.017665455 -0.10192904  2
#>   [3,] -1.114791923  0.83164232  0.710290504 -0.59009094  3
#>   [4,]  0.005168835 -0.55597074  0.104554242 -1.45251693  1
#>   [5,]  0.101721757 -0.44040222  2.623813473 -2.04145762  0
#>   [6,]  1.049718358  1.10642833 -0.644855820  0.37424564  1
#>   [7,]  0.116457331  0.67325641 -0.581581472 -0.51568085 11
#>   [8,] -1.682861068  2.18352814  0.522073288 -0.54225222  7
#>   [9,]  1.149691644  0.06081458  0.204054495 -1.13531587  0
#>  [10,] -1.457781788  2.40072933 -1.358887065 -0.29225294  2
#>  [11,] -2.437567820  0.20086971  1.363310588  0.34151834  3
#>  [12,] -0.725487519  0.67309154  0.331754854  0.11047202 31
#>  [13,] -0.540366739  0.22493002 -0.346220283  0.50485623  2
#>  [14,] -0.398239765  0.68776474 -1.593835838  0.47830653  4
#>  [15,] -2.732286678 -0.12829051  0.054308827  0.25336824  7
#>  [16,]  0.128521317  0.45781532 -0.333420200  0.49124769  9
#>  [17,] -0.531999247 -0.95150711  0.022005627 -0.11980983  2
#>  [18,] -0.496593086 -1.71290586  0.123065782  0.01025369  4
#>  [19,]  3.437699707 -0.84665286 -0.576568891  0.57671640  2
#>  [20,]  1.527893297  1.19380237 -0.510262874  0.86194430  2
#>  [21,]  0.369539038  1.40164373 -0.640640853 -0.41570785  4
#>  [22,] -0.869152648 -0.20724867  0.057028028 -0.39496801  5
#>  [23,]  0.683786469  1.13389218 -0.103348881 -0.31965529  0
#>  [24,]  0.524201608 -1.76524757  1.022556865 -0.97063470  0
#>  [25,]  0.644806619  0.26066097  0.873215130  0.70144450  0
#>  [26,]  0.069525997 -0.79132584  1.582704075  2.06982798  0
#>  [27,]  1.525700455  0.40412324 -0.944178975  0.19689130  8
#>  [28,]  0.782937879 -0.78904381  1.237038804 -0.81341215  0
#>  [29,] -0.382364485  0.15775201  2.219753887 -0.39338455 19
#>  [30,] -0.920624859  0.23440358 -1.267400403 -0.60926324  2
#>  [31,] -0.373604210  0.42831628 -1.466219768 -0.30087815  2
#>  [32,]  0.209102679  0.44403697 -0.390038157  0.47440405 15
#>  [33,] -0.795613245 -1.01486918 -0.240712375 -1.44911398  0
#>  [34,] -0.082694069  1.40454999 -0.374000619  1.31030673  0
#>  [35,]  0.549229874  0.48983364 -1.465492247 -0.04983264  5
#>  [36,]  1.191235074  0.88524346 -2.092159622  0.58453481  4
#>  [37,] -0.487094357 -0.79848453 -1.209265868  0.55119823  0
#>  [38,]  0.479417657  1.86490364  0.650632861 -1.36920808  0
#>  [39,] -0.318920921 -0.33999457  0.470475505  0.47936466  2
#>  [40,]  0.605668926  1.17970987  2.216084040 -0.79986661  9
#>  [41,]  0.770933186 -0.12305867 -0.729877476 -0.07593159  2
#>  [42,] -0.793086063 -1.02485290  0.245785508 -0.27881451  7
#>  [43,] -0.037191536  1.05349339 -0.482499222  0.67004953  0
#>  [44,] -0.440854294  0.54068781  0.853823374 -1.39197566  0
#>  [45,] -1.176353432 -0.67314820  0.748887518 -0.80444717  1
#>  [46,] -0.413591640  1.50165998 -0.144481328 -0.36482168  2
#>  [47,] -0.218612415  0.97427335 -0.340114872 -0.83060249  0
#>  [48,]  1.258353883  1.70238622  0.765346752 -0.22002186 32
#>  [49,] -0.898789357  0.11115959 -0.067490398  1.26570648  0
#>  [50,]  0.438188272  0.10689587  0.214721201 -1.23913729  0
#>  [51,]  0.362342913 -1.14387627  1.465297910  0.73714679  7
#>  [52,]  2.205008908 -0.63216632 -1.085678742  1.41182610  0
#>  [53,] -0.465599844 -0.96767119 -0.054620084 -1.90158826  0
#>  [54,]  0.762968492  1.52280444 -0.372043165  0.53611345  1
#>  [55,]  1.850669150 -0.45638474  0.008407059  0.74530719  3
#>  [56,] -0.888303895 -0.83352400 -0.790327880  0.52791054  1
#>  [57,]  0.111295927 -1.25874995 -0.468082869  1.21844969  1
#>  [58,]  1.709610909 -1.46225641  0.487220297 -1.11152680  0
#>  [59,] -0.366037836  0.13422458 -0.334662145 -0.04486532  0
#>  [60,] -2.251070544 -1.85460309 -0.249191051  0.93426929  0
#>  [61,] -1.601163079  2.03506584  0.863395855  0.90413724  3
#>  [62,]  0.777236040  0.87217255  0.855049549  0.71118598  1
#>  [63,]  1.194392898  1.04528122 -0.186481383  1.67847108  0
#>  [64,]  1.663998261  0.14432474  0.177384733 -0.10068691  0
#>  [65,] -0.461001463 -0.10294196  1.082327662 -0.04437622  4
#>  [66,] -0.234089334  0.81902761  0.005430237  1.67434457  0
#>  [67,] -0.724915242  0.04740599  0.050945595 -0.88244611  1
#>  [68,] -0.566872455 -0.65542595 -0.511675446  1.65154812  0
#>  [69,] -1.424818220  0.13288805 -0.729554796 -0.54474834  3
#>  [70,]  0.221586839  0.60069894 -0.796811223 -0.52588045 15
#>  [71,] -1.595440984  1.67309892 -0.303067553  0.27794293 15
#>  [72,] -0.476149931 -0.79726107  0.491335053 -1.44045312  0
#>  [73,] -0.649390242 -0.03436186  0.544306570  0.63694891  1
#>  [74,] -0.240059096 -0.30302277 -0.500557758 -0.02525575  3
#>  [75,] -0.595215910 -1.01765017 -1.101545542 -0.15782678  5
#>  [76,] -0.912467697  0.12926543 -0.223867022 -1.48740130  0
#>  [77,]  0.173472580 -0.58477314 -0.924862521  0.69378356  9
#>  [78,]  1.192757773 -0.79220373  1.447534146 -0.18700066 18
#>  [79,]  0.275935331 -0.65822023  0.105049822 -1.84962089  0
#>  [80,] -0.058687983 -0.02820495 -0.205451785  0.87725163  1
#>  [81,]  0.936634196  0.57687791 -0.856965013 -0.45484035  3
#>  [82,]  0.084175850  0.56809959 -0.478417957 -1.01342337  4
#>  [83,]  0.069699049  1.08501654  0.097186390 -0.26362533  2
#>  [84,] -0.394591127 -0.85753044  0.753906272 -0.23235930  2
#>  [85,]  1.410040950 -0.94560194 -0.707236222  2.08134662  0
#>  [86,]  0.743423580 -0.82184562 -0.239752863 -1.21791191  0
#>  [87,]  1.806392048 -0.27777419 -1.514220997  0.76684650  1
#>  [88,] -1.851837231  0.31976706  0.636503277  0.40943833  5
#>  [89,] -2.162036017  0.25204515 -0.427147755  0.98423526  0
#>  [90,] -0.330355248  0.07431353  0.425532686  0.63888192  0
#>  [91,] -0.883518608 -0.69097585 -0.502337013  0.24037872  3
#>  [92,]  0.517877792 -0.68337965  0.148044313  0.66902058  0
#>  [93,] -1.311499203 -0.09569945  0.379882061 -1.62115314  0
#>  [94,] -1.157930542 -1.16298928 -1.476236364  0.60390865  0
#>  [95,]  0.485992924  2.62108706 -0.535036711 -1.80464878  0
#>  [96,] -0.492915931  0.30799632  0.079774715  0.13090539  6
#>  [97,] -0.339953300 -1.35196417  0.310298666  0.53500499  3
#>  [98,]  0.592723835  1.09266327  0.832555796 -1.74085570  0
#>  [99,]  1.565299337  1.19589229  0.314477143 -0.27786245  2
#> [100,] -0.542911173  1.75482790 -0.359407592 -0.06342875  6
#> 
#> $eta
#>   [1]  0.17342966  1.84137594  0.44411415 -2.08432943 -0.23118326  0.25130683
#>   [7]  2.19162213  1.20428820 -1.27252053  1.10210681  1.22469952  3.35341151
#>  [13]  0.41294890  0.80473011  2.02610412  2.24786114  0.59890885  1.16699659
#>  [19]  0.04655994 -1.17344658  2.03359418  2.07340388 -0.70066447 -0.48479730
#>  [25] -0.23522909 -5.88672589  1.36873305 -0.60198812  2.81710670  0.28554446
#>  [31]  0.77933833  2.44662029 -2.75525383 -2.20129769  1.62819702  0.95593935
#>  [37] -0.28985137 -2.07394047  1.14803620  1.96421476 -0.10332942  1.52721069
#>  [43] -0.86919744 -1.32858272 -0.36037203  1.44589578 -1.86472423  3.47847433
#>  [49] -2.21658500 -2.18173626  2.23978613 -1.65306479 -7.02405763  0.64126129
#>  [55]  0.46045292  0.41508960 -0.54815772 -0.82218852 -0.40520619 -2.34962085
#>  [61] -0.05391218 -0.32081697 -4.33987070 -0.31854471  1.52268481 -5.13345045
#>  [67] -0.12855042 -3.71342185 -0.80008332  2.60251871  2.53755460 -3.36053364
#>  [73]  0.93021485  0.98320614  0.82719658 -3.53394562  1.59512090  2.87446052
#>  [79] -4.33454231 -1.35319617  1.36817234  0.69582947  0.72648114  0.20426369
#>  [85] -7.91870742 -3.46527663 -0.40148236  2.16124132 -1.99571385 -1.07886031
#>  [91]  1.66596543  0.37523313 -5.05013733 -0.52198989 -4.22517901  1.21926573
#>  [97]  1.03749598 -5.68830172  0.24990657  1.99014151
#> 
#> $mu
#>   [1] 1.189377e+00 6.305208e+00 1.559108e+00 1.243905e-01 7.935940e-01
#>   [6] 1.285705e+00 8.949719e+00 3.334385e+00 2.801247e-01 3.010502e+00
#>  [11] 3.403143e+00 2.860014e+01 1.511268e+00 2.236093e+00 7.584481e+00
#>  [16] 9.467465e+00 1.820132e+00 3.212330e+00 1.047661e+00 3.092991e-01
#>  [21] 7.641502e+00 7.951844e+00 4.962554e-01 6.158220e-01 7.903898e-01
#>  [26] 2.776051e-03 3.930368e+00 5.477216e-01 1.672838e+01 1.330486e+00
#>  [31] 2.180029e+00 1.154925e+01 6.359288e-02 1.106595e-01 5.094681e+00
#>  [36] 2.601113e+00 7.483748e-01 1.256895e-01 3.151997e+00 7.129312e+00
#>  [41] 9.018298e-01 4.605313e+00 4.192879e-01 2.648524e-01 6.974168e-01
#>  [46] 4.245654e+00 1.549389e-01 3.241024e+01 1.089806e-01 1.128454e-01
#>  [51] 9.391323e+00 1.914622e-01 8.902060e-04 1.898874e+00 1.584792e+00
#>  [56] 1.514506e+00 5.780137e-01 4.394688e-01 6.668393e-01 9.540533e-02
#>  [61] 9.475153e-01 7.255560e-01 1.303821e-02 7.272066e-01 4.584517e+00
#>  [66] 5.896181e-03 8.793692e-01 2.439391e-02 4.492915e-01 1.349769e+01
#>  [71] 1.264870e+01 3.471673e-02 2.535054e+00 2.673013e+00 2.286899e+00
#>  [76] 2.918952e-02 4.928925e+00 1.771586e+01 1.310787e-02 2.584130e-01
#>  [81] 3.928165e+00 2.005372e+00 2.067792e+00 1.226622e+00 3.638724e-04
#>  [86] 3.126436e-02 6.693271e-01 8.681908e+00 1.359166e-01 3.399828e-01
#>  [91] 5.290779e+00 1.455331e+00 6.408453e-03 5.933387e-01 1.462272e-02
#>  [96] 3.384702e+00 2.822141e+00 3.385337e-03 1.283905e+00 7.316569e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                  x1          x2           x3           x4            y
#>   [1,]  0.815557727 -1.06176714  1.498841710  0.210287506   0.15482017
#>   [2,]  0.188725969  0.74473025  0.275431772 -1.025858082   1.09079859
#>   [3,]  0.170241485 -0.29356125 -0.258701004  0.247206529   3.70629777
#>   [4,]  0.294818119  0.56556072 -0.439534429  0.026184161   6.02175621
#>   [5,] -0.086201456 -1.26136784  1.074285328  0.283647395   4.09454294
#>   [6,]  0.311655567 -0.55258542  0.248381245  1.084874982   0.56404712
#>   [7,]  0.398767349 -1.06452542  1.220020285 -0.502429684   5.79369165
#>   [8,]  0.482357589 -0.56819086  1.249998349 -0.322712039   7.99009096
#>   [9,]  1.129507415 -1.95510046  1.399047098  1.094314697   0.06855958
#>  [10,] -0.608676888 -0.39458793  0.240586064  0.702433871   5.64451286
#>  [11,]  1.159743109 -0.66750865  2.304475210  0.449836179  15.30038896
#>  [12,] -0.521874809 -1.17260501  1.717702383 -2.747772900 -27.21635902
#>  [13,] -1.343031264 -0.88608122 -0.439902798 -0.015389891 -12.92388846
#>  [14,] -0.507432971 -0.96245781 -0.780217401 -0.284907520  -9.16305624
#>  [15,]  1.332819807  0.27463887 -1.317409395 -0.292721284  -7.85267812
#>  [16,] -0.547465154 -0.07996268 -0.407292147  1.902627860 -23.06580005
#>  [17,] -0.719601009 -1.40087980  1.348864521  1.161814638   8.44048182
#>  [18,] -0.506650763  0.84324371 -1.998749489  0.102106634 -13.49034473
#>  [19,]  1.578004135  1.18501178 -1.024580447 -0.949075449 -18.04403281
#>  [20,]  0.525312986  0.18371794 -1.400705916 -0.502430444 -16.16577572
#>  [21,] -1.179960312  2.69391135  0.062324339  0.324489604  -2.62521967
#>  [22,]  0.506872332 -0.80303022  1.747488351  0.240640235   5.28495339
#>  [23,] -0.814853892 -0.65563576  0.410001135 -1.171128747   1.16410703
#>  [24,]  1.074500700 -2.36922353 -0.536286380  0.629871067  -1.54007157
#>  [25,]  1.117899316  0.18936948 -0.708951793  0.704695953  -7.50539132
#>  [26,] -1.423453475 -0.84048910  1.387187395 -0.811668618  -2.47986657
#>  [27,] -0.840730187  0.50682816 -1.597917310 -1.535370775 -14.88166864
#>  [28,] -0.833467954 -1.72509592  0.782886714 -0.968224659   1.00233508
#>  [29,] -1.008396549 -0.25093472  1.621990716 -2.073430620 -10.44697415
#>  [30,] -0.036424902 -0.22061972 -1.267938990  0.257846925  -8.80986053
#>  [31,] -0.146813203  0.79565045 -0.110376263 -0.133565135  -4.10483672
#>  [32,] -0.848228370  2.10647641  0.464281288 -0.960465425  -2.90233039
#>  [33,]  0.316718448  0.55564360  0.105783350  0.792570601   4.19253960
#>  [34,]  0.347146914 -0.51973979  0.826434368 -1.177943226   2.57633633
#>  [35,]  1.065310465  2.70701442 -0.333880367 -0.688131294  -0.72674336
#>  [36,] -0.231585093 -1.25025412 -0.403897259  1.227811062 -14.58278196
#>  [37,]  1.983289893 -0.33477239  0.725546922  1.368803769  -4.25471108
#>  [38,]  0.494239258  1.65587601  0.833238179  0.624731017   5.70182032
#>  [39,] -1.352099727 -0.54194943  0.384080787 -1.548940596 -12.54422437
#>  [40,] -0.598395782  3.18627989 -1.240178474 -0.869133981 -12.29985028
#>  [41,]  0.476479313 -0.31106806  0.725354931  0.933652630   3.08124011
#>  [42,] -1.083135072  1.49406903  1.643106593  1.935038658  -9.05210055
#>  [43,]  0.367153228 -0.02446007 -0.868463636  0.320279818  -6.40111955
#>  [44,] -0.505784111  0.93352724 -0.258476362  0.293587769  -5.55190139
#>  [45,] -0.089485229  0.05694463 -0.391091328  0.003912389 -10.09816961
#>  [46,] -0.884374400  0.60417678  0.514441376 -0.603971269   6.96818841
#>  [47,]  1.375946295  0.16760116 -0.681088928 -0.688475847  -8.01669035
#>  [48,] -0.435308817 -0.60776818  0.666090661  0.889587601  -1.82485118
#>  [49,] -0.365115532  0.17440836 -0.817498511 -0.044417047 -14.42364239
#>  [50,] -1.174485065 -1.31244563 -0.804697612  0.964607474 -12.07969749
#>  [51,]  0.009242023  0.22842999  0.058789816 -1.641459692 -16.24876578
#>  [52,]  1.240803037  0.51739865  0.866604915  0.199651000  13.07757373
#>  [53,]  1.152461243 -1.65452418  0.117332887 -0.026426716   3.43764328
#>  [54,]  1.254695706  0.87812932 -0.285972131  1.209541191  -6.08681624
#>  [55,]  1.743589972 -1.48471301 -0.468081325 -0.123067784  -6.22700476
#>  [56,]  1.144799514  0.09167401  0.030859063 -0.383259084  -1.77141679
#>  [57,]  1.947970576  0.08284718 -0.650844641  0.387074622 -13.91820119
#>  [58,]  0.959271368  0.12914314  1.532572759 -1.494205359  -9.37020239
#>  [59,] -1.258704608  0.04101476  0.707519052 -1.260691488 -12.71147699
#>  [60,] -0.564181321 -0.07154397  1.311956730  1.114619998  -0.94290780
#>  [61,] -1.511368638  1.13312329 -1.283915102  0.094481433 -13.16208273
#>  [62,] -1.963501678  1.49097321  0.184035539  1.762539117 -12.09097960
#>  [63,] -2.843647995 -0.31718825 -0.734815439 -0.987879360  -3.39601101
#>  [64,]  0.009168629 -0.11214102 -0.660264412 -0.590273540  -9.27583483
#>  [65,]  0.784068855 -0.64184736 -0.951907932 -2.143715284 -33.63758750
#>  [66,]  0.785737300 -0.08502075 -0.325364453 -1.549844077 -22.01413504
#>  [67,]  1.069538561  0.42638828  0.215699631 -0.571063138   1.81031093
#>  [68,] -1.791290658  1.94338868  0.831540008  2.034338947 -14.28798134
#>  [69,] -0.467384305 -0.61602750 -1.825703541 -0.260912434 -11.24459792
#>  [70,] -0.784972067 -1.08558929  0.138192948  0.402352845   1.93441560
#>  [71,]  1.848927342  3.40780605 -0.273578780  0.516073057  -5.88293610
#>  [72,]  0.403398922 -0.64444236  1.616748047  0.056451586  11.56483141
#>  [73,] -1.644931973 -0.45781376 -1.634170609  1.275170277 -10.05748347
#>  [74,] -1.344087799 -0.71738006  0.425591913  0.750927019  -8.40863450
#>  [75,] -2.292973260  0.41360867  1.848666577 -1.176438540   0.42215913
#>  [76,]  0.267593389  0.18862952 -1.270036485  0.101244798  -7.54507602
#>  [77,] -0.490758411  0.66247527 -1.915638452 -0.232071454  -8.07941333
#>  [78,] -0.200115587  1.15005887 -1.455456529 -0.271061063 -19.44666438
#>  [79,]  0.844529826  0.56607541  0.034055288 -0.720331134  -4.87789674
#>  [80,]  1.390667051  0.71952494 -0.814043615 -1.787023642 -16.84263230
#>  [81,] -0.990592050  0.12289579 -2.177846657  0.689404031 -21.89207785
#>  [82,]  0.027600872 -2.06400384 -0.161713061  0.676691663  -6.79464613
#>  [83,]  0.150252902 -0.64379920 -0.285403512 -0.348242672  -0.65526285
#>  [84,] -0.868833519 -0.08980017  2.553274029 -0.273041490  15.32421517
#>  [85,] -1.325056172 -1.06043174  0.524794227  1.301615960 -13.58751213
#>  [86,] -0.593502732 -1.63861020 -1.486114804  0.336030418  -8.85550556
#>  [87,]  0.095152265  1.88653878  2.606111950 -0.425170616  17.24672814
#>  [88,]  0.221640590 -0.55665138  0.636010230  0.622589855   8.80872210
#>  [89,]  0.649918183  0.36324242 -1.476251703 -0.554088175 -13.85650563
#>  [90,]  0.553366552 -0.07015647 -0.502610116 -1.501541600 -19.41595301
#>  [91,]  0.108286529 -0.33040984 -0.551295954  0.334408825   0.65581197
#>  [92,] -0.532931493  0.45271087 -1.300382998  0.034423251  -6.82158226
#>  [93,] -1.444001224  0.38004987  0.772279485  0.174453659   1.40855547
#>  [94,]  0.271374449 -0.71217677  0.367646731  0.557497465   3.38414851
#>  [95,]  1.518218043  0.34473043  0.683265703 -0.186502962   2.16096064
#>  [96,]  0.673109177  1.30894964 -0.004768416  0.369885623  -5.32351219
#>  [97,] -0.872331612 -0.70500720  1.381605361 -0.671988731   8.04438997
#>  [98,] -0.022717915  0.34954771 -0.894334154  0.125907340  -6.35511919
#>  [99,]  0.977829685 -2.22019422  0.745159724  0.107141422   3.17858301
#> [100,] -0.737284058 -1.71263366 -0.556168890  0.713241131  -5.98699797
#> 
#> $eta
#>   [1]   0.1665931   1.5575258   3.8225784   4.6469222   4.1291388   1.4333141
#>   [7]   5.0587060   8.4286775   0.6354668   4.5788498  15.6056278 -28.2051594
#>  [13] -12.2176962 -10.2150322  -6.8482140 -22.7276177   8.7230757 -13.6787948
#>  [19] -17.6587561 -15.1321140  -3.1646364   5.5192127  -0.9360141  -0.6506348
#>  [25]  -7.9477445  -2.2900724 -15.0756680   1.4908107 -11.5890061  -9.3336370
#>  [31]  -5.3906513  -3.1922396   4.5157810   3.8640957  -1.0914297 -14.5042476
#>  [37]  -3.6042150   5.2780157 -12.3080900 -12.9542511   2.3221999  -9.4213917
#>  [43]  -6.6935870  -5.7308725 -10.4717128   7.0363343  -7.8457436  -1.9028014
#>  [49] -13.4737189 -12.3139750 -16.7034383  12.1905808   3.6894588  -6.7343610
#>  [55]  -5.7020976  -2.0908238 -11.7589511  -8.5410047 -14.0772825  -0.2063201
#>  [61] -12.6670175 -11.3959249  -3.4416690  -8.3480749 -34.3605852 -22.3815995
#>  [67]   2.5652310 -14.9935240 -12.4714569   1.3669367  -6.6500478  11.6864869
#>  [73] -11.5957381  -7.0804717  -0.1317111  -7.2991740 -10.0837247 -18.9706533
#>  [79]  -4.0481260 -16.6298710 -20.5008339  -6.3767731   0.8969650  15.1760102
#>  [85] -13.8260039  -7.4156646  14.6459670   7.9208057 -15.0963551 -19.7906628
#>  [91]   1.1556778  -5.4859550   2.5056821   3.0694380   2.2348500  -7.1574220
#>  [97]   7.2637158  -7.1559605   3.4499207  -3.4450159
#> 
#> $mu
#>   [1]   0.1665931   1.5575258   3.8225784   4.6469222   4.1291388   1.4333141
#>   [7]   5.0587060   8.4286775   0.6354668   4.5788498  15.6056278 -28.2051594
#>  [13] -12.2176962 -10.2150322  -6.8482140 -22.7276177   8.7230757 -13.6787948
#>  [19] -17.6587561 -15.1321140  -3.1646364   5.5192127  -0.9360141  -0.6506348
#>  [25]  -7.9477445  -2.2900724 -15.0756680   1.4908107 -11.5890061  -9.3336370
#>  [31]  -5.3906513  -3.1922396   4.5157810   3.8640957  -1.0914297 -14.5042476
#>  [37]  -3.6042150   5.2780157 -12.3080900 -12.9542511   2.3221999  -9.4213917
#>  [43]  -6.6935870  -5.7308725 -10.4717128   7.0363343  -7.8457436  -1.9028014
#>  [49] -13.4737189 -12.3139750 -16.7034383  12.1905808   3.6894588  -6.7343610
#>  [55]  -5.7020976  -2.0908238 -11.7589511  -8.5410047 -14.0772825  -0.2063201
#>  [61] -12.6670175 -11.3959249  -3.4416690  -8.3480749 -34.3605852 -22.3815995
#>  [67]   2.5652310 -14.9935240 -12.4714569   1.3669367  -6.6500478  11.6864869
#>  [73] -11.5957381  -7.0804717  -0.1317111  -7.2991740 -10.0837247 -18.9706533
#>  [79]  -4.0481260 -16.6298710 -20.5008339  -6.3767731   0.8969650  15.1760102
#>  [85] -13.8260039  -7.4156646  14.6459670   7.9208057 -15.0963551 -19.7906628
#>  [91]   1.1556778  -5.4859550   2.5056821   3.0694380   2.2348500  -7.1574220
#>  [97]   7.2637158  -7.1559605   3.4499207  -3.4450159
#>