Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                  x1           x2          x3           x4 y
#>   [1,]  0.870264242  0.708685657  1.19281283  1.380437054 0
#>   [2,]  1.104160912 -0.933288719 -0.56199785  0.704840325 0
#>   [3,]  0.832712196 -0.008893242  1.61544260 -0.420965228 1
#>   [4,] -0.109664038 -0.924336177 -0.98229165  0.934132121 0
#>   [5,]  2.260079729  1.903193655 -0.36393039 -0.018790565 0
#>   [6,] -0.299171548 -1.868010698  1.00971396  0.767072511 0
#>   [7,] -0.565335992  0.166497492  1.98281421  1.720460628 0
#>   [8,] -0.377000916  0.327119574  1.66704366  0.838522554 1
#>   [9,]  1.319019755  1.551425787  0.19627803  0.503360865 1
#>  [10,]  1.686234541 -1.639091323 -1.69807266  1.012695573 0
#>  [11,]  1.037106984  0.837036269 -0.58757630 -0.255836863 0
#>  [12,]  0.240857035  1.222159882  0.56040205  1.395570322 0
#>  [13,] -1.519506407 -0.675887966  1.95929398 -0.013288862 1
#>  [14,]  0.325199974 -0.152680237  0.27065144 -0.483483365 1
#>  [15,]  0.763688905 -1.415962730 -0.20609375 -1.552804495 0
#>  [16,] -0.342065370  0.126350624  1.34772241  0.492052006 0
#>  [17,]  0.270020033 -0.574938227  0.19773745  0.339983716 1
#>  [18,] -0.038178911  0.092630169 -0.98896392 -1.054143764 0
#>  [19,]  1.182493128 -0.094257901  1.45739245 -0.584290340 1
#>  [20,] -0.550680424  0.719376140 -0.94208641 -0.283880537 0
#>  [21,]  0.453745040  0.291176646 -0.28440835 -0.823049800 0
#>  [22,]  1.524461703  0.746070619 -1.41461283 -0.321484947 0
#>  [23,]  0.254130688  0.609180954 -0.54739143 -0.935410031 0
#>  [24,]  0.009788791  0.150539885  0.93007854  0.578935866 0
#>  [25,] -1.006190182  0.454192535 -1.44741767 -0.361779542 0
#>  [26,]  0.992241644 -0.256328982 -0.20223975 -1.835375365 0
#>  [27,] -0.115690189  0.426825326  0.64167029  0.997819152 0
#>  [28,] -0.084362207  0.308224568 -1.00795346 -1.186226151 0
#>  [29,]  1.884660888  0.411505012 -0.27441215  0.354356106 0
#>  [30,]  0.391438530  0.010521998  1.14194477  0.900825908 1
#>  [31,] -0.266065553  1.662297887  1.56919236  0.205292477 1
#>  [32,] -0.582128427  0.111232991  0.91851159 -0.892955673 0
#>  [33,]  1.979625505 -0.931769450  0.78875853  1.192834909 0
#>  [34,]  1.248758108  0.908586632 -0.39117854 -0.890377736 0
#>  [35,] -1.443872398 -0.600215184 -1.13729395  0.528574285 0
#>  [36,]  1.199230954  0.456103648 -0.24801028 -0.258588597 1
#>  [37,] -0.313996357 -0.831188580 -0.39931004 -1.152616202 0
#>  [38,]  0.667776081 -0.109038855 -0.18904681  0.726248960 0
#>  [39,] -0.986311771  0.341103235 -0.57623620 -0.669887342 0
#>  [40,]  1.317534818 -0.663550591  0.45886074 -0.351278538 1
#>  [41,] -0.957880947 -0.735849521  0.07255912 -0.205124161 0
#>  [42,]  0.045773405  0.484009529  0.25611747 -0.015216070 1
#>  [43,]  0.260045842 -1.420759515  2.01351257  0.347828322 1
#>  [44,]  1.301698040  0.394812513  1.13886646 -0.281248377 1
#>  [45,] -0.125839190 -1.118920433  0.27481359  0.790882798 0
#>  [46,]  0.817336447  1.365983920 -0.68809259 -0.053866141 0
#>  [47,]  0.751307229  0.923193816 -1.03421895  0.609899993 0
#>  [48,] -1.014280261  0.363556579 -1.97460300  0.376876475 0
#>  [49,] -1.314430689  0.057574809  1.25183222 -0.499792418 0
#>  [50,] -0.209139113 -0.087609772  0.06355420  1.294207946 0
#>  [51,]  1.104193991  0.244278207  1.46157562 -1.190446572 1
#>  [52,] -0.669929710  0.582095553 -1.01938486 -0.143730333 1
#>  [53,] -1.715690688  0.496874721  1.92658729  0.102356403 1
#>  [54,] -0.693800314 -0.801214120  1.18478493 -0.695827562 1
#>  [55,] -1.565589560  0.786270143  0.10831887 -0.622539794 0
#>  [56,] -0.245373735 -0.011649070 -0.63355244  0.762856302 0
#>  [57,] -1.135836779 -1.023308420  0.40724149 -0.194661681 0
#>  [58,] -1.047844253 -0.952931733  0.09075342 -0.024203551 0
#>  [59,]  1.247211706 -0.980257815 -1.31866388 -0.421211496 0
#>  [60,] -0.229822155 -0.491154895 -0.60737737 -0.761433782 0
#>  [61,] -0.750097508 -0.875449769  0.50187491 -0.053867695 1
#>  [62,] -0.627510640 -0.868953910 -1.14523595 -1.373296991 0
#>  [63,] -0.959900289  0.181571007  0.37048846 -0.075507462 0
#>  [64,]  0.521185874  2.320636197  1.17301143  0.534212082 0
#>  [65,]  2.364546728  1.598815020 -0.59291755 -0.328402586 1
#>  [66,] -0.494682853 -0.307656540  0.99313916  0.320921064 1
#>  [67,] -1.047886091 -0.853490845  0.77791084  0.331904944 0
#>  [68,] -0.895116040  0.439213941  0.01365205 -0.026939893 1
#>  [69,] -1.072337701  1.026164420 -1.87930473 -1.124340847 0
#>  [70,]  2.198409819  1.950578644 -0.19330558 -1.643648492 0
#>  [71,] -0.709607645  0.276593378  0.77397597  0.007867483 1
#>  [72,] -0.072732601  1.129765495 -0.32163596  1.585506153 0
#>  [73,]  0.399514100 -1.807786510 -0.40926169  0.519477708 0
#>  [74,]  0.471730750 -0.861154419 -3.49840134  0.815769780 0
#>  [75,] -1.622874439 -0.693540790  0.43752903 -0.302270207 1
#>  [76,]  0.302142117  0.658531199  0.75157578 -1.539786952 0
#>  [77,]  1.812291962 -0.765401095 -0.16808578 -0.912469233 0
#>  [78,]  2.414602093 -0.954138272 -2.93133867 -0.796270276 0
#>  [79,] -0.891692527 -0.445263504 -1.44565789  1.526785115 0
#>  [80,] -0.521854512 -0.165148192 -1.21938955  1.866609199 0
#>  [81,]  0.651886712 -0.992295872 -0.72980519  0.902120905 0
#>  [82,] -1.550727224 -0.138445982  0.14136390  1.159334326 0
#>  [83,] -0.726027058  0.883103654  0.07470019  0.988124597 0
#>  [84,]  0.002890413  1.751905567 -0.27271315 -0.336691411 0
#>  [85,]  1.549151295  0.234896286  0.11154009  2.425243586 0
#>  [86,] -0.666010526  0.095928669 -0.88387790 -1.049472881 0
#>  [87,]  0.699468391  1.185332394 -0.93754990  1.777129996 0
#>  [88,]  0.218686949  0.332576257  0.55517277 -1.550126154 0
#>  [89,] -0.721661815 -2.556165628  1.53313458  0.724314813 1
#>  [90,]  0.058395593  0.708415611 -0.37294513 -0.257746038 0
#>  [91,]  0.177565195  1.281058762 -0.85748122 -0.261374634 0
#>  [92,] -0.253168623 -0.280290705 -0.72420680 -0.627515901 0
#>  [93,] -0.810467968  2.592220422  0.97998068 -0.065255751 1
#>  [94,] -0.638512567 -0.150093998  0.98327156 -0.188473793 1
#>  [95,] -1.356916105  0.968848866 -0.48787895  0.916214518 0
#>  [96,] -1.547298759  0.418658016  1.12989600 -0.959427333 1
#>  [97,] -1.547924499  0.292457315 -0.59448736  0.522055622 0
#>  [98,]  0.837109262  2.379248747  1.51445359  0.156939122 1
#>  [99,]  0.201088786  0.296296713 -1.11861884  0.535793638 0
#> [100,]  0.751095385  0.328991306 -0.88187159  0.134238611 0
#> 
#> $eta
#>   [1]  -4.7558265  -8.2992595  -0.4310039 -18.9345351  -0.4802774  -6.9320368
#>   [7]  -5.8257704   1.4340838   6.8855571 -20.6378727  -4.3848943  -5.3832723
#>  [13]  10.7288252   2.9623153 -15.7756589  -2.6392627   6.3273648 -18.1643898
#>  [19]   8.2523131  -3.4364659  -6.1165140  -9.2661292  -0.2299879  -0.8969756
#>  [25]  -8.2117716 -17.7443499  -3.6561529 -17.6463717  -5.1195328   0.8170489
#>  [31]   6.2016592  -1.4297729  -8.5098996  -4.7162412  -9.7398639   4.2807784
#>  [37] -18.5272736 -11.1236925  -6.4822431   5.3654045  -1.5656417   4.7540967
#>  [43]  18.9661867  11.2702343  -7.8046397  -9.2951567 -16.7845222 -13.6583080
#>  [49]  -3.7956524 -16.3183235   1.1854906   0.6603673  18.3634094   4.9498758
#>  [55]   0.2899080 -16.5696451  -6.2068111  -6.3557880  -9.0204741  -9.3916085
#>  [61]   2.6553742 -18.0029330  -1.2654579   3.2229453   2.1340255   6.8141493
#>  [67]  -2.5165976   4.1825869 -25.0685021 -14.1846040   6.6762077 -22.0079835
#>  [73]  -5.3929960 -28.6602055   3.5773555  -0.5933890 -13.5412980 -23.6128857
#>  [79] -14.7319641 -25.6855969 -17.6265927  -6.6035623  -2.3104065  -1.7367480
#>  [85] -34.0011988 -11.9693589 -30.8153054  -5.2045160  12.6278083   0.4832315
#>  [91]  -3.6960404 -10.2659886  12.9894287   7.5351061 -15.6084622   4.2647901
#>  [97]  -5.8711877   4.4221783  -5.7117002 -11.8959430
#> 
#> $mu
#>   [1] 8.528079e-03 2.486391e-04 3.938866e-01 5.981855e-09 3.821866e-01
#>   [6] 9.750592e-04 2.941850e-03 8.075368e-01 9.989786e-01 1.089145e-09
#>  [11] 1.231076e-02 4.571771e-03 9.999781e-01 9.508423e-01 1.408375e-07
#>  [16] 6.665389e-02 9.982165e-01 1.292129e-08 9.997394e-01 3.117505e-02
#>  [21] 2.201277e-03 9.456494e-05 4.427551e-01 2.896724e-01 2.713658e-04
#>  [26] 1.966648e-08 2.518122e-02 2.169093e-08 5.943281e-03 6.936095e-01
#>  [31] 9.979780e-01 1.931341e-01 2.014235e-04 8.869382e-03 5.888508e-05
#>  [36] 9.863568e-01 8.988916e-09 1.475826e-05 1.528036e-03 9.953462e-01
#>  [41] 1.728386e-01 9.914573e-01 1.000000e+00 9.999873e-01 4.076721e-04
#>  [46] 9.185966e-05 5.135404e-08 1.170231e-06 2.197451e-02 8.185441e-08
#>  [51] 7.659336e-01 6.593429e-01 1.000000e+00 9.929655e-01 5.719736e-01
#>  [56] 6.366409e-08 2.011600e-03 1.733655e-03 1.208942e-04 8.341421e-05
#>  [61] 9.343415e-01 1.518538e-08 2.200358e-01 9.616887e-01 8.941666e-01
#>  [66] 9.989031e-01 7.470279e-02 9.849704e-01 1.296844e-11 6.913604e-07
#>  [71] 9.987410e-01 2.767287e-10 4.527732e-03 3.572983e-13 9.728104e-01
#>  [76] 3.558576e-01 1.315493e-06 5.559734e-11 3.999347e-07 6.996599e-12
#>  [81] 2.212422e-08 1.353696e-03 9.026476e-02 1.497265e-01 2.220446e-16
#>  [86] 6.335352e-06 2.220446e-16 5.461714e-03 9.999967e-01 6.185106e-01
#>  [91] 2.422043e-02 3.479547e-05 9.999977e-01 9.994663e-01 1.664681e-07
#>  [96] 9.861400e-01 2.811595e-03 9.881344e-01 3.296145e-03 6.817963e-06
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                 x1           x2          x3          x4        y
#>   [1,]  0.73516093 -0.815847621 -0.05484856  0.70290429        0
#>   [2,] -1.58360521 -1.204781413 -0.20724597 -0.79891497        1
#>   [3,] -0.20227900  2.012625451  0.16433141 -1.31269620        0
#>   [4,]  1.73604388 -0.458493875  0.87327922  0.94221657        1
#>   [5,] -1.32591406 -1.697907484 -2.12633057  0.19414097        1
#>   [6,]  0.85194617 -1.650175754  0.34983907 -0.41046626        2
#>   [7,]  1.56032883  0.995434328  0.37508249 -1.47518385        0
#>   [8,] -1.35782452  0.193357159 -1.85607686  0.16093213        1
#>   [9,]  0.96959097 -0.600815899  0.26137500 -0.32971719        2
#>  [10,] -0.19803182  0.505887980  1.10038941  0.94325307        1
#>  [11,]  2.35268340 -0.908046806  0.43994853 -0.64341612        3
#>  [12,] -0.09442522  0.024988528 -0.21287082 -0.09631098        1
#>  [13,]  1.48006668  0.250151509  0.39984712  0.33630402        6
#>  [14,] -1.18865078  1.144624003  1.87755347 -0.02636216        5
#>  [15,]  0.06663404 -0.324211985 -0.14633300 -1.07310704        1
#>  [16,]  1.40941292  1.938328703 -0.28539755  0.27696195        6
#>  [17,]  0.98963523  0.408674902 -1.06008023  0.40078938        6
#>  [18,] -1.10284832  1.187859829 -0.40180440  0.19956062        0
#>  [19,] -0.55005517  0.622632748  0.89551325  0.76881218        7
#>  [20,] -1.31096915 -0.105656336 -1.27532456 -0.38334223        0
#>  [21,]  0.47902508  0.480256092 -0.47795457 -0.08202926        8
#>  [22,] -0.01962671  0.773124879 -0.46512864 -1.54281709        0
#>  [23,]  1.23469154 -1.822004929 -0.21399823 -0.03365258        3
#>  [24,] -0.18354610 -0.869041164  1.49031426 -0.67270063        2
#>  [25,] -1.09007138  0.551924500  0.05517711 -0.40189672        0
#>  [26,] -0.87960086 -1.477892672 -1.41835157  1.49243995        0
#>  [27,] -0.30307198  0.222925263  0.34564959  0.85265871        1
#>  [28,]  0.48485928 -1.131041988  2.01570026 -0.54489777       30
#>  [29,] -0.69515765  0.146224465  0.01061405 -2.25521984        0
#>  [30,]  1.94437430  0.341586123  0.14860511  0.26724136        3
#>  [31,] -1.09366659 -1.462209049 -0.83338996  0.83774256        0
#>  [32,] -0.25551573 -0.544289226  0.12001028  0.41319519        3
#>  [33,]  0.24340742  0.957173344  0.03852883  0.80361166        2
#>  [34,] -0.58395034  0.449077722  0.12710243 -0.37632358        8
#>  [35,]  1.38296468 -0.213133426  0.34730246 -1.43850168        1
#>  [36,]  1.23101968 -0.000400751  3.65488825 -0.18554510 14841500
#>  [37,] -1.92031370 -2.212716700  0.45994552  0.41929298        8
#>  [38,]  0.54140576 -0.994695846 -0.60531943  0.49702193        0
#>  [39,]  0.81860139  2.087202512  1.28064473  0.52014107        2
#>  [40,] -0.09927311  0.391728151 -0.80154512  0.12857360        0
#>  [41,]  0.01452941  0.948794235  0.76674151 -1.14907652        0
#>  [42,] -0.15642931 -0.142603034  0.07439291  0.36937106        4
#>  [43,]  0.17606756 -0.251291235  0.71847298  0.35499274       17
#>  [44,]  1.29545690  0.981760809 -0.06791786 -1.31626747        0
#>  [45,] -0.65486715  2.220217355  0.31723445  0.36572356        3
#>  [46,]  0.26256005  0.528899031 -0.68489374 -2.20839653        0
#>  [47,] -0.07829497  0.266609594 -0.93568069 -0.27937250        0
#>  [48,] -0.29879883 -1.032961283 -0.03596294  1.26082204        0
#>  [49,]  0.19339775  0.814261688 -0.23981688  0.58368254        8
#>  [50,]  0.97611922  1.095868953 -1.33277869  0.02296407        0
#>  [51,] -1.55448292  1.169137712 -0.15997495  1.07093802        1
#>  [52,]  0.62291636  1.209588840  0.10268728 -1.03913655        0
#>  [53,]  1.10724828 -0.606353612  0.84559741 -1.62031821        0
#>  [54,]  0.54916313  0.507038840 -0.29425206 -3.29847344        0
#>  [55,] -0.15379258 -0.671662735 -0.30824050  1.37905998        1
#>  [56,] -0.48218161  0.588478942 -1.57260802 -0.07257319        6
#>  [57,]  0.16589427  0.198353746  1.50475949  0.67941362        6
#>  [58,] -0.93926684  0.663516691  0.99665687 -1.41953385        0
#>  [59,] -0.14153192  1.942932371 -0.86726604  0.24794256        0
#>  [60,] -0.19278947 -1.707495076 -0.68444751 -1.27107211        0
#>  [61,] -1.49601965 -0.414121242  0.87257378 -0.09059308       16
#>  [62,]  0.27158087 -1.821891164  0.34340309 -0.60692972        1
#>  [63,]  0.74717615 -0.396043034  1.42100762  0.62061304        8
#>  [64,] -1.04371500 -0.579506712 -1.66272299  0.15450010        4
#>  [65,] -1.08898936 -0.391358395 -0.14427048  0.24037711        7
#>  [66,] -0.78989738  0.674574420  0.05146537 -1.28919408        1
#>  [67,] -1.03121162  0.097020098 -1.11640675  0.83894098        0
#>  [68,]  1.88790691 -1.234339800  0.18239031  0.41015233        1
#>  [69,] -0.07901725  1.101911863 -1.33154447 -0.15952828        0
#>  [70,]  0.77352392  0.191920417  0.50904597  0.11605459        1
#>  [71,]  2.30087062 -0.731510688  0.10391379 -0.34066534        6
#>  [72,] -0.83415784 -1.699239346 -1.88072027  0.31587807        4
#>  [73,] -0.72031734  1.324435813  1.52520304  1.51089658        1
#>  [74,]  1.12673584  1.004363963 -0.58687040  0.47334439        1
#>  [75,]  1.95615628 -0.264344034  1.44724339  1.26734903        1
#>  [76,] -0.23599214  0.267113340  1.11510063 -0.37384733        2
#>  [77,] -0.63012057  0.634764728  0.34566178 -0.86782829        4
#>  [78,] -1.03284861 -1.245702442 -0.27647765 -0.15605660        2
#>  [79,] -0.42735025  0.601173048 -0.72383733 -0.91273504        2
#>  [80,] -0.08208398 -1.130137924 -0.26490764  1.02386118        0
#>  [81,] -1.49656181  0.507025584  0.38106706 -1.41747016        0
#>  [82,] -0.88835877  0.224512639  0.83087226  2.11291193        0
#>  [83,]  0.32267079  1.068587329 -0.57878917  0.91609759        0
#>  [84,]  0.81189146 -0.622530310  0.74075340  0.35880939        1
#>  [85,] -0.65024586 -1.366664311 -0.66381188  0.99089518        2
#>  [86,]  0.04398498  0.428073416 -0.62125952 -0.41466230        8
#>  [87,]  0.07691775  0.018578301  0.10440132  1.08671347        0
#>  [88,] -0.15435842 -1.268960627 -0.07559558 -0.44336531        0
#>  [89,]  0.77557073 -0.186051367 -0.38066587 -0.47879381        1
#>  [90,] -0.58472969  0.995646359 -0.61701867  0.99350233        0
#>  [91,] -0.54679486  0.914467544  1.38592255 -0.53012446        6
#>  [92,]  1.23898725 -1.302414185 -0.98408276  0.59359010        8
#>  [93,] -1.20719761 -0.110088457 -0.76481585  0.16156014        1
#>  [94,] -0.77695891 -2.194759949 -0.94468005 -1.15992512        0
#>  [95,]  0.42984059  0.699430759  0.82971790 -1.26432778        1
#>  [96,]  0.85716697  0.445279772  0.20199161 -1.34828986        0
#>  [97,]  1.01919858  2.127982707  0.07766225  0.32888780        1
#>  [98,]  1.54976000  0.180906065  0.47119383 -0.97307433        0
#>  [99,]  1.67364595 -1.242822921 -2.77392543  0.02934465        3
#> [100,] -0.18793567  0.464979350 -1.98589029 -0.56441183        1
#> 
#> $eta
#>   [1]  -1.396943318   0.764286508  -3.843099217   0.157724067  -0.087892220
#>   [6]   0.555672814  -4.003805070  -0.523597914   1.535695382   0.370233778
#>  [11]   0.595096250  -0.196092646   1.016765517   1.769345669   0.271761360
#>  [16]   1.054849124   1.229005591  -0.191520670   2.070270118  -0.505483271
#>  [21]   2.177486583  -3.285042317   1.501231659   0.007916736   1.548387379
#>  [26]  -1.736025449  -1.454882491   3.258635899  -8.740144693   1.054196728
#>  [31]   0.180800489   0.822267268   0.404725724   2.356582829  -1.636475166
#>  [36]  16.512950934   2.024632781  -0.426006938  -0.008895993   0.942962274
#>  [41]  -1.369444798   0.152020729   2.949040929  -1.948900452   1.474112465
#>  [46]  -6.605754671   0.143093397  -3.507083428   1.666844590  -0.039807848
#>  [51]  -1.463076358  -2.098635709  -2.303334944 -19.867909680  -3.340302358
#>  [56]   1.959181448   2.007400663  -0.709949492  -0.385047688  -3.196419429
#>  [61]   2.915020303   1.003581861   1.874643513   1.297555770   1.683616707
#>  [66]  -0.107083001  -1.465516458   0.886976585  -0.411774161  -0.111599214
#>  [71]   1.786575178   1.544783983  -0.776827760   0.598696138  -0.227336820
#>  [76]   0.561821898   1.728886569   1.505298102   0.077809152  -1.533280705
#>  [81]  -1.556621972  -6.978558960  -0.554300719   0.952576593   0.995626559
#>  [86]   1.723992138  -1.380627545   0.733252280  -0.074204425  -1.114836037
#>  [91]   1.818697292   2.179899038  -0.200165201  -0.397321589  -0.123695984
#>  [96]  -2.263357596   0.382199407  -1.364649778   0.575509547   0.244636812
#> 
#> $mu
#>   [1] 2.473519e-01 2.147462e+00 2.142709e-02 1.170843e+00 9.158596e-01
#>   [6] 1.743113e+00 1.824608e-02 5.923854e-01 4.644554e+00 1.448073e+00
#>  [11] 1.813205e+00 8.219361e-01 2.764239e+00 5.867013e+00 1.312274e+00
#>  [16] 2.871542e+00 3.417829e+00 8.257026e-01 7.926964e+00 6.032140e-01
#>  [21] 8.824100e+00 3.743900e-02 4.487212e+00 1.007948e+00 4.703878e+00
#>  [26] 1.762194e-01 2.334278e-01 2.601403e+01 1.600307e-04 2.869669e+00
#>  [31] 1.198176e+00 2.275654e+00 1.498891e+00 1.055482e+01 1.946650e-01
#>  [36] 1.484169e+07 7.573329e+00 6.531118e-01 9.911435e-01 2.567576e+00
#>  [41] 2.542481e-01 1.164184e+00 1.908764e+01 1.424306e-01 4.367158e+00
#>  [46] 1.352562e-03 1.153838e+00 2.998424e-02 5.295432e+00 9.609741e-01
#>  [51] 2.315229e-01 1.226236e-01 9.992504e-02 2.352212e-09 3.542624e-02
#>  [56] 7.093518e+00 7.443943e+00 4.916690e-01 6.804182e-01 4.090842e-02
#>  [61] 1.844919e+01 2.728036e+00 6.518495e+00 3.660339e+00 5.384997e+00
#>  [66] 8.984511e-01 2.309587e-01 2.427778e+00 6.624739e-01 8.944026e-01
#>  [71] 5.968975e+00 4.686959e+00 4.598625e-01 1.819745e+00 7.966524e-01
#>  [76] 1.753865e+00 5.634377e+00 4.505497e+00 1.080916e+00 2.158264e-01
#>  [81] 2.108471e-01 9.316448e-04 5.744738e-01 2.592381e+00 2.706420e+00
#>  [86] 5.606867e+00 2.514207e-01 2.081840e+00 9.284819e-01 3.279690e-01
#>  [91] 6.163824e+00 8.845413e+00 8.185955e-01 6.721178e-01 8.836484e-01
#>  [96] 1.040007e-01 1.465504e+00 2.554701e-01 1.778036e+00 1.277157e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                 x1           x2           x3           x4            y
#>   [1,]  1.49590326 -1.394382440 -0.623587243  0.310954976  -1.91243213
#>   [2,] -0.07245736 -0.203153369 -1.035153459  1.376443226 -17.86642072
#>   [3,] -0.03126508 -0.266337452  1.019273925 -0.055431786   6.43467646
#>   [4,]  1.04926952  0.190562483 -0.533813802 -1.140255300 -12.74906486
#>   [5,] -0.69977909 -1.562546918 -1.532185981  0.076994015  -3.45854669
#>   [6,] -1.37671085  0.616904153 -0.163875137 -0.164513694  -2.66632510
#>   [7,]  0.23287950  1.148124302 -1.732684622  0.403807504  -9.88901770
#>   [8,]  0.71661836 -0.900825394  0.456861653 -0.771241491  -8.70712252
#>   [9,]  0.22288306  0.463856703 -0.282347425  0.386414028   3.83712333
#>  [10,]  0.51002211  0.108577940  0.581878346 -1.647616397 -16.08150143
#>  [11,] -0.39500680 -0.685984068  1.204632554  0.369189434  -0.44373360
#>  [12,]  0.86070170  0.420744543 -0.874327328  3.061049145 -55.14986610
#>  [13,] -1.09890232  0.378766006  1.212241695  0.314819171   3.60167621
#>  [14,]  0.02480847  0.536767637 -1.922274793 -0.917382019 -11.41874567
#>  [15,] -0.07616413  1.450472417 -1.045360289  1.450482427 -16.76324015
#>  [16,] -1.22958047  0.207788101  0.181001740 -0.541762400  -9.38193595
#>  [17,] -0.42955008 -1.061024560 -0.636006915 -2.213649862 -34.51891617
#>  [18,]  0.30209064 -0.746279203 -0.727187902 -0.724698612  -6.33679986
#>  [19,] -0.56965108  1.138077486 -1.670100709  0.183628303 -12.30851236
#>  [20,]  0.68923519 -0.618263713 -1.778207150 -1.198461742 -21.40817702
#>  [21,] -1.28427346 -1.522256914  0.914249734  0.901178074  -1.80096958
#>  [22,] -0.10404132 -1.354632008 -0.394025982  0.954756992  -7.83092627
#>  [23,]  1.54818783  1.417696725  2.317134675  1.085879086   5.67043668
#>  [24,]  1.53935968  0.596246539  0.399394706 -2.619898896 -31.40986175
#>  [25,] -0.31113092  1.006986249 -0.753877968 -1.084394693 -19.95140398
#>  [26,] -1.40823578  0.821063662 -1.715374593 -1.857944879 -30.51122075
#>  [27,] -1.26585575 -0.696338448 -1.877854382 -0.536686647 -19.59018784
#>  [28,]  0.52803894 -0.588042961 -0.376062794  0.172794740  -3.49373269
#>  [29,] -0.14431478  1.232306466  1.142739430 -0.424621975  -2.87180194
#>  [30,] -0.94118851  0.523071185 -1.682478295 -1.177915435 -12.74404170
#>  [31,]  0.60879011 -1.292756895  1.845437716  0.004160243   9.04253471
#>  [32,]  2.26710423 -0.877600226  1.576999258 -1.306433923   0.10395076
#>  [33,]  1.50166577  1.263297337  0.741955612 -0.602459732  -0.81212706
#>  [34,] -0.81140456 -0.336203739  0.120903751  0.521102014   6.51916102
#>  [35,]  1.46482950 -0.289487150  0.044742133  0.606066308   1.98966566
#>  [36,]  0.13901507  0.943330776  0.459810404 -1.072966536  -2.22452981
#>  [37,] -0.87016408 -1.885377138 -0.155368722  1.050608152  -7.61147063
#>  [38,] -1.20026895 -0.804248628  0.114509904  1.400322797 -18.77969683
#>  [39,] -0.40080680 -1.614833148 -0.138246599  1.228050724  -9.72083501
#>  [40,] -0.19384593 -0.407810328 -0.199557941 -2.692495368 -39.32242642
#>  [41,]  0.10799520  0.959260073 -2.109660408 -1.002859227 -19.03505298
#>  [42,] -0.01618488  0.328477674  1.641555206 -0.108130797   7.21411975
#>  [43,]  1.83568259  1.289655283 -1.118306528 -0.150385642 -12.20010469
#>  [44,]  0.33485800 -1.613523923 -0.308750131 -0.320989900   2.16985468
#>  [45,]  1.42555600  1.707538997 -0.063755360 -0.585101884   1.34199562
#>  [46,]  0.17867116 -0.121645147 -0.356301277  0.521065868  -1.64749197
#>  [47,] -0.49739124 -0.477135560 -1.029772597  1.093153657  -8.58029768
#>  [48,]  0.33198678 -1.534627320 -0.161953116  1.453929262  -5.73020185
#>  [49,] -1.44112335 -0.604786249  1.192438358 -0.988964342   0.04143804
#>  [50,] -0.48856243  0.240578464 -0.786476148 -1.562138584 -20.36384602
#>  [51,]  1.02873479 -0.864161664 -0.814966752 -0.752444859 -13.88138684
#>  [52,]  1.68178222  1.163664419 -0.322943751 -0.085402197 -11.38430656
#>  [53,] -1.84798499 -0.109665778 -0.713401024 -0.159146157  -3.49531164
#>  [54,]  0.93103656  0.342542890  1.938562499  0.077558264   8.46678164
#>  [55,] -0.13568348  0.866546407 -0.424229959  2.569867120 -41.34733913
#>  [56,]  0.31859387 -0.186339553 -0.600790137  0.426483994  -0.54069093
#>  [57,]  0.27573052 -0.195999365 -1.660681394  0.481108435  -7.92173314
#>  [58,] -0.03799911 -0.425798878 -0.572882925  1.404177585 -11.53672808
#>  [59,]  0.24678952 -1.832269306 -0.681290858 -0.449124975  -4.73480733
#>  [60,] -1.34455639  0.498165604 -0.100370139  1.215808114  -9.18820595
#>  [61,]  0.68204954  1.115755361  0.089773080  0.783223046 -13.01684853
#>  [62,]  1.55219361  1.722210842  0.555280573  0.900484869  -0.12176135
#>  [63,] -0.60309214  0.080749647  1.383249174 -1.283798001  -2.90426250
#>  [64,] -1.24193203  0.796115649 -0.569345408  0.442042901 -10.50794135
#>  [65,]  1.19350250 -0.082906826  1.325886803 -0.581217508   8.58844209
#>  [66,] -0.80946667 -1.242652287 -1.887160839  0.006532763  -7.62623219
#>  [67,] -0.63596096 -1.411823906  1.184229356 -1.167845400   6.04721722
#>  [68,] -0.16822386  1.204203278 -2.332232093  0.875745349 -26.50758915
#>  [69,] -0.36495649  0.916527469  0.832800557  0.225022957  -3.67790304
#>  [70,]  1.13741278  0.405610911 -0.376563407  0.467589434   0.66318360
#>  [71,] -0.65077425  0.428434149 -1.275516300 -0.615187098  -3.53796708
#>  [72,]  0.16168835 -1.314560382 -1.727088979 -1.022560917 -12.59058085
#>  [73,]  0.18623826  0.861446712 -0.442657903  0.369037557  -0.04204502
#>  [74,]  0.08642451  2.158737614 -0.471513302 -0.384492249  -7.07725474
#>  [75,] -0.71760637 -0.784794713  0.059741533 -0.287711345  -1.87455945
#>  [76,]  0.56532876 -2.406974654  0.325332030  0.658231937  -0.92176854
#>  [77,] -0.18529585  0.963281244  0.007592887  0.751324106 -11.28564332
#>  [78,]  1.51004953 -0.002716268  1.370150060  0.085935585   3.15771212
#>  [79,] -0.63676618  0.566416216  0.145869646 -0.974576919   2.53555734
#>  [80,]  1.83818950  0.947910920 -0.741251764  0.342802921 -12.44545364
#>  [81,] -2.09886990  0.848733331  0.074797894 -0.541270099  -6.77927261
#>  [82,]  0.13806607 -0.593799039 -0.862770629 -1.282199743  -8.90087007
#>  [83,]  0.36173856  0.560938378  0.126788645 -0.922385796   3.20308719
#>  [84,]  0.62508610 -0.404631743 -0.001592205  0.647653755  -4.01162725
#>  [85,]  0.35374963 -0.007295115  0.511071688 -0.641591710  -1.31297813
#>  [86,] -0.26009013  0.593991403 -0.422650179  0.053839996  -4.90343267
#>  [87,] -0.60708051  1.412276212 -0.379052228  0.393195960   0.08557637
#>  [88,]  0.03098571 -1.113473688 -0.450322319  0.735460165  -7.10713500
#>  [89,] -1.10649465 -0.203569732 -0.363621600 -0.268286079  -6.55178545
#>  [90,]  1.71743717  1.248598176 -1.890878953 -0.795851599 -22.21677547
#>  [91,] -0.81468656 -1.575277085 -0.852897395 -1.133744824  -5.73712908
#>  [92,] -0.42752002 -0.445973994 -0.367170260 -0.658952334  -4.28659152
#>  [93,] -2.81076001 -0.213092644  0.108482844 -1.912709077 -14.67194460
#>  [94,]  1.29151567  1.026158099  1.049175427  1.273914913  -0.54550189
#>  [95,] -1.29238339 -0.782752894 -1.797115560  1.096694989 -23.95305973
#>  [96,] -0.30636274 -0.088544958 -0.278672903  0.344592278 -11.71877174
#>  [97,] -0.01525169 -1.045983073 -0.061792703  0.400836209  -5.51497257
#>  [98,] -0.30787098 -1.182147579  0.209674241  0.668835286  -6.66802062
#>  [99,] -0.45675424 -0.603941418  1.155711340 -0.061236390   5.03156591
#> [100,] -0.81216689  1.902915987  1.314330569 -0.191860410   9.38291904
#> 
#> $eta
#>   [1]  -1.64972006 -18.56556388   5.89167855 -12.74676418  -3.46129189
#>   [6]  -2.22540512 -11.37501881 -10.59303461   3.97537182 -15.83035326
#>  [11]   1.59144336 -55.26653545   3.58359255 -12.60920117 -17.83959443
#>  [16]  -9.61122415 -34.49696674  -5.58961048 -13.30304645 -22.70322532
#>  [21]  -1.75086835  -7.79955968   7.11954826 -30.66885686 -20.12961426
#>  [26] -31.56393507 -19.83096047  -2.80278162  -1.77793695 -12.97590679
#>  [31]   9.19545905   0.54321705  -0.53345292   5.95306501   0.74881800
#>  [36]  -3.28631787  -8.21320529 -18.35619900 -11.41952610 -39.64955598
#>  [41] -18.95866407   7.75564204 -13.60853817   1.83805636   1.43543154
#>  [46]  -1.69596095  -9.98679147  -5.67652328   0.55986852 -20.90127783
#>  [51] -12.07550411 -11.47641332  -3.31524009   8.32740845 -41.75198334
#>  [56]  -1.07204464  -7.04656274 -11.98062096  -4.99928257 -10.14839344
#>  [61] -12.47683017  -1.33480535  -2.42412651 -10.05185392   7.36418098
#>  [66]  -6.41449929   6.55067680 -26.51749114  -2.58602825   0.78673566
#>  [71]  -4.65820575  -9.69357247  -1.02842690  -5.99997316   0.70795960
#>  [76]   0.27616122 -11.84813966   2.89137858   2.39956535 -13.49442874
#>  [81]  -5.90648671  -9.23102699   2.79687647  -3.16394789   0.06418612
#>  [86]  -5.05674283   0.55476058  -7.34006823  -6.31604823 -22.85643202
#>  [91]  -6.20024024  -4.12468206 -13.44172749  -2.26518255 -25.71469414
#>  [96] -10.42371748  -5.97096542  -6.85732936   5.69625453   9.04666624
#> 
#> $mu
#>   [1]  -1.64972006 -18.56556388   5.89167855 -12.74676418  -3.46129189
#>   [6]  -2.22540512 -11.37501881 -10.59303461   3.97537182 -15.83035326
#>  [11]   1.59144336 -55.26653545   3.58359255 -12.60920117 -17.83959443
#>  [16]  -9.61122415 -34.49696674  -5.58961048 -13.30304645 -22.70322532
#>  [21]  -1.75086835  -7.79955968   7.11954826 -30.66885686 -20.12961426
#>  [26] -31.56393507 -19.83096047  -2.80278162  -1.77793695 -12.97590679
#>  [31]   9.19545905   0.54321705  -0.53345292   5.95306501   0.74881800
#>  [36]  -3.28631787  -8.21320529 -18.35619900 -11.41952610 -39.64955598
#>  [41] -18.95866407   7.75564204 -13.60853817   1.83805636   1.43543154
#>  [46]  -1.69596095  -9.98679147  -5.67652328   0.55986852 -20.90127783
#>  [51] -12.07550411 -11.47641332  -3.31524009   8.32740845 -41.75198334
#>  [56]  -1.07204464  -7.04656274 -11.98062096  -4.99928257 -10.14839344
#>  [61] -12.47683017  -1.33480535  -2.42412651 -10.05185392   7.36418098
#>  [66]  -6.41449929   6.55067680 -26.51749114  -2.58602825   0.78673566
#>  [71]  -4.65820575  -9.69357247  -1.02842690  -5.99997316   0.70795960
#>  [76]   0.27616122 -11.84813966   2.89137858   2.39956535 -13.49442874
#>  [81]  -5.90648671  -9.23102699   2.79687647  -3.16394789   0.06418612
#>  [86]  -5.05674283   0.55476058  -7.34006823  -6.31604823 -22.85643202
#>  [91]  -6.20024024  -4.12468206 -13.44172749  -2.26518255 -25.71469414
#>  [96] -10.42371748  -5.97096542  -6.85732936   5.69625453   9.04666624
#>