Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                 x1          x2          x3          x4 y
#>   [1,] -0.92935335 -1.80846714 -0.97186349 -0.93007347 0
#>   [2,]  0.45960504  1.03626577  0.24369286 -0.77039416 0
#>   [3,]  0.83369578 -0.18597129  1.85325599 -0.57875956 1
#>   [4,]  0.36189653 -1.56910161  0.87371020 -1.66139146 0
#>   [5,]  2.91827666 -1.35306485 -0.04008846  1.88546131 0
#>   [6,]  0.72182904 -0.60088386  0.06077630  0.19464792 0
#>   [7,] -1.68895715  1.25671096 -1.89112225  1.03890178 0
#>   [8,]  0.42668496  0.37988958 -0.22039477  1.71482161 0
#>   [9,] -1.55490837 -0.84660749  1.34774705  0.84279242 1
#>  [10,] -0.89122800 -0.63117736  0.53645959 -0.13736992 1
#>  [11,]  1.81987810  0.59265729  1.01265580  0.66516209 1
#>  [12,]  0.89384735  0.40418750 -0.19969586 -1.93756194 0
#>  [13,] -1.29253755  0.60293933 -1.03821439 -0.18026836 0
#>  [14,] -0.11691527  0.79107588 -2.14687028  0.59110683 0
#>  [15,] -0.07806530 -1.55115459 -1.99269509  0.18383523 0
#>  [16,]  0.58727678 -0.84225942 -0.08968637 -0.34212774 0
#>  [17,]  0.70607634 -1.82390098  0.20112085 -1.95661154 0
#>  [18,] -1.17810913  0.67388007  1.53093062  0.30539930 1
#>  [19,]  0.88718053 -0.08954574 -1.45213341 -0.86525627 0
#>  [20,] -0.17042769 -0.72284090 -0.25409564 -2.68076867 0
#>  [21,] -0.69310785 -0.30693633 -0.96751241  1.57114165 0
#>  [22,]  0.41596747 -1.00956213 -0.81814997  0.70516257 0
#>  [23,] -0.93670853 -0.72670355  0.46050331  0.64194457 0
#>  [24,]  0.74601651 -0.12848621  0.20174771  0.63203861 0
#>  [25,]  0.86315782 -0.03788483 -0.64428855  1.02466866 0
#>  [26,]  0.35243736 -0.94648650  0.30125681  1.00550027 0
#>  [27,] -2.49711891  1.06140306 -0.49765482 -0.02283742 0
#>  [28,]  0.46800277  3.07466765  0.26105012  0.77252479 0
#>  [29,]  0.72912381  1.38724728 -0.29791354  0.73159052 0
#>  [30,]  0.74436664 -0.52741833  0.39778572  0.52386228 0
#>  [31,]  0.05623650  1.29113337  1.22487602 -0.57561640 1
#>  [32,]  0.18002273  0.47636845 -1.30028111 -1.61795585 0
#>  [33,]  0.83311200  1.32507940  0.34180443 -0.19691336 0
#>  [34,] -0.72832464  0.43917496  0.35271606 -0.04573353 1
#>  [35,] -1.42507491  0.74009144  0.12696674 -1.45307776 0
#>  [36,] -0.20091654 -0.40020050  0.41848510 -0.68788172 0
#>  [37,]  0.31353195  1.90217724  0.02797679 -0.66277817 0
#>  [38,]  1.28327169  0.25022838 -0.06865987 -1.29016079 0
#>  [39,] -1.56278930 -0.92376023  0.45240963 -1.80772562 0
#>  [40,] -0.38063813  0.26016319  1.21803198 -0.34729266 0
#>  [41,] -0.26457801 -0.25731545 -1.46098771 -0.40942103 0
#>  [42,] -0.58818866  1.59852897  0.72071376 -0.12347491 1
#>  [43,] -0.96648455 -0.01018691 -0.05800576  0.87631205 0
#>  [44,] -0.53450791  1.63391749 -1.08494090  0.17945780 0
#>  [45,] -0.27481909  0.96786275 -2.37322354 -0.62133406 0
#>  [46,]  0.21899134  2.01629351 -0.37854974 -1.49425290 0
#>  [47,] -0.73210279  0.15161888 -0.02724033 -1.10628144 0
#>  [48,]  0.55959541  1.99384897  0.70441422 -0.60214361 0
#>  [49,]  1.32264978 -1.12243356  1.13213578 -1.45678128 0
#>  [50,] -1.72989760 -0.50915545  1.88960029  0.15172253 1
#>  [51,]  1.83120243  0.03577074  1.55121400 -1.14500714 0
#>  [52,]  0.56503009  0.36054856 -0.32862545  1.17865226 0
#>  [53,]  0.45915915  1.10487159  0.75103926 -0.32496672 0
#>  [54,] -1.29546829 -0.33314200 -0.49027984 -1.40923125 0
#>  [55,] -0.49577273  0.23230610 -1.40889902  1.34023028 0
#>  [56,]  1.42074924  0.39840617 -0.21964196 -1.36697360 0
#>  [57,] -0.81164315 -0.28627177 -1.72243958  1.10710158 0
#>  [58,]  0.16372283 -0.83418813  0.71145637 -0.07642484 1
#>  [59,] -0.18996101  0.47160786  1.26851765 -0.09635833 1
#>  [60,] -0.03367774  0.01217213  0.95565033  0.36249486 1
#>  [61,] -0.33629525  1.73262764  0.19167702 -0.18126680 0
#>  [62,] -0.59694631 -0.14118518  1.87617746  0.85301310 1
#>  [63,] -0.47107276 -1.20663290  0.93631390 -0.40820321 1
#>  [64,]  0.27692164 -1.94425076 -0.05768384  0.38268555 1
#>  [65,] -1.59347286  0.69959732  1.28196255  1.11705786 1
#>  [66,] -0.97685109 -0.22311295 -1.76310013  0.39368631 0
#>  [67,] -0.51928373  0.38698231  0.41086977 -1.34510734 0
#>  [68,]  0.04459765  1.67222866  1.90125209  0.46185967 1
#>  [69,] -0.54205908  0.59748102 -0.55737696  1.62808912 0
#>  [70,]  0.33533422  0.68297143 -1.06270631 -2.31825224 0
#>  [71,] -1.32507434  2.14430889  1.38748072 -1.71584642 0
#>  [72,] -0.11629614 -0.97333192 -0.26815009  1.04571635 0
#>  [73,] -1.28472008 -1.12427965  0.31805209  1.14117748 0
#>  [74,] -1.24196935 -0.09163673 -0.24444600  0.16974680 0
#>  [75,] -1.12138057  0.63371673 -0.92196538 -0.16176329 0
#>  [76,]  1.18210364  1.31892221 -0.85255753 -0.67552877 0
#>  [77,]  1.57585189  0.11925023  0.78478998  1.27344160 0
#>  [78,] -1.10756389  0.60748117  0.67920582 -0.15991183 1
#>  [79,] -0.19826568  0.18295707 -1.11366953  1.44401015 0
#>  [80,] -1.33890901 -0.50496503  0.63753371  0.68444824 0
#>  [81,] -0.73227058 -0.02992234 -1.08660391 -0.67482937 0
#>  [82,]  2.29401257 -0.14065170 -0.20336479 -1.31711610 0
#>  [83,]  0.88508270  0.59750585  0.38423916  0.71057575 0
#>  [84,]  0.13112751 -1.92030983  0.89732095 -0.98323364 1
#>  [85,] -0.68392721  0.66594232 -0.73111687  0.59174906 0
#>  [86,] -1.17010084  1.35962352 -0.51351442  0.53787771 0
#>  [87,]  0.08672741 -0.53416912 -0.40929894 -0.17768908 1
#>  [88,]  0.37413241  0.95044959  0.02260856  0.65777629 0
#>  [89,] -1.32709347 -0.73003637  0.63465980 -0.02880944 0
#>  [90,]  1.48132882 -0.14016677  0.75506271 -0.41229380 1
#>  [91,] -0.13942465  0.68126665 -0.47840397  1.25859855 0
#>  [92,] -0.02647986  1.06943574  0.94680884 -0.34152563 1
#>  [93,] -0.55590585  0.37393374 -1.37159446 -1.23199218 0
#>  [94,] -0.32064002 -0.75824241 -0.41840140  1.21437551 0
#>  [95,]  0.96065564  0.88900584 -0.23784349  0.72682943 0
#>  [96,]  1.37217979  0.51169699  0.43248372 -0.22870257 1
#>  [97,] -0.67517641 -0.59592604 -0.45081102  1.49711796 0
#>  [98,]  2.10100223  0.98565817  0.25925843  0.55606249 0
#>  [99,]  0.77861678 -1.51840724 -0.09744787  0.32691103 0
#> [100,] -1.02394255 -0.91846071  0.65313543  0.27794877 0
#> 
#> $eta
#>   [1] -12.5595991  -5.6027113   4.0104619  -3.8666140 -18.3243387  -4.5540411
#>   [7] -15.3746199 -14.0592473   1.1485203   5.7580786   1.8244668 -22.7547500
#>  [13]  -8.7572609 -18.5214712 -13.2557295  -8.7571195 -27.5323880   6.0388587
#>  [19] -19.1573036 -44.6388963 -11.9381004  -9.7650214  -0.2462785  -8.3271325
#>  [25] -17.3631986  -4.6944906  -8.5529919  -5.9173043  -9.5027623  -2.3526835
#>  [31]   4.9659353 -14.6781454  -4.0896812   8.1014339 -11.4308402  -2.1513417
#>  [37]  -0.6836277  -7.2560563 -17.2022115   0.1072509 -14.0061241   9.3506427
#>  [43]  -8.0229757  -3.3259377 -25.6633532 -13.7177905  -6.4114810  -4.3488101
#>  [49]  -2.0180797  18.0769404  -6.9605664 -11.6309803  -0.2023535 -15.7552463
#>  [55] -21.3625857  -8.0963856 -10.6721551   3.5956095   4.5305784  -1.1198416
#>  [61]  -2.5083858   8.2489967   3.2831881  -1.6049594   5.6388247 -10.8286658
#>  [67]  -6.0653531  13.5648952 -12.8169153 -27.2641509 -16.0234263 -15.6945035
#>  [73] -12.1281309 -10.0045995  -6.8460564  -4.6127587 -11.0696632   3.2582768
#>  [79] -26.5314240  -0.8082432  -8.4176389  -7.3279991  -1.0491166  -1.2808159
#>  [85]   0.3631284  -9.6600453   1.4992900  -2.7270679  -3.5844511   2.0098348
#>  [91] -12.9208623  -1.2222840 -14.4977337 -17.9963304  -8.2510856   8.0726705
#>  [97]  -9.1494536  -1.8496490  -4.2728558  -2.7165047
#> 
#> $mu
#>   [1] 3.511025e-06 3.674301e-03 9.821976e-01 2.050007e-02 1.101136e-08
#>   [6] 1.041497e-02 2.103233e-07 7.836933e-07 7.592405e-01 9.968528e-01
#>  [11] 8.611012e-01 1.311407e-10 1.572902e-04 9.041224e-09 1.750286e-06
#>  [16] 1.573124e-04 1.103661e-12 9.976214e-01 4.787279e-09 2.220446e-16
#>  [21] 6.536512e-06 5.742224e-05 4.387397e-01 2.418061e-04 2.879113e-08
#>  [26] 9.062642e-03 1.929297e-04 2.685219e-03 7.463978e-05 8.685271e-02
#>  [31] 9.930768e-01 4.220484e-07 1.646881e-02 9.996970e-01 1.085537e-05
#>  [36] 1.042059e-01 3.354521e-01 7.053885e-04 3.382007e-08 5.267871e-01
#>  [41] 8.264512e-07 9.999131e-01 3.277356e-04 3.469201e-02 7.153973e-12
#>  [46] 1.102653e-06 1.639896e-03 1.275733e-02 1.173177e-01 1.000000e+00
#>  [51] 9.476603e-04 8.886394e-06 4.495836e-01 1.437419e-07 5.276512e-10
#>  [56] 3.045454e-04 2.318098e-05 9.732891e-01 9.893404e-01 2.460407e-01
#>  [61] 7.527239e-02 9.997385e-01 9.638475e-01 1.672896e-01 9.964556e-01
#>  [66] 1.982264e-05 2.316559e-03 9.999987e-01 2.714459e-06 1.443211e-12
#>  [71] 1.099295e-07 1.527438e-07 5.405269e-06 4.518955e-05 1.062512e-03
#>  [76] 9.826876e-03 1.557756e-05 9.629694e-01 3.002956e-12 3.082650e-01
#>  [81] 2.208869e-04 6.564554e-04 2.593948e-01 2.174114e-01 5.897975e-01
#>  [86] 6.377757e-05 8.174686e-01 6.139491e-02 2.700252e-02 8.818258e-01
#>  [91] 2.446469e-06 2.275348e-01 5.054917e-07 1.528597e-08 2.609070e-04
#>  [96] 9.996881e-01 1.062666e-04 1.359141e-01 1.375021e-02 6.200645e-02
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                  x1           x2           x3          x4    y
#>   [1,]  2.947968855  2.287707126 -1.191773860  0.93226124    0
#>   [2,]  1.253632456  0.248478154 -0.902802377 -1.68558290    0
#>   [3,] -0.564410235  0.475298533  0.084748499  1.63627703    0
#>   [4,] -0.892141291 -0.965333926  1.018411239  1.66677434    0
#>   [5,] -0.950800928  0.411885930  1.991661789  0.24540984   91
#>   [6,] -0.278382241 -0.106842381  1.089360252  1.06664734    0
#>   [7,]  0.253245891 -1.089827880  0.482680636 -0.65622923    2
#>   [8,]  0.631186299 -1.042905523 -0.356658393  0.39756090    1
#>   [9,]  0.161518506 -0.673593373 -0.453173721 -1.00200705    1
#>  [10,]  0.438197252  0.029675189 -0.225170554 -2.38542290    0
#>  [11,] -1.418977436  0.758797796 -0.615016558 -1.58245828    0
#>  [12,]  0.330301760 -0.766638468  1.259867256  0.10725118   12
#>  [13,] -0.170598152 -1.232528560  0.327552601  0.93239518    3
#>  [14,] -0.235608562 -0.584692587 -0.080690668  0.23525606    3
#>  [15,] -1.411528442 -0.082923430  0.921552141  0.97761004    2
#>  [16,] -0.190842023  0.237434741 -1.229577824  0.71637330    1
#>  [17,]  1.197067904  0.337750794 -0.332937414  1.47800491    0
#>  [18,] -0.654275482 -0.664855787 -0.071892051  0.53498211    7
#>  [19,]  0.008184199 -1.379415326 -1.255306156  0.38563353    7
#>  [20,]  0.426701888  0.006408004  0.746043890  1.38899296    0
#>  [21,]  1.250372372 -0.507583076  0.681219136  0.38311082   31
#>  [22,]  0.332923902  0.837216506 -0.485503413 -0.17314494    8
#>  [23,]  0.035901559  0.933984231 -1.004696366 -0.60893115    1
#>  [24,]  0.045317076  0.521517477  0.273360086 -1.31776037    0
#>  [25,] -0.124356289 -0.340760012 -0.075028764 -1.69931786    0
#>  [26,] -1.245353933 -1.086730863  1.451360958  0.78921440    0
#>  [27,]  0.991628635  0.079448589 -1.449669958 -0.24894784    4
#>  [28,]  0.080577500 -1.655392159 -0.175015244  1.24183309    0
#>  [29,]  0.877281190  0.935193023 -0.709530178  1.46765034    0
#>  [30,] -0.269339658 -0.190569920 -0.107536519 -0.44035943    1
#>  [31,] -1.704325337  0.969114474 -0.628078636 -1.82758142    0
#>  [32,] -1.584895446 -0.735024635  0.047608740 -2.13465349    0
#>  [33,]  0.222961910 -0.594703500  1.333200442  0.97339685    8
#>  [34,] -0.813309362 -0.340964016 -0.921320048 -0.05104144   22
#>  [35,]  0.899888895  0.792544608 -0.586303914  0.25500989    2
#>  [36,]  0.699961348  0.720172715  0.563092709 -1.51160303    0
#>  [37,] -1.583394015 -0.192007940  2.705453184 -0.52969661 1152
#>  [38,]  0.097196153  0.206695759  0.929773232 -0.04766320    8
#>  [39,]  0.455438222  0.001516056  2.563118705  1.80589161    1
#>  [40,] -1.647624986  1.413864436 -0.667232534 -1.32942464    0
#>  [41,]  1.747840863 -0.647156215 -0.489172881 -0.40921000    2
#>  [42,]  0.011738928 -1.495045458 -0.691284123  1.16054814    0
#>  [43,]  0.496098739 -1.431795602  0.274417985  0.54898762    9
#>  [44,]  0.192045363 -0.174216267 -1.984924281 -0.17652471    4
#>  [45,]  0.493427474  1.969097176  0.767439213 -0.47928400    2
#>  [46,]  1.091840843 -1.764538138 -0.692108716 -0.36760369    2
#>  [47,] -0.030157329 -0.653311142 -0.604582276 -0.98872020    1
#>  [48,]  0.788475063 -0.073556856 -1.087833897  1.11932205    0
#>  [49,]  1.885493987  0.191130861  0.663260031  0.78539988    1
#>  [50,]  0.559586603  1.267384923  0.553640373 -0.34487228    4
#>  [51,]  0.926021275  0.039374872  0.589563404 -0.83907237    0
#>  [52,] -2.159786463  0.532339181  1.410993809 -0.21844537   14
#>  [53,]  0.306185826  0.408305062  0.377260606 -0.70272581    4
#>  [54,]  0.824914073  2.277870201  0.175634626  0.69839314    0
#>  [55,]  0.346718871  0.117770395  0.426834189  0.52816130    5
#>  [56,] -0.891341312 -0.467197690 -0.238922602 -1.14006166    0
#>  [57,] -0.404430036  0.453200520 -1.439280526 -0.17962670    4
#>  [58,]  0.536562773  0.636127605 -0.196632800  0.32868818   11
#>  [59,] -1.065123953 -0.896457109 -1.019098401 -0.02890827    1
#>  [60,]  0.332481412  2.175478187 -1.504830654 -0.92384379    0
#>  [61,]  0.704886026  1.371110027 -0.606836018  0.06366939    2
#>  [62,] -0.940193407  1.464814587 -0.437792595  0.18810105    7
#>  [63,]  0.743696907  0.237372075  1.223460785 -0.22760563    1
#>  [64,] -1.087113925  0.328727135 -0.321534446 -0.63445296    1
#>  [65,] -0.500193074  0.152682662 -0.237875928 -0.15724992    2
#>  [66,] -0.969124696  0.919071133 -1.315019319  0.97252610    0
#>  [67,] -0.148683023  1.727944436  1.093039555 -0.48774714    2
#>  [68,] -0.114352846  2.207619329 -0.471191580 -2.08361568    0
#>  [69,]  0.704209246 -0.237361757  1.015256584 -0.75161892    1
#>  [70,] -0.993174562 -1.160433338  1.010622050 -1.17458890    0
#>  [71,] -0.412691389  0.644692214  0.529814859 -0.58395316    2
#>  [72,]  0.045576628  1.502051960 -0.022140245 -0.94471376    2
#>  [73,] -1.188115204 -0.550048457  0.126518473 -0.19085490    2
#>  [74,]  0.807267854 -0.098288553 -2.038184845 -1.68253898    0
#>  [75,] -0.321665248  1.055553892 -0.586504326  0.17030321    2
#>  [76,] -1.065418915 -1.254590610  1.523591882 -2.03065489    0
#>  [77,] -0.575302643  0.279991864 -0.986631435  1.30731534    0
#>  [78,] -0.776572337  1.023003696 -1.108866120  0.09282769    4
#>  [79,] -0.619633077  0.532514231  1.034683389 -1.97098281    0
#>  [80,]  0.866036115 -0.994028839  0.117600434  0.17440698    0
#>  [81,]  0.542543964 -0.956019714 -0.246209467 -1.36945204    0
#>  [82,]  0.018693970 -0.196492809  2.395784791  0.84699234   60
#>  [83,]  0.936620058  0.131472494 -0.825799650 -1.96995400    0
#>  [84,]  0.902576568 -0.247453973  0.602419328 -0.04587663    6
#>  [85,] -0.059299987  0.050356948  0.653968621 -0.13812903    1
#>  [86,] -0.828349949 -3.058660445  0.064808800 -0.61852707    4
#>  [87,] -0.948634456  1.125451194 -0.388653336 -0.75320874    1
#>  [88,] -0.953484318  0.430569920 -1.072883700  0.20084033    4
#>  [89,]  1.369520012  0.785920766 -2.007885650 -0.14594138   11
#>  [90,] -0.468482762 -0.849375651 -0.634367679  0.96207444    0
#>  [91,]  0.333523480 -0.242756098  0.824855222  0.30023247   22
#>  [92,]  0.110688672  0.325544181  0.727652239 -0.09442078    9
#>  [93,]  0.225932888  0.153851156  0.842259702  0.34079434    4
#>  [94,]  0.259804608  0.603299743 -0.017080548 -1.66363393    0
#>  [95,] -0.775111454 -0.201903172  0.599588722  1.65829063    0
#>  [96,]  0.458904096 -0.637708888 -0.001951284  0.04845124    7
#>  [97,] -0.317114485 -0.160173233 -1.274559244 -0.85533855    0
#>  [98,]  0.184423242 -0.266357628 -1.949943090  0.07552326   14
#>  [99,]  0.332236566 -1.339652359  0.480011315 -0.25123361   37
#> [100,] -0.514787982 -0.975546034 -3.059405322  2.03642514    0
#> 
#> $eta
#>   [1]  -1.19793311  -4.00826207  -2.73275970  -3.78220531   4.60926825
#>   [6]  -1.45961019   1.30165891  -0.48835612   0.07568693 -10.63374657
#>  [11]  -3.84926278   2.69398169  -0.68141008   0.77536996  -0.95635497
#>  [16]  -1.38827129  -2.08573663   1.65932382   1.83495913  -2.46687842
#>  [21]   3.35071635   2.00552636   0.01857843  -0.73213720  -4.23119046
#>  [26]  -0.03967308  -0.08266523  -1.16929265  -4.42410285   0.10501043
#>  [31]  -5.27621453  -7.58036332   2.06587804   2.95384904   0.73473235
#>  [36]  -3.18389299   7.05882101   1.88483936  -0.93776327  -1.02415310
#>  [41]   0.05610614  -0.52760041   1.94548882   1.95089946   0.67496928
#>  [46]   0.91802058  -0.81765948  -2.91364683  -0.88049761   0.71134965
#>  [51]  -1.10333404   2.65724143   2.00872989  -0.70915757   0.90466833
#>  [56]   0.30353918   1.16622887   1.82234673  -0.24385020  -0.57334968
#>  [61]   0.50783289   2.30550974   0.70927286  -0.04081836   0.37073004
#>  [66]  -1.12458359   1.50946645  -8.78492561   0.42357145  -0.68024478
#>  [71]   1.39343973   0.77243215   1.14200367  -5.98360022  -0.72445113
#>  [76]  -5.30667439  -2.11987028   1.16334861  -4.00439945  -0.25165557
#>  [81]  -3.67732528   4.23619227  -7.92137899   1.61711803   0.38387919
#>  [86]   0.89696915  -0.50504118   1.93707675   2.01301717  -1.73780990
#>  [91]   3.03676837   2.42566811   1.79188057  -2.15511445  -2.58721781
#>  [96]   1.85398015  -1.32813752   2.52874781   3.26361600  -8.12312409
#> 
#> $mu
#>   [1] 3.018174e-01 1.816494e-02 6.503955e-02 2.277242e-02 1.004106e+02
#>   [6] 2.323268e-01 3.675389e+00 6.136343e-01 1.078625e+00 2.408921e-05
#>  [11] 2.129543e-02 1.479045e+01 5.059031e-01 2.171395e+00 3.842911e-01
#>  [16] 2.495063e-01 1.242156e-01 5.255756e+00 6.264878e+00 8.484931e-02
#>  [21] 2.852316e+01 7.430004e+00 1.018752e+00 4.808802e-01 1.453508e-02
#>  [26] 9.611036e-01 9.206593e-01 3.105866e-01 1.198496e-02 1.110722e+00
#>  [31] 5.111745e-03 5.103758e-04 7.892225e+00 1.917964e+01 2.084924e+00
#>  [36] 4.142408e-02 1.163073e+03 6.585297e+00 3.915025e-01 3.591005e-01
#>  [41] 1.057710e+00 5.900191e-01 6.997051e+00 7.035012e+00 1.963973e+00
#>  [46] 2.504328e+00 4.414637e-01 5.427743e-02 4.145766e-01 2.036738e+00
#>  [51] 3.317631e-01 1.425691e+01 7.453844e+00 4.920585e-01 2.471112e+00
#>  [56] 1.354645e+00 3.209865e+00 6.186359e+00 7.836050e-01 5.636343e-01
#>  [61] 1.661686e+00 1.002929e+01 2.032513e+00 9.600035e-01 1.448792e+00
#>  [66] 3.247877e-01 4.524316e+00 1.530225e-04 1.527407e+00 5.064930e-01
#>  [71] 4.028684e+00 2.165026e+00 3.133040e+00 2.519738e-03 4.845905e-01
#>  [76] 4.958389e-03 1.200472e-01 3.200633e+00 1.823524e-02 7.775125e-01
#>  [81] 2.529053e-02 6.914407e+01 3.629015e-04 5.038548e+00 1.467968e+00
#>  [86] 2.452160e+00 6.034807e-01 6.938439e+00 7.485869e+00 1.759052e-01
#>  [91] 2.083779e+01 1.130978e+01 6.000727e+00 1.158899e-01 7.522905e-02
#>  [96] 6.385183e+00 2.649703e-01 1.253780e+01 2.614390e+01 2.966006e-04
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                  x1            x2          x3           x4           y
#>   [1,] -1.674190170  0.3506438135 -0.83163580 -1.068346449  -8.0062568
#>   [2,] -0.384864266  0.4027542617 -0.42277347  1.490872690 -18.6219910
#>   [3,]  0.948613444 -0.7963044230  0.25811963 -0.600791211  -5.0918154
#>   [4,] -0.209842072  1.6582783764  0.22640534  0.337679491  -2.4875801
#>   [5,]  0.916643858  0.0570174286 -0.86160210 -0.428711143 -13.3289023
#>   [6,]  0.166310901 -0.5969593918  1.17762153 -0.172515260  12.0493628
#>   [7,]  1.246936177 -0.8425605235  1.09660093 -0.072899373   7.0404101
#>   [8,] -0.594885993 -0.8770269282  1.05230708 -1.065914137  -2.0397651
#>   [9,] -2.831277392 -0.4254788677 -0.00549834  2.545288143 -25.9857097
#>  [10,]  0.854856071  1.3076260530 -0.25203572  0.351325691  -8.8241448
#>  [11,]  0.813039122  0.4453272467  2.00051485  1.724263245  -6.1377048
#>  [12,]  1.155697604  0.4882874214 -2.55230610 -1.338663699 -17.4867149
#>  [13,]  0.093953689 -1.3448611998  2.14228191  1.501258405   4.8704074
#>  [14,] -1.458356698  0.8825635307 -0.43591579 -0.683134465  -8.9550091
#>  [15,]  0.731658700 -0.6976082454 -1.87822436 -0.836187945 -22.0896323
#>  [16,] -0.327973825 -0.3673193595 -1.47353322  0.349917842 -12.6481360
#>  [17,] -0.753690542 -0.3368223551 -1.67339641 -0.889216029  -6.4047905
#>  [18,] -0.561119478  1.1155387218  1.68625582 -0.278949610   4.9348716
#>  [19,] -0.042269348 -0.7544958888  1.29759043  0.228087230   3.7248027
#>  [20,]  0.513661855 -0.0563187047 -0.50831401 -1.428291130 -17.4324539
#>  [21,]  1.016836925 -1.0248296986 -0.09676160  0.337720918  -5.1107076
#>  [22,] -1.593307168 -0.3156475389 -0.51508541  1.403921610  -9.1363292
#>  [23,]  0.511465090 -0.1458873116 -0.88848531  0.059246309  -6.4587434
#>  [24,] -0.617148274 -1.3158927975  0.77158737 -0.227217692   9.1560538
#>  [25,]  1.288064971 -0.3501037036 -0.17640940 -0.741845223   2.6969410
#>  [26,] -0.583449195  0.1473576752  1.28016773 -1.571219731  -6.4157787
#>  [27,] -0.792021134 -1.4042302212  1.32001007 -1.322245335   7.1334597
#>  [28,]  0.929494235  1.3407317370  0.15209371 -0.521153398  -5.1054152
#>  [29,]  0.514025401  1.5071327869  1.01930267  0.132758511   9.0925569
#>  [30,] -0.935920847  0.2705424379  0.19640798 -0.368842875  -0.3334932
#>  [31,] -2.329169865  1.4668988789  1.48267205  0.580820195   4.7265586
#>  [32,]  0.656706387 -0.5443999368 -0.46188810 -1.122614288 -12.6289015
#>  [33,]  1.578396208 -0.4664958133 -1.38510453  0.858932681 -12.5532466
#>  [34,]  0.506695081 -0.3989811314 -0.83814873 -0.161853431  -1.2870501
#>  [35,] -0.043892169 -0.8987777600  1.32722975  0.327999694   0.9742294
#>  [36,]  0.726913068  0.2461435749 -1.31355515  0.722775558 -19.9829767
#>  [37,] -1.670889424  0.6755916549  0.72296209  0.639494975  10.5452331
#>  [38,]  1.335886070 -1.3874300366 -2.97358186  0.811615774 -14.8723658
#>  [39,] -1.651906654 -1.5137756302 -0.55798722  0.854265714   0.1342966
#>  [40,] -0.968394717 -0.0007782879 -0.32964822 -1.788977804 -23.0935548
#>  [41,]  0.157830257 -0.1736979621 -2.54694086 -0.519457572 -12.6759272
#>  [42,]  0.869614749  1.2118162454 -0.84002756  0.506891318 -13.2367059
#>  [43,]  1.287651594 -0.5465903676  0.85038881 -0.173181097  12.6701586
#>  [44,] -0.123160525 -0.3057712127 -1.06627212 -0.257001485  -8.5411972
#>  [45,]  0.034775441  0.1518531410  0.15548549  0.975448365  -8.7528744
#>  [46,]  0.620550798  0.1166558212 -0.43990396 -0.837158656 -16.3600262
#>  [47,] -1.553625058  0.4044545541 -1.45238723  1.416826764 -17.7416301
#>  [48,] -0.409526464 -0.9582122113 -0.34862249 -1.860999883 -29.5771409
#>  [49,]  1.450599090 -0.9173693013 -0.10770992 -0.773246743  -6.3167067
#>  [50,]  0.005732857 -1.8239037371  0.68895617  0.734727734  -2.7865280
#>  [51,] -0.764042045  1.1349958463  0.46267277 -1.757480680 -14.5046341
#>  [52,] -0.348390347 -0.8014252595 -1.08688924 -0.636133199 -17.5681162
#>  [53,]  0.583481639  0.1628359490  2.29920897 -0.945186078   0.1640832
#>  [54,]  0.588325954 -1.0117729793 -1.22191650 -0.510060631 -16.3703661
#>  [55,] -0.213783499 -0.9674590082 -0.08708008 -0.527021869 -12.4698164
#>  [56,] -1.311920289 -0.0014598252  0.56627094 -2.039706000 -25.8596900
#>  [57,]  0.018045881  0.5898823822 -0.72065179 -1.804037524 -17.3013818
#>  [58,]  0.900332356 -0.3818110510  0.56920860 -1.020693845  -2.6365445
#>  [59,]  0.616193115 -0.9512568161 -0.58003983  0.209042722 -11.5811112
#>  [60,] -0.406530456  0.3984629135  0.94480578  1.236887674  -5.1268932
#>  [61,] -0.296391899 -1.6064613181 -0.59596000  0.243532919  -7.4829975
#>  [62,] -1.046907459 -0.4410220221  0.35568028 -0.049503200   5.3366455
#>  [63,] -0.050009025 -0.7550762768 -1.86965501 -0.168264073 -18.1295420
#>  [64,]  1.721412984 -0.2757846701  1.03210419  0.532991525   1.9408862
#>  [65,]  1.395808807  0.4902763314  0.09090007 -1.140123464  -2.1195193
#>  [66,] -0.272137352  1.3812169361 -1.25183584 -1.733736909 -27.6281952
#>  [67,] -0.068549849 -0.9905230264  0.63936013 -0.218875956  -3.2759365
#>  [68,]  0.607018510 -0.3590280441  1.46620467 -0.826799365   3.5692021
#>  [69,]  1.827660814 -0.3524103508  0.27977590 -0.339436244  -3.0180551
#>  [70,] -0.751369041 -0.0613039465  2.02981778  0.917778512   8.5888532
#>  [71,]  0.312835047 -0.0040482003 -0.15285442  1.566652032 -13.9530684
#>  [72,]  1.250637250  0.7608472531  0.70654843  0.717761010   7.6612058
#>  [73,]  2.024394393 -0.4303650866  0.08326398  0.862519981   0.4070582
#>  [74,]  0.343672082  1.2750230973  2.80788622 -0.731937279  14.7532386
#>  [75,] -0.330504114 -0.7861454343  0.28918370 -0.866883925 -12.1457257
#>  [76,] -1.013707991  0.1217326772  0.16026047 -0.501521255  -6.3139977
#>  [77,]  1.788613811  0.0401698269 -1.54386990  1.818702084 -35.6870542
#>  [78,]  0.501336737  0.6801976599 -1.61016670 -1.876586646 -27.7454555
#>  [79,]  0.498626569  0.1045248362  0.82584380 -2.080798563 -23.1936170
#>  [80,]  0.332624026  0.9861607775  1.73942848  0.674552486   8.4265823
#>  [81,] -1.470273832  0.6703165690 -0.82141018 -0.745784849  -6.5534124
#>  [82,] -0.561720871  1.2511618809  0.61688761 -0.935593974  -1.3918488
#>  [83,] -0.003947737  0.0887249759  0.08845947  0.539578778  -5.6248307
#>  [84,]  0.756907437 -0.2015586310 -2.19224798 -0.395752580 -18.2400090
#>  [85,]  1.787622084  0.1105049248 -0.57343055  2.137581774 -37.0305565
#>  [86,]  1.681662732 -0.9564095300 -2.05039202 -0.394839415 -22.2018522
#>  [87,] -1.379341044 -0.7889964723 -0.92879008 -0.508529644 -15.0818354
#>  [88,] -0.306133006  0.3938848359 -0.98655371 -0.405433926 -11.3037727
#>  [89,] -0.449052658  0.8809942859  0.41006696 -1.077655888  -8.5977844
#>  [90,] -1.931516796  1.0355212083  0.76325961 -0.955714722  -4.0855221
#>  [91,] -1.737492765  0.9867622552  1.29081484 -1.018988257   4.2851727
#>  [92,] -0.187543848 -0.1788766257 -0.62247046 -0.920635374 -13.8789199
#>  [93,] -0.206378373 -0.4858746216 -1.44001049  0.624122442 -14.4962862
#>  [94,] -1.088276791  1.5119525638  0.68314331 -0.023190749   4.9167246
#>  [95,] -1.317748785  0.5522757411 -0.22363556 -0.107309891  -3.0703709
#>  [96,] -0.869981509 -0.2353900693 -1.43977369 -3.329620659 -60.3397627
#>  [97,]  1.944662396 -1.1452360518  1.20651744 -0.966343840  -1.2372368
#>  [98,] -1.052933733 -1.4251951211 -0.57069504  0.006819878  -1.5624999
#>  [99,] -0.916839021  0.2513031833  1.63705824  1.144022893   2.5758843
#> [100,]  0.882237292  0.8613322085 -0.38356567 -1.135704499 -13.4296706
#> 
#> $eta
#>   [1]  -7.22530045 -16.68488866  -6.18654263  -2.03254578 -14.04908108
#>   [6]  11.52291165   6.52131148  -1.91179751 -25.56540769  -8.13285960
#>  [11]  -6.31586615 -18.26642934   6.42557617  -9.04946543 -21.91216652
#>  [16] -11.58585100  -7.03450482   6.20342136   2.69696368 -17.80694535
#>  [21]  -5.29261521  -8.51513591  -7.82253525   9.39611667   3.16343668
#>  [26]  -7.30220872   6.48444787  -3.80859681   7.68036270  -0.49254313
#>  [31]   4.35394107 -11.89139781 -12.03038662  -2.90638940   0.60089308
#>  [36] -18.94302839   8.48185770 -14.43125602  -1.08536107 -21.99461148
#>  [41] -13.86378709 -14.17487927  11.12070689  -8.71608495  -7.83534147
#>  [46] -14.97086567 -16.78511936 -27.48779418  -7.60790188  -3.17203963
#>  [51] -12.94705154 -17.05233191   1.85929443 -16.12589426 -12.10892648
#>  [56] -27.02330670 -17.53933516  -2.34143702 -12.45848173  -4.57341215
#>  [61]  -8.59678920   3.13597191 -16.43351916   0.81147839  -1.02289824
#>  [66] -27.71675482  -3.60919136   4.47090980  -1.17359422   8.93454652
#>  [71] -13.53418224   6.83753358   0.03545287  15.48295522 -10.55506689
#>  [76]  -6.08667924 -36.03317595 -25.58400016 -22.10136352   7.67216628
#>  [81]  -6.77056914  -2.17659905  -6.54237352 -19.65432656 -36.57252986
#>  [86] -23.02615275 -14.51550161 -11.36083270  -7.68955590  -3.24943595
#>  [91]   2.69889911 -13.88022013 -13.23458519   3.60722183  -3.51189393
#>  [96] -60.33742720  -1.26402875  -2.55867362   2.38482171 -14.41447892
#> 
#> $mu
#>   [1]  -7.22530045 -16.68488866  -6.18654263  -2.03254578 -14.04908108
#>   [6]  11.52291165   6.52131148  -1.91179751 -25.56540769  -8.13285960
#>  [11]  -6.31586615 -18.26642934   6.42557617  -9.04946543 -21.91216652
#>  [16] -11.58585100  -7.03450482   6.20342136   2.69696368 -17.80694535
#>  [21]  -5.29261521  -8.51513591  -7.82253525   9.39611667   3.16343668
#>  [26]  -7.30220872   6.48444787  -3.80859681   7.68036270  -0.49254313
#>  [31]   4.35394107 -11.89139781 -12.03038662  -2.90638940   0.60089308
#>  [36] -18.94302839   8.48185770 -14.43125602  -1.08536107 -21.99461148
#>  [41] -13.86378709 -14.17487927  11.12070689  -8.71608495  -7.83534147
#>  [46] -14.97086567 -16.78511936 -27.48779418  -7.60790188  -3.17203963
#>  [51] -12.94705154 -17.05233191   1.85929443 -16.12589426 -12.10892648
#>  [56] -27.02330670 -17.53933516  -2.34143702 -12.45848173  -4.57341215
#>  [61]  -8.59678920   3.13597191 -16.43351916   0.81147839  -1.02289824
#>  [66] -27.71675482  -3.60919136   4.47090980  -1.17359422   8.93454652
#>  [71] -13.53418224   6.83753358   0.03545287  15.48295522 -10.55506689
#>  [76]  -6.08667924 -36.03317595 -25.58400016 -22.10136352   7.67216628
#>  [81]  -6.77056914  -2.17659905  -6.54237352 -19.65432656 -36.57252986
#>  [86] -23.02615275 -14.51550161 -11.36083270  -7.68955590  -3.24943595
#>  [91]   2.69889911 -13.88022013 -13.23458519   3.60722183  -3.51189393
#>  [96] -60.33742720  -1.26402875  -2.55867362   2.38482171 -14.41447892
#>