Simulate Spline Data

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n
Integer, sample size

p
Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family
Family object, one of `binomial` (default), `gaussian` , `poisson`. Please see family for more detail

dispersion
Double, required for gaussian distribution, match to sd parameter.

Value

a list contains `dat`, `eta`, `mu`

dat a numeric matrix, contains predictors `x1` to `xp` where p is replaced by the integer value. The outcome variable is named `y`

etathe linear predictor of the simulated data

muthe mean of the simulated data

Examples

# Binomial Outcome
sim_Bai(100, 4)

#> $dat
#>                  x1           x2          x3           x4 y
#>   [1,] -0.356124416 -0.429380087 -0.38721358 -1.400043517 0
#>   [2,] -1.064464209  1.360461327 -0.78543266  0.255317055 0
#>   [3,]  1.077116538 -0.070857431 -1.05673687 -2.437263611 0
#>   [4,]  1.181575567 -0.272153684 -0.79554143 -0.005571287 0
#>   [5,]  0.198392095 -2.446680029 -1.75627543  0.621552721 0
#>   [6,] -0.400405249  0.065486641 -0.69053790  1.148411606 0
#>   [7,]  0.616154281 -1.098508902 -0.55854199 -1.821817661 0
#>   [8,]  1.974156748 -0.633178176 -0.53666333 -0.247325302 0
#>   [9,]  1.884662324 -2.063654451  0.22712713 -0.244199607 0
#>  [10,] -1.588620547  2.648932029  0.97845492 -0.282705449 1
#>  [11,] -0.539923164 -1.153398386 -0.20888265 -0.553699384 0
#>  [12,] -1.169461464 -0.340637876 -1.39941046  0.628982042 0
#>  [13,]  0.559105989  0.786362576  0.25853729  2.065024895 0
#>  [14,] -1.819347247 -1.270513110 -0.44179945 -1.630989402 0
#>  [15,]  0.393343972  0.542141549  0.56859986  0.512426950 1
#>  [16,]  0.042134106  0.075105900  2.12685046 -1.863011492 0
#>  [17,]  1.179664177  0.558514422  0.42485844 -0.522012515 1
#>  [18,] -0.256921176  0.415406399 -1.68428153 -0.052601910 0
#>  [19,] -1.056336098 -1.452299769  0.24940178  0.542996343 1
#>  [20,]  0.198777205  0.941206122  1.07283825 -0.914074827 1
#>  [21,]  0.650533552 -0.338935872  2.03936926  0.468154420 1
#>  [22,]  0.343913337 -0.075574247  0.44945378  0.362951256 1
#>  [23,]  1.477532312  0.040204392  1.39181405 -1.304543545 0
#>  [24,]  0.072025698  0.124301066  0.42656655  0.737776321 0
#>  [25,]  2.126444534 -0.998432551  0.10758399  1.888504929 0
#>  [26,] -1.476196906  1.233390065  0.02229473 -0.097445104 0
#>  [27,]  0.407888500  0.340424488  0.60361101 -0.935847354 1
#>  [28,]  1.393977798 -0.472702482 -0.26265057 -0.015950311 1
#>  [29,]  0.360278296  0.708753061 -0.52826408 -0.826788954 1
#>  [30,]  0.654550251 -1.528958715  0.19214942 -1.512399651 0
#>  [31,]  1.052155422  0.237425345 -1.14619967  0.935363190 0
#>  [32,] -1.979555125 -1.312814246  0.84618466  0.176488611 1
#>  [33,]  1.208385605  0.747028587  0.08171963  0.243685465 1
#>  [34,] -0.169280084 -1.562518435 -1.30511701  1.623548883 0
#>  [35,]  0.295029753  0.071053360 -0.94491206  0.112038083 0
#>  [36,]  1.266340587 -0.639534770  0.45434159 -0.133997013 1
#>  [37,] -1.135343257 -0.845195739 -0.85520250 -1.910087468 0
#>  [38,] -1.131053798  0.675244698 -0.28689522 -0.279237242 0
#>  [39,]  0.109993384  1.153375794  0.89496163 -0.313445978 1
#>  [40,]  0.852905410 -1.686504742  0.06730444  1.067307879 0
#>  [41,] -0.234337862 -0.902814949 -0.16267634  0.070034850 0
#>  [42,]  2.086688567  1.317633698 -0.82731017 -0.639123324 0
#>  [43,] -0.110919371  1.100189745  1.87650562 -0.049964899 1
#>  [44,] -1.392847056  1.203767839  0.76644020 -0.251483443 0
#>  [45,] -1.142290768 -1.431270777  0.97995670  0.444797116 1
#>  [46,]  1.704608737  1.382910861  1.32178099  2.755417575 0
#>  [47,] -0.080073634  0.003125940 -1.11971083  0.046531380 0
#>  [48,] -0.437281240 -0.077886824  0.51459982  0.577709069 0
#>  [49,] -0.119215094  0.441428226 -1.50909984  0.118194874 0
#>  [50,]  0.786462865  0.128922896  1.53274148 -1.911720491 0
#>  [51,] -0.578945246 -0.830214260  0.42914737  0.862086482 0
#>  [52,] -0.145426885 -0.503592910  0.12210341 -0.243236740 0
#>  [53,]  0.526457991 -1.193641182 -1.13801240 -0.206087195 0
#>  [54,]  1.733578110 -0.751723323 -0.55801513  0.019177592 0
#>  [55,]  1.448657220  1.455841403  1.05253854  0.029560754 1
#>  [56,]  1.518193149 -0.828603533  0.67768364  0.549827542 0
#>  [57,] -0.384007254  0.289774460  0.03849955 -2.274114857 0
#>  [58,]  1.827125177 -0.480053484 -0.35638119  2.682557184 0
#>  [59,] -0.551491750 -0.604829354  0.78284410 -0.361221255 1
#>  [60,] -0.865753541  1.460110180  0.80441162  0.213355750 1
#>  [61,] -0.343831481  0.149679354 -1.90006082  1.074345882 0
#>  [62,]  1.062876458 -1.433321100  0.93578429 -0.665088249 1
#>  [63,]  0.813058204 -0.010303319 -0.30905150  1.113952419 0
#>  [64,]  1.803483361 -0.212236035  0.26306668 -0.245896412 0
#>  [65,] -0.105068694 -0.906340179 -1.79059186 -1.177563309 0
#>  [66,]  0.982453362 -2.102152479 -0.78825884 -0.975850616 0
#>  [67,] -1.713302622  1.893360464 -1.13302167  1.065057320 0
#>  [68,] -0.832019527 -0.968125837  0.36365257  0.131670635 1
#>  [69,]  1.100491882 -0.102603036 -0.28588791  0.488628809 0
#>  [70,] -0.173820106  0.239959572  0.51766913 -1.699450568 0
#>  [71,]  0.178812018  0.060898893 -0.10290867 -1.470736306 0
#>  [72,] -0.698429449 -2.177576028 -0.97406959  0.284150344 0
#>  [73,] -0.960449159 -0.117860143  1.27067230  1.337320413 0
#>  [74,] -0.975423042  0.112294787  0.96086479  0.236696283 1
#>  [75,] -0.338576503  0.007886198  0.76872137  1.318293384 0
#>  [76,]  1.152347074  1.877743872  1.03593077  0.523909788 1
#>  [77,]  0.405101183  2.158756554 -0.47388707  0.606748047 0
#>  [78,] -0.470922497  0.709714522 -1.27533487 -0.109935672 0
#>  [79,] -0.133251019  0.766983379 -0.30562067  0.172181715 0
#>  [80,]  1.226682356 -0.308211421  2.21176949 -0.090327287 1
#>  [81,]  0.332943995  1.012001849 -1.04166838  1.924343341 0
#>  [82,] -0.347088466 -0.919051597 -1.14652385  1.298392759 0
#>  [83,] -0.098550690  0.563380077 -1.67532730  0.748791268 0
#>  [84,]  0.034766060  0.322482749  1.52593866  0.556224329 1
#>  [85,]  0.386127022  0.366674363  0.55418551 -0.548257264 1
#>  [86,]  0.020831228  1.129835153  1.99311026  1.110534893 1
#>  [87,]  0.007586777 -0.941498076 -0.15412074 -2.612334333 0
#>  [88,]  0.930844030  0.217837643  2.56440834 -0.155693776 1
#>  [89,] -0.684749941  1.415412293  1.06199914  0.433889790 1
#>  [90,]  0.337401513 -0.383733048  1.14269488 -0.381951112 1
#>  [91,] -0.412137704 -0.174086374  1.12383884  0.424187575 1
#>  [92,]  0.934261130 -0.221744517 -0.39700149  1.063101996 0
#>  [93,]  1.840316741 -1.009528722 -0.82326115  1.048712620 0
#>  [94,] -0.704819663  0.480725266 -0.57888462 -0.038102895 1
#>  [95,]  0.008510312  1.604407328  1.76378938  0.486148920 1
#>  [96,]  2.034189886 -1.515024529  0.13299215  1.672882611 0
#>  [97,] -1.341686068 -1.416023914  0.37649933 -0.354361164 1
#>  [98,]  1.158979182  0.876777327  1.13870765  0.946347886 1
#>  [99,] -0.203208958  0.624132413  1.24126308  1.316826356 0
#> [100,] -0.378028555  2.112277288  0.61209094 -0.296640025 0
#> 
#> $eta
#>   [1] -13.55925032  -7.73747633 -37.55498447   0.65778514  -5.28417488
#>   [6] -19.15042362 -26.52653355  -4.90662293  -6.23650185  10.23974631
#>  [11]  -3.47214760 -12.74008164 -22.01208338 -10.56561788   7.59402455
#>  [16]  -8.78310343   8.17132020 -14.56072695   0.70909808   2.92153605
#>  [21]   9.07253201   2.44737089  -4.83300487  -3.31897234 -18.64831940
#>  [26]  -4.43167404   0.74992379   3.79880017  -1.48724136 -12.80593865
#>  [31] -13.33115636   7.18191584   5.50636042 -24.37196122  -6.42529178
#>  [36]   6.88137349 -32.19249850  -3.97979668   2.98785210 -11.18616972
#>  [41] -11.45193778  -6.21542869   2.57072916  -3.17805403   3.48319059
#>  [46] -35.01350081 -14.18720217  -4.19774593 -11.44380501 -19.25555746
#>  [51]  -3.63590707  -1.55603309  -8.78633247 -11.77893578   9.24681125
#>  [56]  -2.86835609 -31.44684742 -40.32066796   5.73651054   9.75874750
#>  [61] -26.44560291   7.81420304 -17.55033345  -3.90541442 -26.22742015
#>  [66] -13.20477995 -10.66266565   1.12417963  -3.11107659 -19.40792116
#>  [71] -12.39714421  -2.82859381  -2.88547151   0.83832598 -13.42359777
#>  [76]   6.23900249  -6.89137074  -7.38518522  -5.66303176  19.72258974
#>  [81] -25.57409756 -24.90628247 -13.27807176   7.01399623   4.52602802
#>  [86]   1.14256844 -40.27290267   9.07713811  10.57053776  12.78573730
#>  [91]   1.81324122 -10.27233537 -19.54415219   3.07225749  12.11968441
#>  [96] -10.31492926  -0.06341158   3.88979407  -3.66376609  -5.65092092
#> 
#> $mu
#>   [1] 1.292087e-06 4.359808e-04 2.220446e-16 6.587627e-01 5.045627e-03
#>   [6] 4.820329e-09 3.017678e-12 7.343108e-03 1.952867e-03 9.999643e-01
#>  [11] 3.011519e-02 2.931241e-06 2.755965e-10 2.578691e-05 9.994968e-01
#>  [16] 1.532781e-04 9.997174e-01 4.746314e-07 6.702018e-01 9.489008e-01
#>  [21] 9.998852e-01 9.203690e-01 7.899657e-03 3.492603e-02 7.964120e-09
#>  [26] 1.175474e-02 6.791621e-01 9.780930e-01 1.843361e-01 2.744418e-06
#>  [31] 1.623124e-06 9.992404e-01 9.959556e-01 2.602506e-11 1.617440e-03
#>  [36] 9.989743e-01 2.220446e-16 1.834655e-02 9.520223e-01 1.386444e-05
#>  [41] 1.062875e-05 1.994374e-03 9.289538e-01 3.999999e-02 9.702057e-01
#>  [46] 2.220446e-16 6.895664e-07 1.480688e-02 1.071554e-05 4.339280e-09
#>  [51] 2.568301e-02 1.742166e-01 1.527840e-04 7.664253e-06 9.999036e-01
#>  [56] 5.374019e-02 2.220446e-16 2.220446e-16 9.967844e-01 9.999422e-01
#>  [61] 3.272055e-12 9.995962e-01 2.387740e-08 1.973529e-02 4.069829e-12
#>  [66] 1.841773e-06 2.340200e-05 7.547632e-01 4.265266e-02 3.726035e-09
#>  [71] 4.130350e-06 5.579844e-02 5.287645e-02 6.981125e-01 1.479806e-06
#>  [76] 9.980520e-01 1.015487e-03 6.199911e-04 3.459962e-03 1.000000e+00
#>  [81] 7.821869e-12 1.525243e-11 1.711615e-06 9.991016e-01 9.892923e-01
#>  [86] 7.581509e-01 2.220446e-16 9.998858e-01 9.999743e-01 9.999972e-01
#>  [91] 8.597531e-01 3.457534e-05 3.251491e-09 9.557338e-01 9.999945e-01
#>  [96] 3.313361e-05 4.841524e-01 9.799602e-01 2.499502e-02 3.501972e-03
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson)

#> Error: object of type 'closure' is not subsettable

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian, dispersion = 1 )

#> Error: object of type 'closure' is not subsettable