Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                  x1           x2           x3          x4 y
#>   [1,] -0.512439402 -1.150634696  0.881057375 -0.05781247 1
#>   [2,]  1.474912478  1.712555405 -0.051556809 -0.61846660 0
#>   [3,] -1.475296792 -0.481916123 -1.486014027  0.30013157 0
#>   [4,] -1.024554322  0.463224528  0.106187286 -0.89111616 0
#>   [5,]  0.435103856 -1.156364664 -0.540906750 -1.44111784 0
#>   [6,] -0.156473211 -0.347104694  0.523728699  0.36504961 0
#>   [7,]  0.295974325  0.590739031 -0.289251252 -0.06892203 1
#>   [8,] -1.783792224 -0.405652247 -0.464411862  1.80873949 0
#>   [9,] -0.089736094 -0.458726014 -2.398529726 -0.57396915 0
#>  [10,]  0.950789318 -0.934751950 -1.914554545  1.97915874 0
#>  [11,]  0.580522042 -0.023491501  0.714702707  0.19109684 1
#>  [12,]  1.119390055 -0.608335880  0.706007307 -2.48192083 0
#>  [13,] -1.894646043 -1.506898277 -1.538199013 -1.20325715 0
#>  [14,]  0.838353447  0.830798443 -0.117581247  0.97166855 0
#>  [15,] -2.137205340 -0.700633097  1.296907146 -0.96221196 0
#>  [16,] -1.477947969 -0.733288751 -0.833302115  1.94988001 0
#>  [17,] -0.701871064  0.921853086  0.065050340 -0.48923487 1
#>  [18,] -0.833179667 -1.021970974 -0.634193759  0.06379726 0
#>  [19,]  0.168495012 -0.003075572  1.082015555  0.40778667 1
#>  [20,] -0.288366442  1.509709201  0.392921249  1.43324528 0
#>  [21,]  0.311889087 -1.328087425 -0.825724028 -0.42325381 1
#>  [22,] -1.439670364  2.111845006  0.008775605  0.52876553 0
#>  [23,] -0.760244399 -1.009061252 -1.069806256  0.62715827 0
#>  [24,]  0.958751905  0.109032756  0.958976288  0.27184498 0
#>  [25,] -0.661880978  1.799826830 -1.341293346 -1.67266284 0
#>  [26,] -0.660068855  1.111700356  2.832636873  0.77268149 1
#>  [27,] -1.591952026 -0.741985417 -0.332850003  0.71357237 0
#>  [28,] -1.559411356 -0.634849890  0.948520472  0.08903855 1
#>  [29,] -3.556538234 -0.465382230  0.687177219 -1.47044145 0
#>  [30,] -0.522833563 -0.962721567 -1.089663345  0.25831279 0
#>  [31,]  1.235784265  1.299228713 -0.531237258 -0.84090625 0
#>  [32,] -0.857469114 -2.133017795  0.889547500 -1.33364005 0
#>  [33,]  1.656205404  1.470535738  0.502013567  1.02180942 0
#>  [34,]  0.233939724 -0.435520151  1.140048549 -0.83665850 1
#>  [35,] -1.446679802 -1.474921456  0.089475594 -1.22630707 0
#>  [36,]  0.108983505  0.113847777  1.183750710  1.21547283 0
#>  [37,] -1.940126541  0.346284805  0.916468910 -0.02060691 1
#>  [38,] -0.420489099 -0.888942197 -1.067941238 -0.53190223 0
#>  [39,]  1.710773008  1.094964403 -0.591805353 -0.10324211 0
#>  [40,]  0.333196356  0.918999759  0.419337320  1.70739955 0
#>  [41,]  1.203198437  0.325689824 -1.173781323  1.00426078 0
#>  [42,] -0.796480769 -0.482163811 -0.616032063 -0.79622173 1
#>  [43,]  0.257670214 -0.317381364 -0.567507082 -1.39506757 0
#>  [44,]  1.508410537  0.400490569 -1.261313586 -0.97563602 0
#>  [45,]  0.365894950  0.016622626  2.831229311 -1.17270359 1
#>  [46,] -0.395750282  0.012094717 -1.921733483 -0.63844412 0
#>  [47,] -1.578230406 -0.319853023 -1.815371411 -0.79862715 0
#>  [48,]  0.789458239  0.633968955 -1.421256479  0.97896225 0
#>  [49,]  0.294534599 -1.693250983 -2.714362840  0.32785920 0
#>  [50,] -0.256028381  0.654826276  1.396203231 -0.67572200 1
#>  [51,]  2.043953907  0.449592887 -0.571876604  1.40615218 0
#>  [52,]  0.662286932  0.258818648 -0.385233548  0.11544711 0
#>  [53,]  1.101451663  2.364683121  0.327066466  2.67113441 0
#>  [54,] -0.916173907 -1.409384121  1.612744073  1.01566926 1
#>  [55,]  1.164438578  0.080910881  0.174692015 -0.67304218 0
#>  [56,] -0.443116849 -0.732643648 -1.448813519  0.17761992 0
#>  [57,] -1.122186282 -0.615204256 -1.958629186 -0.07085047 0
#>  [58,]  0.283440633 -0.922366422  1.777817139  0.36608429 1
#>  [59,]  0.390369419 -0.926280588  1.322103554 -0.21090688 1
#>  [60,] -0.602157403 -0.251856411  0.738744638  1.91682528 0
#>  [61,] -0.540617213 -0.964226958  1.767748923  1.33197389 0
#>  [62,]  1.050259262  0.287387376 -0.741082149  1.44571931 0
#>  [63,]  0.441505653  0.857706498 -0.804743230 -0.19449692 0
#>  [64,] -1.277039919  0.121661096 -1.180699982  0.59914427 0
#>  [65,] -0.120291469  1.556097345  1.829584752  0.60278797 1
#>  [66,] -0.593328636 -0.991563771 -0.728797598  1.12833206 0
#>  [67,]  0.150603203  0.188330036 -0.184354567 -1.15369737 0
#>  [68,]  1.289072010  0.503549672  0.604658160  0.50973975 1
#>  [69,] -0.082221827 -1.011393502  0.259520273 -0.56962993 0
#>  [70,]  0.321900720 -0.916975781 -0.415729230  0.73496250 0
#>  [71,] -0.600537422 -0.711846445  1.364272611 -0.94069647 1
#>  [72,] -0.923602911 -0.998191997  0.565491342  0.26891808 0
#>  [73,] -1.192363367 -0.841154974  0.507174321  0.15259611 1
#>  [74,] -1.083233171 -1.541373478 -0.417248075 -1.12158500 0
#>  [75,]  0.991011194  0.814537738 -0.063483711 -0.88135083 0
#>  [76,]  1.654628639  0.107474812 -0.328100061  0.05127251 0
#>  [77,] -0.075362965  0.599510342 -0.836420379 -0.63395709 0
#>  [78,]  0.983944362 -0.110660092  0.616718177 -0.31695495 0
#>  [79,] -0.276660782 -2.189438972  0.367476497  0.51919976 0
#>  [80,] -1.016793474 -0.559380948 -0.193412668 -1.23939963 0
#>  [81,] -1.153383628 -2.992293470  0.673751124  1.13594618 0
#>  [82,]  0.419106393 -0.526329323 -1.877969217  1.04561672 0
#>  [83,]  0.001953708 -1.326776877 -0.277318220  0.01401081 1
#>  [84,] -0.660826605 -0.009315116  2.009105756 -0.34623589 1
#>  [85,] -1.327392064 -1.624992859 -0.792498662  0.54398899 0
#>  [86,]  1.886139332 -0.785699572  1.055029477  0.33293408 0
#>  [87,]  1.421571816  1.285154371 -2.428207750  1.73078310 0
#>  [88,]  0.591748967 -0.019500046 -1.335515408 -1.24707713 0
#>  [89,]  0.266306773  1.749444867 -1.703427203 -1.27407999 0
#>  [90,]  1.520622974 -0.071677296  0.814785787 -0.68235314 0
#>  [91,]  1.603627148  0.345897416 -0.432351717  1.66496868 0
#>  [92,] -0.170164079  0.004785323 -2.614188847  0.86151511 0
#>  [93,]  1.407577977  0.970156879  1.133079645 -0.11989900 1
#>  [94,]  1.140552488  1.075872542  1.165298828  0.17209039 1
#>  [95,] -1.087024497  0.792995806 -1.343928080  0.90731253 0
#>  [96,] -1.869419907 -0.010995748  0.872681306 -1.32118733 0
#>  [97,] -0.265691993  0.428859725  0.133388558  1.08515065 0
#>  [98,] -0.035119818  0.963566140 -0.228601235  0.75378687 0
#>  [99,] -1.312478218  1.960120666 -0.294109918 -0.46291403 0
#> [100,]  0.784545870 -0.420545412 -0.713317188  0.42417734 0
#> 
#> $eta
#>   [1]   3.6639882  -0.2538039  -7.9343170  -2.6238072 -13.4996525   0.3978465
#>   [7]   5.5237476 -11.7736566 -16.3257954 -38.0779551  -3.0076332 -23.1337385
#>  [13] -11.4710786 -11.2060028  -2.8985508 -27.7397766   0.2752906  -4.2060148
#>  [19]   5.0465850 -10.6481393   0.6221481  -8.6136306  -8.2913494  -1.3287210
#>  [25] -18.5403892  12.8146339  -5.0508941   6.8615122  -2.6073918 -11.5162877
#>  [31]  -4.0011056  -2.0843884  -4.7676119  10.7968849  -6.3584648  -2.5295439
#>  [37]   6.2527465 -15.6395471 -13.9343457 -11.3587990  -8.8977131   0.1250180
#>  [43]  -6.4513616 -12.3674580   8.6519819 -21.7601235  -9.9937173 -15.8978186
#>  [49] -14.0898846   3.1589875 -10.7157421  -9.8580950 -31.3615334   9.5223220
#>  [55]  -2.4219313 -13.6241143 -15.3892417   9.3871487   5.3910911 -10.4866249
#>  [61]  -2.3515070 -15.8921628  -5.4243281 -19.1680580   8.1878858 -13.0786198
#>  [67]  -8.3069632   9.2310842  -7.3991412  -6.1958981   4.1842731   0.3084465
#>  [73]  -6.7631410  -9.8910655  -5.6696128 -11.5492868  -6.8394670  -0.2724860
#>  [79]  -5.0961950  -8.2978257 -11.6150471 -12.6544551   0.2238199  10.7985922
#>  [85] -11.0286956  -1.6541410 -29.7839514 -23.2641036 -12.9332492  -1.9109479
#>  [91] -20.5756100 -28.8465836   4.8783505   5.2084079 -15.3680693  -4.7038681
#>  [97]  -9.2266219  -9.7905449 -11.8809689  -7.5622126
#> 
#> $mu
#>   [1] 9.750104e-01 4.368874e-01 3.581083e-04 6.762186e-02 1.371434e-06
#>   [6] 5.981701e-01 9.960250e-01 7.704821e-06 8.124509e-08 2.220446e-16
#>  [11] 4.708222e-02 8.977285e-11 1.042724e-05 1.359217e-05 5.222525e-02
#>  [16] 8.969497e-13 5.683913e-01 1.468674e-02 9.936098e-01 2.374442e-05
#>  [21] 6.507069e-01 1.815804e-04 2.506132e-04 2.093710e-01 8.871790e-09
#>  [26] 9.999973e-01 6.362860e-03 9.989538e-01 6.866421e-02 9.966335e-06
#>  [31] 1.796669e-02 1.106235e-01 8.429004e-03 9.999795e-01 1.729029e-03
#>  [36] 7.381282e-02 9.980785e-01 1.613730e-07 8.879534e-07 1.166625e-05
#>  [41] 1.366825e-04 5.312139e-01 1.575884e-03 4.254802e-06 9.998253e-01
#>  [46] 3.545671e-10 4.568397e-05 1.246422e-07 7.600471e-07 9.592614e-01
#>  [51] 2.219232e-05 5.231919e-05 2.220446e-16 9.999268e-01 8.151554e-02
#>  [56] 1.210938e-06 2.072703e-07 9.999162e-01 9.954637e-01 2.790645e-05
#>  [61] 8.694607e-02 1.253491e-07 4.388681e-03 4.736071e-09 9.997221e-01
#>  [66] 2.089425e-06 2.467315e-04 9.999021e-01 6.114039e-04 2.033628e-03
#>  [71] 9.849953e-01 5.765060e-01 1.154260e-03 5.062241e-05 3.437345e-03
#>  [76] 9.642825e-06 1.069529e-03 4.322969e-01 6.082762e-03 2.489957e-04
#>  [81] 9.029116e-06 3.193292e-06 5.557226e-01 9.999796e-01 1.622898e-05
#>  [86] 1.605501e-01 1.161433e-13 7.880036e-11 2.416352e-06 1.288744e-01
#>  [91] 1.159114e-09 2.965432e-13 9.924479e-01 9.945594e-01 2.117055e-07
#>  [96] 8.978814e-03 9.837536e-05 5.597525e-05 6.920823e-06 5.194541e-04
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                 x1           x2          x3          x4    y
#>   [1,]  0.15868790  0.406140803 -1.62097901 -0.89038526    1
#>   [2,]  1.20489460  0.013859685 -0.12642797  0.67049907    3
#>   [3,] -0.72846950 -0.475995768 -1.49291986  0.08202943   19
#>   [4,] -0.07390385 -0.482643477 -0.80314366 -2.21814137    0
#>   [5,] -2.42544073 -0.861414413 -0.81673828 -0.36738101    0
#>   [6,] -1.24758285  2.339840568 -0.30401623  0.84194741    1
#>   [7,] -0.73368485  1.178222398  0.71381452  0.83033149    2
#>   [8,] -0.09879964 -1.823601160 -0.16540216 -1.24595664    0
#>   [9,]  1.87929449 -1.000836847 -1.56457002  0.36049087    0
#>  [10,]  0.76786661 -0.127708178 -0.18819583  0.33226350    1
#>  [11,]  0.87342662 -0.366598122 -1.57804685  1.27738139    0
#>  [12,]  0.82219359  0.085157131  0.63497304 -0.07850652    0
#>  [13,] -0.15549901  0.622903287 -0.28024346  1.30450819    0
#>  [14,] -0.96050215 -0.266471565 -1.94700066 -0.39942271    4
#>  [15,]  0.63853536 -0.643607685  1.02127812  2.07276807    0
#>  [16,]  1.46062308  0.858311557 -0.92744654  1.07254048    0
#>  [17,]  0.33967972 -0.114983420 -0.38969660 -1.06354864    2
#>  [18,]  2.67388046 -0.477132273 -0.04571060 -0.12301236    2
#>  [19,] -0.81551725 -2.025322729  1.64828709 -1.62805279    0
#>  [20,]  1.36366267 -0.118950891 -1.73631238  1.74225207    0
#>  [21,] -0.53318665  0.208122132  0.28029780 -1.11667869    0
#>  [22,]  0.08890309 -1.543896374 -2.04594296 -2.04156125    0
#>  [23,]  0.72727246  0.463418247  1.00696912  0.99236288    0
#>  [24,] -0.24299934 -0.163936324 -0.95566786  0.38652241    1
#>  [25,] -0.15533878 -0.008341307 -1.02488559 -1.38126599    0
#>  [26,]  0.22622443  1.642513108  0.73745465  0.63358485   22
#>  [27,] -0.33386691  0.978895939 -0.56827908  0.03221488    0
#>  [28,]  0.89269946  0.410257811  1.41750132  0.31586653    7
#>  [29,] -1.47414995  2.474154673  1.02118995  0.91245422    4
#>  [30,] -1.10743547 -0.553988186  0.04493909 -0.87271950    1
#>  [31,] -0.14909626 -0.936519127 -0.73105994  0.55693157    0
#>  [32,]  1.21885458 -0.011187525 -0.71056474  2.34399556    0
#>  [33,] -1.52154606  0.901252232 -0.43245971  1.78814772    0
#>  [34,] -0.70203948 -0.211853610 -1.69051644  0.33989925    6
#>  [35,]  0.83506885 -1.614120066  0.48019641  0.06198897    1
#>  [36,] -0.89221698 -1.854919330 -0.16914573 -0.89365792    1
#>  [37,] -0.87297333 -0.039575134  0.52492821  0.93470550    0
#>  [38,] -0.74215812 -0.340905741 -1.66667398  0.19906995   18
#>  [39,] -1.30630145 -1.862326131 -0.30617882 -0.72321220    0
#>  [40,]  0.99823758  1.034672415  0.73897116  1.09843400    0
#>  [41,] -0.73449631 -0.611147063  0.04115134 -1.40151539    0
#>  [42,] -0.48820682 -0.288888596  0.36084161  1.02092705    2
#>  [43,] -0.78200952 -1.292596728  0.12414759  1.24186260    1
#>  [44,]  0.06382485  0.632389098  1.04393561 -0.38362784   27
#>  [45,] -0.77105048  0.275418156  0.26880073 -0.10577970   11
#>  [46,]  1.59927718  0.559511591  0.19823122  2.17811789    0
#>  [47,]  1.35160221 -1.589654695 -1.97361681 -1.79411480    0
#>  [48,]  0.43657373 -0.548640501  0.86810832  2.26554875    0
#>  [49,]  0.19569773 -0.005190046 -0.48209413  0.13972499    2
#>  [50,] -1.20677113 -0.600878953 -0.95996317 -1.05474539    0
#>  [51,] -0.17475698  0.756538588 -1.78735940 -0.07255185    3
#>  [52,] -1.44460749 -0.562744301  0.75720328  1.53821113    0
#>  [53,]  0.08856267  0.474311759  0.14275366 -1.12220035    2
#>  [54,] -2.41946123  1.420189539 -1.91691599 -2.19868277    0
#>  [55,]  1.13621532 -1.099268908 -0.71860886  2.14215710    0
#>  [56,]  0.61818693 -1.613595563 -0.41842532  1.67199018    0
#>  [57,] -1.45906462  1.419774279  0.65942236 -0.19350452    9
#>  [58,] -1.84167026  1.123651639  0.03074486 -1.04793894    0
#>  [59,] -0.48138112 -0.505421724  1.24687921 -0.18021407   24
#>  [60,] -0.08790852 -1.183222282 -1.74984792 -1.84175285    0
#>  [61,] -1.79489009  1.113105390  1.06169811  0.98088769    0
#>  [62,]  0.01526129 -1.431409662 -0.14299011 -0.36476850    8
#>  [63,] -0.69721919 -0.255624098  0.76188431 -1.11679651    2
#>  [64,]  0.60695382 -0.167701601  3.02724943  0.23377154 5009
#>  [65,]  0.06794322  1.801092251  1.65751683 -1.86818734    0
#>  [66,]  0.98363696  1.206324111 -0.52523973  2.58307658    0
#>  [67,]  0.02049723 -2.184354908  1.16405343 -1.05514549    1
#>  [68,] -0.07415357  0.292021740  0.24590356  1.67046692    0
#>  [69,] -0.21547064 -0.180524476 -0.47332263 -1.08646547    0
#>  [70,]  1.38872634  0.118296242 -0.08004443 -0.75128868    0
#>  [71,]  0.55096243  0.506102910  0.54352122 -0.27952634    6
#>  [72,] -0.23618679 -1.338932392 -0.84400876  0.08069326    2
#>  [73,]  0.23824671  0.556868437 -0.79428884  1.18195212    2
#>  [74,] -0.86394224  1.281908054 -0.21054090 -0.90923386    3
#>  [75,] -0.48285540  0.967515102  0.51234338  0.17264127    0
#>  [76,] -0.64138483 -2.513505364 -0.74612771 -0.54550529    8
#>  [77,] -1.01396187 -1.176735953 -1.50120738  0.65446761    2
#>  [78,]  1.63489512 -0.262205514 -0.21185659  0.11548663    7
#>  [79,] -0.39469114  0.074192302  0.01978892  0.58278435    1
#>  [80,] -0.02210018  0.817507885  0.08882790 -0.58113774    3
#>  [81,] -0.09498902 -2.044834275  0.33605423  0.31308860    1
#>  [82,]  0.43645297 -0.725981419 -0.72784124 -1.23339715    0
#>  [83,] -0.31848073 -0.967005181  0.29925267  0.19705663    0
#>  [84,]  0.60483251 -0.624017878 -1.78143737  0.17724345    1
#>  [85,] -0.58578703  0.135633026 -0.79077241 -1.29154658    0
#>  [86,]  0.88096670 -0.640676897 -0.09719920 -0.35561567    2
#>  [87,]  1.88651511  1.643600142  0.19775934  0.83349485    2
#>  [88,] -0.54046007 -1.267469720  0.93625505  0.86456192    7
#>  [89,] -0.82514306  0.176504137 -3.10112412 -0.93685553    1
#>  [90,] -1.79257182 -2.312822077 -1.25722816  0.13707977   14
#>  [91,]  3.28968033 -0.500181617 -0.25189790 -1.70304186    0
#>  [92,]  0.59521185 -0.660517182  1.47601446  0.15513466    9
#>  [93,]  0.08344670  1.018367843  1.41392318 -0.18029153   13
#>  [94,] -0.08098844 -1.699090294  0.64880028 -0.94790986    1
#>  [95,]  0.60714926  0.126495398 -1.26915268 -0.11898140    0
#>  [96,]  0.66604929 -1.168124745 -1.23006611 -1.90990391    0
#>  [97,]  1.11888414  0.077929239  0.31748070 -1.03886310    0
#>  [98,]  0.53851857 -0.656877436 -0.58865973 -0.24027935    1
#>  [99,] -0.63441276 -0.359786228  1.61250789 -0.90535333   18
#> [100,]  0.62465042 -0.476106219  1.92349361  0.64795463   38
#> 
#> $eta
#>   [1]   0.79611259   0.49751160   3.00694678  -8.18442888  -0.47676340
#>   [6]  -1.05805575   0.61856337  -3.16805579  -0.73905754  -0.14829452
#>  [11]  -1.95156065  -0.15579647  -1.96690689   1.55115889  -7.07896644
#>  [16]  -1.07712747  -0.48729449   1.40125988  -2.09596451  -4.56257333
#>  [21]  -1.44870714  -6.04039175  -0.11919810  -0.18475255  -4.35179403
#>  [26]   2.94507463  -0.44631263   2.31240621   1.07226978  -0.06177407
#>  [31]  -1.22834238  -9.65291310  -5.43569224   2.05503389   1.12477603
#>  [36]  -0.54935953  -0.08776333   2.87284368  -1.62376210  -1.54713577
#>  [41]  -1.15253967   0.09268911   0.04271215   3.17761883   2.15848160
#>  [46]  -7.59273966  -4.09307986  -7.23174236   1.28882148  -1.61196906
#>  [51]   0.60909811  -2.59420626   0.26002155  -8.55523546  -7.65988976
#>  [56]  -4.65455207   1.98107440  -0.90710519   3.07085809  -6.15893584
#>  [61]   0.21482275   2.17395367   0.81809830   8.50719835  -3.28966616
#>  [66] -12.90993905   0.27035203  -4.75244028  -2.02477056  -0.08662335
#>  [71]   2.11310839   1.03449229   0.37385156   0.13007799   0.73836589
#>  [76]   2.20434920   0.18453412   1.09253469  -0.88514768   0.69906397
#>  [81]   0.09197519  -1.80264369  -0.40444461   0.67990810  -2.58010930
#>  [86]   0.61572946   0.36407523   1.34939424  -0.66701482   2.82066870
#>  [91]  -2.64365787   2.20106857   2.15347733  -0.69810770  -0.49626852
#>  [96]  -7.05707282  -0.92953493   0.95189999   3.04278327   3.13210656
#> 
#> $mu
#>   [1] 2.216906e+00 1.644624e+00 2.022555e+01 2.789637e-04 6.207894e-01
#>   [6] 3.471301e-01 1.856259e+00 4.208534e-02 4.775638e-01 8.621771e-01
#>  [11] 1.420522e-01 8.557333e-01 1.398889e-01 4.716933e+00 8.426436e-04
#>  [16] 3.405724e-01 6.142861e-01 4.060312e+00 1.229516e-01 1.043517e-02
#>  [21] 2.348738e-01 2.380626e-03 8.876319e-01 8.313100e-01 1.288368e-02
#>  [26] 1.901208e+01 6.399837e-01 1.009870e+01 2.922004e+00 9.400953e-01
#>  [31] 2.927775e-01 6.423816e-05 4.358217e-03 7.807102e+00 3.079527e+00
#>  [36] 5.773194e-01 9.159776e-01 1.768724e+01 1.971556e-01 2.128568e-01
#>  [41] 3.158336e-01 1.097121e+00 1.043637e+00 2.398956e+01 8.657981e+00
#>  [46] 5.040981e-04 1.668776e-02 7.232596e-04 3.628508e+00 1.994944e-01
#>  [51] 1.838772e+00 7.470515e-02 1.296958e+00 1.925344e-04 4.713594e-04
#>  [56] 9.518176e-03 7.250529e+00 4.036911e-01 2.156040e+01 2.114502e-03
#>  [61] 1.239642e+00 8.792980e+00 2.266186e+00 4.950275e+03 3.726629e-02
#>  [66] 2.473345e-06 1.310426e+00 8.630608e-03 1.320241e-01 9.170224e-01
#>  [71] 8.273920e+00 2.813677e+00 1.453321e+00 1.138917e+00 2.092513e+00
#>  [76] 9.064351e+00 1.202658e+00 2.981822e+00 4.126532e-01 2.011869e+00
#>  [81] 1.096338e+00 1.648625e-01 6.673473e-01 1.973696e+00 7.576572e-02
#>  [86] 1.851006e+00 1.439182e+00 3.855090e+00 5.132384e-01 1.678807e+01
#>  [91] 7.110072e-02 9.034662e+00 8.614763e+00 4.975259e-01 6.087981e-01
#>  [96] 8.612956e-04 3.947373e-01 2.590627e+00 2.096351e+01 2.292222e+01
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                   x1           x2          x3          x4           y
#>   [1,] -1.2312711723  0.242929738 -0.74882842  0.55220637 -16.1623290
#>   [2,]  0.7102365905  1.680701229 -0.40599555 -0.88471320 -10.1914446
#>   [3,]  0.5720496443 -1.798832742  0.18466360  1.44377056 -14.2624456
#>   [4,]  0.8341720278  1.226549352  0.62113362  0.60547411  -7.5267127
#>   [5,] -1.1095024384 -0.182504339 -1.08667100  1.25145336 -20.4340991
#>   [6,]  1.1888774896  0.971344810  0.99217798 -0.40203234   7.1650092
#>   [7,]  1.4518022718  1.800102395  0.13100725  1.22541365  -6.5332433
#>   [8,] -0.0119594617  1.071595908  0.73519218  0.26473044  -1.4298276
#>   [9,]  1.5470326165  0.715775323  0.12636649  0.02785610   0.7165308
#>  [10,] -0.0114540555  0.479559613 -0.52679997  0.23052374  -0.8841637
#>  [11,]  0.7877377387 -2.508967087  0.50498220 -1.06920910  -6.5952225
#>  [12,] -0.1023979173  0.312242473 -0.25420681  0.34232641  -7.9356546
#>  [13,] -0.5763850827 -0.605828857  0.57871710 -0.51987885   4.4922962
#>  [14,] -1.1425275221 -0.353358609 -0.45183505 -1.31519950 -14.7290591
#>  [15,]  0.5489113593 -0.882627545 -0.63240199  0.46936807 -11.8261989
#>  [16,]  1.5185524415  0.522907186 -1.35633198  1.13112875 -11.8880194
#>  [17,]  0.3303183683  1.125177472  0.30177141 -0.75773117  -2.2661481
#>  [18,] -0.4698757704  2.261721052  1.04039807 -0.38428761   2.8650427
#>  [19,]  0.0003053936  0.312909277 -1.58215502  0.11405821 -12.7178156
#>  [20,] -1.1262507200 -0.410696483 -0.86696360 -0.67748485  -7.8038664
#>  [21,]  0.0252410298 -0.373608641 -0.21781362 -0.14226919   0.6497829
#>  [22,]  0.1458911142  0.278648505 -0.22692830 -0.69148937  -1.9011001
#>  [23,] -0.6976782893  0.054532552  0.02459590  0.40685872  -1.8474891
#>  [24,] -0.9144558554 -1.633583726 -0.46028615  2.12349519 -24.3616375
#>  [25,]  1.7689444355  1.473722455 -0.05583439 -0.88119204  -7.2182670
#>  [26,]  1.1399546382  0.287780693 -2.32491094 -2.64531104 -46.9471012
#>  [27,]  1.2282663061  0.640542738 -0.12294341  0.46196401   6.0062702
#>  [28,]  1.4630383838 -0.460762985  0.46972369  0.61364469   4.4676929
#>  [29,]  1.3554233636 -0.037534936  0.78037935 -0.29181764   5.1082112
#>  [30,] -0.2437274785  0.869263119 -0.26789112 -0.08601871  -8.5552958
#>  [31,] -1.1057033873  1.686007764 -1.10386389  0.25802293  -9.0797543
#>  [32,] -1.5864227730  0.169248668  0.90999950  0.22197908   1.2467137
#>  [33,]  0.0533168130  0.671206531  0.15243732  0.27485958   4.0895859
#>  [34,]  0.5883968689 -0.732275705  0.43999258 -0.17287930  -3.0618555
#>  [35,] -0.3214558075 -0.357892967  1.10728736  0.51600486   0.9345115
#>  [36,]  0.8036030439  0.491580676 -0.39686667 -0.86528670  -8.3770152
#>  [37,] -1.1834605412  1.501955325 -0.43291641  0.20310030  -5.0130199
#>  [38,] -0.1775704081  1.673990942  1.35033461 -0.26033951   5.5870585
#>  [39,] -0.6569547337 -1.422881867  0.98946718 -1.32235080   5.0583799
#>  [40,] -0.8342160518  0.632465558 -1.03245414  0.86593954  -3.7974738
#>  [41,]  0.5092604902  0.774867690 -0.77672440  0.76363273  -7.3115209
#>  [42,]  2.4770377466  0.235408098 -1.04433197  0.87874565 -13.3412309
#>  [43,]  0.0662700958  1.426648129 -0.41694064  0.03456301   0.7893056
#>  [44,]  0.4028314831  1.152230278 -0.95390475  0.10648348  -7.5462939
#>  [45,]  1.7093601964  1.615855712  2.24478003  0.79029805   9.3101394
#>  [46,] -0.2156272578 -0.285130727 -1.15138300 -0.80962956 -12.3889948
#>  [47,] -0.3698336177  0.932935047  0.09363148  0.01204633  -6.3327859
#>  [48,]  1.1168935398 -0.688878131 -0.14087922  0.20826426   0.3800417
#>  [49,] -1.4037673252 -0.667437550  0.48977901 -0.21147431  -1.2558225
#>  [50,] -1.0905287643 -0.262597757  1.31780822 -0.09766570   6.5142018
#>  [51,] -2.1207189789 -0.145315089 -0.24675951 -0.36849936  -7.9182351
#>  [52,]  0.0227330000  1.537392818 -0.69723111 -0.43121659  -2.0363682
#>  [53,]  0.0573605616  0.215911228 -1.10682934  1.50117987 -21.3294892
#>  [54,] -1.1501319714  0.570135049 -1.42977505 -0.65386014 -12.8286116
#>  [55,] -0.1663758695 -0.229956049  0.47427742 -1.15872482  -8.4130243
#>  [56,] -1.1730900079 -1.042155363 -0.50713385  0.92068155 -15.8665500
#>  [57,] -0.4404740669 -2.308691277  0.54462014  0.84294268  -0.5099668
#>  [58,] -1.1970602805 -0.618117808 -1.00896010 -2.03452066 -32.2218640
#>  [59,] -0.9169354437 -0.099570469  0.80499994 -0.31336355   5.1997649
#>  [60,] -2.1579004559 -2.632921615  0.26097679 -0.98168140  -9.1842716
#>  [61,] -1.0182835697 -0.133691142 -1.36555448  1.10698688 -17.2585624
#>  [62,] -2.1479681078  1.158408365 -1.81782705  1.31129859 -27.7592838
#>  [63,]  1.5106201817  0.094985628  0.97502868 -0.03651284   1.3439794
#>  [64,] -1.8333558411 -0.870968173 -0.23909820  0.49309631  -4.3876011
#>  [65,] -1.0298605642 -2.797454498  0.93159638  1.75128545 -15.1173345
#>  [66,] -0.9732610427 -0.711806589  0.66359164  0.22059084   2.5395598
#>  [67,] -0.2952357400 -0.892764150 -1.07139950 -0.06206320 -15.6827050
#>  [68,]  1.8367701010  0.471268710 -1.07979699 -2.01898751 -28.5091304
#>  [69,]  0.2701350554  1.279498941 -0.50484494 -0.81978715  -4.2413961
#>  [70,]  0.4558480990 -0.732536699 -0.04502040  0.18042078  -3.6551924
#>  [71,]  0.1109091255 -0.709976170 -0.06411821  1.01553355  -3.9980667
#>  [72,]  1.2134796103 -1.224209659 -1.13154064 -0.56825680  -4.3453762
#>  [73,] -0.3560882245  0.874768982 -0.90973730  0.19432079 -11.1895836
#>  [74,] -0.0537369408  1.921850124 -0.07105274 -0.57474688  -6.5467235
#>  [75,] -1.1825099027  0.667622389 -0.74639236 -1.26499609 -15.4791518
#>  [76,] -0.3065219331  0.643725386  0.19201952  1.55540539 -13.0024263
#>  [77,]  0.2071984610  1.320897959  0.60552027  0.53038106   5.1955114
#>  [78,] -0.7906318806  2.547341638 -1.04214100 -0.69528117  -0.7565500
#>  [79,] -0.3470608365  0.192885813  0.04345369  0.82597937  -9.9455453
#>  [80,] -0.0639541613 -0.358586081  0.28603165  1.20437802  -4.5135554
#>  [81,]  0.1188902318 -0.854545082  0.01324516  0.42536084  -1.3031776
#>  [82,]  0.0314178571  0.801124416  0.84530334  0.96224005  -0.2429788
#>  [83,] -1.0783658243 -0.355372543 -0.67857498 -0.37482414  -5.0507765
#>  [84,]  0.3101351655  0.941042213  1.23526999 -0.03434930   8.9473833
#>  [85,] -0.1406108982  0.030220088  1.97785052 -0.52930389   1.3651475
#>  [86,] -1.0019861809 -1.559842399 -1.30665781 -0.92334096 -11.7142657
#>  [87,]  0.7142451928 -2.540376998 -0.65977813 -1.46573961 -19.1326973
#>  [88,]  0.4293437529  0.084100609 -1.77935468  0.59999222 -16.1644126
#>  [89,]  1.2419126183 -1.899668397  0.33581112  1.24486620  -6.1543024
#>  [90,] -0.8501222712  0.739982793  0.22964321 -0.46227477   4.1348335
#>  [91,]  1.9724384538  0.844758116 -0.92422235  0.48179204 -10.7048125
#>  [92,]  0.4810690263  0.002792642 -0.37737909  0.57642652  -7.4789002
#>  [93,] -0.5154508084 -0.727340360 -1.77854791  0.90460963 -16.3135667
#>  [94,] -0.9494323432 -0.406055005 -1.08806521 -0.94986470  -6.3281727
#>  [95,]  0.4500589812  1.510047550  0.24623504 -2.11033098 -17.6309577
#>  [96,] -0.8415558286  1.251602697  0.20374347  0.52462338   1.6314697
#>  [97,]  1.1535963110 -0.421060838 -0.57166694  0.01051965   3.7273745
#>  [98,] -0.4310065984 -2.236220687  0.37141030  0.73660432  -3.4665741
#>  [99,] -0.3883479009  1.608899569  0.09772077  0.11236532   0.3751391
#> [100,] -1.9573463539 -1.179807372 -1.01758849  0.07888889  -5.8857491
#> 
#> $eta
#>   [1] -14.55475127  -9.47493284 -15.88974496  -6.33536750 -18.72240637
#>   [6]   5.17052744  -5.99444931  -1.80959313   0.42074846  -2.05023087
#>  [11]  -5.64947759  -7.04552875   4.50878366 -13.11234190 -11.79620847
#>  [16] -12.86767991  -2.02038835   1.40904615 -11.46653433  -8.56738534
#>  [21]   1.70977293  -2.54273763  -1.39843316 -23.33035479  -7.53472021
#>  [26] -47.62459496   5.35884574   4.45851264   3.73402605  -9.12756544
#>  [31]  -8.39956786   2.91546913   4.03592504  -3.16165809   3.00317984
#>  [36]  -8.93987373  -5.33947344   5.08859929   3.86287491  -3.34661564
#>  [41]  -8.01869723 -13.13911795   0.30586286  -8.00404130   7.90448035
#>  [46] -13.92990087  -7.00623774   0.32599996  -1.55227807   5.85527606
#>  [51]  -7.72708673  -2.04105969 -20.26541576 -12.27087396  -8.23151216
#>  [56] -16.09443653  -0.55767197 -31.68504078   3.60923073  -8.00475799
#>  [61] -17.30854638 -28.53100544   0.02884147  -3.62897704 -15.04233548
#>  [66]   2.04591728 -16.68680417 -29.52649940  -4.16722941  -2.08530117
#>  [71]  -4.81606667  -3.70864896 -12.19754138  -7.42560201 -15.13318912
#>  [76] -13.45283457   5.32884821  -1.41375248 -11.77754426  -5.29011727
#>  [81]  -2.59015538   0.63421381  -4.96704950   7.97508570   1.29233299
#>  [86] -11.10422231 -18.15833025 -15.82693341  -6.20608523   4.98689456
#>  [91]  -9.44845508  -8.37571701 -17.17844406  -6.99385791 -17.12242854
#>  [96]   0.65930436   3.17986834  -2.47772677  -0.27143823  -6.07689756
#> 
#> $mu
#>   [1] -14.55475127  -9.47493284 -15.88974496  -6.33536750 -18.72240637
#>   [6]   5.17052744  -5.99444931  -1.80959313   0.42074846  -2.05023087
#>  [11]  -5.64947759  -7.04552875   4.50878366 -13.11234190 -11.79620847
#>  [16] -12.86767991  -2.02038835   1.40904615 -11.46653433  -8.56738534
#>  [21]   1.70977293  -2.54273763  -1.39843316 -23.33035479  -7.53472021
#>  [26] -47.62459496   5.35884574   4.45851264   3.73402605  -9.12756544
#>  [31]  -8.39956786   2.91546913   4.03592504  -3.16165809   3.00317984
#>  [36]  -8.93987373  -5.33947344   5.08859929   3.86287491  -3.34661564
#>  [41]  -8.01869723 -13.13911795   0.30586286  -8.00404130   7.90448035
#>  [46] -13.92990087  -7.00623774   0.32599996  -1.55227807   5.85527606
#>  [51]  -7.72708673  -2.04105969 -20.26541576 -12.27087396  -8.23151216
#>  [56] -16.09443653  -0.55767197 -31.68504078   3.60923073  -8.00475799
#>  [61] -17.30854638 -28.53100544   0.02884147  -3.62897704 -15.04233548
#>  [66]   2.04591728 -16.68680417 -29.52649940  -4.16722941  -2.08530117
#>  [71]  -4.81606667  -3.70864896 -12.19754138  -7.42560201 -15.13318912
#>  [76] -13.45283457   5.32884821  -1.41375248 -11.77754426  -5.29011727
#>  [81]  -2.59015538   0.63421381  -4.96704950   7.97508570   1.29233299
#>  [86] -11.10422231 -18.15833025 -15.82693341  -6.20608523   4.98689456
#>  [91]  -9.44845508  -8.37571701 -17.17844406  -6.99385791 -17.12242854
#>  [96]   0.65930436   3.17986834  -2.47772677  -0.27143823  -6.07689756
#>