Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n

Integer, sample size

p

Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family

Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion

Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)
#> $dat
#>                 x1          x2           x3           x4 y
#>   [1,]  1.72516555  0.97692295 -0.820882821  1.694930895 0
#>   [2,] -0.85764852 -0.30475833 -0.633754568  1.443142408 0
#>   [3,]  2.63047679 -0.27637812 -2.235119499 -0.465002046 0
#>   [4,]  0.49740795 -0.69695181  0.317974723 -1.944792623 0
#>   [5,] -1.51565282  0.47572903  0.243934517  0.657898030 1
#>   [6,] -1.60006951 -1.17464152 -0.305716792 -0.283349875 0
#>   [7,]  0.56178848  0.36467934  1.512865406  1.340543968 0
#>   [8,] -0.39525132  0.36465969  1.242002529  0.517920681 0
#>   [9,]  0.17518782 -0.21045276 -0.701516986  1.726293936 0
#>  [10,] -0.11001147  0.18493349 -0.721060863  0.387489430 0
#>  [11,]  0.66899726 -0.74627433 -1.242140459 -0.920613626 0
#>  [12,] -1.30658119 -1.07683888  0.304522660 -1.001714218 0
#>  [13,]  0.66742429 -0.39126781 -0.506134683 -0.043454442 0
#>  [14,]  0.96534226  0.26343720 -0.461848646 -1.104181244 0
#>  [15,] -0.94860858  0.05574804  1.037499582  3.172084571 0
#>  [16,]  0.69193946  0.82216906  0.589961238 -0.720023860 0
#>  [17,]  0.04839636  0.22795571  0.373621118  2.036859545 0
#>  [18,]  0.73337282  1.23549556 -0.433454700  1.971631279 0
#>  [19,] -0.56856849 -1.87994024 -1.509500982 -0.568846520 0
#>  [20,] -1.30483272 -0.06602089  1.564194087  0.690965246 1
#>  [21,] -1.14213827 -0.44626494  1.393907606  1.698089337 0
#>  [22,]  0.40788332 -0.63377264  1.479252758 -2.054406697 0
#>  [23,]  0.23577144  0.17914108  0.721583328 -0.386595915 1
#>  [24,] -0.75071050 -0.04477752  0.318151582  0.065791499 0
#>  [25,]  0.33749178  1.46400061 -0.606706461 -0.152830773 1
#>  [26,] -0.73037923 -0.95260325  0.080969546  0.576322195 0
#>  [27,] -0.00444295  0.05603272  0.255673943 -2.863482401 0
#>  [28,]  1.02243557  0.95154500 -1.098760292  1.491434726 0
#>  [29,]  1.67457002 -2.46648493  1.764891028  1.597339486 0
#>  [30,] -0.24034919 -1.97315452 -0.626008355  2.445027821 0
#>  [31,] -1.42294906 -1.34527354  0.934976737  0.218916028 1
#>  [32,] -0.10629517 -1.75891223 -0.244953102  2.425203131 0
#>  [33,] -0.86140896 -0.56677528 -0.982073636  0.798099082 0
#>  [34,]  0.06947630  1.95775476 -0.521707632 -1.547937625 0
#>  [35,] -0.82788175 -2.29408736  1.771654185 -0.406530989 1
#>  [36,]  2.47249321  0.92690910 -0.074982972  0.033199046 0
#>  [37,]  1.36829273  0.70562468 -0.005666305  1.504518131 0
#>  [38,] -1.43098908  1.56462795 -1.155175469 -0.280541093 0
#>  [39,]  0.53755293 -0.90015275 -1.222049873 -1.025362566 0
#>  [40,] -0.91562056  2.12350101  0.253619499 -1.201527126 0
#>  [41,]  0.54666077  2.16188422  0.899422472  0.663413975 0
#>  [42,]  1.46940964  0.91340889 -0.971520155 -0.527618041 0
#>  [43,]  0.44182859 -1.24700719  0.280906443  0.172154365 1
#>  [44,] -0.23203068  0.68978818 -1.049462266  0.646111110 0
#>  [45,]  0.03393513  1.00608841 -0.260732932  1.369415961 0
#>  [46,]  1.55284173 -0.71512509 -0.985857489 -0.150089175 0
#>  [47,]  1.45642177 -0.35470754 -1.020327822 -1.777218965 0
#>  [48,] -0.47948028  0.90501669  1.618315379 -0.768274360 1
#>  [49,] -1.51557256 -0.05243406  1.026096582 -0.631296263 0
#>  [50,]  0.11278943  0.86584315  0.446950579  1.034485373 1
#>  [51,] -0.43391380 -0.20866111  1.180541672  1.136040293 1
#>  [52,]  0.06259757 -0.13643707  0.980049016 -0.636834082 1
#>  [53,] -0.86162205  0.58630009 -0.758652977 -1.574076353 0
#>  [54,]  0.48284459  1.14113107 -0.588850828  1.522196862 0
#>  [55,]  0.09123596  0.05898591  0.694664626  1.522315653 0
#>  [56,]  1.83567770 -0.23880786  0.311278187 -0.162527120 0
#>  [57,] -0.81680113 -1.19736825  0.537625869  0.724547148 1
#>  [58,]  0.26372557  0.95652687  0.854050356  0.364204759 1
#>  [59,]  0.49151418  2.33861894 -0.253094707  1.090691570 0
#>  [60,] -1.81313573 -1.06795451 -0.032136040 -0.326765657 1
#>  [61,] -0.17208877 -0.74883854 -0.114655000  0.664721108 0
#>  [62,] -2.17334506 -0.68715490 -0.218186668 -1.076634721 0
#>  [63,]  1.03220671  0.71434287 -0.930772171  1.222655354 0
#>  [64,] -0.98758381 -1.08530084 -2.956373035  1.246224231 0
#>  [65,]  0.56400477  0.51115033 -1.273963438 -0.328287486 0
#>  [66,]  0.93064781 -0.02741348  0.828738003  2.165597096 0
#>  [67,] -0.08178161  0.41651623 -0.753257360 -1.350342673 0
#>  [68,]  0.08778693  0.02338821  1.838600302 -1.190392695 1
#>  [69,] -0.33496142 -1.69202051  0.278659378  1.268653759 0
#>  [70,]  0.21848560  0.81230811 -0.275768803  1.363835898 0
#>  [71,]  1.45145589 -1.74435078 -0.491300767  1.034291465 0
#>  [72,] -0.77553137 -0.79302376  0.578796801  0.618779072 1
#>  [73,]  0.98529658 -0.83728293  0.277349155  0.003190198 0
#>  [74,]  0.92063201  0.19545832 -0.279185177 -0.383836651 0
#>  [75,] -1.77081171 -0.07592557  1.850101471 -0.900792207 1
#>  [76,]  1.18680896  0.63027729 -0.168796066 -1.236632903 0
#>  [77,] -0.01635361  0.85562320  1.080486921 -0.029303200 0
#>  [78,]  1.68239512 -0.38845651 -0.386296715 -0.522872107 0
#>  [79,] -0.70482809  0.96020549 -0.210856585  1.044057450 0
#>  [80,]  0.68313324 -0.21904323 -0.330470171  0.397246000 0
#>  [81,]  0.44351560 -0.12974244  0.430704922  0.383334411 1
#>  [82,]  0.41757601  0.70970842 -0.640243037 -1.561018121 0
#>  [83,] -1.23344368 -1.07783186  0.231062916  1.276145326 0
#>  [84,] -0.20481778  1.13309236  0.382527034 -0.898129053 0
#>  [85,]  1.50682001  0.54787718  1.656637067  0.705397144 1
#>  [86,]  0.57997706  0.23119551 -0.218945095  1.392096532 0
#>  [87,]  0.17171441 -0.59788752 -1.989947438  0.172923464 0
#>  [88,]  0.62598729 -0.95794622 -0.266830866 -0.889156020 0
#>  [89,]  0.03904088 -1.12514556  0.525491191 -1.111408620 0
#>  [90,]  0.01676711 -0.52803149  0.738082134  0.286961372 1
#>  [91,] -0.60190479  0.56626580 -0.336310888  1.614017409 0
#>  [92,]  0.11773609 -0.20900242  0.080920063 -0.991781496 0
#>  [93,]  1.98154922  0.10502567 -0.410258068  0.800461773 0
#>  [94,] -0.45500023  0.13074312  1.011133401 -0.012717481 1
#>  [95,] -0.16081366  1.37361712  0.304013634 -0.605980862 0
#>  [96,]  0.37853466  1.17387419 -2.087247146  1.736956490 0
#>  [97,] -1.17764747 -1.61775817 -0.095632041 -0.270653700 0
#>  [98,]  0.47878261  0.60233893  0.388492788  1.016276777 0
#>  [99,] -0.19424898  0.39170848 -1.339216687 -0.776672486 0
#> [100,] -0.69045508 -0.12003644  0.147207562 -0.786578755 1
#> 
#> $eta
#>   [1] -28.92490868  -8.82721568 -17.17637108 -19.08548554   1.46868938
#>   [6]  -0.68902143  -2.42577343   2.42710260 -15.15635184 -12.91963106
#>  [11] -19.39177104 -11.59616973  -4.95515387 -13.06224484 -47.95425511
#>  [16]  -5.03724451 -20.88915742 -30.73932878 -13.95811937  -1.64790985
#>  [21]  -7.50174202 -10.07938380   3.81974757   2.54735049   1.99714762
#>  [26]  -1.63137704 -45.05948453 -21.28597261  -4.28336484 -43.92011457
#>  [31]   5.10539182 -37.58551610  -4.32837566 -17.38112697  15.52139132
#>  [36]  -3.39211059  -6.36076759  -6.94294235 -19.22558810  -8.51684776
#>  [41]  -3.22596709  -9.77673797   3.67570014 -11.77356204 -14.96395903
#>  [46] -10.26630287 -18.39865052   2.79001249   0.45123342  -1.81853631
#>  [51]  -1.93487714   3.42108336 -10.62418277 -19.78516658 -10.17396795
#>  [56]  -2.39092809   4.34090909   5.07963670  -8.47303044  -0.01601548
#>  [61] -10.70200296 -13.45577637 -10.90448422 -28.65273927  -8.00368794
#>  [66] -25.21679485 -15.51638083   1.31330895 -12.72093443  -7.11938033
#>  [71] -10.08699846   2.20866807  -3.76375422  -9.28887249   8.26292308
#>  [76]  -1.74928070   3.81408525  -5.82082527  -6.34211709  -7.63395671
#>  [81]   1.07568950 -12.31433873 -15.42732318 -11.81388736   9.65862297
#>  [86] -17.22959198  -7.42142472 -14.50071571  -4.43160388   6.16254789
#>  [91]  -9.56998264  -1.60193708 -11.69257005  -0.62402808  -4.65449704
#>  [96] -28.97064903  -5.63265654   0.29061902 -15.72559831  -0.30145187
#> 
#> $mu
#>   [1] 2.742027e-13 1.466646e-04 3.470538e-08 5.143739e-09 8.128581e-01
#>   [6] 3.342508e-01 8.122834e-02 9.188708e-01 2.616254e-07 2.449483e-06
#>  [11] 3.786699e-09 9.201179e-06 6.997683e-03 2.123920e-06 2.220446e-16
#>  [16] 6.449742e-03 8.471381e-10 2.220446e-16 8.670924e-07 1.613916e-01
#>  [21] 5.518171e-04 4.193349e-05 9.785374e-01 9.273953e-01 8.804973e-01
#>  [26] 1.636418e-01 2.220446e-16 5.696650e-10 1.360842e-02 2.220446e-16
#>  [31] 9.939726e-01 2.220446e-16 1.301727e-02 2.827955e-08 9.999998e-01
#>  [36] 3.254294e-02 1.725059e-03 9.644937e-04 4.471293e-09 2.000291e-04
#>  [41] 3.820014e-02 5.675342e-05 9.752942e-01 7.705550e-06 3.171283e-07
#>  [46] 3.478454e-05 1.022275e-08 9.421337e-01 6.109325e-01 1.396096e-01
#>  [51] 1.262117e-01 9.683570e-01 2.432011e-05 2.555120e-09 3.814919e-05
#>  [56] 8.386710e-02 9.871428e-01 9.938163e-01 2.089868e-04 4.959962e-01
#>  [61] 2.249932e-05 1.432946e-06 1.837531e-05 3.599759e-13 3.341161e-04
#>  [66] 1.118112e-11 1.825246e-07 7.880663e-01 2.987907e-06 8.086137e-04
#>  [71] 4.161541e-05 9.010252e-01 2.267061e-02 9.243869e-05 9.997422e-01
#>  [76] 1.481379e-01 9.784182e-01 2.956391e-03 1.757477e-03 4.835091e-04
#>  [81] 7.456774e-01 4.486924e-06 1.995256e-07 7.401005e-06 9.999361e-01
#>  [86] 3.290662e-08 5.979384e-04 5.039866e-07 1.175556e-02 9.978976e-01
#>  [91] 6.978772e-05 1.677111e-01 8.355601e-06 3.488659e-01 9.428948e-03
#>  [96] 2.619431e-13 3.566291e-03 5.721477e-01 1.480674e-07 4.252026e-01
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())
#> $dat
#>                 x1          x2           x3           x4  y
#>   [1,]  1.94543032 -0.39221105 -0.897856529 -0.465137867  6
#>   [2,] -0.53771934 -0.19868212  1.082318070 -0.823952449  6
#>   [3,]  1.32972182  0.13576838 -0.348178380  1.256897913  1
#>   [4,]  1.58085659  1.64014444 -0.367541198  0.313275537  5
#>   [5,] -1.17621725 -0.15846091 -0.222405023 -2.170979063  0
#>   [6,]  0.84015000  1.91322249  0.836596001  1.374856066  0
#>   [7,] -0.50457032 -0.16104419  0.110013938 -0.048447344  6
#>   [8,] -0.52690953 -0.33506100  2.384948960  1.161905731 32
#>   [9,] -0.08321815  0.75997032 -1.384928032  0.557435770  1
#>  [10,]  0.63014219 -0.50133703 -1.172103375  0.882795466  0
#>  [11,] -1.54309328  0.94605756 -0.078954078 -0.858127000  1
#>  [12,] -1.67736835  0.59506358  1.454673991 -0.475754700 59
#>  [13,] -1.36804377  0.41208697  1.558529267  0.661071476  5
#>  [14,]  0.69174488 -1.15267223 -1.446796725 -0.356466495  0
#>  [15,] -1.72730304  0.23511833  1.221870822 -2.217923640  0
#>  [16,] -1.56723865 -0.11329750 -2.167027488  0.894498540  1
#>  [17,] -3.05223366 -0.29768168  0.568625096 -0.598346194  6
#>  [18,]  1.35705345  0.99655011 -0.272972861  0.380403662  6
#>  [19,] -0.41468180  0.05861199  1.580720162  1.163521466  0
#>  [20,] -0.59636815 -0.10423478  0.068952360 -0.584734906  3
#>  [21,] -0.78720703  0.26604457  0.448442106 -1.017356641  1
#>  [22,]  1.88772868 -0.56843603 -0.725702861 -1.007236488  1
#>  [23,] -0.74160317  0.12735733 -0.617484508  0.233546384  6
#>  [24,] -1.76907076 -0.13647026 -2.163146714 -1.089578829  1
#>  [25,] -0.24928270  1.10933814 -1.296047945 -0.121936442  1
#>  [26,]  0.32838563 -0.56592658  0.258183393  0.439910732 15
#>  [27,]  0.87994033 -1.57205796  2.032474593  0.343482146 37
#>  [28,]  0.57932850 -1.30546586 -0.145501928 -0.226638144  3
#>  [29,]  0.08826500 -1.76179526 -0.292773986  1.388599664  0
#>  [30,] -0.39748060 -2.37323834 -1.211010945 -0.055088246  7
#>  [31,]  0.06446319 -0.66074431  1.021169105 -0.179647570 10
#>  [32,] -1.58789732 -1.89917602  0.324376047  0.922638136  1
#>  [33,] -1.01693089 -0.18593490 -0.256771698 -0.927408256  0
#>  [34,] -0.73653481 -0.74384831  0.068060031  0.494832618  1
#>  [35,]  0.68332776  0.31171414 -0.680704843 -0.661877594  0
#>  [36,]  0.55005602  0.53812303  0.376306975 -1.321148818  2
#>  [37,] -1.35671395 -0.60459071  0.610650399  0.739616666  0
#>  [38,]  1.10074239 -2.08002510 -0.321686608  0.237943111  2
#>  [39,] -1.13762447 -1.17570450 -1.427626289 -0.970947215  0
#>  [40,]  1.65625341  0.30994647 -0.542156531  0.188507770  1
#>  [41,] -0.29122244  0.28112748  1.183695680 -0.782861658  0
#>  [42,]  0.71496070 -0.76434161 -1.139410040 -0.981990945  0
#>  [43,]  0.30517885 -0.85388830  0.467631344  0.502802768  1
#>  [44,] -0.63528412 -0.18158464 -1.234919532 -0.798136131  0
#>  [45,]  0.75721563 -1.02296973 -0.702053607  0.338211643  0
#>  [46,] -0.32783063 -1.54448038  0.740502232 -1.142632293  0
#>  [47,] -0.49516684  2.20072373 -1.370790196 -0.280272942  0
#>  [48,]  0.79906805 -0.52181628  1.706606696 -0.234940900 28
#>  [49,]  1.79295824 -0.23727857 -1.791608009 -1.264104416  0
#>  [50,] -0.07981346 -1.69055152 -0.513351406 -0.001646742  4
#>  [51,] -0.50271929 -1.45285011 -0.281092357  0.574867704  5
#>  [52,] -1.73228673 -0.91985533  0.533477124  0.858295881  2
#>  [53,] -0.38438824 -0.25846280  0.006496552 -0.019501654  3
#>  [54,] -1.63133286  0.23333350  0.048643348 -1.528624738  0
#>  [55,]  0.81854821  0.92544217 -0.938436841 -0.257901001  0
#>  [56,]  0.88095832  0.27076268 -1.176620315 -0.584705458  1
#>  [57,] -1.26238638  0.18339461 -0.311658493  0.113764207  1
#>  [58,]  0.73755925  1.73670423  0.042749846  1.253246630  0
#>  [59,]  0.34928385  0.21064323 -0.016892736  1.580550130  0
#>  [60,]  0.30656806  2.35698394  1.080331107 -0.451261204 20
#>  [61,] -1.26276574 -0.90836390  0.473988972 -1.229637287  0
#>  [62,]  0.43318204 -1.22356078  0.033591020  0.593970959  6
#>  [63,]  0.43134949 -0.03634975 -1.049403490  0.742568964  1
#>  [64,]  0.06905373 -0.08010288  0.593547192 -1.546483372  0
#>  [65,] -1.19466292 -1.47163936  1.736295992  1.632446059  0
#>  [66,]  0.53505141  0.96854847  0.060677430  0.220269007  1
#>  [67,]  0.11481488 -1.36059854  0.250216835  0.866222969  4
#>  [68,]  0.62057039  0.88102150  1.724843449  0.353990096  9
#>  [69,] -0.86291392 -0.08359894 -1.189208295  1.087063129  0
#>  [70,] -1.50460493  0.50662515  0.638086654 -0.349320872 21
#>  [71,]  0.74595972 -0.82731479  0.850212997  1.321094396  0
#>  [72,]  1.23978827  0.02187065  1.140090655  0.850352349  1
#>  [73,]  0.56613674  0.64465346  0.514477376  0.413848162  7
#>  [74,]  0.91403359  0.64349054 -1.127281221 -0.752028707  1
#>  [75,] -0.16516343 -0.12466408 -0.585582656 -1.308515155  0
#>  [76,]  0.45088263  0.88045153 -1.482136078 -0.881635144  0
#>  [77,] -0.15297564 -0.42027400 -0.059316650 -0.515393095  2
#>  [78,] -0.73086984 -0.45289749  0.513143887 -1.338226135  1
#>  [79,]  1.26497705 -0.67579521  0.757339610 -0.619160817  6
#>  [80,]  0.77100851  1.95776782  0.067840676 -0.405947680  0
#>  [81,] -1.29336393  0.15129811 -0.470419989  0.335269037  0
#>  [82,]  1.05554827  1.08388246 -1.252825713  0.751276049  1
#>  [83,] -0.33236652  0.82593967 -0.561072091 -1.715587570  0
#>  [84,] -0.14542262 -1.59746315 -0.284257631  0.107466413  1
#>  [85,]  0.28499809  1.54163211  0.547353099 -0.535118544 16
#>  [86,] -0.16666641 -0.79624749  0.602421318  0.085742110  2
#>  [87,] -1.56126455  0.19861940  0.050106467 -0.357950680  1
#>  [88,] -1.30789271  0.41178516 -0.620859599 -0.910024607  0
#>  [89,] -1.18290183 -1.38457356 -1.130856621 -2.479906593  0
#>  [90,] -1.74659205 -0.96412470 -1.161189483 -1.488485714  0
#>  [91,]  0.69304130 -1.21210004  0.039957499  0.696756735  2
#>  [92,] -0.61598268 -0.55276907  1.710766709 -0.854503290 21
#>  [93,] -0.11877785  0.96245832  0.620642295 -0.757839595  0
#>  [94,]  1.57013851 -0.23609920 -0.086938841 -0.373510048  4
#>  [95,] -0.96632117 -0.27787292 -2.402472630 -1.212686335  0
#>  [96,] -0.29807032  0.05607692  0.093989408 -1.502853277  0
#>  [97,]  0.89328957 -1.06050313 -2.176690276 -1.972397600  0
#>  [98,] -0.05814834 -0.24123120  1.029606061  1.289734980  0
#>  [99,] -0.65839597  0.62344956  0.539247049 -0.163782447 40
#> [100,]  1.03954831  1.23016612 -2.343787136 -0.427827508  1
#> 
#> $eta
#>   [1]   1.37840849   1.23422405  -1.93836798   1.52268694  -9.07560273
#>   [6]  -3.20257783   1.41164364   3.67121017   0.40283895  -0.29082558
#>  [11]  -0.50392312   4.15348292   1.78488313  -0.20506134  -7.05037928
#>  [16]  -0.49570101   1.38503731   0.93022808  -1.26628045   0.90991083
#>  [21]   0.14044602  -0.88597481   1.14996599  -0.54726338  -0.90285036
#>  [26]   2.62233725   3.71184141   1.51534458  -2.57444010   1.46894586
#>  [31]   2.50085179  -0.61573251  -0.41962076   1.48972097  -0.69919418
#>  [36]  -1.25119230   0.26101607   1.23035535  -1.51725802   0.20701256
#>  [41]  -0.18555994  -2.33292544   1.15947822   0.66366730  -0.82902662
#>  [46]  -0.94393316   0.19007414   2.94468727  -2.68521349   0.63252561
#>  [51]   1.63743761   0.20241572   1.26304614  -3.08730370  -0.45177216
#>  [56]  -0.60176922  -0.52465340  -2.16974861  -3.47097447   2.87983644
#>  [61]  -3.37537603   1.44337531  -0.29091235  -3.17619150  -2.05074361
#>  [66]   0.34315117   1.60365907   1.98925063  -1.00178339   2.43879537
#>  [71]  -3.38790532   0.86758090   1.84043731   0.40465764  -3.34653974
#>  [76]  -0.47419770   1.02583832   0.05855516   2.04814218  -0.61394784
#>  [81]  -0.83819669  -0.67220995  -5.50443438   0.92868065   3.06639707
#>  [86]   0.13912395   0.80671993  -0.88200320 -11.10847085  -3.26870004
#>  [91]  -0.23238923   3.09258516  -0.82476145   1.06215345  -1.06998478
#>  [96]  -5.02156821  -7.99373503  -1.13375891   3.43336831   0.33346362
#> 
#> $mu
#>   [1] 3.968581e+00 3.435712e+00 1.439387e-01 4.584527e+00 1.144237e-04
#>   [6] 4.065726e-02 4.102693e+00 3.929944e+01 1.496066e+00 7.476461e-01
#>  [11] 6.041558e-01 6.365532e+01 5.958883e+00 8.145973e-01 8.670800e-04
#>  [16] 6.091437e-01 3.994975e+00 2.535087e+00 2.818781e-01 2.484101e+00
#>  [21] 1.150787e+00 4.123121e-01 3.158085e+00 5.785309e-01 4.054124e-01
#>  [26] 1.376786e+01 4.092910e+01 4.550989e+00 7.619647e-02 4.344653e+00
#>  [31] 1.219288e+01 5.402450e-01 6.572960e-01 4.435858e+00 4.969856e-01
#>  [36] 2.861634e-01 1.298249e+00 3.422445e+00 2.193124e-01 1.229998e+00
#>  [41] 8.306390e-01 9.701153e-02 3.188269e+00 1.941901e+00 4.364739e-01
#>  [46] 3.890945e-01 1.209339e+00 1.900472e+01 6.820663e-02 1.882359e+00
#>  [51] 5.141977e+00 1.224357e+00 3.536177e+00 4.562481e-02 6.364992e-01
#>  [56] 5.478415e-01 5.917604e-01 1.142063e-01 3.108672e-02 1.781136e+01
#>  [61] 3.420525e-02 4.234966e+00 7.475812e-01 4.174434e-02 1.286392e-01
#>  [66] 1.409382e+00 4.971189e+00 7.310054e+00 3.672240e-01 1.145923e+01
#>  [71] 3.377936e-02 2.381144e+00 6.299292e+00 1.498789e+00 3.520597e-02
#>  [76] 6.223842e-01 2.789433e+00 1.060303e+00 7.753483e+00 5.412100e-01
#>  [81] 4.324897e-01 5.105790e-01 4.068689e-03 2.531167e+00 2.146443e+01
#>  [86] 1.149267e+00 2.240547e+00 4.139529e-01 1.498485e-05 3.805587e-02
#>  [91] 7.926375e-01 2.203397e+01 4.383395e-01 2.892593e+00 3.430137e-01
#>  [96] 6.594178e-03 3.375709e-04 3.218213e-01 3.098082e+01 1.395794e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )
#> $dat
#>                  x1           x2          x3          x4           y
#>   [1,]  0.610795904 -0.919387831  0.40230310 -0.29259165  -6.1764743
#>   [2,] -1.013161348  0.488360244 -0.73132634  1.10759023  -8.3997615
#>   [3,] -0.302910468 -1.782977722  0.59541393  0.73530974  -8.5630737
#>   [4,]  1.544316703  0.642074829  1.75346751  0.58562110   9.7855852
#>   [5,] -1.115556539 -1.193369500 -1.50050930 -2.21288544 -37.7222351
#>   [6,]  0.246302617 -0.013479090 -0.18029749  0.29972495  -1.8817874
#>   [7,]  0.709807854 -0.092230526  1.68257834 -0.32546920   4.5793054
#>   [8,]  0.461302222  0.045608205 -0.08934324  1.05651353 -10.0723035
#>   [9,]  0.244404570  0.594053210  0.11962770 -1.27303393  -1.1949909
#>  [10,]  0.007854124  0.397666257 -1.26063586 -1.32605126 -17.0908377
#>  [11,]  0.353409022  0.357308094 -1.46433109 -1.96684032 -24.5059220
#>  [12,] -0.372742873 -1.170072329 -0.46833681  0.15291063  -7.0964305
#>  [13,]  0.263383119 -0.514772168 -1.35596193 -0.80947846  -5.6503783
#>  [14,] -0.356934365 -0.238605973 -1.10036803 -0.41363976 -10.6599901
#>  [15,] -0.661658068 -1.326378635  0.53799365  1.88223417  -8.7732226
#>  [16,] -1.204253419  1.173076637 -0.75467858 -0.18208050 -15.7870617
#>  [17,] -0.825551236  0.056234916  0.49917721  0.06810384   1.2775793
#>  [18,]  1.546452139  0.532346619 -0.64207360  0.71641945  -7.9448248
#>  [19,]  0.116068557  0.722036043  0.07768941 -2.22456265 -18.4126678
#>  [20,]  0.577518613  1.806106829 -0.43070526 -0.19329598  -5.8851628
#>  [21,]  0.605922187  1.954647961  1.03490961 -0.73047725  -3.3143089
#>  [22,] -0.081898253  0.346256812 -0.79462973 -0.74826852 -10.6225277
#>  [23,]  1.147676621 -0.848638452  0.73753252 -1.13637072  -2.7232832
#>  [24,]  0.979025391 -1.205795726  0.50065851 -0.23963101   1.1701813
#>  [25,]  0.153318756  0.747206345  0.24546916 -1.63628051  -5.4723590
#>  [26,]  0.469911492 -1.133206406  0.28401433  0.26951118  -0.1875834
#>  [27,] -1.828671337  0.558582034 -1.96351555  0.43478145  -7.1155571
#>  [28,] -0.727332471 -1.437619942 -0.18100645 -1.00899138   2.9434790
#>  [29,]  0.030427573 -0.002359203 -1.53948927  2.92650809 -53.2243282
#>  [30,] -0.037790816 -1.369329334 -0.27612044 -0.62668127  -2.7031952
#>  [31,]  1.293323479  0.225825276  1.33745027 -0.85800868   3.7245552
#>  [32,]  0.827152943 -1.036834441  1.16295169  1.33436586 -13.0430645
#>  [33,] -0.441385180 -1.777713127 -0.51987783  0.33512436 -10.5676719
#>  [34,]  0.243876220 -0.987165467 -0.18894813 -0.37567529  -2.1815964
#>  [35,]  0.200643401 -1.702441598  0.76755681  0.87759503   3.2873660
#>  [36,] -1.098653649 -1.101148252  0.27300706 -0.40476786  -5.2133320
#>  [37,]  0.503579160 -1.231693647 -0.57534843 -0.13472855  -3.3735731
#>  [38,] -0.560599055  0.181007737  1.95452031 -0.12595143  10.7703036
#>  [39,] -0.494921875 -0.399271095  0.32508845  0.94136121   0.9240496
#>  [40,]  1.177836559 -0.013887174  0.39359024  0.16841208   0.7899097
#>  [41,] -0.931191214  0.717099673  0.16463345  0.07027487   4.2841526
#>  [42,]  1.403596974 -0.108026369 -0.06896552 -1.07094009  -7.4613285
#>  [43,]  1.163867113  0.892027151 -0.82084040  1.46734746 -14.4495133
#>  [44,]  0.287130776 -0.979323011 -0.50968294 -0.49083030  -5.4631520
#>  [45,] -0.929113352  0.116986940 -0.61773631  1.57146061 -18.9681809
#>  [46,] -0.873145646  0.319701740  0.40703016  0.24934668   6.4382128
#>  [47,] -1.629264393 -0.495364133 -0.62531304  0.79027643  -0.2512835
#>  [48,]  1.244327915 -0.203271748  0.08350965  0.99186710   0.5477714
#>  [49,] -0.970370718 -0.372132089 -0.57140685  0.04776553  -0.7441376
#>  [50,] -0.087592952 -1.044004938 -1.08640134 -0.34607726 -13.0425078
#>  [51,]  0.060664930 -0.610926174 -0.79099996 -0.75963533  -6.7505388
#>  [52,] -1.746905341 -0.769157759  0.85341178 -0.31116056   6.6384952
#>  [53,] -2.706913683  0.661379164  0.98517589 -0.02506397  14.7414394
#>  [54,] -1.326318093 -1.818587486  0.04156930  0.44809678  -9.5148072
#>  [55,] -1.280301795 -0.499824263  0.84523864  1.18209234  -4.9220840
#>  [56,]  1.433591175  1.116084790  1.10186035  1.49164868  -7.5407794
#>  [57,]  0.469453877 -1.317861591 -0.81007409 -0.15604527  -4.1083821
#>  [58,] -0.301680238  1.667418966  1.05447977  1.07810540  -3.5793589
#>  [59,]  0.039262000  1.176687820  1.18652801 -0.61515487   2.1665313
#>  [60,] -0.604129358  0.239224584  1.11227717  1.23409779  -1.9387381
#>  [61,] -1.447895696  0.140247082 -0.32000236 -0.18710482  -9.2477783
#>  [62,] -0.092495552  1.808065113  0.13464286 -0.74626188  -6.3243186
#>  [63,] -0.535519959  0.521056238 -0.80352030  1.49690736 -10.3926698
#>  [64,] -0.065056740 -0.435239398 -1.53424503 -0.14713614  -9.6468956
#>  [65,]  0.613938595 -1.381476599  1.34223426 -0.82742600   3.3033163
#>  [66,] -0.449168848 -0.599930908  0.34293327  0.05236593   0.6271758
#>  [67,]  1.098183036  0.530588503 -1.36793257  1.90761818 -20.9370333
#>  [68,]  0.097864913  2.040645698  0.85298678 -0.07714551   0.9140557
#>  [69,]  0.220909904  0.438671115 -1.80285543  0.28627152  -2.9208259
#>  [70,]  0.071662564 -2.399832796  2.17543284 -0.53361984  15.6022050
#>  [71,]  2.116277696 -1.066028534 -0.18963331  0.01980993  -0.5951479
#>  [72,] -1.260433428  0.641648621 -0.08399013 -0.11686083  -2.9568242
#>  [73,]  0.331030608 -0.985108033 -1.34059178  0.48272472  -9.8402977
#>  [74,] -0.371084000 -0.597425421 -0.82501911 -1.27397073 -15.5202282
#>  [75,]  0.651625327 -0.922198614  1.23843638 -0.67032734  -3.8116197
#>  [76,] -0.689405524  0.327276420 -1.43008316  0.35443321  -5.4247469
#>  [77,] -0.822904339 -1.365589526 -0.23478105 -0.14354479   4.1098529
#>  [78,]  0.486100684  0.862008990 -0.46153387  0.07594701  -4.8525575
#>  [79,]  0.327232210 -0.615611060  0.28826123 -0.29632356   4.4602797
#>  [80,]  0.154190160 -1.247460263  0.44768956  1.75365402  -9.0129301
#>  [81,] -1.029692000  0.429449405  1.27831926  0.71221173   5.7195603
#>  [82,] -0.019484134  0.547159427  0.13205426 -0.17698723   1.3047211
#>  [83,]  0.274790590 -0.810052331 -0.61990092  0.61262070  -6.1608896
#>  [84,] -0.210237726 -1.110598253 -1.08520821 -0.22287183 -14.4103017
#>  [85,] -2.649746700  0.066993182 -0.14956310  2.57505567 -34.1644909
#>  [86,]  0.471436871  0.753967026 -1.31661578  1.28640580 -14.2411628
#>  [87,] -0.680748318 -0.671815881  1.24188757 -1.28074093   0.8720995
#>  [88,] -0.943841706 -0.984112954  0.29865740 -0.88518100  -6.3044584
#>  [89,] -1.770282809 -2.882583451 -1.45812008  1.35148626 -18.2272409
#>  [90,]  2.504657447  0.386866566  0.39065362  1.22964490  -5.1657526
#>  [91,]  0.898912113  0.935353174  1.09711422  0.52750465  -1.8869854
#>  [92,] -1.036969530  1.396205349  1.31626753  0.23653796   6.6732991
#>  [93,] -2.383759766 -2.408410239  0.28105978  2.65549384 -35.6378315
#>  [94,] -0.395373059  0.729580763 -0.49624373 -2.07694980 -26.6742046
#>  [95,]  1.327712850 -0.037893537 -0.39805320  0.32464876  -3.2073859
#>  [96,]  0.667406178  0.476784500  0.04897091 -0.85960437  -5.9862976
#>  [97,]  0.194314090 -2.452094719 -0.85988097  2.03905769 -17.3170746
#>  [98,]  0.463671239 -1.311157037  0.88058227  0.04377713   9.4570908
#>  [99,] -0.220954095 -0.279558486  0.43980697  0.69966748  -4.1072474
#> [100,] -0.787428296 -1.122057503 -0.88609009  0.50782446  -2.7482923
#> 
#> $eta
#>   [1]  -6.72049218  -9.07392253  -7.95599937   9.62725952 -37.73229564
#>   [6]  -1.36783868   1.33733607 -10.32261676   0.09506578 -15.94969748
#>  [11] -24.95365373  -8.02104810  -4.61223322 -11.20923808  -8.41497521
#>  [16] -14.31808673   1.96196101  -5.53227594 -19.94380972  -6.02346828
#>  [21]  -3.87558249 -11.10973133  -3.12973441   1.44152903  -7.32974936
#>  [26]  -0.14630011  -5.85953276   1.24968844 -56.09060465  -2.50083225
#>  [31]   5.23292999 -10.82573034  -9.47776259  -1.99645082   3.21665606
#>  [36]  -5.27234139  -3.65312841   8.79110224  -0.17505285   1.88465022
#>  [41]   4.15545150  -6.32991606 -14.43466634  -4.62731503 -19.28980324
#>  [46]   3.95975356  -1.18088574  -0.10137711  -0.19710136 -14.08574495
#>  [51]  -5.81047810   5.81007031  12.70899629  -9.89936356  -4.81248490
#>  [56]  -7.88286447  -2.33062222  -2.31731529   1.66984997  -1.54923684
#>  [61]  -8.91575421  -5.68265434 -12.68550130  -8.81841230   3.72959892
#>  [66]   0.68828032 -21.19807392   2.59163955  -5.27807414  16.27145973
#>  [71]  -1.74277610  -3.36185387 -10.01697326 -16.41560059  -4.39071285
#>  [76]  -6.12446826   5.23776590  -4.67140163   5.56378023  -8.20788199
#>  [81]   5.05482400   2.44242731  -5.23257357 -14.57276728 -35.49903725
#>  [86] -14.95174765   2.24891531  -5.58186602 -18.30932532  -5.45356857
#>  [91]  -1.74213111   6.59289446 -34.41895201 -26.75025684  -2.95862277
#>  [96]  -6.04815711 -18.82994899   7.99911516  -3.69286939  -4.44051717
#> 
#> $mu
#>   [1]  -6.72049218  -9.07392253  -7.95599937   9.62725952 -37.73229564
#>   [6]  -1.36783868   1.33733607 -10.32261676   0.09506578 -15.94969748
#>  [11] -24.95365373  -8.02104810  -4.61223322 -11.20923808  -8.41497521
#>  [16] -14.31808673   1.96196101  -5.53227594 -19.94380972  -6.02346828
#>  [21]  -3.87558249 -11.10973133  -3.12973441   1.44152903  -7.32974936
#>  [26]  -0.14630011  -5.85953276   1.24968844 -56.09060465  -2.50083225
#>  [31]   5.23292999 -10.82573034  -9.47776259  -1.99645082   3.21665606
#>  [36]  -5.27234139  -3.65312841   8.79110224  -0.17505285   1.88465022
#>  [41]   4.15545150  -6.32991606 -14.43466634  -4.62731503 -19.28980324
#>  [46]   3.95975356  -1.18088574  -0.10137711  -0.19710136 -14.08574495
#>  [51]  -5.81047810   5.81007031  12.70899629  -9.89936356  -4.81248490
#>  [56]  -7.88286447  -2.33062222  -2.31731529   1.66984997  -1.54923684
#>  [61]  -8.91575421  -5.68265434 -12.68550130  -8.81841230   3.72959892
#>  [66]   0.68828032 -21.19807392   2.59163955  -5.27807414  16.27145973
#>  [71]  -1.74277610  -3.36185387 -10.01697326 -16.41560059  -4.39071285
#>  [76]  -6.12446826   5.23776590  -4.67140163   5.56378023  -8.20788199
#>  [81]   5.05482400   2.44242731  -5.23257357 -14.57276728 -35.49903725
#>  [86] -14.95174765   2.24891531  -5.58186602 -18.30932532  -5.45356857
#>  [91]  -1.74213111   6.59289446 -34.41895201 -26.75025684  -2.95862277
#>  [96]  -6.04815711 -18.82994899   7.99911516  -3.69286939  -4.44051717
#>