Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n
Integer, sample size

p
Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family
Family object, one of `binomial` (default), `gaussian` , `poisson`. Please see family for more detail

dispersion
Double, required for gaussian distribution, match to sd parameter.

Value

a list contains `dat`, `eta`, `mu`

dat a numeric matrix, contains predictors `x1` to `xp` where p is replaced by the integer value. The outcome variable is named `y`

etathe linear predictor of the simulated data

muthe mean of the simulated data

Examples

# Binomial Outcome
sim_Bai(100, 4)

#> $dat
#>                  x1           x2           x3           x4 y
#>   [1,]  0.385994964  1.143623543 -0.110119494  0.578376820 0
#>   [2,] -0.838842632 -1.381713624 -1.051575816  0.113186948 1
#>   [3,] -0.530027590 -0.844069552 -0.891401342  0.440206334 0
#>   [4,] -2.293182084 -1.836583479 -0.619138043 -2.002320938 0
#>   [5,]  1.855263642  1.044170831  0.255199046  1.041633836 0
#>   [6,] -0.624714964  0.215210107 -0.439315917 -1.013748921 0
#>   [7,]  0.175872582  0.800349614  1.468149212 -0.504802858 1
#>   [8,] -1.365788853  0.714933031  0.832506126  0.089953043 1
#>   [9,] -1.345518474 -0.317171209 -0.913787407 -1.292000906 0
#>  [10,]  0.947265271 -2.252045124 -0.272595426 -1.730530080 0
#>  [11,] -0.115739671  1.505248197  0.751086555  1.132424775 0
#>  [12,] -0.429404494 -0.419147872 -0.076856255  0.955398518 0
#>  [13,]  0.079631976 -0.208179987  0.147101528 -0.714008278 1
#>  [14,]  1.006463994  1.025008622 -0.321238201 -0.130393939 0
#>  [15,]  1.754518747 -2.109352248 -1.703140073  0.950804885 0
#>  [16,] -0.119361311  0.581155983 -0.761038444  0.470794223 0
#>  [17,]  1.643998625  0.427476044 -0.759025393  0.334838125 0
#>  [18,] -1.310997765  0.309712183  0.292608648 -1.549423258 0
#>  [19,]  0.362701424 -0.348680801  0.332236386 -0.621150422 1
#>  [20,]  1.369665491  0.472684011 -0.231934768 -1.615618228 0
#>  [21,] -0.348512400  0.702742649 -0.815807168  0.853465374 0
#>  [22,] -0.545470652 -1.543797498  0.216074797  1.888819161 0
#>  [23,] -0.020519678 -1.139419424 -1.138865785 -0.867441159 0
#>  [24,] -0.338849889  0.370802634  0.140105007  0.714420496 0
#>  [25,]  0.185592100  0.249291812 -0.662133906  0.614427022 0
#>  [26,]  0.345862657  1.240415284 -0.534391766  0.305098743 0
#>  [27,] -0.793332374 -0.371594045 -1.412830881 -0.925999799 0
#>  [28,] -0.748631612  0.908843016  1.217203500  1.359534396 1
#>  [29,] -1.846666644  0.772073385  0.425531888  0.203698336 1
#>  [30,] -1.556824415  0.366647135 -0.120928613  1.288543508 0
#>  [31,]  0.002354941  0.234063054  0.651430723 -0.212971834 0
#>  [32,]  1.470206284 -0.824637494  0.490205565  0.487549814 0
#>  [33,] -0.777484495  0.793613760  0.705192456 -1.067709165 1
#>  [34,]  0.704613433  0.815375576  0.805650671 -1.180488666 0
#>  [35,]  0.696606398 -0.640412559 -0.179990583  2.159175760 0
#>  [36,]  0.931972929  0.154668564  1.072861877 -0.417812648 0
#>  [37,]  2.616178688 -1.319062104 -0.082523465  0.087654958 1
#>  [38,] -1.477492275 -0.559444239  1.072598758 -0.543709068 1
#>  [39,]  0.074677229 -1.278683713 -0.002079741 -0.119867022 1
#>  [40,] -0.640770175 -0.161854635 -0.859905321  1.800624858 0
#>  [41,]  0.114925465 -1.212698055  0.748074130 -0.794247350 1
#>  [42,]  0.747021300  1.371345622  1.356515448 -0.360360559 1
#>  [43,] -1.118479256  0.082627704  0.400001599  0.631460646 0
#>  [44,]  0.341004819 -1.265853773  0.636256611 -1.663279352 0
#>  [45,]  0.125031186  1.174275795 -1.866035907 -0.488005245 0
#>  [46,]  0.311369721  0.695991675 -0.580927235 -0.020752219 1
#>  [47,]  0.426443804  0.800070376  1.093896134  0.933871038 1
#>  [48,] -1.269579318 -1.002196891 -0.532691097 -0.232670878 0
#>  [49,]  0.869811052 -0.677115667  0.683984092 -1.019630064 0
#>  [50,] -0.023574630  0.720963599 -0.684263667 -1.680328321 0
#>  [51,]  0.646068260  0.960168144  1.625943271  2.281394917 0
#>  [52,]  1.232672999 -0.145665667 -0.631734454 -1.042285303 0
#>  [53,] -1.280410543  1.505255540 -1.353779986  0.712307192 0
#>  [54,] -0.176086170  0.678180406  0.387122228  1.220747981 0
#>  [55,]  1.148912321 -0.265552196 -0.967667515  0.336018683 0
#>  [56,]  0.106409425 -0.458132537 -0.358785321 -0.279147937 1
#>  [57,]  1.052599358  1.738431533 -0.586397434  0.076422131 0
#>  [58,]  0.063508668 -0.013105949  1.612913478 -0.354509030 1
#>  [59,]  0.760288668 -0.794353354  0.505671343  1.432658547 0
#>  [60,] -0.583391234  0.255333393  0.024106090  0.002671474 0
#>  [61,] -0.833176971 -0.878777648  0.295359647  0.344012511 0
#>  [62,] -0.955279258 -0.502384014  1.275350360 -0.830323380 1
#>  [63,]  0.468197456  0.918235108  0.204477180  0.482115591 0
#>  [64,]  1.489361598 -0.415556110  0.186747913 -0.688660369 1
#>  [65,]  0.845637803 -0.497454127  0.709749842 -1.093359720 0
#>  [66,] -0.090330359  0.122983974  0.375846729  0.841060978 0
#>  [67,] -0.760974092 -0.994093841  0.405413237  1.454136238 0
#>  [68,] -0.422703485  0.485937859  0.330985753  0.137240272 1
#>  [69,] -1.116990551 -0.551041758  1.593760511  0.325544925 1
#>  [70,] -1.373205437  0.166600822  0.555075757  1.223728019 0
#>  [71,] -0.464627202  0.095564919 -1.209478318 -1.053314136 0
#>  [72,] -0.751982369  0.937361737 -0.225557951  1.840332640 0
#>  [73,] -0.407974543 -0.778592244 -0.342381835 -2.231484275 0
#>  [74,]  0.177182569 -0.346014963 -1.736559496 -0.595844719 0
#>  [75,]  0.999326344  0.145927947  0.163560903 -0.230472754 0
#>  [76,]  0.116269256 -0.679378277  0.298293513  1.756315261 0
#>  [77,] -1.835065315 -0.613698253  0.175870029  0.475705566 1
#>  [78,] -0.298695682  2.202935147 -1.405601011  0.782644366 0
#>  [79,]  0.314652720 -0.328772531 -0.215255891 -1.146245649 0
#>  [80,] -1.082466160  1.751776850 -0.502680266  0.950180682 0
#>  [81,]  0.221793014 -0.971651737  1.683881702 -0.345227990 1
#>  [82,] -0.049851303 -1.689546894 -1.725036625 -2.473551488 0
#>  [83,]  1.107534869 -1.481386178 -0.252554758  1.376264164 0
#>  [84,] -0.993360619  2.426894783 -1.604868007  2.477145067 0
#>  [85,]  1.924843474 -0.228460020  1.153376302 -1.337234198 0
#>  [86,] -0.556762404  1.113283521 -1.266363058  2.024155203 0
#>  [87,]  0.329474662 -0.437063852 -0.250127880  0.859593181 0
#>  [88,]  1.758286351  1.264884610 -0.538988584  0.728471344 0
#>  [89,]  1.273909790  1.436597731 -0.611159385  1.702971138 0
#>  [90,] -0.583885195 -0.918664322  0.807837347 -1.349671966 0
#>  [91,] -0.327692378 -1.914602512  1.357967557  0.085960972 1
#>  [92,]  0.145241262  0.114969531  1.000928446  0.204207089 1
#>  [93,]  0.214101138  1.510348121 -1.258705305 -0.979050445 0
#>  [94,]  0.011554669 -0.009052613 -1.395429138  0.594650797 0
#>  [95,]  1.219699084 -0.135856744  0.326775093  1.039317096 0
#>  [96,] -1.030643641 -0.336250668 -0.566269784 -1.278834279 0
#>  [97,]  0.764734975  0.357796527 -0.040093207 -0.319978820 0
#>  [98,] -0.514982604  0.340498580  1.007425936 -1.872828606 0
#>  [99,]  1.740986972  3.097230704 -1.373520862  0.091322657 0
#> [100,]  1.005845241  0.677352741  0.330858231  0.712363770 1
#> 
#> $eta
#>   [1]  -4.230741658   0.249773820 -10.428018172 -33.536091547 -13.240929055
#>   [6]  -8.378433473  11.232630670   2.358996377 -16.267916223 -17.773156370
#>  [11]  -3.157716829  -4.671006832   0.172272398  -7.076605379 -22.734449788
#>  [16]  -6.582744658  -8.237941313 -16.883136099   5.962649842  -9.159460844
#>  [21] -11.248390309 -13.775433587 -13.512600179  -6.427688706  -4.697580369
#>  [26]  -3.177166684  -5.635321046  -0.349451253   6.365468174  -7.856527803
#>  [31]  -0.003235563  -2.113942902   2.852964993  -8.075470450 -29.861630772
#>  [36]  -1.571303810  -2.151949519   5.321340728   3.192466026 -19.217967987
#>  [41]   4.189152498   2.028409317  -8.480721476  -5.053020921 -13.663312462
#>  [46]   2.621974249   3.669599280 -13.412563576  -6.411394139 -17.938254787
#>  [51] -24.665426163  -6.371503386 -13.496860233  -7.842233674  -1.594613932
#>  [56]   2.945315800  -1.257477841   6.151055956 -16.527443088  -0.652837915
#>  [61]   0.149170927   7.572658043  -2.557589258   1.581318663  -3.802229699
#>  [66]  -9.321883880  -8.231024974   1.385161513   6.906280730 -12.649074898
#>  [71] -19.386201913 -17.294398998 -33.698627353  -5.638008824  -4.462793341
#>  [76] -12.022054702   4.122051658 -20.621514871  -1.412925220  -9.628274923
#>  [81]   9.134491663 -44.462637958  -5.648653505 -39.305661635  -4.368523670
#>  [86] -31.749152906   1.672345007 -13.962012959 -12.232276642  -7.245804485
#>  [91]  -1.802463802   4.350978302  -5.344072261 -14.673575579  -0.892513032
#>  [96] -10.577949185  -6.525565880 -12.248619792 -18.750228286   1.400046994
#> 
#> $mu
#>   [1] 1.433317e-02 5.621208e-01 2.959078e-05 2.220446e-16 1.776384e-06
#>   [6] 2.297168e-04 9.999868e-01 9.136467e-01 8.608624e-08 1.910804e-08
#>  [11] 4.078829e-02 9.275988e-03 5.429619e-01 8.439227e-04 1.338301e-10
#>  [16] 1.382132e-03 2.643582e-04 4.653151e-08 9.974335e-01 1.052085e-04
#>  [21] 1.302808e-05 1.040890e-06 1.353791e-06 1.613574e-03 9.034937e-03
#>  [26] 4.003408e-02 3.556835e-03 4.135155e-01 9.982830e-01 3.870662e-04
#>  [31] 4.991911e-01 1.077490e-01 9.454717e-01 3.109801e-04 1.074629e-13
#>  [36] 1.720306e-01 1.041492e-01 9.951376e-01 9.605498e-01 4.505495e-09
#>  [41] 9.850672e-01 8.837478e-01 2.073860e-04 6.349428e-03 1.164390e-06
#>  [46] 9.322625e-01 9.751467e-01 1.496225e-06 1.640039e-03 1.620000e-08
#>  [51] 1.940625e-11 1.706669e-03 1.375268e-06 3.926365e-04 1.687357e-01
#>  [56] 9.500416e-01 2.214084e-01 9.978733e-01 6.640834e-08 3.423503e-01
#>  [61] 5.372237e-01 9.994859e-01 7.191828e-02 8.293912e-01 2.183360e-02
#>  [66] 8.943725e-05 2.661924e-04 7.998187e-01 9.989995e-01 3.210518e-06
#>  [71] 3.807847e-09 3.084167e-08 2.220446e-16 3.547322e-03 1.139868e-02
#>  [76] 6.010151e-06 9.840474e-01 1.107108e-09 1.957731e-01 6.583620e-05
#>  [81] 9.998921e-01 2.220446e-16 3.509894e-03 2.220446e-16 1.251141e-02
#>  [86] 2.220446e-16 8.418882e-01 8.637229e-07 4.870658e-06 7.126520e-04
#>  [91] 1.415514e-01 9.872700e-01 4.753675e-03 4.239815e-07 2.905915e-01
#>  [96] 2.547088e-05 1.463345e-03 4.791703e-06 7.192491e-09 8.021913e-01
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())

#> $dat
#>                 x1           x2           x3          x4        y
#>   [1,] -0.80520647  0.486414289  0.532207247  0.11299534    15894
#>   [2,] -1.93205307 -0.857244463 -1.296930815 -0.36876161        0
#>   [3,]  0.71708201 -0.208889627 -0.116459881 -1.81381842        0
#>   [4,]  1.79248656 -0.021828351 -1.791435814 -0.72799437        0
#>   [5,]  1.12319487  0.371728862  1.037638450 -0.32334346     6339
#>   [6,]  1.60994768 -2.300614875 -0.751737315  0.11817806        0
#>   [7,]  1.73279950 -0.686391490  1.443320207  0.52272118        4
#>   [8,]  1.27570372  0.823954712  0.734189087 -0.39471204     1322
#>   [9,]  0.42400799  0.671560953 -1.054557986 -0.73826236        0
#>  [10,] -0.13777813 -0.136811085  0.030424727  1.31970010        0
#>  [11,] -0.59793732 -0.324855419 -0.620591656 -0.54771747        0
#>  [12,] -0.72207327  1.165364420 -0.193502015  2.11653074        0
#>  [13,]  0.07276248 -0.903741440  0.393606874  0.23888968        0
#>  [14,]  1.47336251  0.985612115 -0.033226158 -0.51596937        0
#>  [15,] -0.31620608 -0.126323527 -0.528722573  1.15875218        0
#>  [16,] -1.41472141 -1.443967173  3.071007147 -0.63435781 12207302
#>  [17,] -1.98617904  0.032154077 -1.424723774  0.86271083        0
#>  [18,]  0.70573403 -2.117716447  0.330701264 -1.68157303        0
#>  [19,] -0.79042763  1.665816928 -1.759479023 -1.63679212        0
#>  [20,]  0.75848016  0.575671055 -1.016384201  1.43029016        0
#>  [21,]  2.07837562 -0.002235961  1.036862040  1.53937650        0
#>  [22,]  1.91913490 -0.722014992  0.034712279 -0.07721086        0
#>  [23,]  0.07552982  0.448672361 -1.693691356  0.72223242        0
#>  [24,]  0.26775153  0.051736240 -0.374082127  0.76201239        0
#>  [25,]  0.43629074 -1.253649454 -1.258362453 -1.10521674        0
#>  [26,]  1.20974878  0.558374959 -0.608105630 -1.46116859        0
#>  [27,] -1.25305556  0.298702402 -1.368707861  0.29556190        0
#>  [28,]  1.76502863 -0.415528828 -0.942163446  0.65852651        0
#>  [29,] -0.15932203 -1.625374469  0.739857730  0.64514272        0
#>  [30,]  0.89887422  0.299647475  0.022094341  0.62365346        0
#>  [31,] -0.70499533 -0.146007948  0.548882732  2.31584930        0
#>  [32,]  1.12689657  0.330470436 -1.350698058 -0.38159888        0
#>  [33,] -1.97432221 -0.344358479 -0.644467633  0.03168816        0
#>  [34,] -1.24477283  0.891762344  1.047844734  0.77662540        0
#>  [35,]  1.03125039 -0.136294335 -0.321393364 -1.98304783        0
#>  [36,]  0.57909705  0.281191185  0.034325883 -0.00544671        0
#>  [37,] -1.03789001 -0.110812460  1.063948370  0.70241171        0
#>  [38,]  1.56162232 -0.036247432 -1.587608154  1.22316402        0
#>  [39,]  0.10195962  0.024977216 -1.583462404 -0.14139745        0
#>  [40,]  2.32423830  0.131053564  0.110531327 -0.49537901        0
#>  [41,]  2.32585155  2.595523922  0.246298837  0.84052263      122
#>  [42,] -1.29314996  0.419283992  0.057149917  0.56005014        0
#>  [43,] -0.04501885 -0.281698421 -1.083303495  1.13704594        0
#>  [44,]  0.19569423  0.544863114  0.216301111  0.47380131     1483
#>  [45,]  0.48062339  0.868230689 -0.823685525 -0.15044036        0
#>  [46,]  2.01734246 -0.206675268 -1.500780397  2.13899523        0
#>  [47,]  1.19560865  0.074947320  1.226567010  0.21799655      579
#>  [48,]  0.65657552 -2.250901330 -0.215867461 -1.89596918        0
#>  [49,]  1.02629785  1.500961160  1.733784762 -0.14273839   512792
#>  [50,]  0.52019266 -0.482768943 -1.432535684  1.46242573        0
#>  [51,]  1.12061565  0.605210639 -0.176478952  1.07465712        0
#>  [52,]  0.39989766  0.800685483  0.756314868  0.65505667      109
#>  [53,] -0.98452766  0.318362191 -0.980007712 -0.12313366        0
#>  [54,] -0.50256218  0.103472107 -0.386424379 -0.92972180        0
#>  [55,]  0.98714844 -0.657636600  0.381516254  1.51655653        0
#>  [56,]  2.19148101  0.033874701  1.345768539 -0.07679296     1557
#>  [57,] -0.16504221 -0.649756559  0.983748677 -0.36979194        2
#>  [58,] -0.68604080  0.911239650  0.086637298  0.60623267        2
#>  [59,]  0.94139941 -0.047272921  0.469669925  1.52673800        0
#>  [60,] -0.16404284 -1.178515013 -0.892900170 -0.08151085        0
#>  [61,] -1.30233941  2.242660967  0.002229617  1.42835885        0
#>  [62,] -0.72348400  1.467485607  0.150747679  1.01357184        9
#>  [63,]  1.39008874  0.552061735 -0.574554552 -0.67657783        2
#>  [64,]  0.68184063  0.359622515  0.235160435 -0.23656567        0
#>  [65,]  0.46826028  1.837521669  1.265858408  0.50559912      273
#>  [66,]  0.42092109  0.883273304 -1.390298039 -2.52775293        0
#>  [67,] -0.80033126 -1.713201462  0.115795724 -0.70066867        1
#>  [68,] -0.48845751  0.293728600  0.892577296  0.09160758        9
#>  [69,]  0.53900452  1.330807535  0.038752879  0.85217358        0
#>  [70,]  1.43517102  0.001025496 -0.038372698  1.12390932        0
#>  [71,] -0.26183872  0.944012124  0.811453346  0.82786930        0
#>  [72,] -1.41862372 -0.163110155  0.033486905 -1.36811281        0
#>  [73,] -0.51379272  0.020936334  0.663318866  0.18975968        0
#>  [74,]  0.77230122  0.551059233 -0.220091795  1.96075593        0
#>  [75,]  1.40368234  0.754814726 -0.015736545 -0.80601772        1
#>  [76,] -0.01580491 -1.017904646  2.087713501  0.24735027     1040
#>  [77,] -1.02328956 -0.320525550  1.496586647  0.50782823     6019
#>  [78,] -2.11639216 -0.457239570 -0.309764025 -0.49580175        0
#>  [79,]  0.15037144 -2.168675688  1.584290933 -0.90210237     2815
#>  [80,] -0.52087120  0.066409650 -1.147278380  1.01253254        0
#>  [81,] -0.90545484  1.513114967  0.937994138  1.63654057        0
#>  [82,] -0.74998690  0.242563643 -1.176473378 -0.82552051        0
#>  [83,] -0.78657602  0.196097401  1.402133012 -1.30128401        2
#>  [84,] -0.55968187 -0.656023781 -0.177175377 -0.48404261        1
#>  [85,] -1.39085131 -0.478355050 -0.650604263 -0.19156194        0
#>  [86,] -0.28263060  0.568878744 -1.229364408 -1.20399135        0
#>  [87,] -0.48028500 -2.102626810  0.251325895 -1.55456583        0
#>  [88,] -0.96762101 -0.003301496 -1.563215351 -0.14928970        0
#>  [89,]  1.77794098  0.416918015  0.813079991 -0.98824895        0
#>  [90,] -0.06908904  0.473867599 -0.110848678 -1.30053093        0
#>  [91,] -0.42618825 -0.062925119 -0.761172061  0.76400192        0
#>  [92,] -2.25140373 -1.636219349  0.829512582 -0.76422002        0
#>  [93,] -0.91417095  1.168229652 -1.055171964  1.05984222        0
#>  [94,] -0.80052458 -0.408686195  0.222967963  0.57440044     1041
#>  [95,] -0.75330590 -0.094117059  2.441195038  1.45435482      289
#>  [96,]  0.66064917  0.467589309  2.000641700  1.92660650        0
#>  [97,]  1.72769564 -0.828195377 -0.754805825  0.37769313        0
#>  [98,]  0.78267858  0.396813229 -1.303238961  0.45781732        0
#>  [99,] -0.78181841  0.244475509  0.246026158 -0.31447817        8
#> [100,] -0.12579755  0.627437586  0.548893707 -0.19633722        3
#> 
#> $eta
#>   [1]   9.66565596 -11.58011316 -22.58448155 -22.93656665   8.74532180
#>   [6]  -6.34692593   0.42223491   7.20236751  -4.90448909 -14.67167459
#>  [11]  -0.53697626 -23.56915813  -1.19465172  -5.52040740 -17.05366332
#>  [16]  16.31756716 -17.15977747 -19.76265129 -17.18575227 -18.82106391
#>  [21]  -8.24634617  -5.02815236  -9.26391420  -5.61782948 -11.21155154
#>  [26]  -7.01819721 -15.91922704 -10.30088230  -2.23496371  -7.02741455
#>  [31] -20.83226764  -6.73242125  -1.01682450  -4.74938380 -22.96646690
#>  [36]  -4.87549762  -0.22574008 -23.38090773 -11.87914699  -1.39056718
#>  [41]   4.86816417  -5.40566434 -13.28539108   7.29234200  -6.90998275
#>  [46] -31.93416125   6.33432871 -22.99417027  13.14510170 -17.72433622
#>  [51]  -0.94827163   4.55607303  -7.25218060 -12.01497346 -10.79292617
#>  [56]   7.37472458  -0.06768921  -0.17962246 -14.93165307 -10.97564797
#>  [61] -18.49720230   2.24688972  -0.11077443  -4.16666956   5.60795792
#>  [66] -40.73015895   0.42989526   2.55782269  -6.08070104  -9.58790325
#>  [71]  -7.67886407 -13.28802400  -0.99890398 -23.05840735  -0.18753738
#>  [76]   6.95180269   8.69860695  -4.03346943   7.94483302 -16.84228802
#>  [81]  -2.80850746  -9.04582713   0.33959604  -1.54182535  -5.01565301
#>  [86] -17.02847689 -14.34602992 -13.45031300  -4.34126969  -9.57530854
#>  [91] -13.72307519  -3.58987312 -14.27510987   6.92742181   5.72521743
#>  [96]  -9.23778271 -15.04006544 -13.62238811   2.34371479   1.86791520
#> 
#> $mu
#>   [1] 1.576671e+04 9.350197e-06 1.554835e-10 1.093391e-10 6.281234e+03
#>   [6] 1.752125e-03 1.525367e+00 1.342606e+03 7.413230e-03 4.247884e-07
#>  [11] 5.845130e-01 5.808242e-11 3.028094e-01 4.004216e-03 3.923631e-08
#>  [16] 1.220757e+07 3.528607e-08 2.613302e-09 3.438133e-08 6.700633e-09
#>  [21] 2.622149e-04 6.550903e-03 9.478360e-05 3.632517e-03 1.351715e-05
#>  [26] 8.954383e-04 1.220022e-07 3.360343e-05 1.069960e-01 8.872227e-04
#>  [31] 8.967288e-10 1.191644e-03 3.617418e-01 8.657028e-03 1.061183e-10
#>  [36] 7.631296e-03 7.979255e-01 7.011336e-11 6.933492e-06 2.489341e-01
#>  [41] 1.300819e+02 4.491070e-03 1.699135e-06 1.469007e+03 9.977750e-04
#>  [46] 1.352602e-14 5.635909e+02 1.032188e-10 5.114994e+05 2.006405e-08
#>  [51] 3.874100e-01 9.520886e+01 7.086275e-04 6.052898e-06 2.054432e-05
#>  [56] 1.595152e+03 9.345509e-01 8.355856e-01 3.275408e-07 1.711341e-05
#>  [61] 9.263329e-09 9.458272e+00 8.951406e-01 1.550381e-02 2.725870e+02
#>  [66] 2.220446e-16 1.537097e+00 1.290768e+01 2.286573e-03 6.855301e-05
#>  [71] 4.625000e-04 1.694667e-06 3.682829e-01 9.679678e-11 8.289981e-01
#>  [76] 1.045032e+03 5.994556e+03 1.771277e-02 2.820961e+03 4.847159e-08
#>  [81] 6.029492e-02 1.178819e-04 1.404380e+00 2.139901e-01 6.633299e-03
#>  [86] 4.023708e-08 5.882993e-07 1.440799e-06 1.301999e-02 6.942188e-05
#>  [91] 1.096842e-06 2.760183e-02 6.315366e-07 1.019861e+03 3.064999e+02
#>  [96] 9.729308e-05 2.938885e-07 1.213031e-06 1.041987e+01 6.474784e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )

#> $dat
#>                 x1          x2          x3          x4            y
#>   [1,]  0.41212408 -0.66755520  0.91915273 -0.95113841   0.65877891
#>   [2,]  0.59570991  1.00404451 -0.32515048 -0.36794402 -10.04920479
#>   [3,] -0.08125807  0.40857769 -0.97961270 -0.12367872  -7.35150382
#>   [4,]  0.88126900 -0.46764781 -0.95210323 -0.34417383  -6.41138886
#>   [5,]  0.94490127  1.22074106 -0.16323012  0.72827331  -9.80425596
#>   [6,]  2.37091477  0.33442452 -0.39997239  1.32197390 -11.17707732
#>   [7,] -0.30419244 -0.85803491  0.19201184  1.39144477 -19.23230338
#>   [8,]  0.45842063  0.08051104 -0.28115131  0.47441695  -8.75785564
#>   [9,] -0.43256653  0.81650754 -2.05983549 -0.40256406 -17.50806946
#>  [10,] -0.67300936 -0.24982678 -1.03487900  0.79179127  -4.16097816
#>  [11,] -0.68249662  0.71420802  2.99475552 -0.96306534  18.30678529
#>  [12,] -0.39675974  0.68988008  0.65410422 -0.77408936  -1.48249762
#>  [13,]  0.14676967  0.98939456 -0.31564225 -2.67452586 -39.00166581
#>  [14,]  0.66893248  1.07297249  1.00088061  0.16752571  -3.54312951
#>  [15,] -0.56528319 -1.26885749 -0.48727941 -0.49524216  -2.29197793
#>  [16,]  0.17007117  1.32201695 -0.01595245  0.49661664   0.32234789
#>  [17,] -0.29117703 -0.56387952  2.02241943  1.21732856   1.39123436
#>  [18,]  1.86553257  1.76166101 -0.82557185 -0.53641377  -9.60051947
#>  [19,] -0.50717361 -0.64997134 -0.96112862  0.51084872  -7.47527964
#>  [20,] -2.34053098  0.63878506  0.24881942 -0.17939574   0.02130833
#>  [21,]  0.56017442  0.14152028 -0.64709300 -0.42276263 -10.90800840
#>  [22,]  0.85967083 -0.07129532  0.41816741  0.13444519  -5.69063652
#>  [23,] -0.84764967  0.82937151  0.56381769  2.04440812 -14.87938103
#>  [24,]  2.02262308  0.18699052 -0.86967011 -0.75852344 -10.21545786
#>  [25,] -0.02478192  1.46749574 -0.56207730 -1.04710637  -7.60762876
#>  [26,]  0.04273221  0.90453664  1.23407764  0.29094018   4.44134796
#>  [27,]  0.56328774 -0.49551504  0.78348548 -1.52357936  -6.98628806
#>  [28,] -0.64165838  0.24849737 -0.28963790 -0.52028192  -1.24461517
#>  [29,] -0.24059417 -1.02757136  0.49528606  1.69667178 -21.96500531
#>  [30,]  0.01744152 -0.04396826 -0.53403248  2.04796436 -28.32743516
#>  [31,]  0.18482853 -0.92988060  1.20949738  1.98098662 -14.09403232
#>  [32,]  0.85458112 -0.07647743  0.37381260 -0.50981686  -6.37979554
#>  [33,] -2.14941880 -0.35995922  0.48360543 -1.05415158  -3.77310630
#>  [34,]  0.56724777  0.19604791 -2.04501924  0.99546597 -22.30844003
#>  [35,]  0.31102021  0.99691125  0.99245161 -0.62715361   4.79179362
#>  [36,]  0.27441875  0.36628567 -0.73042160  0.44564626  -0.84906355
#>  [37,]  0.11607404  0.69089766  0.63374342 -1.37384585  -0.74936925
#>  [38,] -0.77136270 -0.87613525  0.30130406  0.73360571  -2.26388680
#>  [39,]  1.72291963 -0.50910335  0.53540384  0.57286060  -1.67589539
#>  [40,]  2.08954243 -0.62946316 -0.50840230 -0.15387745  -1.61755084
#>  [41,]  2.52713527 -0.05134831  1.95536073  1.56696424  -5.60118433
#>  [42,]  0.54316887 -1.99258928 -0.35029394 -0.05814550  -9.56074006
#>  [43,]  1.44959051  0.48024268  1.43845252  1.49053004   0.38128795
#>  [44,] -0.99568925  0.18226741 -0.32589152  0.26757687  -6.65337106
#>  [45,]  0.01684000  1.46305754 -1.02519855 -0.31059035  -3.99151864
#>  [46,]  0.37528387 -1.50110155 -0.04390200 -0.15146275   5.72847276
#>  [47,]  1.08248947  1.11979077 -0.32434552  0.08898270  -6.12680660
#>  [48,]  1.18400856 -1.07878348  0.31934744 -1.45037365  -8.89476852
#>  [49,] -1.88052053  0.33874400  0.19391764  0.84298116  -0.28202002
#>  [50,]  1.72953938 -0.15415391  0.43564150  1.36025230 -13.17441804
#>  [51,] -2.84539729 -1.06370457  2.04509405 -1.48198211   0.77408874
#>  [52,]  1.03076352 -0.19415249 -0.99308722  1.30346859 -14.34652444
#>  [53,] -0.37251595 -0.49651922  0.72410853  0.33291375   3.45247549
#>  [54,]  0.93382913  1.65133257  0.20801872 -2.20710605 -23.32842253
#>  [55,] -1.19113241  1.04585205 -0.59416315  0.89534412 -18.19756073
#>  [56,] -1.72019223  0.57291928 -0.56590379 -0.57103034   3.32445452
#>  [57,] -0.96608851 -1.57845898 -2.32113855  0.80307389 -15.01168236
#>  [58,]  0.15613649 -0.69969700  1.68277464 -0.15117294  14.24924585
#>  [59,] -0.32166439 -1.04901833  1.96927191 -0.11658391   2.80516437
#>  [60,]  0.42452723  2.03110961  0.50153864 -0.31220221  -0.55469581
#>  [61,]  0.43928272  0.67371806 -0.06052325 -0.08643189   3.55340027
#>  [62,]  1.65183408  1.26665727  0.17757223 -0.31138401  -4.04071954
#>  [63,]  0.47469849 -0.01397260 -0.72381599 -0.59945530 -10.81170376
#>  [64,]  0.09513949 -0.93069782 -0.68976867  0.84327504 -10.44325422
#>  [65,]  1.25971301 -0.80748218 -1.60657709  0.91241433 -12.93870258
#>  [66,]  0.39730883 -0.55598739  2.09277438  0.09778511  17.75359417
#>  [67,] -0.30969780 -0.46384001 -0.95783909 -0.53678456  -9.82058784
#>  [68,]  0.82591536 -2.01952392  0.15861449  0.10178658  -8.02570026
#>  [69,]  0.30449601 -0.21489672  1.81618883  0.14769492  15.26429169
#>  [70,] -1.40896159 -1.32004024  0.42402218 -1.33520701  -8.28064104
#>  [71,] -0.32305666 -0.81192753  0.75353450  1.29192650 -12.39367510
#>  [72,]  1.10475613 -1.21443945  1.28912411  2.01723437 -10.11905965
#>  [73,] -0.77370097 -1.07744912  1.02151744 -0.36189542   7.24803239
#>  [74,]  0.89568062  1.01421809  0.34032502 -0.14310390  -7.56157321
#>  [75,]  0.98024952  0.53169112  0.10249559 -0.71130105   0.82628120
#>  [76,] -1.00618307  0.74188351 -0.92318861  0.23496616  -5.63607762
#>  [77,] -0.69546606 -0.16025589  1.84860068 -0.42953123  13.83109242
#>  [78,] -0.48620686 -0.14496339  0.17954935 -0.07216856  -1.59094252
#>  [79,] -0.35724726  1.84383570  0.80626799  0.01589134  -1.12492674
#>  [80,]  1.07821141 -1.26750352  2.18546792  0.65669226  13.83551373
#>  [81,] -1.15937534  0.03121845 -1.75716887  0.89479368 -23.28082864
#>  [82,]  0.24031660 -0.48682666  1.48203488  0.40974996  15.96496211
#>  [83,]  1.19133929 -0.26056529  0.03561794 -0.03701812   6.75336412
#>  [84,]  0.94689258  1.15150374 -0.22886523  0.18611168  -8.73758195
#>  [85,]  1.17958131 -0.71795193 -1.14973765 -2.12431160 -26.15656709
#>  [86,]  1.36533384 -0.80949978  0.26695992 -1.47167575 -12.40618347
#>  [87,]  0.54395009 -0.58757878  0.65822671 -0.40972493   1.10555414
#>  [88,]  0.72462563 -0.75841782  1.16813576 -0.23075934  -0.47930756
#>  [89,]  0.70793034  1.25606537 -1.03729483  0.13432327 -13.14362189
#>  [90,]  0.98330637 -0.45300889 -1.04967722 -0.50702660  -4.47090468
#>  [91,]  0.60664997 -0.64639512  0.66986085  2.02372629 -22.85395040
#>  [92,]  0.54674412 -0.28969276  1.40121241  1.12277235   2.32078701
#>  [93,]  0.66984579 -0.25195250  0.52342579  0.94612265  -5.56498469
#>  [94,] -0.44353073 -1.11170959 -0.31820181 -2.13546587 -29.58969313
#>  [95,] -0.43333304  0.24384304  1.12657695 -0.23576870   2.55825793
#>  [96,]  0.02739657  0.05719409 -0.36408292 -0.45991705  -8.19442356
#>  [97,] -0.20518852  0.39119382 -0.78366728  1.68058276 -22.74988944
#>  [98,] -0.35024803  1.18829834 -0.89026968  0.22731656 -13.14887119
#>  [99,] -0.11994916  0.41710581 -1.06471645  1.36108966 -17.49257226
#> [100,] -0.73970114  0.19514458  0.93839510  0.53433165   5.63535742
#> 
#> $eta
#>   [1]   2.1862101 -10.5145766  -7.9925282  -7.3732664  -9.3549933  -8.8909983
#>   [7] -18.4380570  -7.0207504 -16.3968545  -4.5037673  19.0402369  -0.5149129
#>  [13] -38.5495375  -3.9968750  -1.3368929   1.3638097   0.8736397  -9.9760083
#>  [19]  -7.8337332  -0.6542844 -11.8227702  -5.2738822 -14.9218971 -12.0095630
#>  [25]  -8.0815868   4.9925876  -6.7788449  -2.6564860 -21.0402626 -27.9822928
#>  [31] -13.2708635  -6.7418736  -4.4740672 -23.7511022   3.6526354  -1.0198475
#>  [37]  -0.7422099  -1.2741698  -1.4612587  -0.9837242  -5.6142568  -8.5540774
#>  [43]   0.8254974  -6.8737966  -4.6304198   5.1480787  -4.8816528  -7.7034936
#>  [49]  -0.2485314 -13.5044850   1.4319827 -14.3990258   2.2516390 -23.0276101
#>  [55] -17.6455749   2.3832635 -15.4093415  11.9061414   2.9822397  -0.5108919
#>  [61]   3.2814784  -6.5385354 -10.1767703 -10.3718023 -13.3456951  16.6503203
#>  [67]  -9.4880108  -8.2913104  15.1010755  -7.3459653 -12.9126054  -9.9607332
#>  [73]   6.7128983  -6.2754776  -0.2133551  -5.4152262  13.1231204  -1.5186599
#>  [79]  -0.9286354  14.1073587 -24.0741520  15.2003430   5.5173982  -8.2824388
#>  [85] -27.6202730 -10.0509953   2.2360709  -1.7946209 -14.4228837  -5.6904774
#>  [91] -20.4619200   1.8814871  -5.2537618 -29.3897043   0.8958041  -7.8461781
#>  [97] -23.6167211 -14.2429274 -18.4820935   4.7013548
#> 
#> $mu
#>   [1]   2.1862101 -10.5145766  -7.9925282  -7.3732664  -9.3549933  -8.8909983
#>   [7] -18.4380570  -7.0207504 -16.3968545  -4.5037673  19.0402369  -0.5149129
#>  [13] -38.5495375  -3.9968750  -1.3368929   1.3638097   0.8736397  -9.9760083
#>  [19]  -7.8337332  -0.6542844 -11.8227702  -5.2738822 -14.9218971 -12.0095630
#>  [25]  -8.0815868   4.9925876  -6.7788449  -2.6564860 -21.0402626 -27.9822928
#>  [31] -13.2708635  -6.7418736  -4.4740672 -23.7511022   3.6526354  -1.0198475
#>  [37]  -0.7422099  -1.2741698  -1.4612587  -0.9837242  -5.6142568  -8.5540774
#>  [43]   0.8254974  -6.8737966  -4.6304198   5.1480787  -4.8816528  -7.7034936
#>  [49]  -0.2485314 -13.5044850   1.4319827 -14.3990258   2.2516390 -23.0276101
#>  [55] -17.6455749   2.3832635 -15.4093415  11.9061414   2.9822397  -0.5108919
#>  [61]   3.2814784  -6.5385354 -10.1767703 -10.3718023 -13.3456951  16.6503203
#>  [67]  -9.4880108  -8.2913104  15.1010755  -7.3459653 -12.9126054  -9.9607332
#>  [73]   6.7128983  -6.2754776  -0.2133551  -5.4152262  13.1231204  -1.5186599
#>  [79]  -0.9286354  14.1073587 -24.0741520  15.2003430   5.5173982  -8.2824388
#>  [85] -27.6202730 -10.0509953   2.2360709  -1.7946209 -14.4228837  -5.6904774
#>  [91] -20.4619200   1.8814871  -5.2537618 -29.3897043   0.8958041  -7.8461781
#>  [97] -23.6167211 -14.2429274 -18.4820935   4.7013548
#>