Simulate Spline Data

Simulation functions for binomial, gaussian and poisson outcomes. The simulation equations follow Bai

sim_Bai(n, p, family = binomial(), dispersion = 1)

Arguments

n
Integer, sample size

p
Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message

family
Family object, one of binomial (default), gaussian , poisson. Please see family for more detail

dispersion
Double, required for gaussian distribution, match to sd parameter.

Value

a list contains dat, eta, mu

dat a numeric matrix, contains predictors x1 to xp where p is replaced by the integer value. The outcome variable is named y

etathe linear predictor of the simulated data

muthe mean of the simulated data

Details

Examples

# Binomial Outcome
sim_Bai(100, 4)

#> $dat
#>                 x1           x2           x3           x4 y
#>   [1,] -1.10735295  0.385994964  1.143623543 -0.110119494 1
#>   [2,] -0.62788279 -0.838842632 -1.381713624 -1.051575816 0
#>   [3,] -1.47950195 -0.530027590 -0.844069552 -0.891401342 0
#>   [4,]  0.15434317 -2.293182084 -1.836583479 -0.619138043 0
#>   [5,] -2.29948441  1.855263642  1.044170831  0.255199046 0
#>   [6,] -0.43872768 -0.624714964  0.215210107 -0.439315917 0
#>   [7,] -0.64968061  0.175872582  0.800349614  1.468149212 0
#>   [8,]  0.42404095 -1.365788853  0.714933031  0.832506126 1
#>   [9,]  0.04704973 -1.345518474 -0.317171209 -0.913787407 1
#>  [10,] -1.28504407  0.947265271 -2.252045124 -0.272595426 0
#>  [11,]  0.04553236 -0.115739671  1.505248197  0.751086555 1
#>  [12,] -1.44942970 -0.429404494 -0.419147872 -0.076856255 0
#>  [13,] -1.63370440  0.079631976 -0.208179987  0.147101528 0
#>  [14,] -0.51728496  1.006463994  1.025008622 -0.321238201 0
#>  [15,] -0.99115143  1.754518747 -2.109352248 -1.703140073 0
#>  [16,] -0.72848504 -0.119361311  0.581155983 -0.761038444 1
#>  [17,]  0.45195163  1.643998625  0.427476044 -0.759025393 1
#>  [18,]  0.88718133 -1.310997765  0.309712183  0.292608648 0
#>  [19,]  1.26457340  0.362701424 -0.348680801  0.332236386 1
#>  [20,] -1.29141939  1.369665491  0.472684011 -0.231934768 0
#>  [21,]  0.18251031 -0.348512400  0.702742649 -0.815807168 1
#>  [22,] -0.08980967 -0.545470652 -1.543797498  0.216074797 0
#>  [23,] -0.11262928 -0.020519678 -1.139419424 -1.138865785 0
#>  [24,]  0.16280134 -0.338849889  0.370802634  0.140105007 1
#>  [25,]  0.55586101  0.185592100  0.249291812 -0.662133906 0
#>  [26,]  0.85339468  0.345862657  1.240415284 -0.534391766 1
#>  [27,] -0.19489936 -0.793332374 -0.371594045 -1.412830881 0
#>  [28,] -0.04341191 -0.748631612  0.908843016  1.217203500 0
#>  [29,] -0.52943832 -1.846666644  0.772073385  0.425531888 1
#>  [30,] -0.59851060 -1.556824415  0.366647135 -0.120928613 1
#>  [31,]  0.14929768  0.002354941  0.234063054  0.651430723 0
#>  [32,]  1.25582092  1.470206284 -0.824637494  0.490205565 0
#>  [33,] -0.23436953 -0.777484495  0.793613760  0.705192456 0
#>  [34,]  1.42234028  0.704613433  0.815375576  0.805650671 1
#>  [35,]  0.29212754  0.696606398 -0.640412559 -0.179990583 1
#>  [36,] -0.73688675  0.931972929  0.154668564  1.072861877 0
#>  [37,] -0.12999655  2.616178688 -1.319062104 -0.082523465 0
#>  [38,]  2.20088565 -1.477492275 -0.559444239  1.072598758 0
#>  [39,]  1.61308129  0.074677229 -1.278683713 -0.002079741 0
#>  [40,] -0.26749958 -0.640770175 -0.161854635 -0.859905321 0
#>  [41,]  0.12760317  0.114925465 -1.212698055  0.748074130 0
#>  [42,] -0.37089274  0.747021300  1.371345622  1.356515448 0
#>  [43,] -0.45240228 -1.118479256  0.082627704  0.400001599 0
#>  [44,]  1.19747825  0.341004819 -1.265853773  0.636256611 0
#>  [45,]  0.96133970  0.125031186  1.174275795 -1.866035907 0
#>  [46,]  1.09924455  0.311369721  0.695991675 -0.580927235 1
#>  [47,]  0.05353454  0.426443804  0.800070376  1.093896134 0
#>  [48,] -0.07108256 -1.269579318 -1.002196891 -0.532691097 0
#>  [49,] -0.26521262  0.869811052 -0.677115667  0.683984092 0
#>  [50,] -1.08747279 -0.023574630  0.720963599 -0.684263667 0
#>  [51,]  0.11299534  0.646068260  0.960168144  1.625943271 0
#>  [52,] -0.36876161  1.232672999 -0.145665667 -0.631734454 0
#>  [53,] -1.81381842 -1.280410543  1.505255540 -1.353779986 1
#>  [54,] -0.72799437 -0.176086170  0.678180406  0.387122228 1
#>  [55,] -0.32334346  1.148912321 -0.265552196 -0.967667515 0
#>  [56,]  0.11817806  0.106409425 -0.458132537 -0.358785321 0
#>  [57,]  0.52272118  1.052599358  1.738431533 -0.586397434 1
#>  [58,] -0.39471204  0.063508668 -0.013105949  1.612913478 0
#>  [59,] -0.73826236  0.760288668 -0.794353354  0.505671343 1
#>  [60,]  1.31970010 -0.583391234  0.255333393  0.024106090 1
#>  [61,] -0.54771747 -0.833176971 -0.878777648  0.295359647 0
#>  [62,]  2.11653074 -0.955279258 -0.502384014  1.275350360 0
#>  [63,]  0.23888968  0.468197456  0.918235108  0.204477180 1
#>  [64,] -0.51596937  1.489361598 -0.415556110  0.186747913 0
#>  [65,]  1.15875218  0.845637803 -0.497454127  0.709749842 0
#>  [66,] -0.63435781 -0.090330359  0.122983974  0.375846729 0
#>  [67,]  0.86271083 -0.760974092 -0.994093841  0.405413237 0
#>  [68,] -1.68157303 -0.422703485  0.485937859  0.330985753 1
#>  [69,] -1.63679212 -1.116990551 -0.551041758  1.593760511 0
#>  [70,]  1.43029016 -1.373205437  0.166600822  0.555075757 1
#>  [71,]  1.53937650 -0.464627202  0.095564919 -1.209478318 0
#>  [72,] -0.07721086 -0.751982369  0.937361737 -0.225557951 0
#>  [73,]  0.72223242 -0.407974543 -0.778592244 -0.342381835 0
#>  [74,]  0.76201239  0.177182569 -0.346014963 -1.736559496 0
#>  [75,] -1.10521674  0.999326344  0.145927947  0.163560903 0
#>  [76,] -1.46116859  0.116269256 -0.679378277  0.298293513 0
#>  [77,]  0.29556190 -1.835065315 -0.613698253  0.175870029 0
#>  [78,]  0.65852651 -0.298695682  2.202935147 -1.405601011 0
#>  [79,]  0.64514272  0.314652720 -0.328772531 -0.215255891 0
#>  [80,]  0.62365346 -1.082466160  1.751776850 -0.502680266 1
#>  [81,]  2.31584930  0.221793014 -0.971651737  1.683881702 0
#>  [82,] -0.38159888 -0.049851303 -1.689546894 -1.725036625 0
#>  [83,]  0.03168816  1.107534869 -1.481386178 -0.252554758 0
#>  [84,]  0.77662540 -0.993360619  2.426894783 -1.604868007 0
#>  [85,] -1.98304783  1.924843474 -0.228460020  1.153376302 0
#>  [86,] -0.00544671 -0.556762404  1.113283521 -1.266363058 1
#>  [87,]  0.70241171  0.329474662 -0.437063852 -0.250127880 0
#>  [88,]  1.22316402  1.758286351  1.264884610 -0.538988584 1
#>  [89,] -0.14139745  1.273909790  1.436597731 -0.611159385 0
#>  [90,] -0.49537901 -0.583885195 -0.918664322  0.807837347 0
#>  [91,]  0.84052263 -0.327692378 -1.914602512  1.357967557 0
#>  [92,]  0.56005014  0.145241262  0.114969531  1.000928446 0
#>  [93,]  1.13704594  0.214101138  1.510348121 -1.258705305 1
#>  [94,]  0.47380131  0.011554669 -0.009052613 -1.395429138 0
#>  [95,] -0.15044036  1.219699084 -0.135856744  0.326775093 0
#>  [96,]  2.13899523 -1.030643641 -0.336250668 -0.566269784 0
#>  [97,]  0.21799655  0.764734975  0.357796527 -0.040093207 1
#>  [98,] -1.89596918 -0.514982604  0.340498580  1.007425936 1
#>  [99,] -0.14273839  1.740986972  3.097230704 -1.373520862 1
#> [100,]  1.46242573  1.005845241  0.677352741  0.330858231 0
#> 
#> $eta
#>   [1]  3.566920e+00 -1.520628e+01 -8.091215e+00 -7.524059e+00 -9.636358e-01
#>   [6] -1.918429e+00 -6.725951e+00  5.387654e+00 -2.556269e+00 -2.295863e+01
#>  [11]  4.774393e+00 -1.612904e+00 -3.133760e+00  1.090738e+00 -2.656442e+01
#>  [16]  2.784016e+00  3.347689e+00 -2.334241e-01  1.661562e+00 -2.767596e+00
#>  [21]  7.544507e+00 -1.084274e+01 -2.130655e+01  8.380823e+00 -7.060397e+00
#>  [26]  9.417124e-01 -2.120439e+01 -6.689291e+00 -4.314916e-01  6.146424e+00
#>  [31] -1.723675e+00  4.394264e-01 -6.693076e+00  5.318759e+00  2.287883e+00
#>  [36] -3.743693e+00 -9.198501e+00 -2.099614e+00 -1.643133e+01 -1.039718e+01
#>  [41] -1.188313e+01 -4.115123e+00 -4.559800e+00 -6.102076e+00 -1.690541e+01
#>  [46]  3.451647e+00  1.254653e+00 -8.757637e+00 -1.387949e+01 -5.315938e+00
#>  [51] -2.042414e+00 -1.032856e+01  5.520785e+00  6.914360e+00 -1.588003e+01
#>  [56] -5.454931e+00  2.557224e+00 -2.163739e+01 -8.713020e-01  6.637175e+00
#>  [61] -9.146020e+00 -1.320887e+01  1.184801e+01 -2.924578e-01 -3.205899e+00
#>  [66]  3.397386e-01 -1.423909e+01  9.412414e+00 -1.503687e+01  3.904961e+00
#>  [71] -5.618563e+00 -4.231045e-01 -7.694003e+00 -2.691113e+01 -7.330288e+00
#>  [76] -1.112319e+01 -3.978717e+00  5.283787e-01 -8.032383e+00  2.141964e+00
#>  [81] -1.943563e+01 -3.265062e+01 -1.365673e+01 -8.334087e+00 -1.124444e+01
#>  [86] -3.898617e-01 -9.215936e+00  1.129828e+01  8.773994e-04 -8.350817e+00
#>  [91] -2.308076e+01 -1.132592e+01  7.789280e-01 -1.411127e+01 -9.449591e+00
#>  [96] -4.366950e+00  7.122953e+00  1.823401e+00  5.857070e+00 -3.921022e-01
#> 
#> $mu
#>   [1] 9.725330e-01 2.488848e-07 3.061237e-04 5.396450e-04 2.761508e-01
#>   [6] 1.280368e-01 1.197942e-03 9.954481e-01 7.200647e-02 1.069529e-10
#>  [11] 9.916275e-01 1.661858e-01 4.173599e-02 7.485206e-01 2.905475e-12
#>  [16] 9.418059e-01 9.660291e-01 4.419075e-01 8.404476e-01 5.910053e-02
#>  [21] 9.994713e-01 1.954565e-05 5.580612e-10 9.997708e-01 8.577009e-04
#>  [26] 7.194454e-01 6.180877e-10 1.242619e-03 3.937702e-01 9.978634e-01
#>  [31] 1.513984e-01 6.081223e-01 1.237930e-03 9.951251e-01 9.078685e-01
#>  [36] 2.311939e-02 1.011807e-04 1.091343e-01 7.310764e-08 3.051753e-05
#>  [41] 6.905908e-06 1.606174e-02 1.035579e-02 2.233218e-03 4.550637e-08
#>  [46] 9.692802e-01 7.781043e-01 1.572311e-04 9.380267e-07 4.888654e-03
#>  [51] 1.148212e-01 3.268503e-05 9.960133e-01 9.990076e-01 1.268796e-07
#>  [56] 4.256970e-03 9.280573e-01 4.008661e-10 2.949835e-01 9.986910e-01
#>  [61] 1.066321e-04 1.834257e-06 9.999928e-01 4.274023e-01 3.894432e-02
#>  [66] 5.841270e-01 6.547013e-07 9.999183e-01 2.948285e-07 9.802559e-01
#>  [71] 3.616724e-03 3.957741e-01 4.553436e-04 2.054212e-12 6.549558e-04
#>  [76] 1.476563e-05 1.836601e-02 6.291049e-01 3.246680e-04 8.949154e-01
#>  [81] 3.624197e-09 2.220446e-16 1.172076e-06 2.401307e-04 1.307961e-05
#>  [86] 4.037506e-01 9.943216e-05 9.999876e-01 5.002193e-01 2.361477e-04
#>  [91] 9.465679e-11 1.205619e-05 6.854490e-01 7.439656e-07 7.871552e-05
#>  [96] 1.253087e-02 9.991943e-01 8.609737e-01 9.971485e-01 4.032113e-01
#> 

# Logistic Outcome
sim_Bai(100, 4, family = poisson())

#> $dat
#>                  x1          x2           x3           x4   y
#>   [1,]  1.015236070 -0.80520647  0.486414289  0.532207247   2
#>   [2,]  0.294676272 -1.93205307 -0.857244463 -1.296930815   0
#>   [3,] -0.296458790  0.71708201 -0.208889627 -0.116459881   5
#>   [4,] -0.822289822  1.79248656 -0.021828351 -1.791435814   0
#>   [5,]  1.691546077  1.12319487  0.371728862  1.037638450   0
#>   [6,] -1.476545705  1.60994768 -2.300614875 -0.751737315   2
#>   [7,]  0.929693563  1.73279950 -0.686391490  1.443320207   0
#>   [8,] -0.613832093  1.27570372  0.823954712  0.734189087   6
#>   [9,]  0.617598091  0.42400799  0.671560953 -1.054557986   0
#>  [10,] -0.830615177 -0.13777813 -0.136811085  0.030424727   7
#>  [11,] -1.133375927 -0.59793732 -0.324855419 -0.620591656   0
#>  [12,] -0.156378319 -0.72207327  1.165364420 -0.193502015   6
#>  [13,] -0.243091747  0.07276248 -0.903741440  0.393606874   0
#>  [14,] -1.129264413  1.47336251  0.985612115 -0.033226158   9
#>  [15,] -0.062192485 -0.31620608 -0.126323527 -0.528722573   7
#>  [16,]  0.487082631 -1.41472141 -1.443967173  3.071007147   0
#>  [17,] -0.054636953 -1.98617904  0.032154077 -1.424723774   0
#>  [18,] -0.198624209  0.70573403 -2.117716447  0.330701264   1
#>  [19,] -1.453375000 -0.79042763  1.665816928 -1.759479023   0
#>  [20,]  0.198434624  0.75848016  0.575671055 -1.016384201   3
#>  [21,] -1.391001690  2.07837562 -0.002235961  1.036862040   0
#>  [22,] -2.244235104  1.91913490 -0.722014992  0.034712279   0
#>  [23,] -0.231792440  0.07552982  0.448672361 -1.693691356   0
#>  [24,] -0.686551360  0.26775153  0.051736240 -0.374082127   7
#>  [25,] -0.482915280  0.43629074 -1.253649454 -1.258362453   0
#>  [26,] -2.243508266  1.20974878  0.558374959 -0.608105630   1
#>  [27,] -0.381078571 -1.25305556  0.298702402 -1.368707861   0
#>  [28,]  0.116047733  1.76502863 -0.415528828 -0.942163446   0
#>  [29,]  0.892492885 -0.15932203 -1.625374469  0.739857730   0
#>  [30,]  1.808520401  0.89887422  0.299647475  0.022094341   3
#>  [31,]  1.079663512 -0.70499533 -0.146007948  0.548882732   1
#>  [32,]  1.452055280  1.12689657  0.330470436 -1.350698058   0
#>  [33,]  2.556923589 -1.97432221 -0.344358479 -0.644467633   1
#>  [34,]  1.072728070 -1.24477283  0.891762344  1.047844734   0
#>  [35,] -1.178914989  1.03125039 -0.136294335 -0.321393364   0
#>  [36,]  1.075417132  0.57909705  0.281191185  0.034325883  23
#>  [37,]  0.053855088 -1.03789001 -0.110812460  1.063948370   0
#>  [38,] -0.234343833  1.56162232 -0.036247432 -1.587608154   0
#>  [39,]  0.950759734  0.10195962  0.024977216 -1.583462404   0
#>  [40,]  0.073021934  2.32423830  0.131053564  0.110531327   3
#>  [41,]  0.663527998  2.32585155  2.595523922  0.246298837 210
#>  [42,] -0.816198742 -1.29314996  0.419283992  0.057149917  31
#>  [43,]  0.295544263 -0.04501885 -0.281698421 -1.083303495   0
#>  [44,]  1.346357662  0.19569423  0.544863114  0.216301111   5
#>  [45,] -0.280015436  0.48062339  0.868230689 -0.823685525   1
#>  [46,] -0.063043628  2.01734246 -0.206675268 -1.500780397   0
#>  [47,] -0.132833831  1.19560865  0.074947320  1.226567010   0
#>  [48,]  0.629361770  0.65657552 -2.250901330 -0.215867461   2
#>  [49,] -0.641231532  1.02629785  1.500961160  1.733784762   0
#>  [50,] -0.104018599  0.52019266 -0.482768943 -1.432535684   0
#>  [51,] -1.388668827  1.12061565  0.605210639 -0.176478952   1
#>  [52,]  0.437214206  0.39989766  0.800685483  0.756314868   7
#>  [53,]  0.315862087 -0.98452766  0.318362191 -0.980007712   1
#>  [54,]  0.194610535 -0.50256218  0.103472107 -0.386424379  19
#>  [55,] -0.455762127  0.98714844 -0.657636600  0.381516254   0
#>  [56,]  0.812534989  2.19148101  0.033874701  1.345768539   0
#>  [57,]  0.275042593 -0.16504221 -0.649756559  0.983748677   1
#>  [58,]  0.006009411 -0.68604080  0.911239650  0.086637298   8
#>  [59,]  2.010186412  0.94139941 -0.047272921  0.469669925   1
#>  [60,]  0.313808823 -0.16404284 -1.178515013 -0.892900170   0
#>  [61,] -0.846162712 -1.30233941  2.242660967  0.002229617 528
#>  [62,] -0.134641045 -0.72348400  1.467485607  0.150747679   7
#>  [63,]  1.470832200  1.39008874  0.552061735 -0.574554552  10
#>  [64,] -1.475125736  0.68184063  0.359622515  0.235160435  10
#>  [65,]  0.204499683  0.46826028  1.837521669  1.265858408   6
#>  [66,] -0.341768421  0.42092109  0.883273304 -1.390298039   0
#>  [67,]  1.842157648 -0.80033126 -1.713201462  0.115795724   0
#>  [68,] -0.205909535 -0.48845751  0.293728600  0.892577296   1
#>  [69,]  1.502231278  0.53900452  1.330807535  0.038752879  27
#>  [70,]  0.242188060  1.43517102  0.001025496 -0.038372698  19
#>  [71,]  0.053455720 -0.26183872  0.944012124  0.811453346   6
#>  [72,] -0.125146310 -1.41862372 -0.163110155  0.033486905   5
#>  [73,]  0.247787238 -0.51379272  0.020936334  0.663318866   6
#>  [74,] -0.626612578  0.77230122  0.551059233 -0.220091795  16
#>  [75,]  0.682877534  1.40368234  0.754814726 -0.015736545   6
#>  [76,]  0.589606056 -0.01580491 -1.017904646  2.087713501   0
#>  [77,] -0.814967320 -1.02328956 -0.320525550  1.496586647   0
#>  [78,] -0.345798987 -2.11639216 -0.457239570 -0.309764025   1
#>  [79,]  0.056529047  0.15037144 -2.168675688  1.584290933   0
#>  [80,] -0.566766181 -0.52087120  0.066409650 -1.147278380   1
#>  [81,] -0.042837457 -0.90545484  1.513114967  0.937994138   1
#>  [82,] -1.281559297 -0.74998690  0.242563643 -1.176473378   1
#>  [83,]  0.967582242 -0.78657602  0.196097401  1.402133012   0
#>  [84,]  1.029048853 -0.55968187 -0.656023781 -0.177175377   4
#>  [85,] -2.166425254 -1.39085131 -0.478355050 -0.650604263   0
#>  [86,] -0.303348223 -0.28263060  0.568878744 -1.229364408   0
#>  [87,]  0.179268039 -0.48028500 -2.102626810  0.251325895  14
#>  [88,]  1.427395820 -0.96762101 -0.003301496 -1.563215351   0
#>  [89,] -0.619946595  1.77794098  0.416918015  0.813079991   1
#>  [90,] -0.131842450 -0.06908904  0.473867599 -0.110848678   1
#>  [91,]  0.267551749 -0.42618825 -0.062925119 -0.761172061   9
#>  [92,] -1.456778518 -2.25140373 -1.636219349  0.829512582   0
#>  [93,]  0.234340520 -0.91417095  1.168229652 -1.055171964   2
#>  [94,] -0.951528054 -0.80052458 -0.408686195  0.222967963   5
#>  [95,]  1.898439401 -0.75330590 -0.094117059  2.441195038   0
#>  [96,] -1.042770546  0.66064917  0.467589309  2.000641700   0
#>  [97,] -1.333746286  1.72769564 -0.828195377 -0.754805825   1
#>  [98,]  1.764724810  0.78267858  0.396813229 -1.303238961   0
#>  [99,]  0.506063430 -0.78181841  0.244475509  0.246026158   1
#> [100,] -0.655292217 -0.12579755  0.627437586  0.548893707   1
#> 
#> $eta
#>   [1]   0.430472630  -2.230799084   0.989198278  -3.897568277  -2.469280911
#>   [6]   0.744874774  -2.821223667   1.157449705  -0.555616910   2.017765053
#>  [11]   0.215316537   1.057168741  -1.145973690   2.129228872   1.205702003
#>  [16] -16.687342282  -3.899052327   0.582995709  -4.048887289   1.025234484
#>  [21]  -2.383592588  -0.340921818  -6.103828156   1.684179276  -1.538530891
#>  [26]  -0.723383277  -2.390885814   0.549432834  -0.711011255   0.183841324
#>  [31]   1.007492777  -2.751538133  -0.840491889   0.671387896  -0.713708528
#>  [36]   2.983845843  -1.313737413  -3.656761080  -4.899190959   1.840918286
#>  [41]   5.360711210   3.213492264  -0.794856423   1.672979977   0.426754060
#>  [46]  -4.489722441  -3.064654095   1.107943541  -3.386684513  -2.534318637
#>  [51]   0.059579008   1.558743012  -0.375263739   2.954758437  -0.194321150
#>  [56]  -3.895033851   0.227193586   1.919723817   0.358108024   0.409180015
#>  [61]   6.327446571   1.614004851   1.587858202   2.107437686   2.032209456
#>  [66]  -2.193722160  -0.520264604  -0.108678038   3.464298142   3.012110661
#>  [71]   1.584107366   1.629735909   2.361175391   2.663549397   1.386986869
#>  [76]  -8.932017794  -3.069697927  -0.095943910  -4.529154238   0.009634946
#>  [81]  -0.204823309  -2.718131895  -3.314308029   1.967292422   0.518871543
#>  [86]  -1.612452903   2.755478221  -4.004424328   1.287795170   0.286515683
#>  [91]   2.210108796  -0.079392224   0.055676530   0.736843995 -11.648035253
#>  [96]  -5.759147180  -0.232144387  -2.507327183   0.707477156   1.697937231
#> 
#> $mu
#>   [1] 1.537984e+00 1.074425e-01 2.689078e+00 2.029119e-02 8.464570e-02
#>   [6] 2.106178e+00 5.953305e-02 3.181808e+00 5.737182e-01 7.521496e+00
#>  [11] 1.240254e+00 2.878210e+00 3.179142e-01 8.408380e+00 3.339102e+00
#>  [16] 5.659516e-08 2.026110e-02 1.791397e+00 1.744177e-02 2.787749e+00
#>  [21] 9.221868e-02 7.111145e-01 2.234298e-03 5.388027e+00 2.146963e-01
#>  [26] 4.851082e-01 9.154855e-02 1.732270e+00 4.911473e-01 1.201825e+00
#>  [31] 2.738726e+00 6.382961e-02 4.314982e-01 1.956951e+00 4.898243e-01
#>  [36] 1.976368e+01 2.688135e-01 2.581599e-02 7.452610e-03 6.302323e+00
#>  [41] 2.128763e+02 2.486577e+01 4.516461e-01 5.328022e+00 1.532276e+00
#>  [46] 1.122376e-02 4.666998e-02 3.028125e+00 3.382062e-02 7.931574e-02
#>  [51] 1.061390e+00 4.752843e+00 6.871080e-01 1.919708e+01 8.233934e-01
#>  [56] 2.034269e-02 1.255073e+00 6.819075e+00 1.430620e+00 1.505583e+00
#>  [61] 5.597255e+02 5.022887e+00 4.893257e+00 8.227134e+00 7.630928e+00
#>  [66] 1.115010e-01 5.943633e-01 8.970192e-01 3.195402e+01 2.033026e+01
#>  [71] 4.874938e+00 5.102527e+00 1.060341e+01 1.434712e+01 4.002771e+00
#>  [76] 1.320912e-04 4.643518e-02 9.085150e-01 1.078980e-02 1.009682e+00
#>  [81] 8.147913e-01 6.599793e-02 3.635920e-02 7.151288e+00 1.680131e+00
#>  [86] 1.993979e-01 1.572856e+01 1.823478e-02 3.624786e+00 1.331779e+00
#>  [91] 9.116708e+00 9.236776e-01 1.057256e+00 2.089331e+00 8.736200e-06
#>  [96] 3.153800e-03 7.928316e-01 8.148574e-02 2.028866e+00 5.462668e+00
#> 

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian(), dispersion = 1 )

#> $dat
#>                  x1          x2          x3          x4            y
#>   [1,] -0.243414776  0.81251610  1.32380523  0.50592356   0.33700072
#>   [2,]  0.708157194  0.25661233  1.03062147  0.41527205  -1.25347633
#>   [3,] -0.480039667 -2.14876584  0.31737389 -1.17741390  -7.14397437
#>   [4,] -0.316551268  0.69010855 -1.11190379  0.03344533  -9.34278674
#>   [5,]  0.510471244 -1.77244383  0.62121117 -0.67987338  -2.63172497
#>   [6,] -0.266264354  0.54629404  1.80910855  0.97781403   3.93962769
#>   [7,]  0.636432097  0.01894582  1.11398601  1.77247739 -18.83155257
#>   [8,] -1.013955594 -0.74870221  0.46535053  0.27793329  -0.76965728
#>   [9,]  0.932230747 -0.79432104 -1.06862973 -0.48101931 -13.02275701
#>  [10,] -0.650788332  0.30078439  0.25594028 -1.79106060 -12.27386244
#>  [11,]  0.436178180  0.05466585 -0.77721893  0.53066370  -8.61399362
#>  [12,] -0.164089402  0.36422911 -0.95031824 -1.12476618 -17.12242879
#>  [13,]  1.268656386 -1.16518670  1.23051633  0.04022615  10.32914943
#>  [14,]  0.415432504  0.26055583 -0.29032133 -1.18796740  -9.30138693
#>  [15,] -0.048104283 -2.41711628 -1.24524959  2.24883111 -32.46518743
#>  [16,]  0.198292951  1.14908502 -0.92511062  1.55316175 -17.79223556
#>  [17,] -0.388687305  1.20657925 -0.35175827  0.32302195 -10.51101639
#>  [18,] -0.452232286 -0.43198213 -0.04163083  0.02275970   1.54728374
#>  [19,] -0.139421427  0.10685674  0.72929742 -0.24942787  -5.52858160
#>  [20,] -1.325197658  0.48619572  0.29340303 -2.17231912 -25.31477243
#>  [21,]  0.754069518 -1.55345832  3.58694548 -1.33965912   9.81722911
#>  [22,]  1.357852830 -1.08314239  0.75574301 -0.98048575  -0.74715179
#>  [23,] -1.220784861 -0.51601203  0.40012730 -1.79278068 -18.00653404
#>  [24,] -0.567918662 -0.20777098  0.82934571  0.52281281   4.05234615
#>  [25,] -1.188849569  0.90728141 -0.30513624 -1.59132848 -23.50104356
#>  [26,] -1.012275885  0.81142513  0.71010332  0.80626313  -1.09233267
#>  [27,] -0.432531486 -0.99119325  0.43861382 -0.68381307  -6.20873878
#>  [28,]  0.345706392 -0.45430696  0.35822370  0.33438806  10.35891871
#>  [29,]  0.190927596 -0.31572854 -0.12423765  0.94300642   0.53132939
#>  [30,]  1.287488578 -1.00633308 -0.69321609 -0.89838767  -8.55574550
#>  [31,] -0.213504255 -1.01659466  0.57434653  0.76733587  -9.99970963
#>  [32,]  0.593561305  2.02996478 -0.21715952  0.09233868  -7.39403279
#>  [33,] -1.815799874 -0.68763843  3.03758116 -0.35169603  20.68330441
#>  [34,] -0.460524573 -1.58790685 -1.59166713  0.11537300  -9.55084358
#>  [35,]  0.746337982  1.01693810 -0.67368924 -0.14326135 -14.51190722
#>  [36,]  0.898995877 -0.72310219  0.72857280 -2.41982035 -30.98691628
#>  [37,] -1.155462161 -1.45348504 -1.15156224 -1.16379404 -15.90400464
#>  [38,] -0.282423821 -2.14081548 -1.15077521  0.10342665 -13.90377889
#>  [39,] -0.003256871  0.13485443 -0.13207528 -0.12614333  -7.33922735
#>  [40,]  1.515028276  0.66199778 -0.58104636 -0.14174341  -2.18905537
#>  [41,]  0.309380270  0.49461317  1.66135003  0.91043704  13.88205400
#>  [42,] -0.460958821 -1.52664365  1.16794898  1.47761791  -3.78908813
#>  [43,]  0.293447573 -0.11021600  0.76988546  1.64624260  -6.33655589
#>  [44,]  0.070972006  2.46874128 -0.16311581 -0.17999302   0.87359368
#>  [45,] -2.313759365 -0.56354343  1.27526535  0.97685639  -0.13968001
#>  [46,]  0.192706032 -1.44585342  0.18088798 -0.18660273   5.97180729
#>  [47,] -0.352441451 -0.47655578  1.47436052 -0.03409839   7.00886241
#>  [48,] -1.533292787  1.01460047 -1.51743472  1.48873220 -24.28687520
#>  [49,] -1.734069209 -0.11802003  0.30440036 -0.18983930   4.11751560
#>  [50,]  0.141892735 -0.35432184  0.10746739 -1.08930962   1.57090500
#>  [51,] -0.219599601  0.28725584 -1.22479257  0.37108948 -15.17126111
#>  [52,]  0.591564494  0.01801262  0.44851998 -1.02777011 -12.21924564
#>  [53,]  0.240247909 -0.89842398  0.50807095  0.01491585   3.10158379
#>  [54,] -0.372564160  1.44473649  0.35021264 -0.52029856  -2.49629541
#>  [55,]  0.264545865 -0.53779915  0.75821367 -0.05224443  10.73772159
#>  [56,]  0.651544488 -0.48236000  0.29092872 -1.43507957  -9.56885439
#>  [57,] -0.418319774 -0.73561961  0.02285959 -0.32797907  -6.41820696
#>  [58,] -2.078760627 -0.01523526 -0.20814311 -1.28534762 -16.57697969
#>  [59,]  0.590955041 -0.28506084 -0.26361162 -0.42553724  -2.80563437
#>  [60,] -0.705640744 -0.77089599 -1.98413580  0.25747566 -11.97839266
#>  [61,]  0.986688400 -1.91152294 -0.35762022  0.20131940  -9.64339528
#>  [62,]  0.032359045  0.17789966 -0.45429735  0.48207740  -7.68178328
#>  [63,] -0.496043573  0.19632069  0.28615770  0.50663659  -4.37126487
#>  [64,]  1.673249059  0.05004416 -0.28917939 -0.52669136 -14.45020931
#>  [65,] -1.051266614 -1.04527444  0.51190457 -0.56477398  -2.20158006
#>  [66,]  0.028015821 -0.13668239 -1.10433455  0.04039804  -8.72535245
#>  [67,]  0.526890091  1.22750061 -0.22247479 -0.20078138  -7.43826013
#>  [68,]  1.270786920 -0.19386519 -1.45215149 -0.90421719  -8.60695904
#>  [69,] -0.587592446 -1.99307110  0.39929877  0.16665552   0.64274493
#>  [70,] -0.430471444 -0.42990567  1.42857119 -0.52428478   6.52826673
#>  [71,] -0.022966263 -0.65191632 -0.48214099  0.01756390  -4.47706923
#>  [72,]  0.074873923 -1.32275532  0.99238818  0.94851719   5.18548733
#>  [73,] -2.089995616 -0.61100236 -1.16370065  0.27026578  -9.16415479
#>  [74,]  1.023897173 -0.43348477  1.62078912 -0.16094140  13.81494257
#>  [75,]  0.623240132  0.22113162 -0.10744815  1.96519951 -25.82660867
#>  [76,] -1.367330008  0.41463530 -0.77320677  0.39554804  -7.77130327
#>  [77,]  0.044589523  1.00351963 -0.73627880  0.87371937 -11.07879743
#>  [78,]  0.366273799 -2.62604224  0.81570668  0.47232416   6.75798648
#>  [79,] -0.178469697 -0.87353493  2.20303644 -1.48522645  -7.23124815
#>  [80,]  2.101677593 -0.77371083 -0.67104663 -0.83616230  -9.10142608
#>  [81,] -0.242409658  0.59305855 -0.42158960 -0.64626921  -6.18445349
#>  [82,]  0.011717215  0.45356334 -0.77774912 -0.04743704  -2.53003243
#>  [83,] -2.778433689 -0.48962671  0.56154237  0.16775999   9.61462267
#>  [84,]  0.274491395 -0.66608550 -2.22250706 -0.09973867 -10.14434855
#>  [85,] -0.622731928 -1.08965423 -0.69541367  0.08599646  -3.37400520
#>  [86,] -1.027831717 -1.04090695 -0.38354996 -0.37257403  -9.18764413
#>  [87,]  0.728990109  0.50160668  2.05856698  0.18133254   8.79448725
#>  [88,] -0.392461437  0.04135376 -1.10798826 -0.90680641 -18.45210165
#>  [89,]  0.357845575 -0.05308168 -0.95593877  0.59115715  -6.25566618
#>  [90,] -0.755189860 -1.52739908 -0.63576842  0.79251323  -0.06236973
#>  [91,]  0.066689713 -0.16356852  0.76675745  0.07687238   5.55220268
#>  [92,]  0.257355650 -2.13509492  0.68695810 -0.06783720   7.18387440
#>  [93,] -0.090963056 -0.13683359  1.72142669  0.43328805   4.81045220
#>  [94,] -0.730508027 -2.90422064  0.85130925  0.04909540   4.41889929
#>  [95,] -0.395841246  0.33701264  0.42990828 -0.03280005  -1.81842660
#>  [96,] -1.415808515  0.76451066 -1.05527774 -0.51092478 -10.64475057
#>  [97,] -0.108938918 -0.66937819 -0.04588097  0.35643054  -4.70026690
#>  [98,]  0.074299910 -0.16096887 -0.13188138  0.41794614  -1.73357207
#>  [99,]  0.450190145  1.50726103 -0.31946356  0.57920526  -0.46299199
#> [100,]  0.387345818 -0.12169384  0.47360616 -1.47515865  -6.40547113
#> 
#> $eta
#>   [1]   0.59529574  -2.77696693  -7.69531781  -9.67268851  -2.80464766
#>   [6]   3.51298616 -18.01785299  -1.42101238 -13.94273880 -12.66396423
#>  [11]  -9.57071772 -16.95245523  10.70912735  -9.44559488 -31.74690334
#>  [16] -17.99834421 -10.14368202   0.75884087  -5.21762880 -24.45563233
#>  [21]   8.73194944   0.03762753 -16.78557380   5.10903758 -22.50415300
#>  [26]  -0.41972758  -6.87400715   8.12284103   1.13312756  -8.26412776
#>  [31]  -9.15873476  -9.42573230  20.25393228 -10.36186110 -14.33760727
#>  [36] -30.67054480 -15.41100275 -14.09455736  -6.23548237  -2.08264205
#>  [41]  12.42306129  -3.48161383  -7.09373730   2.58441907  -0.73498961
#>  [46]   8.03977394   7.09460216 -24.31916175   3.82782920   0.76196693
#>  [51] -15.49752359 -10.51538816   2.57607920  -1.69326156  10.97537731
#>  [56] -10.42420345  -5.94871991 -16.69499248  -4.05307593 -11.28999536
#>  [61]  -8.25982570  -7.63885086  -4.16203064 -12.98321586  -4.43735844
#>  [66]  -8.10020748  -6.27034634  -8.75648788  -0.21857497   7.57447013
#>  [71]  -4.64932924   5.49250808 -10.06750040  12.82969388 -27.47043716
#>  [76]  -8.58586299 -11.90391060   7.10376040  -7.64239549  -8.45821169
#>  [81]  -7.12651297  -2.99962421  10.27673490  -9.83487744  -4.16857947
#>  [86]  -8.27279317   9.26000756 -19.27177101  -7.77670192  -0.32772452
#>  [91]   4.93313574   6.71279356   4.34730799   4.57906066  -1.78274113
#>  [96] -10.07006575  -5.54561692  -1.15180595   0.03936511  -7.49354739
#> 
#> $mu
#>   [1]   0.59529574  -2.77696693  -7.69531781  -9.67268851  -2.80464766
#>   [6]   3.51298616 -18.01785299  -1.42101238 -13.94273880 -12.66396423
#>  [11]  -9.57071772 -16.95245523  10.70912735  -9.44559488 -31.74690334
#>  [16] -17.99834421 -10.14368202   0.75884087  -5.21762880 -24.45563233
#>  [21]   8.73194944   0.03762753 -16.78557380   5.10903758 -22.50415300
#>  [26]  -0.41972758  -6.87400715   8.12284103   1.13312756  -8.26412776
#>  [31]  -9.15873476  -9.42573230  20.25393228 -10.36186110 -14.33760727
#>  [36] -30.67054480 -15.41100275 -14.09455736  -6.23548237  -2.08264205
#>  [41]  12.42306129  -3.48161383  -7.09373730   2.58441907  -0.73498961
#>  [46]   8.03977394   7.09460216 -24.31916175   3.82782920   0.76196693
#>  [51] -15.49752359 -10.51538816   2.57607920  -1.69326156  10.97537731
#>  [56] -10.42420345  -5.94871991 -16.69499248  -4.05307593 -11.28999536
#>  [61]  -8.25982570  -7.63885086  -4.16203064 -12.98321586  -4.43735844
#>  [66]  -8.10020748  -6.27034634  -8.75648788  -0.21857497   7.57447013
#>  [71]  -4.64932924   5.49250808 -10.06750040  12.82969388 -27.47043716
#>  [76]  -8.58586299 -11.90391060   7.10376040  -7.64239549  -8.45821169
#>  [81]  -7.12651297  -2.99962421  10.27673490  -9.83487744  -4.16857947
#>  [86]  -8.27279317   9.26000756 -19.27177101  -7.77670192  -0.32772452
#>  [91]   4.93313574   6.71279356   4.34730799   4.57906066  -1.78274113
#>  [96] -10.07006575  -5.54561692  -1.15180595   0.03936511  -7.49354739
#>